Woman's health seriously endangered

abhlth

Source

General Social Survey https://gss.norc.org

Details

Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: c. The woman's own health is seriously endangered by the pregnancy?

Overview

For further details see the official GSS documentation.

Counts by year:

year	don't know	no	no answer	yes	iap	skipped on web	Total

1972	66	201	8	1338	-	-	1613

1973	26	114	2	1362	-	-	1504

1974	32	110	-	1342	-	-	1484

1975	38	135	3	1314	-	-	1490

1976	28	135	7	1329	-	-	1499

1977	34	141	8	1347	-	-	1530

1978	36	140	4	1352	-	-	1532

1980	37	141	2	1288	-	-	1468

1982	55	173	5	1627	-	-	1860

1983	47	156	32	1364	-	-	1599

1984	31	150	11	1281	-	-	1473

1985	40	159	3	1332	-	-	1534

1986	-	-	-	-	1470	-	1470

1987	66	211	7	1535	-	-	1819

1988	32	106	3	836	504	-	1481

1989	24	99	6	904	504	-	1537

1990	27	73	3	814	455	-	1372

1991	31	81	7	874	524	-	1517

1993	43	105	4	923	531	-	1606

1994	59	181	3	1753	996	-	2992

1996	65	156	8	1694	981	-	2904

1998	83	218	3	1578	950	-	2832

2000	72	205	4	1580	956	-	2817

2002	21	76	2	825	1841	-	2765

2004	27	121	8	742	1914	-	2812

2006	62	233	16	1692	2507	-	4510

2008	57	139	10	1146	671	-	2023

2010	35	158	22	1066	763	-	2044

2012	58	146	21	1083	666	-	1974

2014	46	195	21	1452	824	-	2538

2016	53	202	25	1610	977	-	2867

2018	43	151	14	1366	774	-	2348

2021	9	141	1	1290	2574	17	4032

2022	21	80	10	1062	2367	4	3544

2024	38	118	9	1437	1694	13	3309

Total	1442	4950	292	43538	25443	34	75699

Values

1 yes

2 no

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook