Married–wants no more children

abnomore

Source

General Social Survey https://gss.norc.org

Details

Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: b. She is married and does not want any more children?

Overview

For further details see the official GSS documentation.

Counts by year:

year	don't know	no	no answer	yes	iap	skipped on web	Total

1972	80	921	5	607	-	-	1613

1973	49	760	2	693	-	-	1504

1974	73	749	-	662	-	-	1484

1975	62	774	2	652	-	-	1490

1976	46	779	6	668	-	-	1499

1977	61	782	7	680	-	-	1530

1978	46	885	3	598	-	-	1532

1980	58	744	3	663	-	-	1468

1982	93	955	4	808	-	-	1860

1983	49	928	30	592	-	-	1599

1984	43	814	10	606	-	-	1473

1985	41	888	5	600	-	-	1534

1986	-	-	-	-	1470	-	1470

1987	68	1028	9	714	-	-	1819

1988	26	570	2	379	504	-	1481

1989	37	550	5	441	504	-	1537

1990	33	483	5	396	455	-	1372

1991	36	526	7	424	524	-	1517

1993	51	539	5	480	531	-	1606

1994	65	995	6	930	996	-	2992

1996	91	972	7	853	981	-	2904

1998	86	1033	5	758	950	-	2832

2000	85	1050	6	720	956	-	2817

2002	29	493	2	400	1841	-	2765

2004	30	502	6	360	1914	-	2812

2006	65	1107	13	818	2507	-	4510

2008	40	716	17	579	671	-	2023

2010	36	636	22	587	763	-	2044

2012	44	669	18	577	666	-	1974

2014	50	862	22	780	824	-	2538

2016	43	964	24	859	977	-	2867

2018	38	736	22	778	774	-	2348

2021	10	583	3	841	2574	21	4032

2022	38	438	10	684	2367	7	3544

2024	61	646	18	876	1694	14	3309

Total	1763	26077	311	22063	25443	42	75699

Values

1 yes

2 no

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook