Not married

absingle

Source

General Social Survey https://gss.norc.org

Details

Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: f. She is not married and does not want to marry the man?

Overview

For further details see the official GSS documentation.

Counts by year:

year	don't know	no	no answer	yes	iap	skipped on web	Total

1972	103	848	8	654	-	-	1613

1973	49	738	5	712	-	-	1504

1974	64	709	-	711	-	-	1484

1975	71	732	5	682	-	-	1490

1976	56	715	5	723	-	-	1499

1977	64	733	7	726	-	-	1530

1978	55	867	4	606	-	-	1532

1980	62	724	2	680	-	-	1468

1982	91	961	5	803	-	-	1860

1983	73	905	33	588	-	-	1599

1984	42	792	11	628	-	-	1473

1985	43	874	5	612	-	-	1534

1986	-	-	-	-	1470	-	1470

1987	65	1045	9	700	-	-	1819

1988	41	565	4	367	504	-	1481

1989	43	538	5	447	504	-	1537

1990	41	477	4	395	455	-	1372

1991	36	524	7	426	524	-	1517

1993	56	526	5	488	531	-	1606

1994	52	1015	6	923	996	-	2992

1996	80	1010	11	822	981	-	2904

1998	87	1033	5	757	950	-	2832

2000	88	1076	6	691	956	-	2817

2002	27	519	2	376	1841	-	2765

2004	20	512	12	354	1914	-	2812

2006	57	1159	10	777	2507	-	4510

2008	38	765	13	536	671	-	2023

2010	36	704	17	524	763	-	2044

2012	45	713	17	533	666	-	1974

2014	44	938	19	713	824	-	2538

2016	30	1051	27	782	977	-	2867

2018	38	816	20	700	774	-	2348

2021	11	612	1	813	2574	21	4032

2022	29	467	10	666	2367	5	3544

2024	57	695	18	830	1694	15	3309

Total	1794	26358	318	21745	25443	41	75699

Values

1 yes

2 no

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook