Number of cohabiting family member

acqfmcoh

Source

General Social Survey https://gss.norc.org

Details

Question Next, we are going to ask questions about people in your family, including relatives and in-laws. O. How many women in your family live in a romantic relationship with a man to whom they are not married?

Values

1 0

2 1

3 2 - 5

4 6 - 10

5 more than 10

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2006	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for acqfmcoh.

See also