Favor preference in hiring black people

affrmact

Source

General Social Survey https://gss.norc.org

Details

Question 153. Some people say that because of past discrimination, blacks should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of blacks is wrong because it discriminates against whites. What about your opinion - are you for or against preferential hiring and promotion of blacks? If favors: Do you favor preference in hiring and promotion strongly or not strongly? If opposes: Do you oppose preference in hiring and promotion strongly or not strongly?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	don't know	no answer	not strongly favors	not strongly opposes	strongly favors	strongly opposes	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	1606

1994	1522	79	13	100	354	135	789	-	2992

1996	944	117	22	127	485	183	1026	-	2904

1998	961	145	18	117	425	142	1024	-	2832

2000	940	119	14	130	480	201	933	-	2817

2002	1857	22	3	58	249	84	492	-	2765

2004	1906	38	4	60	244	86	474	-	2812

2006	2518	75	13	136	525	188	1055	-	4510

2008	694	65	9	87	349	133	686	-	2023

2010	614	66	14	114	377	160	699	-	2044

2012	672	63	18	99	354	139	629	-	1974

2014	863	67	16	130	475	173	814	-	2538

2016	979	87	24	151	545	268	813	-	2867

2018	789	88	24	132	443	230	642	-	2348

2021	1378	22	4	299	818	349	1162	-	4032

2022	1168	119	15	248	667	395	919	13	3544

2024	1126	115	17	220	573	333	884	41	3309

Total	48319	1287	228	2208	7363	3199	13041	54	75699

Values

1 strongly favors

2 not strongly favors

3 not strongly opposes

4 strongly opposes

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook