Asians discriminated in jobs

asnjobs

Source

General Social Survey https://gss.norc.org

Details

Question How much discrimination is there that hurts the chances of Hispanic Americans to get good paying jobs? Would you say there is a lot, some, only a little, or none at all? C. Asian Americans?

Values

1 a lot

2 some

3 only a little

4 none at all

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1990	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for asnjobs.

See also