Is 10th person staying somewhere else now?

away10

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what? 5. Is the tenth person staying somewhere else right now?

Values

0 no

1 yes

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1975	-/-/-/-	full

1976	-/-/-/-	full

1977	-/-/-/-	full

1978	-/-/-/-	full

1980	-/-/-/-	full

1982	-/-/-/-	full

1983	-/-/-/-	full

1984	-/-/-/-	full

1985	-/-/-/-	full

1986	-/-/-/-	full

1987	-/-/-/-	full

1988	A/B/C/-	full

1989	A/B/C/-	full

1990	A/B/C/-	full

1991	A/B/C/-	full

1993	A/B/C/-	full

1994	A/B/C/-	full

1996	A/B/C/-	full

1998	A/B/C/-	full

2000	A/B/C/-	full

2002	A/B/C/-	full

2004	A/B/C/-	full

2006	A/B/C/D	full

2008	A/B/C/-	full

2010	A/B/C/-	full

2012	A/B/C/-	full

2014	A/B/C/-	full

2016	A/B/C/-	full

2018	A/B/C/-	full

2021	A/B/C/-	full

2022	A/B/C/-	full

Link at the GSS

For details see the GSS Data Explorer page for away10.

See also