Household members less than 6 yrs old

babies

Source

General Social Survey https://gss.norc.org

Details

Question Household Size and Composition (see Appendix D: Recodes, for information about these variables). A. NUMBER OF MEMBERS UNDER 6 YEARS OLD: As of 2024, BABIES has been recoded to only account for 0 to 2+ children younger than 6 in the household. This has been retroactively applied to all previous years of BABIES. The original BABIES containing data from 1972 to 2022 has been renamed BABIES_7222.

Values

[0] 0 household members under 6

[8] 8 or more

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C/D	2006

Full/-	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for babies.

Counts by year:

year	0 household members under 6	1	2	3	4	5	no answer	6	not available in this release	iap	skipped on web	Total

1972	1215	228	129	34	7	-	-	-	-	-	-	1613

1973	1150	216	110	22	5	1	-	-	-	-	-	1504

1974	1127	227	101	24	2	1	2	-	-	-	-	1484

1975	1167	228	83	10	1	1	-	-	-	-	-	1490

1976	1160	224	88	21	4	-	2	-	-	-	-	1499

1977	1248	197	62	13	1	-	9	-	-	-	-	1530

1978	1232	195	78	11	2	-	14	-	-	-	-	1532

1980	1179	200	66	10	2	-	11	-	-	-	-	1468

1982	1507	240	90	18	-	-	5	-	-	-	-	1860

1983	1275	212	90	14	3	-	5	-	-	-	-	1599

1984	1203	176	74	13	1	1	5	-	-	-	-	1473

1985	1244	200	66	17	2	-	5	-	-	-	-	1534

1986	1189	194	70	10	2	-	5	-	-	-	-	1470

1987	1509	215	76	16	-	1	2	-	-	-	-	1819

1988	1200	202	58	10	-	-	11	-	-	-	-	1481

1989	1233	192	77	13	1	-	21	-	-	-	-	1537

1990	1134	157	67	4	-	-	10	-	-	-	-	1372

1991	1231	183	87	8	1	1	6	-	-	-	-	1517

1993	1321	198	75	9	-	-	3	-	-	-	-	1606

1994	2485	369	108	23	4	-	3	-	-	-	-	2992

1996	2408	340	132	22	2	-	-	-	-	-	-	2904

1998	2401	308	102	15	-	-	5	1	-	-	-	2832

2000	2322	315	115	16	2	-	47	-	-	-	-	2817

2002	2379	259	74	22	2	-	29	-	-	-	-	2765

2004	2439	241	96	15	1	-	20	-	-	-	-	2812

2006	3850	421	167	29	5	-	38	-	-	-	-	4510

2008	1697	187	93	15	2	-	29	-	-	-	-	2023

2010	1759	201	60	11	3	2	8	-	-	-	-	2044

2012	1686	180	75	12	1	-	20	-	-	-	-	1974

2014	2275	177	68	11	3	-	4	-	-	-	-	2538

2016	2550	192	86	13	2	-	24	-	-	-	-	2867

2018	2073	185	53	7	-	1	29	-	-	-	-	2348

2021	-	-	-	-	-	-	-	-	4032	-	-	4032

2022	1577	130	56	4	1	-	8	-	-	1764	4	3544

Total	55425	7389	2832	492	62	9	380	1	4032	1764	4	72390

See also