2nd mention, what nervous breakdown is, imputed

brkdimp2

Source

General Social Survey https://gss.norc.org

Details

Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it? What is it like? What happens to a person who has one? How does he act?) 2. R's imputed references for what a nervous breakdown is – Second mention.

Values

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for brkdimp2.

See also