How close are persons number 1 and 2?

close12

Source

General Social Survey https://gss.norc.org

Details

Question IF LESS THAN 5 NAMES MENTIONED CROSS OUT UNUSED BOXES. Please think about the relations between the people you just mentioned. Some of them may be total strangers in the sense that they wouldn't recognize each other if they bumped into each other on the street. Others may be especially close, as close or closer to each other as they are to you. First, think about NAME 1 and NAME 2. ASK CLOSE1 FOR FIRST PAIR. A. Are ____ and ____ total strangers? IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DoWN) IF NO........(ASK CLOSE12, CLOSE13, CLOSE14, CLOSE15, CLOSE 23, CLOSE24, CLOSE25, CLOSE34, CLOSE35, CLOSE45) B. Are they especially close? PROBE: As close or closer to each other as they are to you? IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN) IF NO........(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN)

Closeness of First and Second Persons

Values

[1] especially close

[2] know each other

[3] total strangers

[7] refused

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2004

Full	1985

Overview

For further details see the GSS Data Explorer page for close12.

Counts by year:

year	iap	especially close	know each other	total strangers	no answer	Total

1985	364	540	471	147	12	1534

2004	1977	405	283	100	47	2812

Total	2341	945	754	247	59	4346

See also