Travel time to work

commute

Source

General Social Survey https://gss.norc.org

Details

Question A. What kind of work do you (did you normally) do? That is, what (is/was) your job called? OCCUPATION: (See REMARKS) B. What (do/did) you actually do in that job? Tell me, what (are/were) some of your main duties? C. What kind of place (do/did) you work for? INDUSTRY: (See REMARKS) D. What (do/did) they (make/do)? E. (Are/Were) you self employed or (do/did) you work for someone else? F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)? G. About how much time (does/did) it usually take you to travel to work - about how many minutes?

Values

97 97+ minutes

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-	1986

Overview

For further details see the GSS Data Explorer page for commute.

Counts by year:

year	iap	0	1	10	11	12	13	14	15	16	17	2	20	22	24	25	28	3	30	32	35	4	40	43	45	46	5	50	55	6	60	7	70	75	8	9	90	97+ minutes	don't know	no answer	Total

1986	100	54	14	190	1	16	1	2	203	1	3	34	180	2	1	60	1	23	155	2	30	6	22	1	63	1	153	5	3	2	55	13	4	4	8	1	8	8	8	32	1470

Total	100	54	14	190	1	16	1	2	203	1	3	34	180	2	1	60	1	23	155	2	30	6	22	1	63	1	153	5	3	2	55	13	4	4	8	1	8	8	8	32	1470

See also