Confidence in military

conarmy

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. M. Military

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1973	-/-	full

1974	-/-	full

1975	-/-	full

1976	-/-	full

1977	-/-	full

1978	-/-	full

1980	-/-	full

1982	-/-	full

1983	-/-	full

1984	-/-	full

1986	-/-	full

1987	-/-	full

1988	B/C	partial

1989	B/C	partial

1990	B/C	partial

1991	B/C	partial

1993	B/C	partial

1994	B/C	partial

1996	B/C	partial

1998	B/C	partial

2000	B/C	partial

2002	B/C	partial

2004	B/C	partial

2006	B/C	partial

2008	B/C	partial

2010	B/C	partial

2012	B/C	partial

2014	B/C	partial

2016	B/C	partial

2018	B/C	partial

2021	B/C	partial

2022	B/C	partial

2024	B/C	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	a great deal	don't know	hardly any	no answer	only some	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	-	475	41	241	6	741	-	1504

1974	-	587	39	199	1	658	-	1484

1975	-	524	69	213	3	681	-	1490

1976	-	584	92	199	8	616	-	1499

1977	-	554	48	157	4	767	-	1530

1978	-	451	56	196	4	825	-	1532

1980	-	405	61	237	1	764	-	1468

1982	-	545	47	269	8	991	-	1860

1983	-	468	48	207	3	873	-	1599

1984	484	352	27	125	12	473	-	1473

1985	1534	-	-	-	-	-	-	1534

1986	-	461	37	200	4	768	-	1470

1987	-	610	60	232	5	912	-	1819

1988	484	338	32	134	2	491	-	1481

1989	502	335	42	139	3	516	-	1537

1990	473	293	26	121	-	459	-	1372

1991	500	605	18	60	3	331	-	1517

1993	549	441	18	116	5	477	-	1606

1994	981	737	46	246	11	971	-	2992

1996	979	716	62	213	2	932	-	2904

1998	921	689	53	237	5	927	-	2832

2000	921	727	68	184	10	907	-	2817

2002	1853	492	14	63	-	343	-	2765

2004	1936	495	4	74	1	302	-	2812

2006	2521	929	31	256	1	772	-	4510

2008	658	681	17	153	2	512	-	2023

2010	667	720	19	129	2	507	-	2044

2012	636	726	20	109	1	482	-	1974

2014	851	843	14	157	3	670	-	2538

2016	911	1027	16	147	3	763	-	2867

2018	785	935	17	103	4	504	-	2348

2021	1360	1254	3	255	-	1147	13	4032

2022	1173	1039	17	294	2	1017	2	3544

2024	1116	914	25	271	1	970	12	3309

Total	24408	20952	1187	5936	120	23069	27	75699

See also