Confidence in major companies

conbus

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. B. Major companies

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for conbus.

Counts by year:

year	iap	a great deal	only some	hardly any	don't know	no answer	skipped on web	Total

1973	-	439	799	162	100	4	-	1504

1974	-	465	750	215	53	1	-	1484

1975	-	286	801	314	82	7	-	1490

1976	-	328	764	324	75	8	-	1499

1977	-	415	862	188	61	4	-	1530

1978	-	331	885	245	68	3	-	1532

1980	-	399	779	208	80	2	-	1468

1982	-	393	1085	278	97	7	-	1860

1983	-	379	937	213	66	4	-	1599

1984	484	301	561	85	34	8	-	1473

1986	-	353	908	150	55	4	-	1470

1987	-	502	1052	176	85	4	-	1819

1988	484	245	598	105	45	4	-	1481

1989	502	244	623	106	59	3	-	1537

1990	473	223	544	97	33	2	-	1372

1991	500	205	629	132	50	1	-	1517

1993	549	221	661	130	43	2	-	1606

1994	981	510	1232	199	64	6	-	2992

1996	979	447	1137	260	80	1	-	2904

1998	921	500	1078	243	87	3	-	2832

2000	921	528	1077	207	77	7	-	2817

2002	1853	158	572	163	18	1	-	2765

2004	1936	146	545	167	18	-	-	2812

2006	2521	335	1205	396	50	3	-	4510

2008	658	207	882	244	32	-	-	2023

2010	667	174	848	317	32	6	-	2044

2012	636	216	814	273	33	2	-	1974

2014	851	312	1030	313	28	4	-	2538

2016	911	340	1246	340	27	3	-	2867

2018	785	294	1012	222	31	4	-	2348

2021	1360	450	1628	578	7	2	7	4032

2022	1173	326	1463	529	42	6	5	3544

Total	20145	10672	29007	7579	1712	116	12	69243

See also