Confid. in exec branch of fed govt

confed

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. E. Executive branch of the federal government

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for confed.

Counts by year:

year	iap	a great deal	only some	hardly any	don't know	no answer	skipped on web	Total

1973	-	439	755	275	29	6	-	1504

1974	-	202	630	618	32	2	-	1484

1975	-	198	813	439	38	2	-	1490

1976	-	201	874	374	45	5	-	1499

1977	-	426	830	221	48	5	-	1530

1978	-	191	908	380	49	4	-	1532

1980	-	177	737	502	49	3	-	1468

1982	-	323	984	495	49	9	-	1860

1983	-	206	868	471	52	2	-	1599

1984	484	181	493	280	23	12	-	1473

1986	-	302	779	346	41	2	-	1470

1987	-	316	904	530	65	4	-	1819

1988	484	164	530	266	34	3	-	1481

1989	502	202	557	225	47	4	-	1537

1990	473	210	451	208	30	-	-	1372

1991	500	263	515	212	24	3	-	1517

1993	549	124	556	340	32	5	-	1606

1994	981	226	1034	701	42	8	-	2992

1996	979	194	862	800	68	1	-	2904

1998	921	265	909	671	60	6	-	2832

2000	921	249	924	632	82	9	-	2817

2002	1853	243	455	193	19	2	-	2765

2004	1936	179	412	271	12	2	-	2812

2006	2521	282	870	789	44	4	-	4510

2008	658	133	647	552	32	1	-	2023

2010	667	227	620	499	26	5	-	2044

2012	636	196	618	493	28	3	-	1974

2014	851	199	723	731	29	5	-	2538

2016	911	238	851	827	36	4	-	2867

2018	785	182	659	692	25	5	-	2348

2021	1360	336	1140	1182	2	-	12	4032

2022	1173	243	962	1125	30	4	7	3544

Total	20145	7517	23870	16340	1222	130	19	69243

See also