Should ethnic congressmen represent ethnic groups

congeth

Source

General Social Survey https://gss.norc.org

Details

Question Some people say that it is only fair that in a society like ours, people holding public office should have roughly the same racial or ethnic background as the population as a whole. So, for example, if 30pct of the people were Hispanic, then 30pct of the public officials should be Hispanic as well. Others say that in choosing people for public office they should be considered purely on the basis of ability without regard to their ethnic or racial background. Let me ask you about these ideas in several different areas. A. What about people in Congress? Should they have the same racial or ethnic background as their constituents, or should they be considered purely on the basis of their ability without regard to their racial or ethnic background?

Values

1 consider ethnic background

2 ability, not background

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1994	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	ability, not background	consider ethnic background	don't know	no answer	not available in this year	Total

1972	1613	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	1606

1994	1518	1279	131	53	11	-	2992

1996	2904	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	4032

2022	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	3309	3309

Total	67372	1279	131	53	11	6853	75699

See also