Confid. in united states supreme court

conjudge

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. J. U.S. Supreme Court

Values

[1] a great deal

[2] only some

[3] hardly any

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for conjudge.

Counts by year:

year	iap	a great deal	only some	hardly any	don't know	no answer	skipped on web	Total

1973	-	471	745	231	50	7	-	1504

1974	-	492	710	213	67	2	-	1484

1975	-	457	688	276	64	5	-	1490

1976	-	528	650	229	84	8	-	1499

1977	-	543	752	165	62	8	-	1530

1978	-	429	806	223	69	5	-	1532

1980	-	361	734	286	87	-	-	1468

1982	-	543	1003	237	67	10	-	1860

1983	-	434	875	225	61	4	-	1599

1984	484	325	497	121	35	11	-	1473

1986	-	434	765	206	55	10	-	1470

1987	-	622	924	190	79	4	-	1819

1988	485	344	499	105	44	4	-	1481

1989	502	356	517	111	49	2	-	1537

1990	473	315	431	114	39	-	-	1372

1991	500	373	470	124	45	5	-	1517

1993	549	322	546	141	45	3	-	1606

1994	981	604	1003	328	69	7	-	2992

1996	979	543	959	318	101	4	-	2904

1998	921	592	950	267	96	6	-	2832

2000	921	601	932	239	116	8	-	2817

2002	1853	321	454	104	32	1	-	2765

2004	1936	266	451	141	18	-	-	2812

2006	2521	632	977	307	69	4	-	4510

2008	658	405	716	204	40	-	-	2023

2010	667	400	706	229	38	4	-	2044

2012	636	378	700	217	38	5	-	1974

2014	851	395	912	334	43	3	-	2538

2016	911	488	1038	389	37	4	-	2867

2018	785	462	828	229	41	3	-	2348

2021	1360	689	1437	536	3	1	6	4032

2022	1173	373	1060	894	29	9	6	3544

Total	20146	14498	24735	7933	1772	147	12	69243

See also