Confidence in organized labor

conlabor

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. F. Organized labor

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for conlabor.

Counts by year:

year	iap	a great deal	only some	hardly any	don't know	no answer	skipped on web	Total

1973	-	232	817	384	62	9	-	1504

1974	-	270	793	377	41	3	-	1484

1975	-	150	807	436	95	2	-	1490

1976	-	173	710	493	118	5	-	1499

1977	-	225	757	483	59	6	-	1530

1978	-	168	707	575	78	4	-	1532

1980	-	220	727	434	85	2	-	1468

1982	-	241	1006	520	81	12	-	1860

1983	-	127	774	630	66	2	-	1599

1984	484	84	517	353	24	11	-	1473

1986	-	117	693	576	79	5	-	1470

1987	-	211	941	556	106	5	-	1819

1988	484	99	498	344	52	4	-	1481

1989	502	97	524	345	66	3	-	1537

1990	473	96	476	279	47	1	-	1372

1991	500	112	491	345	67	2	-	1517

1993	549	87	553	340	73	4	-	1606

1994	981	198	1045	648	111	9	-	2992

1996	979	215	979	567	162	2	-	2904

1998	921	210	986	556	155	4	-	2832

2000	921	239	962	500	185	10	-	2817

2002	1853	101	550	210	50	1	-	2765

2004	1936	103	478	249	44	2	-	2812

2006	2521	223	1087	568	111	-	-	4510

2008	658	165	781	358	60	1	-	2023

2010	667	146	781	373	74	3	-	2044

2012	636	152	719	385	78	4	-	1974

2014	851	196	929	488	72	2	-	2538

2016	911	249	1185	440	78	4	-	2867

2018	785	211	967	310	70	5	-	2348

2021	1360	297	1655	696	8	-	16	4032

2022	1173	311	1383	582	84	4	7	3544

Total	20145	5725	26278	14400	2541	131	23	69243

See also