Confidence in press

conpress

Source

General Social Survey https://gss.norc.org

Details

Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? READ EACH ITEM; CODE ONE FOR EACH. G. Press

Values

[1] a great deal

[2] only some

[3] hardly any

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for conpress.

Counts by year:

year	iap	a great deal	only some	hardly any	don't know	no answer	skipped on web	Total

1973	-	346	911	220	23	4	-	1504

1974	-	383	821	259	18	3	-	1484

1975	-	354	823	265	42	6	-	1490

1976	-	424	776	263	27	9	-	1499

1977	-	383	874	236	33	4	-	1530

1978	-	307	893	301	27	4	-	1532

1980	-	322	849	253	43	1	-	1468

1982	-	330	1111	364	45	10	-	1860

1983	-	215	974	375	31	4	-	1599

1984	484	166	574	219	19	11	-	1473

1986	-	268	797	373	27	5	-	1470

1987	-	342	1016	411	47	3	-	1819

1988	484	182	529	250	29	7	-	1481

1989	502	171	556	275	29	4	-	1537

1990	473	132	516	219	31	1	-	1372

1991	500	166	545	282	21	3	-	1517

1993	549	114	512	406	21	4	-	1606

1994	981	195	992	783	31	10	-	2992

1996	979	205	918	744	54	4	-	2904

1998	921	176	878	808	46	3	-	2832

2000	921	191	879	767	48	11	-	2817

2002	1853	92	425	382	13	-	-	2765

2004	1936	80	415	378	3	-	-	2812

2006	2521	208	947	810	23	1	-	4510

2008	658	117	616	615	17	-	-	2023

2010	667	140	621	594	19	3	-	2044

2012	636	119	578	613	28	-	-	1974

2014	851	137	781	745	24	-	-	2538

2016	911	159	815	963	17	2	-	2867

2018	785	199	636	713	13	2	-	2348

2021	1360	306	1093	1255	2	-	16	4032

2022	1173	203	929	1210	21	1	7	3544

Total	20145	7132	24600	16351	872	120	23	69243

See also