Confidence in scientific community

consci

Source

General Social Survey https://gss.norc.org

Details

Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? k. Scientific community.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	a great deal	don't know	hardly any	no answer	only some	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	-	552	142	97	9	704	-	1504

1974	-	667	157	99	3	558	-	1484

1975	-	560	159	96	3	672	-	1490

1976	-	638	172	112	13	564	-	1499

1977	-	624	119	83	8	696	-	1530

1978	-	553	126	111	5	737	-	1532

1980	-	607	140	93	6	622	-	1468

1982	-	645	179	133	17	886	-	1860

1983	-	662	101	85	6	745	-	1599

1984	484	436	56	54	13	430	-	1473

1985	1534	-	-	-	-	-	-	1534

1986	-	569	84	113	7	697	-	1470

1987	-	754	135	130	7	793	-	1819

1988	484	385	78	56	5	473	-	1481

1989	502	414	93	63	5	460	-	1537

1990	473	335	79	59	1	425	-	1372

1991	500	411	73	63	3	467	-	1517

1993	549	395	90	71	5	496	-	1606

1994	981	769	105	141	10	986	-	2992

1996	979	752	155	144	7	867	-	2904

1998	921	756	145	149	6	855	-	2832

2000	921	779	149	141	8	819	-	2817

2002	1853	336	55	81	-	440	-	2765

2004	1936	359	30	52	1	434	-	2812

2006	2521	807	98	140	1	943	-	4510

2008	658	524	63	76	1	701	-	2023

2010	667	544	65	79	6	683	-	2044

2012	636	529	73	92	1	643	-	1974

2014	851	680	59	132	5	811	-	2538

2016	911	791	69	117	3	976	-	2867

2018	785	681	55	100	2	725	-	2348

2021	1360	1337	4	176	-	1141	14	4032

2022	1173	930	51	296	3	1088	3	3544

2024	1116	756	61	276	-	1089	11	3309

Total	24408	20537	3220	3710	170	23626	28	75699

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook