Person #2 is r's coworker

cowork2

Source

General Social Survey https://gss.norc.org

Details

Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer. When I read you a name, please tell me all the ways that person is connected to you. How is (NAME) connected to you? PROBE: What other ways? ASK FOR EACH NAME. F2. Second Person is Co-Worker

Values

1 mentioned

2 not mentioned

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1985	-/-/-	full

1987	-/-/-	full

2004	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for cowork2.

See also