Defense spending–r's position

defspdr

Source

General Social Survey https://gss.norc.org

Details

Question Some people believe that we should spend much less money for defense. Suppose these people are at one end of the scale at point number 1. Others feel that defense spending should be greatly increased. Suppose these people are at the other end, at point 7. And, of course, some other people have opinions somewhere in between at points 2, 3, 4, 5, or 6. ENTER CODE IN BOXES. A. Where would you place yourself on this scale, or haven't you thought much about this?

Values

[0] havent thought much

[1] greatly decrease

[7] greatly increase

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-	1980

Overview

For further details see the GSS Data Explorer page for defspdr.

Counts by year:

year	iap	havent thought much	greatly decrease	2	3	4	5	6	greatly increase	don't know	no answer	Total

1980	-	135	47	49	97	308	335	189	278	25	5	1468

Total	0	135	47	49	97	308	335	189	278	25	5	1468

See also