Denomination in which r was raised

denom16

Source

General Social Survey https://gss.norc.org

Details

Question In what religion were you raised? A. What specific denomination is that, if any? As of 2024, DENOM16 has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of DENOM16. The original DENOM16 containing data 1973 to 2022 has been renamed DENOM16_7322.

Values

[1] baptist

[2] methodist

[3] lutheran

[4] presbyterian

[5] episcopal

[6] other denomination

[7] no denomination or non-denominational church

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for denom16.

Counts by year:

year	not available in this year	baptist	methodist	lutheran	presbyterian	episcopal	other denomination	no denomination or non-denominational church	iap	no answer	don't know	skipped on web	Total

1973	-	354	198	131	73	41	151	26	523	7	-	-	1504

1974	-	366	217	119	81	31	146	22	500	2	-	-	1484

1975	-	343	235	149	67	39	167	30	458	1	1	-	1490

1976	-	340	196	107	95	38	167	27	526	3	-	-	1499

1977	-	367	221	145	78	36	163	31	482	7	-	-	1530

1978	-	371	222	127	63	29	171	32	508	9	-	-	1532

1980	-	330	224	105	91	37	168	25	483	5	-	-	1468

1982	-	550	272	128	81	49	181	33	556	10	-	-	1860

1983	-	340	183	143	62	40	201	26	594	10	-	-	1599

1984	-	327	197	92	79	26	189	36	499	28	-	-	1473

1985	-	348	201	96	72	39	199	23	551	5	-	-	1534

1986	-	346	182	119	47	35	193	26	515	5	2	-	1470

1987	-	588	221	84	82	31	252	26	500	34	1	-	1819

1988	-	363	184	88	82	31	171	30	529	3	-	-	1481

1989	-	321	169	121	74	32	186	61	567	6	-	-	1537

1990	-	272	164	111	67	30	190	40	487	11	-	-	1372

1991	-	343	179	109	71	31	193	37	548	6	-	-	1517

1993	-	332	185	116	86	41	237	37	557	15	-	-	1606

1994	-	706	319	220	139	68	309	69	1139	21	2	-	2992

1996	-	666	338	193	120	73	271	60	1165	16	2	-	2904

1998	-	635	317	179	124	68	251	60	1147	48	3	-	2832

2000	-	645	298	159	105	47	272	91	1172	25	3	-	2817

2002	-	567	240	175	104	70	317	107	1159	20	6	-	2765

2004	-	574	256	144	89	57	329	152	1196	10	5	-	2812

2006	-	605	239	154	104	61	356	100	2863	2	26	-	4510

2008	-	416	155	76	69	40	269	86	891	12	9	-	2023

2010	-	434	145	95	86	48	182	143	888	10	13	-	2044

2012	-	399	156	99	48	38	198	122	905	4	5	-	1974

2014	-	475	171	113	78	44	293	156	1185	12	11	-	2538

2016	-	561	200	134	75	54	302	191	1316	11	23	-	2867

2018	-	477	155	107	62	37	237	151	1090	12	20	-	2348

2021	-	566	309	208	147	74	291	289	2108	4	29	7	4032

2022	-	569	201	145	95	63	281	324	1749	18	78	21	3544

2024	-	559	199	174	92	56	250	208	1703	12	34	22	3309

Total	0	15455	7348	4465	2888	1534	7733	2877	31059	404	273	50	74086

See also