Whites hurt by aff. action

discaff

Source

General Social Survey https://gss.norc.org

Details

Question What do you think the chances are these days that a white person won”t get a job or promotion while an equally or less qualified black person gets one instead? Is this very likely, somewhat likely, or not very likely to happen these days?

Values

[1] very likely

[2] somewhat likely

[3] not very likely

[4] very unlikely

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1990, 1994, 2021, 2022, 2024

A/C/-	1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018

Overview

For further details see the GSS Data Explorer page for discaff.

Counts by year:

year	iap	very likely	somewhat likely	not very likely	don't know	no answer	skipped on web	Total

1990	-	343	534	467	16	12	-	1372

1994	2235	201	310	210	26	10	-	2992

1996	1932	236	423	270	33	10	-	2904

1998	950	359	861	577	57	28	-	2832

2000	956	351	816	629	34	31	-	2817

2002	1841	185	437	270	31	1	-	2765

2004	1914	140	414	310	33	1	-	2812

2006	2507	353	915	648	82	5	-	4510

2008	671	215	591	490	51	5	-	2023

2010	763	187	558	488	43	5	-	2044

2012	666	200	539	503	61	5	-	1974

2014	824	240	718	699	54	3	-	2538

2016	978	263	795	759	61	11	-	2867

2018	774	210	615	679	60	10	-	2348

2021	4	476	1549	1923	21	3	56	4032

2022	-	399	1406	1595	109	16	19	3544

Total	17015	4358	11481	10517	772	156	75	44374

See also