A man won't get a job or promotion

discaffm

Source

General Social Survey https://gss.norc.org

Details

Question A. What do you think the chances are these days that a man won't get a job or promotion while an equally or less qualified woman gets one instead? Is this very likely, somewhat likely, somewhat unlikely, or very unlikely these days?

Values

1 very likely

2 somewhat likely

3 not very likely

4 very unlikely

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/-	2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C	1996

Overview

For further details see the GSS Data Explorer page for discaffm.

Counts by year:

year	iap	very likely	somewhat likely	not very likely	very unlikely	don't know	no answer	skipped on web	Total

1996	2199	98	259	194	103	49	2	-	2904

2000	1868	117	302	270	156	49	55	-	2817

2002	2327	63	168	123	74	10	-	-	2765

2004	2359	59	170	138	70	15	1	-	2812

2006	3516	122	366	322	164	19	1	-	4510

2008	1375	81	270	192	87	17	1	-	2023

2010	1313	79	282	244	102	21	3	-	2044

2012	1314	72	247	216	104	17	4	-	1974

2014	1712	92	314	262	133	25	-	-	2538

2016	1931	106	324	307	176	23	-	-	2867

2018	1554	89	273	264	153	14	1	-	2348

2021	2689	119	422	511	284	2	1	4	4032

2022	2339	108	403	411	258	20	1	4	3544

Total	26496	1205	3800	3454	1864	281	70	8	37178

See also