Doctors unable to tell me all the options

doc12

Source

General Social Survey https://gss.norc.org

Details

Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement. L. I worry that my doctor is being prevented from telling me the full range of options for my treatment.

Values

[1] strongly agree

[2] agree

[3] uncertain

[4] disagree

[5] strong disagree

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1998

Overview

For further details see the GSS Data Explorer page for doc12.

Counts by year:

year	iap	strongly agree	agree	uncertain	disagree	strong disagree	don't know	no answer	Total

1998	1445	45	254	237	690	114	41	6	2832

Total	1445	45	254	237	690	114	41	6	2832

See also