How many in family earned money?

earnrs

Source

General Social Survey https://gss.norc.org

Details

Question 36. Just thinking about your family now – those people in the househol who are related to you – how many persons in the family, including yourself, earned any money last year from any job or employment?

Overview

For further details see the official GSS documentation.

Counts by year:

year	0	1	2	3 or more	no answer	iap	skipped on web	don't know	Total

1972	186	682	506	206	33	-	-	-	1613

1973	162	608	506	218	10	-	-	-	1504

1974	192	595	477	205	15	-	-	-	1484

1975	201	593	483	205	8	-	-	-	1490

1976	233	627	449	182	8	-	-	-	1499

1977	214	613	494	205	4	-	-	-	1530

1978	229	629	479	187	8	-	-	-	1532

1980	249	544	483	173	19	-	-	-	1468

1982	291	752	577	187	53	-	-	-	1860

1983	234	621	544	183	17	-	-	-	1599

1984	233	611	441	170	18	-	-	-	1473

1985	276	578	499	170	11	-	-	-	1534

1986	297	506	478	169	20	-	-	-	1470

1987	291	709	585	218	16	-	-	-	1819

1988	275	560	452	184	10	-	-	-	1481

1989	286	553	475	211	12	-	-	-	1537

1990	248	508	456	152	8	-	-	-	1372

1991	320	581	459	146	11	-	-	-	1517

1993	298	577	556	155	20	-	-	-	1606

1994	493	1190	981	273	55	-	-	-	2992

1996	432	1213	920	276	63	-	-	-	2904

1998	429	1149	949	282	23	-	-	-	2832

2000	471	1162	942	235	7	-	-	-	2817

2002	466	1097	918	265	19	-	-	-	2765

2004	394	1131	996	284	7	-	-	-	2812

2006	645	1890	1502	452	21	-	-	-	4510

2008	328	791	668	233	3	-	-	-	2023

2010	354	825	632	219	14	-	-	-	2044

2012	369	770	637	186	12	-	-	-	1974

2014	467	1023	797	237	14	-	-	-	2538

2016	571	1104	891	282	19	-	-	-	2867

2018	423	891	762	260	12	-	-	-	2348

2021	832	1421	1216	426	-	58	79	-	4032

2022	736	1392	971	373	29	6	37	-	3544

2024	718	1314	842	370	9	-	44	12	3309

Total	12843	29810	24023	8179	608	64	160	12	75699

Values

3 3 or more

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook