Highest year of school completed

educ

Source

General Social Survey https://gss.norc.org

Details

Question ASK ALL PARTS OF QUESTION ABOUT RESPONDENT BEFORE GOING ON TO ASK ABOUT R'S FATHER; AND THEN R'S MOTHER; THEN R'S SPOUSE, IF R IS CURRENTLY MARRIED. A. What is the highest grade in elementary school or high school that (you/your father/ your mother/your (husband/wife)) finished and got credit for? CODE EXACT GRADE. B. IF FINISHED 9th-12th GRADE OR DK*: Did (you/he/she) ever get a high school diploma or a GED certificate? (SEE D BELOW.) (See REMARKS) C. Did (you/he/she) complete one or more years of college for credit–not including schooling such as business college, technical or vocational school? IF YES: How many years did (you/he/she) complete? Do you (Does (he/she)) have any college degrees? (IF YES: What degree or degrees?) CODE HIGHEST DEGREE EARNED.

Values

[0] no formal schooling

[1] 1st grade

[2] 2nd grade

[3] 3rd grade

[4] 4th grade

[5] 5th grade

[6] 6th grade

[7] 7th grade

[8] 8th grade

[9] 9th grade

[10] 10th grade

[11] 11th grade

[12] 12th grade

[13] 1 year of college

[14] 2 years of college

[15] 3 years of college

[16] 4 years of college

[17] 5 years of college

[18] 6 years of college

[19] 7 years of college

[20] 8 or more years of college

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C/D	2006

Full/-	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for educ.

Counts by year:

year	no formal schooling	1	10	11	12	13	14	15	16	17	18	19	2	20	3	4	5	6	7	8	9	no answer	don't know	Total

1972	14	4	116	119	507	92	124	53	119	26	22	12	13	12	19	19	30	38	48	142	79	5	-	1613

1973	7	4	97	108	485	92	126	43	130	30	13	17	7	13	14	18	21	30	42	135	67	5	-	1504

1974	6	1	107	90	491	104	109	43	128	39	28	9	1	17	9	13	17	30	45	132	62	3	-	1484

1975	7	2	101	126	499	115	101	34	132	24	23	9	3	12	8	7	25	28	42	124	65	2	1	1490

1976	4	-	110	93	510	94	103	36	134	31	23	17	4	14	15	19	19	24	46	128	69	4	2	1499

1977	4	-	106	114	507	90	103	44	128	32	33	11	1	15	11	20	24	37	48	120	72	3	7	1530

1978	6	4	96	90	541	105	120	55	119	34	34	17	1	16	8	20	13	19	39	126	63	3	3	1532

1980	1	1	91	83	503	97	109	48	134	37	38	14	6	18	7	13	18	27	38	101	79	1	4	1468

1982	7	2	106	140	620	142	159	66	146	33	36	25	5	16	15	8	21	48	51	119	87	4	4	1860

1983	5	1	73	127	547	97	155	58	179	39	40	16	4	18	10	12	12	18	34	90	62	-	2	1599

1984	3	3	99	91	491	108	121	74	149	40	31	20	4	25	9	11	11	19	26	77	58	3	-	1473

1985	3	2	86	100	511	123	133	74	141	46	44	12	6	32	9	8	9	19	27	86	63	-	-	1534

1986	6	-	90	113	488	103	127	50	163	41	36	17	2	25	9	8	13	24	23	78	53	1	-	1470

1987	9	2	94	98	588	149	182	66	182	62	48	16	4	27	16	10	19	30	32	97	78	6	4	1819

1988	1	1	71	104	455	124	158	66	164	32	42	18	1	28	6	12	12	23	27	75	58	-	3	1481

1989	3	2	70	95	489	125	184	76	152	40	61	18	-	27	4	9	9	12	21	82	51	5	2	1537

1990	4	1	53	81	436	106	167	62	163	44	49	13	5	28	1	9	5	19	14	58	52	-	2	1372

1991	2	-	73	85	461	130	175	73	194	43	45	22	-	30	5	5	6	12	25	68	56	7	-	1517

1993	3	-	58	86	477	154	177	75	224	50	73	25	4	27	-	7	7	20	27	61	47	-	4	1606

1994	4	1	102	168	929	277	321	146	433	97	119	46	3	64	6	12	15	19	29	109	85	4	3	2992

1996	4	-	121	155	849	296	339	153	392	120	125	42	-	86	8	6	13	14	23	80	69	1	8	2904

1998	2	-	113	138	851	270	350	146	412	86	109	41	5	69	10	9	8	23	21	82	75	4	8	2832

2000	1	-	106	153	823	290	374	127	393	90	114	39	3	68	6	6	9	22	20	95	69	9	-	2817

2002	5	2	88	137	818	265	374	157	377	93	128	43	15	70	2	6	8	25	14	66	60	7	5	2765

2004	4	1	72	156	734	265	403	174	450	106	138	51	7	93	3	2	4	17	17	54	59	2	-	2812

2006	22	4	152	215	1204	422	628	212	687	167	208	78	28	112	13	11	23	69	32	85	127	9	2	4510

2008	6	-	59	122	581	158	258	94	322	51	124	38	2	52	5	10	5	29	14	35	53	2	3	2023

2010	5	1	66	108	558	186	229	109	334	71	101	33	5	72	4	9	6	30	12	45	55	4	1	2044

2012	3	2	59	101	540	163	261	99	307	80	92	41	3	71	6	10	4	27	8	48	47	1	1	1974

2014	8	1	90	134	689	193	318	133	452	78	139	52	3	100	6	3	9	25	8	43	53	-	1	2538

2016	2	3	90	118	824	242	359	137	485	108	149	63	3	110	3	2	4	31	18	48	59	8	1	2867

2018	4	2	65	95	657	183	313	127	430	97	119	45	4	72	10	5	3	20	8	35	51	1	2	2348

2021	9	1	52	83	829	277	542	208	942	258	351	113	2	216	3	1	2	15	5	25	32	48	18	4032

2022	3	1	78	116	909	268	506	189	697	167	210	79	4	148	8	6	6	23	12	37	57	20	-	3544

Total	177	49	3010	3942	21401	5905	8208	3307	9994	2392	2945	1112	158	1803	268	326	410	866	896	2786	2172	172	91	72390

See also