Most important environment problem

enprbus

Source

General Social Survey https://gss.norc.org

Details

Question D. Here is a list of some different environmental problems. Which problem, if any, do you think is the most important for America as a whole?

Values

[1] air pollution

[2] chemicals and pesticides

[3] water shortage

[4] water pollution

[5] nuclear waste

[6] domestic waste disposal

[7] climate change

[8] genetically modified foods

[9] using up our natural resources

[10] none of these

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B	2010, 2021

Overview

For further details see the GSS Data Explorer page for enprbus.

Counts by year:

year	iap	air pollution	chemicals and pesticides	water shortage	water pollution	nuclear waste	domestic waste disposal	climate change	genetically modified foods	using up our natural resources	none of these	can't choose	no answer	skipped on web	Total

2010	614	199	112	95	134	88	60	127	61	344	22	181	7	-	2044

2021	2003	141	140	76	131	40	87	642	115	177	93	217	160	10	4032

Total	2617	340	252	171	265	128	147	769	176	521	115	398	167	10	6076

See also