Should govt reduce income differences

eqwlth

Source

General Social Survey https://gss.norc.org

Details

Question A. Some people think that the government in Washington ought to reduce the income differences between the rich and the poor, perhaps by raising the taxes of wealthy families or by giving income assistance to the poor. Others think that the government should not concern itself with reducing this income difference between the rich and the poor. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that the government ought to reduce the income differences between rich and poor, and a score of 7 meaning that the government should not concern itself with reducing income differences. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):

Values

1 the government should reduce income differences

7 the government should not concern itself with reducing income differences

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1978, 1980, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for eqwlth.

Counts by year:

year	iap	the government should reduce income differences	2	3	4	5	6	the government should not concern itself with reducing income differences	don't know	no answer	skipped on web	Total

1978	771	146	82	130	157	81	57	94	11	3	-	1532

1980	-	246	136	238	290	180	106	227	40	5	-	1468

1983	-	315	171	256	278	177	131	230	38	3	-	1599

1984	-	306	179	223	251	191	117	171	24	11	-	1473

1986	-	338	130	247	302	160	89	182	19	3	-	1470

1987	-	408	160	307	369	218	103	221	25	8	-	1819

1988	484	193	95	180	197	121	75	112	21	3	-	1481

1989	502	182	132	199	206	116	72	108	18	2	-	1537

1990	473	188	107	157	190	80	56	93	25	3	-	1372

1991	500	201	126	173	200	123	75	89	27	3	-	1517

1993	549	184	122	201	187	130	81	130	18	4	-	1606

1994	981	289	183	322	420	292	167	303	27	8	-	2992

1996	979	331	200	303	409	237	158	234	45	8	-	2904

1998	921	282	181	331	402	213	157	303	35	7	-	2832

2000	921	301	217	294	374	258	160	245	38	9	-	2817

2002	1853	175	84	164	177	125	66	109	11	1	-	2765

2004	1936	171	80	152	168	107	64	123	10	1	-	2812

2006	2521	415	187	320	410	254	149	221	28	5	-	4510

2008	658	319	116	234	240	178	93	169	15	1	-	2023

2010	667	244	106	219	249	215	99	227	12	6	-	2044

2012	636	324	132	194	240	148	84	202	12	2	-	1974

2014	851	365	165	251	284	232	121	253	15	1	-	2538

2016	911	451	248	297	328	255	147	210	15	5	-	2867

2018	785	347	154	272	260	221	119	166	21	3	-	2348

2021	1360	882	242	325	389	249	153	421	2	-	9	4032

2022	1173	783	198	322	337	237	127	347	12	4	4	3544

Total	20432	8386	3933	6311	7314	4798	2826	5190	564	109	13	59876

See also