Have sex other than spouse while married

evstray

Source

General Social Survey https://gss.norc.org

Details

Question Have you ever had sex with someone other than your husband or wife while you were married?

Values

[1] yes

[2] no

[3] never married

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2021

B/C/-	2016, 2018

Overview

For further details see the GSS Data Explorer page for evstray.

Counts by year:

year	iap	yes	no	never married	no answer	don't know	Total

1991	221	146	853	279	18	-	1517

1993	114	195	982	258	57	-	1606

1994	201	354	1781	586	70	-	2992

1996	258	369	1618	615	43	1	2904

1998	383	324	1481	587	57	-	2832

2000	417	308	1418	602	71	1	2817

2002	502	311	1307	594	49	2	2765

2004	573	291	1360	540	45	3	2812

2006	2096	350	1414	623	26	1	4510

2008	240	214	1046	491	28	4	2023

2010	202	227	1058	523	31	3	2044

2012	238	213	1003	486	32	2	1974

2014	188	285	1375	636	53	1	2538

2016	1096	217	1003	523	28	-	2867

2018	942	159	806	425	14	2	2348

2021	1615	297	1408	646	49	17	4032

Total	9286	4260	19913	8414	671	37	42581

See also