People fair or try to take advantage

fair

Source

General Social Survey https://gss.norc.org

Details

Question Do you think most people would try to take advantage of you if they got a chance, or would they try to be fair?

Values

[1] would take advantage of you

[2] would try to be fair

[3] depends

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1972, 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

Overview

For further details see the GSS Data Explorer page for fair.

Counts by year:

year	would take advantage of you	would try to be fair	depends	don't know	no answer	iap	not available in this year	Total

1972	542	954	96	19	2	-	-	1613

1973	562	861	72	8	1	-	-	1504

1975	454	917	105	12	2	-	-	1490

1976	542	887	62	8	-	-	-	1499

1978	454	983	78	10	7	-	-	1532

1980	505	882	63	18	-	-	-	1468

1983	553	950	82	14	-	-	-	1599

1984	507	913	47	3	3	-	-	1473

1986	490	911	55	9	5	-	-	1470

1987	751	966	75	18	9	-	-	1819

1988	335	598	51	12	1	484	-	1481

1989	376	599	54	6	-	502	-	1537

1990	325	515	53	6	-	473	-	1372

1991	366	579	59	7	6	500	-	1517

1993	384	589	74	8	2	549	-	1606

1994	787	1065	134	16	9	981	-	2992

1996	804	954	132	22	13	979	-	2904

1998	725	997	167	19	3	921	-	2832

2000	718	973	172	24	9	921	-	2817

2002	361	451	97	3	-	1853	-	2765

2004	347	444	80	4	1	1936	-	2812

2006	781	962	236	10	-	2521	-	4510

2008	553	701	105	5	1	658	-	2023

2010	524	720	121	10	2	667	-	2044

2012	553	684	89	11	1	636	-	1974

2014	675	869	137	6	-	851	-	2538

2016	797	1028	114	11	6	911	-	2867

2018	671	767	112	12	1	785	-	2348

2021	-	-	-	-	-	-	4032	4032

2022	566	533	107	12	1	2325	-	3544

Total	16008	23252	2829	323	85	19453	4032	65982

See also