Hispanics committed to fair and equal treatment

fairhsps

Source

General Social Survey https://gss.norc.org

Details

Question In the next statement a score of 1 means that you think almost all of the people in the group have a "commitment to the fair and equal treatment of all groups in society." A score of 7 means that you think almost everyone in the group "lacks a commitment to the fair and equal treatment of all groups in society." A score of 4 means that you think that the group is not towards one end or the other, and of course you may choose any number in between that comes closest to where you think people in the group stand. D. Hispanics or Latin Americans?

Values

[1] commitment to fair and equal treatment

[7] lacks commitment to fair and equal treatment

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2000

Overview

For further details see the GSS Data Explorer page for fairhsps.

Counts by year:

year	iap	commitment to fair and equal treatment	2	3	4	5	6	lacks commitment to fair and equal treatment	don't know	no answer	Total

2000	1419	49	94	211	522	247	80	33	150	12	2817

Total	1419	49	94	211	522	247	80	33	150	12	2817

See also