Gender of 2nd person

gender2

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what? 2. Gender of second person

Values

1 male

2 female

3 other

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/D	2006

Full/-	1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for gender2.

Counts by year:

year	iap	male	female	no answer	not available in this release	skipped on web	Total

1975	203	100	1187	-	-	-	1490

1976	893	53	553	-	-	-	1499

1977	256	128	1146	-	-	-	1530

1978	295	121	1116	-	-	-	1532

1980	285	134	1048	1	-	-	1468

1982	412	199	1241	8	-	-	1860

1983	311	152	1131	5	-	-	1599

1984	327	151	990	5	-	-	1473

1985	339	157	1038	-	-	-	1534

1986	313	169	988	-	-	-	1470

1987	424	242	1149	4	-	-	1819

1988	328	173	978	2	-	-	1481

1989	327	202	989	19	-	-	1537

1990	323	199	843	7	-	-	1372

1991	374	205	937	1	-	-	1517

1993	373	216	1014	3	-	-	1606

1994	752	393	1845	2	-	-	2992

1996	743	463	1671	27	-	-	2904

1998	777	456	1586	13	-	-	2832

2000	741	720	1340	16	-	-	2817

2002	883	668	1212	2	-	-	2765

2004	707	715	1390	-	-	-	2812

2006	1226	1132	2150	2	-	-	4510

2008	527	528	967	1	-	-	2023

2010	595	510	936	3	-	-	2044

2012	533	502	936	3	-	-	1974

2014	678	669	1185	6	-	-	2538

2016	849	793	1223	2	-	-	2867

2018	677	666	1000	5	-	-	2348

2021	-	-	-	-	4032	-	4032

2022	2666	320	557	-	-	1	3544

Total	18137	11136	34346	137	4032	1	67789

See also