Gender of 4th person

gender4

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what? 2. Gender of fourth person

Values

1 male

2 female

3 other

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1975	-/-/-/-	full

1976	-/-/-/-	full

1977	-/-/-/-	full

1978	-/-/-/-	full

1980	-/-/-/-	full

1982	-/-/-/-	full

1983	-/-/-/-	full

1984	-/-/-/-	full

1985	-/-/-/-	full

1986	-/-/-/-	full

1987	-/-/-/-	full

1988	A/B/C/-	full

1989	A/B/C/-	full

1990	A/B/C/-	full

1991	A/B/C/-	full

1993	A/B/C/-	full

1994	A/B/C/-	full

1996	A/B/C/-	full

1998	A/B/C/-	full

2000	A/B/C/-	full

2002	A/B/C/-	full

2004	A/B/C/-	full

2006	A/B/C/D	full

2008	A/B/C/-	full

2010	A/B/C/-	full

2012	A/B/C/-	full

2014	A/B/C/-	full

2016	A/B/C/-	full

2018	A/B/C/-	full

2021	A/B/C/-	full

2022	A/B/C/-	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	female	male	no answer	don't know	other	skipped on web	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	952	254	284	-	-	-	-	-	1490

1976	1255	115	128	1	-	-	-	-	1499

1977	1007	272	250	1	-	-	-	-	1530

1978	1017	242	271	2	-	-	-	-	1532

1980	1040	209	217	2	-	-	-	-	1468

1982	1343	271	243	3	-	-	-	-	1860

1983	1104	244	247	4	-	-	-	-	1599

1984	1062	182	226	3	-	-	-	-	1473

1985	1112	208	213	1	-	-	-	-	1534

1986	1032	199	238	1	-	-	-	-	1470

1987	1282	254	280	3	-	-	-	-	1819

1988	1097	191	193	-	-	-	-	-	1481

1989	1098	198	231	10	-	-	-	-	1537

1990	1032	179	158	3	-	-	-	-	1372

1991	1125	187	203	2	-	-	-	-	1517

1993	1190	192	221	3	-	-	-	-	1606

1994	2274	342	369	7	-	-	-	-	2992

1996	2184	359	342	19	-	-	-	-	2904

1998	2195	316	309	12	-	-	-	-	2832

2000	2127	334	338	18	-	-	-	-	2817

2002	2266	251	247	1	-	-	-	-	2765

2004	2214	270	320	8	-	-	-	-	2812

2006	3502	459	534	3	12	-	-	-	4510

2008	1547	225	238	4	9	-	-	-	2023

2010	1627	182	225	4	6	-	-	-	2044

2012	1526	222	219	5	2	-	-	-	1974

2014	2040	230	253	7	8	-	-	-	2538

2016	2353	231	267	10	6	-	-	-	2867

2018	1899	217	223	6	3	-	-	-	2348

2021	3166	423	430	-	-	11	2	-	4032

2022	3210	168	163	-	-	2	1	-	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	56479	7626	8080	143	46	13	3	3309	75699

See also