Opinion of how people get ahead

getahead

Source

General Social Survey https://gss.norc.org

Details

Question Some people say that people get ahead by their own hard work; others say that lucky breaks or help from other people are more important. Which do you think is most important?

Values

[1] hard work most important

[2] hard work, luck equally important

[3] luck most important

[4] other

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/C/D	2006

Full	1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987

Overview

For further details see the GSS Data Explorer page for getahead.

Counts by year:

year	iap	hard work most important	hard work, luck equally important	luck most important	other	don't know	no answer	not available in this year	Total

1973	-	973	355	150	17	7	2	-	1504

1974	-	902	423	127	19	9	4	-	1484

1976	-	931	357	197	-	11	3	-	1499

1977	-	927	430	159	-	10	4	-	1530

1980	-	927	409	118	-	11	3	-	1468

1982	-	1067	505	253	-	28	7	-	1860

1984	-	976	264	219	-	7	7	-	1473

1985	-	1003	286	220	-	22	3	-	1534

1987	-	1180	330	288	-	14	7	-	1819

1988	504	651	204	116	-	5	1	-	1481

1989	504	676	202	143	-	8	4	-	1537

1990	455	598	190	122	-	5	2	-	1372

1991	524	651	203	125	-	7	7	-	1517

1993	531	703	231	134	-	3	4	-	1606

1994	996	1375	386	213	-	14	8	-	2992

1996	981	1333	344	218	-	20	8	-	2904

1998	950	1247	417	195	-	13	10	-	2832

2000	956	1194	451	193	-	13	10	-	2817

2002	1841	578	248	92	-	5	1	-	2765

2004	1914	582	232	79	-	4	1	-	2812

2006	1771	1868	544	314	-	11	2	-	4510

2008	671	886	294	162	-	10	-	-	2023

2010	763	884	265	120	-	11	1	-	2044

2012	666	910	264	128	-	4	2	-	1974

2014	824	1188	340	172	-	13	1	-	2538

2016	980	1330	316	230	-	7	4	-	2867

2018	774	1127	239	193	-	13	2	-	2348

2021	-	-	-	-	-	-	-	4032	4032

2022	2348	709	323	152	-	11	1	-	3544

Total	18953	27376	9052	4832	36	296	109	4032	64686

See also