Contributed money in education

giveduc

Source

General Social Survey https://gss.norc.org

Details

Question Listed on this card are examples of many different fields in which people and families contribute money or other property for charitable purposes. I mean making a voluntary contribution and not with the intention of making a profit or obtaining goods and/or services for yourself. In which, if any, of the fields listed on this card have you and the members of your family or household contributed some money or other property in 1995? Just read off the letter of each field (DON'T READ. RECORD BELOW UNDER GIVHLTH-GIVOTH.) B. Education

Values

1 yes

2 no

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for giveduc.

See also