Importance of attending church regularly

gochurch

Source

General Social Survey https://gss.norc.org

Details

Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it … A. To attend regularly religious services at church or synagogue.

Values

[1] very important

[5] not very important

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1998

Overview

For further details see the GSS Data Explorer page for gochurch.

Counts by year:

year	iap	very important	2	3	4	not very important	don't know	no answer	Total

1988	-	418	218	298	205	318	13	11	1481

1998	1387	428	183	292	158	345	27	12	2832

Total	1387	846	401	590	363	663	40	23	4313

See also