Environmental threats exaggerated

grnexagg

Source

General Social Survey https://gss.norc.org

Details

Question How much do you agree or disagree with each of these statements? C. Many of the claims about environmental threats are exaggerated.

Values

1 strongly agree

2 agree

3 neither agree nor disagree

4 disagree

5 strongly disagree

NA(d) can't choose / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2000	A/B/C	full

2010	A/B/-	partial

2021	A/B/-	partial

Link at the GSS

For details see the GSS Data Explorer page for grnexagg.

See also