Pay higher taxes to help envir?

grntaxes

Source

General Social Survey https://gss.norc.org

Details

Question B. And how willing would you be to pay much higher taxes in order to protect the environment?

Values

1 very willing

2 fairly willing

3 neither willing nor unwilling

4 fairly unwilling

5 very unwilling

NA(d) can't choose / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/-	2010, 2021

A/B/C	1993, 1994, 2000

Overview

For further details see the GSS Data Explorer page for grntaxes.

Counts by year:

year	iap	very willing	fairly willing	neither willing nor unwilling	fairly unwilling	very unwilling	can't choose	no answer	skipped on web	Total

1993	49	108	484	300	379	200	34	52	-	1606

1994	1606	77	381	286	360	230	26	26	-	2992

2000	1541	72	297	315	225	258	51	58	-	2817

2010	614	82	377	283	283	343	55	7	-	2044

2021	2003	129	479	433	361	373	61	172	21	4032

Total	5813	468	2018	1617	1608	1404	227	315	21	13491

See also