Participated in activity of political party

grppol

Source

General Social Survey https://gss.norc.org

Details

Question People sometimes belong to different kinds of groups or associations. The next few questions contain different types of groups. For each type of group, please select a response to say whether you have participated in the activities of this group in the past 12 months. A. A political party, club or association

Values

1 i have participated more than twice

2 i have participated once or twice

3 i belong to such a group but never participate

4 i do not belong to such a group

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2002	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for grppol.

See also