Data Dictionary for the GSS Cumulative Data File 1972-2024

A tibble containing information on the variables in the GSS Cumulative Data File.

Usage

data(gss_dict)

Format

A tibble.

`pos`

Column-index position of the variable in `gss_all`. (Integer.)

`variable`

The name of the variable. (Character.)

`label`

The variable's label. 'r' means 'Respondent'. (Character.)

`missing`

Number of missing observations. (Integer.)

`var_doc_label`

The variable's label, per the Berkeley SDA web pages. Should correspond directly to `label`. (Character)

`value_labels`

The value labels and associated numeric values for the variable. Numeric values are given in square brackets. (Character.)

`var_text`

The text of the question asked of the respondent. (Character.)

`years`

Which years the question was asked. This information is stored as a tibble in a list column. The tibble contains a column for the `year` and a TRUE/FALSE coumn for `present`. (List.)

`var_yrtab`

Crosstab of responses by year. Stored in a list-column. The tibbles contain `year` values in the rows and value labels across their columns, with response freqencies in the cells as counts. (List.)

`var_ballots`

Crosstab of ballots by year. Stored in a list-column. The tibbles contain `year` values in the rows and `iap` and the ballots labels across the columns. Counts in `iap` mean the question was not asked that year, or the ballot system was not used. Cells are the number of respondents, if any. (Grouped tibble.)

`col_type`

The column type of the variable in `gss_all`. (Character.)

`var_type`

The variable type of the variable in `gss_all`. (Character.)

`var_na_codes`

The missing value codes associated with the variable in `gss_all`. (Character.)

Source

National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).