Felt happy?

hapfeel

Source

General Social Survey https://gss.norc.org

Details

Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS) (IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.") D. Felt happy?

Values

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for hapfeel.

See also