Happiness of marriage

hapmar

Source

General Social Survey https://gss.norc.org

Details

Question A. Taking things all together, how would you describe your marriage? Would you say that your marriage is very happy, pretty happy, or not too happy?

Values

[1] very happy

[2] pretty happy

[3] not too happy

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for hapmar.

Counts by year:

year	iap	very happy	pretty happy	not too happy	don't know	no answer	skipped on web	Total

1973	428	727	317	28	1	3	-	1504

1974	419	732	290	37	2	4	-	1484

1975	488	671	297	27	2	5	-	1490

1976	525	650	300	23	-	1	-	1499

1977	555	632	298	35	5	5	-	1530

1978	571	622	307	25	1	6	-	1532

1980	579	599	258	25	2	5	-	1468

1982	867	625	326	30	4	8	-	1860

1983	634	600	330	31	-	4	-	1599

1984	644	541	259	25	-	4	-	1473

1985	663	488	346	29	1	7	-	1534

1986	644	517	274	28	1	6	-	1470

1987	903	573	305	26	-	12	-	1819

1988	693	492	268	27	-	1	-	1481

1989	690	508	311	21	3	4	-	1537

1990	643	468	240	16	1	4	-	1372

1991	713	506	268	26	1	3	-	1517

1993	747	527	301	25	3	3	-	1606

1994	1451	927	555	47	4	8	-	2992

1996	1514	854	495	30	4	7	-	2904

1998	1486	855	445	37	1	8	-	2832

2000	1538	789	434	41	6	9	-	2817

2002	2161	370	215	18	1	-	-	2765

2004	2130	417	234	25	-	6	-	2812

2006	3094	861	512	37	1	5	-	4510

2008	1046	596	343	30	1	7	-	2023

2010	1152	544	314	28	2	4	-	2044

2012	1074	589	282	26	-	3	-	1974

2014	1376	691	425	39	2	5	-	2538

2016	1654	726	430	48	1	8	-	2867

2018	1348	638	324	30	1	7	-	2348

2021	2031	1211	701	74	-	3	12	4032

2022	2073	881	502	72	1	12	3	3544

Total	36534	21427	11506	1066	52	177	15	70777

See also