Should govt aid black people?

helpblk

Source

General Social Survey https://gss.norc.org

Details

Question Now look at CARD BF. Some people think that Black people have been discriminated against for so long that the government has a special obligation to help improve their living standards. Others believe that the government should not be giving special treatment to Black people.A. Where would you place yourself on this scale, or haven't you made up your mind on this?

Values

1 government should help

3 agree with both

5 no special treatment

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1975	-/-/-	full

1983	-/-/-	full

1984	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	B/C/-	partial

1989	B/C/-	partial

1990	A/B/C	full

1991	B/C/-	partial

1993	B/C/-	partial

1994	B/C/-	partial

1996	B/C/-	partial

1998	B/C/-	partial

2000	B/C/-	partial

2002	B/C/-	partial

2004	B/C/-	partial

2006	B/C/-	partial

2008	B/C/-	partial

2010	B/C/-	partial

2012	B/C/-	partial

2014	B/C/-	partial

2016	B/C/-	partial

2018	B/C/-	partial

2021	B/C/-	partial

2022	B/C/-	partial

2024	B/C/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	2	4	agree with both	don't know	government should help	no answer	no special treatment	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	1484

1975	-	126	179	311	40	240	2	592	-	1490

1976	1499	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	1860

1983	-	151	322	406	66	120	7	527	-	1599

1984	-	137	248	439	51	136	23	439	-	1473

1985	1534	-	-	-	-	-	-	-	-	1534

1986	-	140	260	420	39	120	4	487	-	1470

1987	-	204	279	519	46	278	8	485	-	1819

1988	484	99	172	285	30	74	4	333	-	1481

1989	502	97	190	278	36	90	2	342	-	1537

1990	-	138	220	452	46	139	3	374	-	1372

1991	500	113	189	311	32	104	4	264	-	1517

1993	549	105	217	333	30	75	7	290	-	1606

1994	981	146	465	584	60	164	11	581	-	2992

1996	979	163	443	552	63	155	10	539	-	2904

1998	921	194	399	591	70	124	7	526	-	2832

2000	921	173	361	598	81	176	15	492	-	2817

2002	1853	72	177	282	20	76	2	283	-	2765

2004	1936	62	158	278	12	66	5	295	-	2812

2006	2521	147	365	662	37	209	4	565	-	4510

2008	658	107	253	440	50	125	7	383	-	2023

2010	667	115	266	405	49	124	6	412	-	2044

2012	636	117	231	394	47	93	9	447	-	1974

2014	851	145	293	505	58	150	7	529	-	2538

2016	913	220	371	608	72	246	7	430	-	2867

2018	785	188	253	507	61	235	9	310	-	2348

2021	1382	407	315	679	11	560	3	641	34	4032

2022	1174	287	256	707	60	498	13	535	14	3544

2024	1116	244	240	676	52	410	12	537	22	3309

Total	34353	4097	7122	12222	1219	4787	191	11638	70	75699

See also