Should govt do more or less?

helpnot

Source

General Social Survey https://gss.norc.org

Details

Question Now look at CARD BD. Some people think that the government in Washington is trying to do too many things that should be left to individuals and private businesses. Others disagree and think that the government should Do even more to solve our country's problems. Still others have opinions somewhere in between. A. Where would you place yourself on this scale, or haven't you made up your mind on this?

Values

[1] government should do more

[3] agree with both

[5] government does too much

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1975, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for helpnot.

Counts by year:

year	iap	government should do more	2	agree with both	4	government does too much	don't know	no answer	skipped on web	Total

1975	-	371	161	433	173	245	104	3	-	1490

1983	-	198	175	574	280	258	110	4	-	1599

1984	-	204	184	537	241	204	79	24	-	1473

1986	-	180	185	604	228	189	81	3	-	1470

1987	-	328	260	657	241	220	104	9	-	1819

1988	484	137	138	392	141	128	60	1	-	1481

1989	502	134	147	390	159	131	72	2	-	1537

1990	473	118	135	364	133	95	54	-	-	1372

1991	500	139	149	395	148	111	70	5	-	1517

1993	549	138	156	401	162	155	39	6	-	1606

1994	981	235	271	709	358	325	98	15	-	2992

1996	979	229	239	722	324	304	100	7	-	2904

1998	921	202	231	749	339	282	105	3	-	2832

2000	921	213	226	709	331	313	95	9	-	2817

2002	1853	129	93	431	123	108	27	1	-	2765

2004	1936	137	86	381	123	126	22	1	-	2812

2006	2521	339	199	859	288	250	51	3	-	4510

2008	658	226	174	539	175	194	53	4	-	2023

2010	667	205	165	498	189	257	58	5	-	2044

2012	636	199	147	530	171	242	45	4	-	1974

2014	851	223	174	624	258	342	65	1	-	2538

2016	913	329	210	754	273	292	92	4	-	2867

2018	785	231	219	644	207	190	68	4	-	2348

2021	1377	539	386	977	319	388	10	-	36	4032

2022	1174	520	260	906	255	335	71	7	16	3544

Total	19681	5903	4770	14779	5639	5684	1733	125	52	58366

See also