To help others

helpoth

Source

General Social Survey https://gss.norc.org

Details

Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO HELP OTHERS WHEN THEY NEED HELP

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	fifth	first	fourth	no answer	second	third	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	1534

1986	730	30	100	143	8	216	243	-	1470

1987	-	74	209	371	19	516	630	-	1819

1988	484	44	127	206	3	332	285	-	1481

1989	502	40	123	196	35	312	329	-	1537

1990	473	29	119	179	28	273	271	-	1372

1991	500	33	126	175	35	316	332	-	1517

1993	549	37	135	208	26	337	314	-	1606

1994	981	61	252	381	33	567	717	-	2992

1996	979	50	241	373	30	589	642	-	2904

1998	921	49	253	375	31	567	636	-	2832

2000	921	52	247	344	33	590	630	-	2817

2002	1853	23	127	154	13	279	316	-	2765

2004	1936	21	151	126	8	281	289	-	2812

2006	3508	23	159	170	7	304	339	-	4510

2008	658	24	217	188	14	443	479	-	2023

2010	667	19	237	211	28	440	442	-	2044

2012	636	19	214	207	15	452	431	-	1974

2014	851	26	311	207	17	528	598	-	2538

2016	911	23	357	230	22	656	668	-	2867

2018	785	16	312	173	21	533	508	-	2348

2021	1360	22	533	214	99	878	926	-	4032

2022	1173	31	470	181	77	768	844	-	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	39964	746	5020	5012	602	10177	10869	3309	75699

Values

1 first

2 second

3 third

4 fourth

5 fifth

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook