Beaten as child or adult

hitage

Source

General Social Survey https://gss.norc.org

Details

Question Have you ever been punched or beaten by another person? A. Did this happen to you as a child or as an adult?

Values

[1] child

[2] adult

[3] both

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for hitage.

Counts by year:

year	iap	child	adult	both	no answer	don't know	Total

1973	1091	179	124	109	1	-	1504

1975	1017	214	155	99	3	2	1490

1976	1075	176	151	96	-	1	1499

1978	995	237	178	122	-	-	1532

1980	977	208	183	98	2	-	1468

1983	869	311	237	179	3	-	1599

1984	886	259	194	131	2	1	1473

1986	935	212	197	123	3	-	1470

1987	1161	268	239	139	9	3	1819

1988	1129	143	139	67	3	-	1481

1989	1159	161	135	81	1	-	1537

1990	1042	132	123	74	1	-	1372

1991	1165	135	138	72	7	-	1517

1993	1226	128	149	99	3	1	1606

1994	2796	86	60	49	1	-	2992

Total	17523	2849	2402	1538	39	8	24359

See also