Where did you have your last hiv test

hivtest2

Source

General Social Survey https://gss.norc.org

Details

Question Have you ever been tested for HIV? Do not count tests you may have had as part of a blood donation. Include oral test (where they take a swab from your mouth). B. Where did you have your last HIV test – at a private doctor or HMO office, at a counselling and testing site, at a hospital, at a clinic, in a jail or prison, at home, or somewhere else? As of 2024, HIVTEST2 has been recoded to only account for seven categories by combining jail, home, and drug treatment facility into "Somewhere else." This has been retroactively applied to all previous years of HIVTEST2. The original HIVTEST2 containing data from 2006 to 2022 has been renamed HIVTEST2_0622.

Values

1 private doctor or hmo office

2 counseling and testing site

3 hospital

4 clinic

5 jail or prison

6 home

7 somewhere else

8 emergency room

9 drug treatment facility

10 hospital inpatient

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	B/C/-	partial

2018	B/C/-	partial

2021	B/C/-	partial

2022	B/C/-	partial

2024	B/C/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	clinic	counseling and testing site	don't know	home	hospital	jail or prison	no answer	private doctor or hmo office	somewhere else	drug treatment facility	emergency room	hospital inpatient	skipped on web	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	2904	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	3556	185	48	5	28	145	19	18	458	48	-	-	-	-	-	4510

2008	1295	162	41	3	14	103	13	17	325	50	-	-	-	-	-	2023

2010	1300	149	46	2	17	108	18	24	342	38	-	-	-	-	-	2044

2012	1249	159	25	1	9	106	11	32	343	39	-	-	-	-	-	1974

2014	1510	227	35	3	17	187	19	38	456	46	-	-	-	-	-	2538

2016	2135	174	39	1	8	126	11	26	312	35	-	-	-	-	-	2867

2018	1768	119	16	-	6	94	10	24	290	21	-	-	-	-	-	2348

2021	3294	123	43	-	9	-	8	21	418	55	9	17	35	-	-	4032

2022	2971	31	31	23	13	19	113	11	283	6	38	4	-	1	-	3544

2024	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3309	3309

Total	65588	1329	324	38	121	888	222	211	3227	338	47	21	35	1	3309	75699

See also