Who would r have voted for-1972 election

if72who

Source

General Social Survey https://gss.norc.org

Details

Question B. In 1972, you remember that McGovern ran for President on the Democratic ticket against Nixon for the Republicans. Do you remember for sure whether or not you voted in that election? 2. IF DID NOT VOTE: Who would you have voted for, for President, if you had voted?

Values

1 mcgovern

2 nixon

3 other

4 refused

5 wouldnt vote

6 wldnt vt-relig

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1973	-	full

1974	-	full

1975	-	full

1976	-	full

1977	-	full

Link at the GSS

For details see the GSS Data Explorer page for if72who.

See also