Imp to r of having a fulfilling job

impjob

Source

General Social Survey https://gss.norc.org

Details

Question I'm going to read you a list of some things that different people value. Some people say these things are very important to them. Other people say they are not so important. Please tell me how important each things is to you personally, using the responses on this card (HAND CARD TO RESPONDENT). How about being financially secure? Is it one of the most important values you hold, very important, somewhat important, not too important, or not at all important? G. Having a fulfilling job.

Values

1 one of most important

2 very important

3 somewhat important

4 not too important

5 not at all important

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1993	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for impjob.

See also