Earnings from all sources for this job

inclstjb

Source

General Social Survey https://gss.norc.org

Details

Question Thinking about your last job, in which of these groups did your earnings from all sources for this job fall? that is, before taxes or other deductions.

Values

[1] a. under $1,000

[2] b. $1,000 to 2,999

[3] c. $3,000 to 3,999

[4] d. $4,000 to 4,999

[5] e. $5,000 to 5,999

[6] f. $6,000 to 6,999

[7] g. $7,000 to 7,999

[8] h. $8,000 to 9,999

[9] i. $10,000 to 12,499

[10] j. $12,500 to 14,999

[11] k. $15,000 to 17,499

[12] l. $17,500 to 19,999

[13] m. $20,000 to 22,499

[14] n. $22,500 to 24,999

[15] o. $25,000 to 29,999

[16] p. $30,000 to 34,999

[17] q. $35,000 to 39,999

[18] r. $40,000 to 49,999

[19] s. $50,000 to 59,999

[20] t. $60,000 to 74,999

[21] u. $75,000 to $89,999

[22] v. $90,000 to $109,999

[23] w. $110,000 to $129,999

[24] x. $130,000 to $149,999

[25] y. $150,000 or over

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2008

Overview

For further details see the GSS Data Explorer page for inclstjb.

Counts by year:

year	iap	a. under $1,000	b. $1,000 to 2,999	c. $3,000 to 3,999	d. $4,000 to 4,999	e. $5,000 to 5,999	f. $6,000 to 6,999	g. $7,000 to 7,999	h. $8,000 to 9,999	i. $10,000 to 12,499	j. $12,500 to 14,999	k. $15,000 to 17,499	l. $17,500 to 19,999	m. $20,000 to 22,499	n. $22,500 to 24,999	o. $25,000 to 29,999	p. $30,000 to 34,999	q. $35,000 to 39,999	r. $40,000 to 49,999	s. $50,000 to 59,999	t. $60,000 to 74,999	u. $75,000 to $89,999	v. $90,000 to $109,999	w. $110,000 to $129,999	x. $130,000 to $149,999	y. $150,000 or over	don't know	no answer	Total

2008	1822	5	5	6	4	7	2	6	11	14	10	9	8	10	9	18	11	9	15	6	6	6	4	2	2	1	7	8	2023

Total	1822	5	5	6	4	7	2	6	11	14	10	9	8	10	9	18	11	9	15	6	6	6	4	2	2	1	7	8	2023

See also