Total family income

income06

Source

General Social Survey https://gss.norc.org

Details

Question In which of these groups did your total family income, from all sources, fall last year before taxes, that is. Just tell me the letter.

Values

[1] under $1 000

[2] $1 000 to 2 999

[3] $3 000 to 3 999

[4] $4 000 to 4 999

[5] $5 000 to 5 999

[6] $6 000 to 6 999

[7] $7 000 to 7 999

[8] $8 000 to 9 999

[9] $10000 to 12499

[10] $12500 to 14999

[11] $15000 to 17499

[12] $17500 to 19999

[13] $20000 to 22499

[14] $22500 to 24999

[15] $25000 to 29999

[16] $30000 to 34999

[17] $35000 to 39999

[18] $40000 to 49999

[19] $50000 to 59999

[20] $60000 to 74999

[21] $75000 to $89999

[22] $90000 to $109999

[23] $110000 to $129999

[24] $130000 to $149999

[25] $150000 or over

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	2008, 2010, 2012, 2014

A/B/C/D	2006

Overview

For further details see the GSS Data Explorer page for income06.

Counts by year:

year	iap	under $1 000	$1 000 to 2 999	$3 000 to 3 999	$4 000 to 4 999	$5 000 to 5 999	$6 000 to 6 999	$7 000 to 7 999	$8 000 to 9 999	$10000 to 12499	$12500 to 14999	$15000 to 17499	$17500 to 19999	$20000 to 22499	$22500 to 24999	$25000 to 29999	$30000 to 34999	$35000 to 39999	$40000 to 49999	$50000 to 59999	$60000 to 74999	$75000 to $89999	$90000 to $109999	$110000 to $129999	$130000 to $149999	$150000 or over	don't know	refused	Total

2006	-	43	38	29	27	40	45	48	83	142	145	126	102	157	125	212	231	217	394	332	360	284	229	162	89	213	195	442	4510

2008	-	19	25	23	9	16	19	29	34	61	44	52	47	59	70	95	99	86	144	141	204	129	120	80	46	123	78	171	2023

2010	-	43	24	17	8	19	19	27	37	82	65	72	34	66	78	81	99	89	147	143	174	129	111	69	57	115	103	136	2044

2012	-	24	25	18	10	8	22	24	38	77	63	58	48	66	60	93	89	101	151	118	153	151	104	68	45	144	105	111	1974

2014	-	38	29	20	12	22	18	16	51	84	64	66	50	69	86	123	141	104	173	208	227	172	167	94	73	207	98	126	2538

Total	0	167	141	107	66	105	123	144	243	446	381	374	281	417	419	604	659	597	1009	942	1118	865	731	473	310	802	579	986	13089

See also