R's industry code (1970)

industry

Source

General Social Survey https://gss.norc.org

Details

Question A. What kind of work do you (did you normally) do? That is, what (is/was) your job called? OCCUPATION: (See REMARKS) B. What (do/did) you actually do in that job? Tell me, what (are/were) some of your main duties? C. What kind of place (do/did) you work for? INDUSTRY: (See REMARKS) D. What (do/did) they (make/do)? E. (Are/Were) you self employed or (do/did) you work for someone else? F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)? G. About how much time (does/did) it usually take you to travel to work - about how many minutes?

Values

17 agricultural production

18 agricultural services, except horticultural

19 horticultural services

27 forestry

28 fisheries

47 metal mining

48 coal mining

49 crude petroleum and natural gas extractions

57 nonmetallic mining and quarrying, except fuel

67 general building contractors

68 general contractors, except building

69 special trade contractors

77 not specified construction

107 logging

108 sawmills, planing mills, and mill work

109 miscellaneous wood products

118 furniture fixtures

119 glass and glass products

127 cement concrete, gypsum, and plaster products

128 structural clay products

137 pottery and related products

138 miscellaneous nonmetallic mineral and stone products

139 blast furnaces, steel works, rolling and finishing mills

147 other primary iron and steel industries

148 primary aluminum industries

149 other primary nonferrous industries

157 cutlery, hand tools, and other hardware

158 fabricated structural metal products

159 screw machine products

167 metal stamping

168 miscellaneous fabricated metal products

169 not specified metal industries

177 engines and turbines

178 farm machinery and equipment

179 construction and material handling machines

187 metalworking machinery

188 office and accounting machines

189 electronic computing equipment

197 machinery, except electrical, n.e.c.

198 not specified machinery

199 household appliances

207 radio, t.v., and communication equipment

208 electrical machinery, equipment, and supplies, n.e.c.

209 not specified electrical machinery, equipment, and supplies

219 motor vehicles and motor vehicle equipment

227 aircraft and parts

228 ship and boat building and repairing

229 railroad locomotives and equipment

237 mobile dwellings and campers

238 cycles and miscellaneous transportation equipment

239 scientific and controlling instruments

247 optical and health services supplies

248 photographic equipment and supplies

249 watches, clocks, and clockwork-operated devices

257 not specified professional equipment

258 ordinance

259 miscellaneous manufacturing industries

268 meat products

269 dairy products

278 canning and preserving fruits, vegetables, and sea foods

279 grain-mill products

287 bakery products

288 confectionery and related products

289 beverage industries

297 miscellaneous food preparation and kindred products

298 not specified food industries

299 tobacco manufactures

307 knitting mills

308 dyeing and finishing textiles, except wool and knit goods

309 floor coverings, except hard surface

317 yarn, thread, and fabric mills

318 miscellaneous textile mill products

319 apparel and accessories

327 miscellaneous fabricated textile products

328 pulp, paper and paperboard mills

329 miscellaneous paper and pulp products

337 paperboard containers and boxes

338 newspaper publishing and printing

339 printing, publishing, and allied industries, except newspapers

347 industrial chemicals

348 plastics, synthetics and resins, except fibers

349 synthetic fibers

357 drugs and medicines

358 soaps and cosmetics

359 paints, varnishes, and related products

367 agricultural chemicals

368 miscellaneous chemicals

369 not specified chemicals and allied products

377 petroleum refining

378 miscellaneous petroleum and coal products

379 rubber products

387 miscellaneous plastic products

388 tanned, curried, and finished leather

389 footwear, except rubber

397 leather products, except footwear

398 not specified manufacturing industries

407 railroads and railway express

408 street railways and bus lines

409 taxicab service

417 trucking service

418 warehousing and storage

419 water transportation

427 air transportation

428 pipe lines, except natural gas

429 services incidental to transportation

447 radio broadcasting and television

448 telephone (wire and radio)

449 telegraph and miscellaneous communication services

467 electric light and power

468 electric-gas utilities

469 gas and steam supply systems

477 water supply

478 sanitary services

479 other and not specified utilities

507 motor vehicles and equipment

508 drugs, chemicals, and allied products

509 dry goods and apparel

527 food and related products

528 farm products-raw materials

529 electrical goods

537 hardware, plumbing, and heating supplies

538 not specified electrical and hardware products

539 machinery equipment and supplies

557 metals and minerals, n.e.c.

558 petroleum products

559 scrap and waste materials

567 alcoholic beverages

568 paper and its products

569 lumber and construction materials

587 wholesalers, n.e.c.

588 not specified wholesale trade

607 lumber and building material retailing

608 hardware and farm equipment stores

609 department and mail order establishments

617 limited price variety stores

618 vending machine operators

619 direct selling establishments

627 misc. general merchandise stores

628 grocery stores

629 dairy product stores

637 retail bakeries

638 food stores, n.e.c.

639 motor vehicle dealers

647 tire, battery, and accessory dealers

648 gasoline service stations

649 miscellaneous vehicle dealers

657 apparel and accessories stores, except shoe stores

658 shoe stores

667 furniture and home furnishing stores

668 household appliances, tv, and radio stores

669 eating and drinking places

677 drug stores

678 liquor stores

679 farm and garden supply stores

687 jewelry stores

688 fuel and ice dealers

689 retail florists

697 miscellaneous retail stores

698 not specified retail trade

707 banking

708 credit agencies

709 security, commodity brokerage, and investment companies

717 insurance

718 real estate, incl. real estate-insurance-law offices

727 advertising

728 services to dwellings and other buildings

729 commercial research, development, and testing labs

737 employment and temporary help agencies

738 business management and consulting services

739 computer programming services

747 detective and protective services

748 business services, n.e.c.

749 automobile services, except repair

757 automobile repair and related services

758 electrical repair shops

759 miscellaneous repair services

769 private households

777 hotels and motels

778 lodging places, except hotels and motels

779 laundering, cleaning, and other garment services

787 beauty shops

788 barber shops

789 shoe repair shops

797 dressmaking shops

798 miscellaneous personal services

807 theaters and motion pictures

808 bowling alleys, billiard and pool parlors

809 miscellaneous entertainment and recreation services

828 offices of physicians

829 offices of dentists

837 offices of chiropractors

838 hospitals

839 convalescent institutions

847 offices of health practitioners,n.e.c.

848 health services, n.e.c.

849 legal services

857 elementary and secondary schools

858 colleges and universities

859 libraries

867 educational services, n.e.c.

868 not specified educational services

869 museums, art galleries, and zoos

877 religious organizations

878 welfare services

879 residential welfare facilities

887 nonprofit membership organizations

888 engineering and architectural services

889 accounting, auditing, and bookkeeping services

897 miscellaneous professional and related services

907 postal service

917 federal public administration

927 state public administration

937 local public administration

Question Years and Ballots

year	ballots	availability

1972	-/-/-	full

1973	-/-/-	full

1974	-/-/-	full

1975	-/-/-	full

1976	-/-/-	full

1977	-/-/-	full

1978	-/-/-	full

1980	-/-/-	full

1982	-/-/-	full

1983	-/-/-	full

1984	-/-/-	full

1985	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/C	full

1990	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for industry.

See also