Years of service as an interviewer at norc

intyrs_0418

Source

General Social Survey https://gss.norc.org

Details

Question D. Years of service as an interviewer at NORC As of 2024, INTYRS has been recoded to only account for fewer categories. This has been retroactively applied to all previous years of INTYRS. The original INTYRS containing data from 2004 to 2018 has been renamed INTYRS_0418.

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] i don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	2004, 2008, 2010, 2012, 2014, 2016, 2018

A/B/C/D	2006

Overview

For further details see the GSS Data Explorer page for intyrs_0418.

Counts by year:

year	not available in this year	0	0.1	0.2	0.3	0.4	0.5	0.6	0.7	0.9	1	1.1	1.2	1.3	1.6	1.9	11.5	11.8	12	2.1	2.4	2.7	2.9	3	3.1	3.4	3.7	4.3	4.8	5.2	6.8	6.9	7.5	7.6	7.9	8.1	8.9	9.3	9.5	9.6	9.8	no answer	0.8	1.8	10.7	11.1	11.2	11.3	11.7	11.9	13.1	13.6	17.2	17.8	19.6	2	2.2	2.5	2.6	2.8	3.2	3.5	3.9	4.2	4.4	4.5	4.6	4.7	5	5.1	5.3	5.8	6.3	6.5	7	8.2	8.4	8.5	9.1	9.7	0.12	0.16	0.23	0.26	0.27	0.31	0.35	0.38	0.39	0.42	0.43	0.44	0.61	0.97	1.03	1.06	1.11	1.39	1.92	10.29	11.41	11.74	13.22	13.95	14.97	15.52	15.71	18.72	19.87	2.01	2.03	2.36	2.38	2.54	2.94	2.99	25.64	3.06	3.09	3.16	3.67	3.78	3.82	3.87	4.02	4.05	4.1	4.17	4.29	4.35	4.62	4.65	5.09	5.26	5.28	5.29	5.6	5.88	6.27	6.39	6.54	6.63	6.67	6.68	6.72	6.81	6.89	7.07	7.11	7.39	7.87	8.62	8.67	8.87	8.93	0.517808219178082	0.520547945205479	0.523287671232877	0.83013698630137	0.857534246575342	0.906849315068493	0.986301369863014	1.04931506849315	1.1041095890411	1.11780821917808	1.12602739726027	1.16164383561644	1.19452054794521	1.44383561643836	1.5	1.96259124087591	4.87448653582839	5.09356458238247	8.41131731061758	9	10	11	14	15	17	19	23	4	6	8	13	16	25	20	21	Total

2004	-	5	73	142	131	1005	18	127	66	66	3	11	101	3	16	6	31	16	70	13	12	24	79	114	85	40	34	7	19	33	26	45	35	31	19	58	5	15	44	64	50	70	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	-	-	-	-	21	784	-	-	-	136	258	127	-	-	-	-	11	72	-	-	-	5	-	34	-	-	41	12	34	18	89	7	-	-	-	-	-	-	-	26	-	-	10	167	15	55	21	90	9	5	46	76	42	1	12	653	451	40	64	24	90	138	11	3	60	106	36	37	13	75	121	51	15	67	17	48	39	2	44	81	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510

2008	-	-	-	25	-	24	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4	-	-	-	-	-	-	-	-	-	16	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	23	24	95	9	60	10	33	10	86	11	155	244	14	1	1	2	2	50	24	42	2	17	23	42	13	23	15	12	19	6	24	8	21	1	14	22	10	35	14	33	8	3	8	34	2	59	117	14	48	56	18	10	18	11	16	22	20	5	13	3	16	33	14	1	1	13	9	13	18	17	15	11	30	13	15	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	-	-	-	-	-	-	13	-	-	-	161	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	27	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	4	-	-	-	-	-	-	-	-	18	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	13	35	1279	7	14	28	29	39	5	21	34	199	3	4	3	12	32	10	13	39	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	-	560	-	-	-	-	-	-	-	-	117	-	-	-	-	-	-	-	28	-	-	-	-	106	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	399	-	-	-	-	-	-	-	-	-	-	-	-	23	-	-	-	-	-	79	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	25	61	51	18	3	30	21	13	139	96	205	-	-	-	-	-	1974

2014	-	340	-	-	-	-	-	-	-	-	123	-	-	-	-	-	-	-	67	-	-	-	-	15	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	501	-	-	-	-	-	-	-	-	-	-	-	-	117	-	-	-	-	-	219	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	103	151	41	-	-	33	7	-	523	185	47	25	25	16	-	-	2538

2016	-	751	-	-	-	-	-	-	-	-	166	-	-	-	-	-	-	-	122	-	-	-	-	77	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	336	-	-	-	-	-	-	-	-	-	-	-	-	40	-	-	-	-	-	123	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	209	104	-	10	-	-	14	22	189	346	205	79	19	-	55	-	2867

2018	-	680	-	-	-	-	-	-	-	-	33	-	-	-	-	-	-	-	31	-	-	-	-	27	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	542	-	-	-	-	-	-	-	-	-	-	-	-	94	-	-	-	-	-	33	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	48	21	62	139	34	-	-	30	179	165	136	2	-	8	4	80	2348

Total	0	2336	73	167	152	1813	31	127	66	202	861	138	101	3	16	6	42	88	318	13	12	29	79	373	85	40	75	19	53	51	115	52	35	31	19	58	5	15	44	90	50	101	10	167	15	55	21	90	9	5	46	92	42	1	12	2433	451	40	64	24	90	138	11	3	60	110	36	37	287	75	121	51	15	67	489	48	39	2	44	81	23	24	95	9	60	10	33	10	86	11	155	244	14	1	1	2	2	50	24	42	2	17	23	42	13	23	15	12	19	6	24	8	21	1	14	22	10	35	14	33	8	3	8	34	2	59	117	14	48	56	18	10	18	11	16	22	20	5	13	3	16	33	14	1	1	13	9	13	18	17	15	11	30	13	15	13	35	1279	7	14	28	29	39	5	21	34	199	3	4	3	12	32	10	13	424	337	154	167	37	63	42	65	1030	792	593	106	44	24	59	80	21116

See also