Liking for israel

israel

Source

General Social Survey https://gss.norc.org

Details

Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM. G. Israel

Values

[0] +5

[1] +4

[2] +3

[3] +2

[4] +1

[5] -1

[6] -2

[7] -3

[8] -4

[9] -5

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1974, 1975, 1977, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for israel.

Counts by year:

year	not available in this year	+5	+4	+3	+2	+1	-1	-2	-3	-4	-5	dk, na, iap	iap	Total

1974	-	198	141	231	175	258	128	57	67	33	89	107	-	1484

1975	-	146	123	205	169	275	164	82	77	48	76	125	-	1490

1977	-	180	151	204	173	270	149	71	64	41	69	158	-	1530

1982	-	207	147	250	211	290	215	121	92	62	95	170	-	1860

1983	-	148	123	169	172	285	209	106	111	67	103	106	-	1599

1985	-	168	125	211	201	254	160	94	82	48	75	116	-	1534

1986	-	165	128	175	180	285	144	84	73	53	104	79	-	1470

1988	-	96	60	101	91	155	130	77	77	30	94	570	-	1481

1989	-	87	56	107	95	177	104	70	59	43	118	621	-	1537

1990	-	75	48	86	93	182	128	69	45	30	93	523	-	1372

1991	-	152	104	162	138	177	76	38	35	14	55	566	-	1517

1993	-	105	73	133	104	227	123	71	59	41	66	604	-	1606

1994	-	46	32	52	59	110	55	31	34	14	34	2525	-	2992

Total	0	1773	1311	2086	1861	2945	1785	971	875	524	1071	6270	0	21472

See also