Liking for japan

japan

Source

General Social Survey https://gss.norc.org

Details

Question 103. You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? b. Japan

Overview

For further details see the official GSS documentation.

Counts by year:

year	not available in this year	+1	+2	+3	+4	+5	-1	-2	-3	-4	-5	dk, na, iap	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	1504

1974	-	253	217	314	149	110	101	60	63	52	87	78	1484

1975	-	273	210	277	137	83	124	69	78	64	73	102	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	1499

1977	-	267	203	288	100	102	125	72	78	68	96	131	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	1468

1982	-	294	231	344	169	160	163	95	83	67	113	141	1860

1983	-	307	255	257	121	105	159	84	97	53	83	78	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	1473

1985	-	249	246	298	148	128	136	70	85	40	53	81	1534

1986	-	243	240	319	149	111	109	80	59	31	74	55	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	-	1819

1988	-	185	136	182	83	82	85	44	51	18	68	547	1481

1989	-	194	128	189	97	73	83	56	45	18	79	575	1537

1990	-	179	112	151	67	80	83	40	50	30	88	492	1372

1991	-	158	129	189	92	95	78	46	51	41	89	549	1517

1993	-	203	137	165	77	76	123	58	58	46	92	571	1606

1994	-	93	54	73	35	41	50	32	35	21	50	2508	2992

1996	2904	-	-	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	-	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	-	-	-	3544

2024	3309	-	-	-	-	-	-	-	-	-	-	-	3309

Total	54227	2898	2298	3046	1424	1246	1419	806	833	549	1045	5908	75699

Values

0 +5

1 +4

2 +3

3 +2

4 +1

5 -1

6 -2

7 -3

8 -4

9 -5

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook