Could r find equally good job

jobfind

Source

General Social Survey https://gss.norc.org

Details

Question B. About how easy would it be for you to find a job with another employer with approximately the same income and fringe benefits you now have? Would you say very easy, somewhat easy, or not easy at all?

Values

1 very easy

2 somewhat easy

3 not easy

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1977, 1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for jobfind.

Counts by year:

year	iap	very easy	somewhat easy	not easy	don't know	no answer	skipped on web	Total

1977	609	241	271	379	18	12	-	1530

1978	638	246	283	340	19	6	-	1532

1982	795	217	261	545	29	13	-	1860

1983	654	176	271	473	20	5	-	1599

1985	583	235	292	397	16	11	-	1534

1986	595	238	281	335	16	5	-	1470

1988	862	169	220	212	14	4	-	1481

1989	921	207	166	229	11	3	-	1537

1990	770	190	180	223	7	2	-	1372

1991	907	142	214	241	11	2	-	1517

1993	925	147	220	298	11	5	-	1606

1994	1681	268	419	585	30	9	-	2992

1996	1531	363	437	533	29	11	-	2904

1998	1571	376	437	408	28	12	-	2832

2000	1561	461	402	360	23	10	-	2817

2002	2180	156	211	206	9	3	-	2765

2004	1879	230	305	372	10	16	-	2812

2006	3284	379	433	392	11	11	-	4510

2008	1169	195	273	363	11	12	-	2023

2010	1220	103	269	432	8	12	-	2044

2012	1186	126	275	364	9	14	-	1974

2014	1505	197	349	475	10	2	-	2538

2016	1735	229	446	440	5	12	-	2867

2018	1379	224	370	355	9	11	-	2348

2021	2498	221	647	621	1	33	11	4032

2022	2138	403	577	411	10	2	3	3544

Total	34776	6139	8509	9989	375	238	14	60040

See also