Short working hours

jobhour

Source

General Social Survey https://gss.norc.org

Details

Question A. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? B. Which comes next? C. Which is third most important? D. Which is fourth most important? C. Workings hours are short, lots of free time

Values

1 most impt

2 second

3 third

4 fourth

5 fifth

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1973	-/-	full

1974	-/-	full

1976	-/-	full

1977	-/-	full

1980	-/-	full

1982	-/-	full

1984	-/-	full

1985	-/-	full

1987	-/-	full

1988	A/C	partial

1989	A/C	partial

1990	A/C	partial

1991	A/C	partial

1993	A/C	partial

1994	A/C	partial

2006	D/-	partial

2012	A/B	partial

2014	A/B	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	fifth	fourth	most impt	no answer	second	third	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	-	684	370	70	42	147	191	-	1504

1974	-	621	393	71	30	166	203	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	-	759	325	58	50	133	174	-	1499

1977	-	768	368	66	50	116	162	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	-	740	371	47	32	122	156	-	1468

1982	937	472	179	27	41	73	131	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	-	847	316	42	32	97	139	-	1473

1985	-	884	319	51	34	100	146	-	1534

1986	1470	-	-	-	-	-	-	-	1470

1987	-	935	459	62	47	140	176	-	1819

1988	504	496	249	27	31	75	99	-	1481

1989	504	566	228	28	53	73	85	-	1537

1990	455	486	211	27	37	58	98	-	1372

1991	524	522	216	34	48	64	109	-	1517

1993	531	605	219	37	38	65	111	-	1606

1994	2508	275	94	10	21	39	45	-	2992

1996	2904	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	2812

2006	2992	717	363	74	20	158	186	-	4510

2008	2023	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	2044

2012	1002	437	195	63	27	127	123	-	1974

2014	1288	573	283	70	19	126	179	-	2538

2016	2867	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	4032

2022	-	-	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	46393	11387	5158	864	652	1879	2513	6853	75699

See also