Is r likely to lose job

joblose

Source

General Social Survey https://gss.norc.org

Details

Question A. Thinking about the next 12 months, how likely do you think it is that you will lose your job or be laid off–very likely, fairly likely, not too likely, or not at all likely?

Values

1 very likely

2 fairly likely

3 not too likely

4 not likely

5 leaving labor force

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1977	-/-	full

1978	-/-	full

1982	-/-	full

1983	-/-	full

1985	-/-	full

1986	-/-	full

1988	A/B	partial

1989	A/B	partial

1990	A/B	partial

1991	A/B	partial

1993	A/B	partial

1994	A/B	partial

1996	A/B	partial

1998	A/B	partial

2000	A/B	partial

2002	A/B	partial

2004	A/B	partial

2006	A/B	partial

2008	A/B	partial

2010	A/B	partial

2012	A/B	partial

2014	A/B	partial

2016	A/B	partial

2018	A/B	partial

2021	A/B	partial

2022	A/B	partial

2024	A/B	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	don't know	fairly likely	leaving labor force	no answer	not likely	not too likely	very likely	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	1499

1977	609	16	54	5	4	586	219	37	-	1530

1978	638	9	33	-	5	629	182	36	-	1532

1980	1468	-	-	-	-	-	-	-	-	1468

1982	795	24	79	-	10	593	283	76	-	1860

1983	654	15	72	-	5	564	233	56	-	1599

1984	1473	-	-	-	-	-	-	-	-	1473

1985	583	10	42	-	9	609	218	63	-	1534

1986	595	19	57	-	6	571	191	31	-	1470

1987	1819	-	-	-	-	-	-	-	-	1819

1988	862	7	27	-	2	405	153	25	-	1481

1989	921	5	23	-	3	424	136	25	-	1537

1990	770	9	33	-	1	395	145	19	-	1372

1991	907	6	40	-	1	374	152	37	-	1517

1993	925	9	54	-	4	405	179	30	-	1606

1994	1681	15	58	-	12	807	347	72	-	2992

1996	1531	24	91	-	9	812	378	59	-	2904

1998	1571	20	53	-	9	806	327	46	-	2832

2000	1561	19	53	-	10	860	270	44	-	2817

2002	2180	4	41	-	4	360	139	37	-	2765

2004	1879	11	49	-	15	573	237	48	-	2812

2006	3284	14	77	-	11	763	312	49	-	4510

2008	1169	6	52	-	11	487	250	48	-	2023

2010	1220	8	68	-	15	414	263	56	-	2044

2012	1186	7	47	-	15	460	216	43	-	1974

2014	1505	5	57	-	3	642	289	37	-	2538

2016	1735	5	43	-	11	733	298	42	-	2867

2018	1379	10	41	-	11	682	201	24	-	2348

2021	2498	-	92	-	33	750	619	34	6	4032

2022	2138	9	80	-	2	873	392	46	4	3544

2024	2065	9	89	-	3	716	379	46	2	3309

Total	49191	295	1505	5	224	16293	7008	1166	12	75699

See also