High income

jobpay

Source

General Social Survey https://gss.norc.org

Details

Question On these cards there are various aspects of jobs. How important Do you personally consider these job characteristics? Using the scale from one to seven for your answers again, tell me for each card its letter and the number you've decided on. SHUFFLE THE PINK CARDS BEFORE HANDING TO RESPONDENT. B. High Income

Values

1 unimportant

7 very important

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1982	-	full

Link at the GSS

For details see the GSS Data Explorer page for jobpay.

See also