No danger of being fired

jobsec

Source

General Social Survey https://gss.norc.org

Details

Question 183a. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? b. Which comes next? c. Which is third most important? d. Which is fourth most important? NO DANGER OF BEING FIRED

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	fifth	fourth	most impt	no answer	second	third	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	-	424	438	100	42	210	290	-	1504

1974	-	481	417	111	30	159	286	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	-	376	458	111	50	214	290	-	1499

1977	-	362	467	119	50	208	324	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	-	452	486	87	32	170	241	-	1468

1982	937	169	242	93	41	175	203	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	-	340	512	109	32	192	288	-	1473

1985	-	387	528	109	34	168	308	-	1534

1986	1470	-	-	-	-	-	-	-	1470

1987	-	478	587	120	47	221	366	-	1819

1988	504	264	299	62	31	131	190	-	1481

1989	504	267	342	49	53	116	206	-	1537

1990	455	236	288	55	37	106	195	-	1372

1991	524	251	303	53	48	120	218	-	1517

1993	531	242	348	91	38	146	210	-	1606

1994	2508	98	152	39	21	71	103	-	2992

1996	2904	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	2812

2006	2992	412	403	170	20	225	288	-	4510

2008	2023	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	2044

2012	1002	232	275	81	26	149	209	-	1974

2014	1288	306	355	103	16	215	255	-	2538

2016	2867	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	4032

2022	-	-	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	46393	5777	6900	1662	648	2996	4470	6853	75699

Values

1 most impt

2 second

3 third

4 fourth

5 fifth

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook