Contrast between judge and lover

judgeluv

Source

General Social Survey https://gss.norc.org

Details

Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.

Values

[1] judge

[7] lover

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	1989, 1990, 1993, 1994, 1996

A/B/C	1988, 1991, 1998

B/C/-	2008

Full	1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for judgeluv.

Counts by year:

year	iap	judge	2	3	4	5	6	lover	don't know	no answer	Total

1984	484	373	102	84	218	41	22	83	1	65	1473

1985	-	583	118	113	363	81	51	148	20	57	1534

1986	-	620	133	97	316	53	53	129	13	56	1470

1987	-	706	154	122	449	60	53	169	20	86	1819

1988	-	528	143	99	401	48	54	113	33	62	1481

1989	531	367	90	86	279	31	30	73	6	44	1537

1990	444	345	81	72	229	34	24	84	9	50	1372

1991	158	385	140	116	320	75	55	59	-	209	1517

1993	526	396	103	76	250	54	46	87	9	59	1606

1994	1015	737	197	137	485	77	61	172	31	80	2992

1996	1913	351	85	77	268	32	31	95	12	40	2904

1998	1387	482	135	108	420	56	59	106	22	57	2832

2008	658	448	94	128	355	57	53	82	111	37	2023

Total	7116	6321	1575	1315	4353	699	592	1400	287	902	24560

See also