Allow militarists book in library

libmil

Source

General Social Survey https://gss.norc.org

Details

Question 80. Consider a person who advocates doing away with elections and letting the military run the country. c. Suppose he wrote a book advocating doing away with elections and letting the military run the country. Somebody in your community suggests that the book should be removed from the library. Would you favor removing it, or not?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	don't know	no answer	not remove	remove	skipped on web	not available in this year	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	-	52	6	838	603	-	-	1499

1977	-	41	4	833	652	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	-	49	2	852	565	-	-	1468

1982	-	88	6	977	789	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	-	30	5	869	569	-	-	1473

1985	-	33	5	862	634	-	-	1534

1986	1470	-	-	-	-	-	-	1470

1987	-	65	8	1020	726	-	-	1819

1988	504	36	7	551	383	-	-	1481

1989	504	39	7	610	377	-	-	1537

1990	455	34	4	554	325	-	-	1372

1991	524	22	5	661	305	-	-	1517

1993	531	37	4	721	313	-	-	1606

1994	996	57	3	1274	662	-	-	2992

1996	981	68	4	1234	617	-	-	2904

1998	950	90	3	1253	536	-	-	2832

2000	956	65	6	1187	603	-	-	2817

2002	1841	24	2	642	256	-	-	2765

2004	1914	15	3	615	265	-	-	2812

2006	2507	35	2	1375	591	-	-	4510

2008	671	40	4	916	392	-	-	2023

2010	763	21	2	901	357	-	-	2044

2012	666	31	6	927	344	-	-	1974

2014	824	28	9	1197	480	-	-	2538

2016	977	31	8	1377	474	-	-	2867

2018	774	33	4	1135	402	-	-	2348

2021	2714	1	1	910	392	14	-	4032

2022	-	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	-	3309	3309

Total	30744	1065	120	24291	12612	14	6853	75699

Values

1 remove

2 not remove

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook