Neighborhood half southern white

liveso

Source

General Social Survey https://gss.norc.org

Details

Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening? F. Living in a neighborhood where half your neighbors were raised in the South?

Values

1 strongly favor

2 favor

3 neither favor nor oppose

4 oppose

5 strongly oppose

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	1990

Overview

For further details see the GSS Data Explorer page for liveso.

Counts by year:

year	iap	strongly favor	favor	neither favor nor oppose	oppose	strongly oppose	don't know	no answer	Total

1990	458	45	146	536	132	33	17	5	1372

Total	458	45	146	536	132	33	17	5	1372

See also