Contrast between mother and father

mapa

Source

General Social Survey https://gss.norc.org

Details

Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.

Values

[1] mother

[7] father

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	1989, 1990, 1993, 1994, 1996

A/B/C	1988, 1991, 1998

B/C/-	2008

Full	1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for mapa.

Counts by year:

year	iap	mother	2	3	4	5	6	father	no answer	don't know	Total

1984	484	32	16	30	257	86	80	429	59	-	1473

1985	-	50	15	26	353	142	140	732	58	18	1534

1986	-	62	15	24	347	109	123	735	47	8	1470

1987	-	64	23	22	454	126	133	895	81	21	1819

1988	-	41	26	26	372	105	104	720	59	28	1481

1989	531	33	10	12	257	85	93	471	41	4	1537

1990	444	20	14	14	229	77	91	437	38	8	1372

1991	158	42	26	52	325	114	116	522	162	-	1517

1993	526	33	12	29	251	85	104	506	54	6	1606

1994	1015	48	28	33	504	158	174	940	72	20	2992

1996	1913	36	14	22	231	93	82	467	35	11	2904

1998	1387	42	16	30	354	119	139	681	50	14	2832

2008	658	50	15	43	340	102	118	599	32	66	2023

Total	7116	553	230	363	4274	1401	1497	8134	788	204	24560

See also