Mothers occupational prestige score using threshold method (2010)

mapres105plus

Source

General Social Survey https://gss.norc.org

Details

Question MAPRES105PLUS is based on the 2010 Census occupation classification. This prestige score is calculated using an alternative method, based on the percentage of ratings that is greater than or equal to a threshold (rating five). The effect of individual raters is also removed using hierarchical linear modeling (HLM).

Values

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for mapres105plus.

Counts by year:

year	iap	10	11	12	13	14	15	16	17	18	19	20	21	22	23	24	25	26	27	28	29	30	31	32	33	34	35	36	37	38	39	40	42	43	44	46	47	48	49	5	50	51	52	53	54	55	56	58	59	6	60	61	62	63	64	65	66	67	68	69	7	70	72	73	74	75	76	78	79	8	80	82	83	84	85	86	87	88	89	9	90	91	92	93	94	95	71	77	81	97	57	41	45	not available in this release	Total

1994	1351	15	9	58	10	116	21	99	2	105	8	9	28	69	4	50	7	58	38	66	37	1	1	11	2	3	9	1	4	21	36	13	13	1	8	22	7	56	5	5	88	3	1	2	22	143	8	18	2	2	8	5	13	3	5	10	10	1	9	11	1	1	1	11	2	2	3	2	2	7	3	1	4	7	80	3	52	24	2	7	6	18	1	1	5	3	-	-	-	-	-	-	-	-	2992

1996	1158	6	6	63	10	114	17	99	5	102	4	13	18	74	1	48	8	61	39	76	63	2	-	13	4	-	5	3	4	19	29	14	7	-	8	24	11	57	7	14	91	7	-	4	22	158	16	22	2	-	7	3	15	3	8	15	3	-	16	15	1	2	-	6	2	1	2	3	1	8	5	1	6	6	89	3	59	23	-	6	11	28	1	-	14	3	4	2	3	1	-	-	-	-	2904

1998	1213	14	9	52	19	91	14	92	2	94	12	14	11	70	1	46	5	64	44	81	48	4	-	16	2	4	8	1	5	13	34	18	11	1	7	23	5	41	8	5	89	5	-	2	15	129	10	14	4	6	5	-	11	5	6	15	3	1	27	13	-	-	-	7	3	-	2	-	4	4	8	1	6	2	90	4	64	17	1	4	4	28	2	-	9	5	3	-	2	-	-	-	-	-	2832

2000	1184	6	10	44	20	105	14	80	1	92	7	10	18	56	1	55	5	58	42	81	40	3	-	18	-	1	9	2	2	11	36	17	13	1	8	39	6	44	4	5	85	7	-	2	15	155	11	17	7	5	1	5	10	5	5	18	9	-	21	14	7	2	1	10	2	4	1	5	4	13	3	1	6	2	73	4	60	20	1	8	3	23	2	-	13	1	6	-	-	1	1	-	-	-	2817

2002	1139	4	8	60	13	98	20	95	1	93	4	10	20	50	1	49	9	68	39	76	39	5	-	19	3	3	9	1	3	14	31	15	10	1	10	18	4	28	5	9	113	4	-	2	20	177	19	10	6	4	4	5	15	5	8	10	5	1	20	10	2	1	3	17	3	1	4	6	-	6	6	-	5	1	78	4	45	11	2	7	4	14	2	2	7	7	1	-	1	2	1	-	-	-	2765

2004	1124	4	9	46	11	93	22	93	-	113	5	16	21	41	2	52	9	57	33	81	58	1	1	20	5	2	6	1	9	14	29	8	18	2	16	18	4	39	15	7	88	3	-	2	13	142	16	12	5	5	1	7	5	5	7	14	11	-	18	23	2	4	2	9	6	2	1	5	4	6	7	1	8	8	110	5	48	33	2	10	3	27	-	2	14	6	1	-	2	-	1	1	-	-	2812

2006	2690	4	6	46	12	116	24	101	3	108	10	12	16	57	2	52	7	82	38	86	52	2	-	17	5	2	9	1	-	14	39	13	12	1	14	13	5	40	9	9	117	5	1	2	18	159	14	15	7	2	6	3	11	6	8	18	12	-	27	11	8	2	-	15	4	4	4	11	1	9	5	-	6	10	111	1	58	12	2	5	9	29	2	-	18	6	4	-	2	-	1	-	-	-	4510

2008	730	7	8	42	12	84	10	66	3	76	4	8	13	32	2	45	5	43	33	48	42	1	-	8	2	1	8	2	4	10	15	15	11	2	23	19	2	29	8	9	76	10	-	2	16	109	10	22	3	4	9	5	8	5	8	11	3	-	18	10	3	3	-	8	1	1	3	5	1	5	4	2	5	1	65	2	42	9	1	9	5	23	3	-	4	2	-	-	2	1	2	-	-	-	2023

2010	677	2	13	38	15	66	11	79	3	62	8	16	14	40	1	43	2	50	28	59	43	-	-	15	3	1	2	-	3	12	31	17	8	-	8	21	6	30	6	12	74	4	-	2	15	111	11	18	5	7	2	5	8	5	7	9	4	-	17	14	2	1	4	15	5	1	3	6	1	7	5	-	9	2	84	3	50	16	1	6	5	29	-	1	10	9	2	-	-	2	2	-	-	-	2044

2012	647	6	5	37	9	74	26	69	1	77	6	12	26	38	-	38	7	40	36	79	44	2	-	7	2	2	4	4	2	14	33	6	11	-	11	18	4	33	15	8	48	2	-	3	15	93	10	11	2	3	3	5	10	4	10	13	4	-	13	8	2	1	1	7	3	2	-	3	1	2	8	1	7	5	68	7	49	12	2	7	10	22	1	1	10	26	3	-	2	-	1	-	-	-	1974

2014	866	-	13	39	16	80	23	90	5	75	5	7	21	37	1	52	6	55	45	101	68	3	-	14	6	4	8	2	3	10	44	5	10	1	22	16	8	35	7	11	90	5	-	3	13	150	16	16	7	2	3	12	10	4	9	16	7	1	26	19	2	3	1	14	1	-	3	7	3	9	5	-	7	5	74	10	64	21	3	5	7	33	2	-	19	7	3	-	3	2	1	1	-	-	2538

2016	875	3	12	43	16	128	28	103	1	94	9	24	19	37	3	77	8	65	45	96	63	3	-	25	4	5	12	2	6	20	34	9	18	1	28	23	9	42	18	6	102	5	1	2	30	141	27	24	6	6	8	8	8	6	18	21	4	1	33	14	1	2	2	18	8	4	4	7	4	6	5	-	13	5	94	7	106	17	1	8	2	29	1	2	20	9	2	1	4	5	-	-	1	-	2867

2018	691	-	11	41	25	72	23	89	4	88	7	10	14	23	5	53	6	31	49	97	59	-	-	31	2	4	6	3	2	20	29	4	15	3	11	25	4	29	11	10	92	12	1	3	20	139	22	17	6	8	8	7	14	6	6	32	5	2	22	8	2	-	3	8	8	2	3	5	4	8	10	1	12	4	81	9	69	11	2	5	7	32	1	-	14	7	-	-	2	-	1	-	-	-	2348

2021	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4032	4032

2022	3544	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3544

Total	17889	71	119	609	188	1237	253	1155	31	1179	89	161	239	624	24	660	84	732	509	1027	656	27	2	214	40	32	95	23	47	192	420	154	157	14	174	279	75	503	118	110	1153	72	4	31	234	1806	190	216	62	54	65	70	138	62	105	202	80	7	267	170	33	22	18	145	48	24	33	65	30	90	74	9	94	58	1097	62	766	226	20	87	76	335	18	9	157	91	29	3	23	14	11	2	1	4032	43002

See also