Marital status of 2nd person

mar2

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what? 4. Marital status of second person

Values

[1] married

[2] widowed

[3] divorced

[4] separated

[5] never married

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/D	2006

Full	1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for mar2.

Counts by year:

year	iap	married	widowed	divorced	separated	never married	no answer	don't know	not available in this release	skipped on web	Total

1975	244	1080	15	16	7	127	1	-	-	-	1490

1976	919	502	8	8	1	61	-	-	-	-	1499

1977	295	1050	15	15	4	142	8	1	-	-	1530

1978	341	1013	11	19	5	137	6	-	-	-	1532

1980	336	942	13	24	8	136	9	-	-	-	1468

1982	493	1066	19	39	12	220	11	-	-	-	1860

1983	376	1018	13	37	10	128	16	1	-	-	1599

1984	372	878	11	26	7	170	9	-	-	-	1473

1985	395	921	7	28	14	158	11	-	-	-	1534

1986	351	869	15	38	21	172	4	-	-	-	1470

1987	493	957	21	39	15	268	26	-	-	-	1819

1988	402	830	16	29	10	182	12	-	-	-	1481

1989	387	894	13	33	10	183	17	-	-	-	1537

1990	371	762	13	37	7	163	19	-	-	-	1372

1991	438	840	15	31	14	166	13	-	-	-	1517

1993	433	902	10	48	14	188	10	1	-	-	1606

1994	891	1606	28	79	19	339	28	2	-	-	2992

1996	895	1438	28	98	24	346	74	1	-	-	2904

1998	897	1391	18	83	18	369	55	1	-	-	2832

2000	875	1332	28	95	23	424	40	-	-	-	2817

2002	984	1244	15	88	20	380	32	2	-	-	2765

2004	810	1548	21	83	32	310	2	6	-	-	2812

2006	1366	2307	27	148	49	600	2	11	-	-	4510

2008	616	1016	8	56	17	295	3	12	-	-	2023

2010	674	937	11	72	28	291	5	26	-	-	2044

2012	602	964	16	69	21	286	3	13	-	-	1974

2014	754	1231	19	79	26	382	11	36	-	-	2538

2016	937	1282	26	111	37	442	14	18	-	-	2867

2018	727	1082	19	96	33	369	5	17	-	-	2348

2021	-	-	-	-	-	-	-	-	4032	-	4032

2022	2667	712	6	40	3	110	3	1	-	2	3544

Total	20341	32614	485	1664	509	7544	449	149	4032	2	67789

See also