Close relative marry hispanic

marhisp

Source

General Social Survey https://gss.norc.org

Details

Question What about having a close relative marry a Hispanic American person? Would you be very in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?

Values

1 strongly favor

2 favor

3 neither favor nor oppose

4 oppose

5 strongly oppose

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1990	A/B/C	full

2000	A/B/C	full

2004	A/B/-	partial

2006	A/B/-	partial

2008	A/B/-	partial

2010	A/B/-	partial

2012	A/B/-	partial

2014	A/B/-	partial

2016	A/B/-	partial

2018	A/B/-	partial

2021	A/B/-	partial

2022	A/B/-	partial

2024	A/B/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	don't know	favor	neither favor nor oppose	no answer	oppose	strongly favor	strongly oppose	skipped on web	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	-	1537

1990	-	19	94	636	10	343	62	208	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	-	-	-	2992

1996	2904	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	2832

2000	1419	47	228	597	14	158	243	111	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	2765

2004	1906	3	140	484	4	100	121	54	-	-	2812

2006	2518	8	321	950	3	227	360	123	-	-	4510

2008	694	6	209	651	4	129	250	80	-	-	2023

2010	614	5	230	734	5	124	267	65	-	-	2044

2012	672	15	198	688	2	87	260	52	-	-	1974

2014	863	12	268	880	6	98	356	55	-	-	2538

2016	979	7	299	1013	6	109	412	42	-	-	2867

2018	789	6	241	788	8	87	381	48	-	-	2348

2021	1318	-	257	1794	2	82	517	40	22	-	4032

2022	1168	9	264	1356	4	69	630	39	5	-	3544

2024	-	-	-	-	-	-	-	-	-	3309	3309

Total	52449	137	2749	10571	68	1613	3859	917	27	3309	75699

See also