Number of memberships

memnum

Source

General Social Survey https://gss.norc.org

Details

Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type? READ EACH ITEM. CODE OnE FOR EACH. Q. Any membership at all

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2004

B/C/-	1988, 1989, 1990, 1991, 1993, 1994

Full	1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987

Overview

For further details see the GSS Data Explorer page for memnum.

Counts by year:

year	iap	0	1	10	11	12	13	15	2	3	4	5	6	7	8	9	no answer	14	16	don't know	Total

1974	-	375	381	5	1	2	1	1	281	185	114	67	28	21	11	8	3	-	-	-	1484

1975	-	410	424	4	-	-	-	-	261	169	97	46	38	23	8	6	3	1	-	-	1490

1977	-	433	423	1	2	3	1	-	247	168	106	68	33	24	12	9	-	-	-	-	1530

1978	-	426	418	1	2	1	-	-	271	202	106	46	28	19	4	6	2	-	-	-	1532

1980	-	498	371	3	-	-	-	-	242	164	81	43	30	13	9	5	8	-	1	-	1468

1983	-	435	409	1	-	2	-	-	300	182	119	58	52	18	17	5	1	-	-	-	1599

1984	-	474	347	8	3	-	-	-	256	158	89	53	36	19	12	6	11	1	-	-	1473

1986	-	408	363	-	1	-	-	-	259	168	127	66	37	24	8	6	3	-	-	-	1470

1987	-	606	478	3	-	1	-	-	314	159	97	72	33	25	13	5	11	-	2	-	1819

1988	484	294	256	-	-	-	-	-	187	102	71	36	28	11	7	5	-	-	-	-	1481

1989	502	313	243	-	-	-	-	-	191	120	72	41	20	16	9	5	4	-	1	-	1537

1990	473	286	218	2	2	-	1	-	154	86	69	39	25	8	5	3	1	-	-	-	1372

1991	500	329	242	2	-	1	-	-	182	107	73	37	22	10	7	2	3	-	-	-	1517

1993	549	298	255	1	1	-	-	-	183	124	86	50	25	15	10	5	4	-	-	-	1606

1994	2481	146	125	-	-	-	-	-	82	67	34	27	12	9	6	1	2	-	-	-	2992

2004	1340	555	329	1	-	-	1	-	209	141	102	52	41	22	11	1	6	-	-	1	2812

Total	6329	6286	5282	32	12	10	4	1	3619	2302	1443	801	488	277	149	78	62	2	4	1	27182

See also