Membership in veteran group

memvet

Source

General Social Survey https://gss.norc.org

Details

Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type? READ EACH ITEM. CODE OnE FOR EACH. C. Veterans' groups

Values

1 yes

2 no

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	2004

B/C/-	1988, 1989, 1990, 1991, 1993, 1994

Full	1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987

Overview

For further details see the GSS Data Explorer page for memvet.

Counts by year:

year	iap	yes	no	no answer	don't know	Total

1974	-	132	1332	20	-	1484

1975	-	114	1350	25	1	1490

1977	-	126	1392	12	-	1530

1978	-	99	1416	16	1	1532

1980	-	108	1333	25	2	1468

1983	-	106	1485	8	-	1599

1984	-	101	1350	20	2	1473

1986	-	91	1370	9	-	1470

1987	-	105	1692	21	1	1819

1988	484	84	904	9	-	1481

1989	502	78	929	28	-	1537

1990	473	64	829	6	-	1372

1991	500	71	937	9	-	1517

1993	549	71	980	5	1	1606

1994	2481	40	469	2	-	2992

2004	1340	78	1389	5	-	2812

Total	6329	1468	19157	220	8	27182

See also