Days of poor mental health past 30 days

mntlhlth

Source

General Social Survey https://gss.norc.org

Details

Question B. Now thinking about your mental health, which includes stress, depression, and problems with emotions, for how many days during the past 30 days was your mental health not good?

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2002, 2004, 2006, 2010, 2012, 2014, 2018, 2022

B/C/-	2016

Overview

For further details see the GSS Data Explorer page for mntlhlth.

Counts by year:

year	iap	0	1	10	12	14	15	18	19	2	20	21	22	25	26	28	3	30	4	5	6	7	8	9	don't know	no answer	11	16	17	23	24	29	27	13	skipped on web	Total

2002	969	1050	64	63	3	11	54	1	1	117	32	3	2	9	2	4	84	86	44	87	5	39	4	1	14	16	-	-	-	-	-	-	-	-	-	2765

2004	945	769	125	92	6	16	57	2	-	175	58	1	1	16	-	2	140	51	76	140	15	45	16	2	10	40	1	2	3	3	1	2	-	-	-	2812

2006	2776	1098	62	45	4	6	40	-	-	124	25	-	1	4	-	3	76	58	48	77	11	27	3	-	6	11	-	-	1	-	-	2	2	-	-	4510

2010	857	646	63	40	3	10	41	-	-	77	17	2	-	9	1	-	48	47	43	68	7	22	3	1	8	28	1	-	-	-	-	1	1	-	-	2044

2012	795	518	88	46	10	3	34	2	-	132	31	2	1	6	-	3	67	24	38	88	8	21	5	1	10	30	-	3	1	-	1	5	1	-	-	1974

2014	1288	796	48	37	2	7	29	-	-	69	18	2	-	4	-	1	45	56	29	63	5	21	7	1	3	6	-	-	-	-	1	-	-	-	-	2538

2016	1745	609	47	45	1	12	33	-	-	92	19	2	-	4	1	-	63	52	24	60	4	22	3	1	8	15	-	2	1	-	1	-	1	-	-	2867

2018	900	788	76	60	5	13	41	1	-	107	19	-	2	7	-	-	85	51	32	75	9	26	6	1	7	33	1	2	-	-	-	1	-	-	-	2348

2022	1557	1001	82	89	12	18	65	-	1	126	48	1	1	18	-	5	82	90	51	141	19	38	7	-	16	22	1	2	1	1	2	1	-	1	45	3544

Total	11832	7275	655	517	46	96	394	6	2	1019	267	13	8	77	4	18	690	515	385	799	83	261	54	8	82	201	4	11	7	4	6	12	5	1	45	25402

See also