Degree to which faith is free of doubt

myfaith

Source

General Social Survey https://gss.norc.org

Details

Question Here is another card with contrasting ideas. If you feel that "My faith is completely free of doubts," you would place yourself at 1. If you feel "My faith is mixed with doubts," you would place yourself at 7. If you feel somewhere between these two, you would place yourself at 2, 3, 4, 5 or 6. Where would you place your feelings about your faith… A. From Free of Doubt to Mixed with doubts

Values

[1] faith free of doubt

[7] faith mixed w doubt

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988

Overview

For further details see the GSS Data Explorer page for myfaith.

Counts by year:

year	iap	faith free of doubt	2	3	4	5	6	faith mixed w doubt	don't know	no answer	Total

1988	-	401	267	230	241	138	73	105	7	19	1481

Total	0	401	267	230	241	138	73	105	7	19	1481

See also