National defense:version y

natarmsy

Source

General Social Survey https://gss.norc.org

Details

Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. i. National defense.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	974	192	23	12	87	185	-	1473

1985	751	302	36	2	123	320	-	1534

1986	730	332	29	6	119	254	-	1470

1987	603	475	63	13	211	454	-	1819

1988	718	291	51	10	134	277	-	1481

1989	768	312	45	9	116	287	-	1537

1990	674	294	27	15	65	297	-	1372

1991	756	425	30	15	96	195	-	1517

1993	799	315	31	12	94	355	-	1606

1994	1508	676	73	38	233	464	-	2992

1996	1441	648	86	60	231	438	-	2904

1998	1381	667	85	9	254	436	-	2832

2000	1409	618	87	4	351	348	-	2817

2002	1358	591	57	2	477	280	-	2765

2004	1401	538	37	3	445	388	-	2812

2006	3002	478	61	2	413	554	-	4510

2008	1004	345	36	3	268	367	-	2023

2010	1014	408	49	-	243	330	-	2044

2012	998	395	41	1	241	298	-	1974

2014	1274	436	53	2	392	381	-	2538

2016	1437	487	53	-	525	365	-	2867

2018	1188	449	42	3	341	325	-	2348

2021	1987	915	5	-	512	606	7	4032

2022	1780	684	50	1	550	476	3	3544

2024	1683	624	53	-	514	432	3	3309

Total	46217	11897	1203	222	7035	9112	13	75699

Values

1 too little

2 about right

3 too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook