Law enforcement:version y

natcrimy

Source

General Social Survey https://gss.norc.org

Details

Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. e. Law enforcement.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	974	173	13	12	268	33	-	1473

1985	751	262	33	5	440	43	-	1534

1986	730	284	27	8	373	48	-	1470

1987	603	451	42	13	627	83	-	1819

1988	718	273	28	11	409	42	-	1481

1989	768	221	29	10	460	49	-	1537

1990	674	238	24	16	380	40	-	1372

1991	756	271	22	14	407	47	-	1517

1993	799	251	33	12	466	45	-	1606

1994	1508	399	37	38	909	101	-	2992

1996	1441	436	44	59	801	123	-	2904

1998	1381	511	42	9	776	113	-	2832

2000	1409	534	51	5	706	112	-	2817

2002	1358	568	30	4	663	142	-	2765

2004	1401	506	33	1	738	133	-	2812

2006	3002	529	35	-	816	128	-	4510

2008	1005	365	17	1	543	92	-	2023

2010	1014	389	24	1	492	124	-	2044

2012	998	381	28	1	460	106	-	1974

2014	1274	484	29	1	581	169	-	2538

2016	1437	491	21	1	766	151	-	2867

2018	1188	379	41	3	615	122	-	2348

2021	1987	727	-	-	853	457	8	4032

2022	1780	489	32	-	961	281	1	3544

2024	1683	519	30	2	828	244	3	3309

Total	46218	10131	745	227	15338	3028	12	75699

Values

1 too little

2 about right

3 too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook