Drug rehabilitation:version y

natdrugy

Source

General Social Survey https://gss.norc.org

Details

Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)? F. Drug rehabilitation

Values

1 too little

2 about right

3 too much

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1984	-/-/-	full

1985	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/C	full

1990	A/B/C	full

1991	A/B/C	full

1993	A/B/C	full

1994	A/B/C	full

1996	A/B/C	full

1998	A/B/C	full

2000	A/B/C	full

2002	A/B/C	full

2004	A/B/C	full

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	A/B/C	full

2018	A/B/C	full

2021	A/B/C	full

2022	A/B/C	full

2024	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	974	186	30	10	228	45	-	1473

1985	751	237	60	4	416	66	-	1534

1986	730	234	43	6	389	68	-	1470

1987	603	331	70	14	713	88	-	1819

1988	718	217	45	9	429	63	-	1481

1989	768	197	33	7	468	64	-	1537

1990	674	157	25	16	441	59	-	1372

1991	756	220	41	14	430	56	-	1517

1993	799	234	45	14	435	79	-	1606

1994	1508	437	88	36	757	166	-	2992

1996	1441	401	95	57	743	167	-	2904

1998	1381	419	101	7	761	163	-	2832

2000	1409	422	109	5	730	142	-	2817

2002	1358	425	85	6	715	176	-	2765

2004	1401	477	92	3	685	154	-	2812

2006	3002	463	113	-	777	155	-	4510

2008	1005	346	65	3	475	129	-	2023

2010	1014	334	76	4	499	117	-	2044

2012	998	326	89	2	404	155	-	1974

2014	1274	398	83	2	617	164	-	2538

2016	1437	357	70	2	850	151	-	2867

2018	1188	255	47	2	746	110	-	2348

2021	1987	595	11	-	1225	203	11	4032

2022	1780	435	56	-	1126	142	5	3544

2024	1683	432	54	1	979	159	1	3309

Total	46218	8535	1626	224	16038	3041	17	75699

See also