Improving & protecting environment

natenvir

Source

General Social Survey https://gss.norc.org

Details

Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. b. Improving and protecting the environment.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	-	387	85	6	915	111	-	1504

1974	-	393	98	8	871	114	-	1484

1975	-	459	92	-	795	144	-	1490

1976	-	468	69	5	818	139	-	1499

1977	-	522	110	6	724	168	-	1530

1978	-	501	80	4	801	146	-	1532

1980	-	454	83	3	702	226	-	1468

1982	-	559	124	4	964	209	-	1860

1983	-	498	98	7	865	131	-	1599

1984	983	159	19	5	286	21	-	1473

1985	783	236	30	4	421	60	-	1534

1986	740	213	43	5	431	38	-	1470

1987	1216	148	30	4	395	26	-	1819

1988	763	188	29	4	464	33	-	1481

1989	769	155	32	5	548	28	-	1537

1990	698	127	35	5	476	31	-	1372

1991	761	180	23	5	511	37	-	1517

1993	807	239	36	5	445	74	-	1606

1994	1484	443	51	2	885	127	-	2992

1996	1463	391	61	1	836	152	-	2904

1998	1451	386	59	4	826	106	-	2832

2000	1408	385	52	2	858	112	-	2817

2002	1407	439	33	3	793	90	-	2765

2004	1411	405	38	1	862	95	-	2812

2006	3026	365	38	-	992	89	-	4510

2008	1019	236	28	-	659	81	-	2023

2010	1030	297	36	2	557	122	-	2044

2012	976	306	25	2	559	106	-	1974

2014	1264	394	29	1	731	119	-	2538

2016	1430	409	35	1	878	114	-	2867

2018	1160	284	30	1	790	83	-	2348

2021	2045	413	6	-	1376	190	2	4032

2022	1764	349	27	-	1210	193	1	3544

2024	1626	370	35	-	1085	188	5	3309

Total	33097	11758	1699	105	25329	3703	8	75699

Values

1 too little

2 about right

3 too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook