Parks and recreation

natpark

Source

General Social Survey https://gss.norc.org

Details

Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. o. Parks and recreation.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	484	569	34	15	313	58	-	-	1473

1985	-	889	65	4	468	108	-	-	1534

1986	-	874	67	11	429	89	-	-	1470

1987	-	1021	76	16	593	113	-	-	1819

1988	-	894	72	13	421	81	-	-	1481

1989	-	859	94	9	497	78	-	-	1537

1990	-	783	82	19	412	76	-	-	1372

1991	-	901	70	21	450	75	-	-	1517

1993	-	920	90	15	480	101	-	-	1606

1994	-	1769	116	45	874	188	-	-	2992

1996	-	1616	139	61	920	168	-	-	2904

1998	-	1564	124	8	963	173	-	-	2832

2000	-	1547	116	8	991	155	-	-	2817

2002	-	1598	76	4	940	147	-	-	2765

2004	-	1685	54	3	879	191	-	-	2812

2006	1518	1734	70	2	996	190	-	-	4510

2008	-	1252	45	1	617	108	-	-	2023

2010	-	1201	41	1	672	129	-	-	2044

2012	-	1185	51	1	620	117	-	-	1974

2014	-	1540	47	2	790	159	-	-	2538

2016	-	1692	48	2	952	173	-	-	2867

2018	-	1346	60	-	839	103	-	-	2348

2021	-	2412	10	-	1428	162	20	-	4032

2022	-	1953	58	-	1347	178	8	-	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	17581	31804	1705	261	17891	3120	28	3309	75699

Values

1 too little

2 about right

3 too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook