Assistance to black people: version y

natracey

Source

General Social Survey https://gss.norc.org

Details

Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?H. Assistance to Black people.

Values

1 too little

2 about right

3 too much

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1984	-/-/-	full

1985	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/C	full

1990	A/B/C	full

1991	A/B/C	full

1993	A/B/C	full

1994	A/B/C	full

1996	A/B/C	full

1998	A/B/C	full

2000	A/B/C	full

2002	A/B/C	full

2004	A/B/C	full

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	A/B/C	full

2018	A/B/C	full

2021	A/B/C	full

2022	A/B/C	full

2024	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	974	203	48	13	119	116	-	1473

1985	751	314	74	2	200	193	-	1534

1986	730	331	45	7	166	191	-	1470

1987	603	451	101	11	433	220	-	1819

1988	718	301	90	10	191	171	-	1481

1989	768	308	79	8	207	167	-	1537

1990	674	291	53	15	197	142	-	1372

1991	756	299	76	18	227	141	-	1517

1993	799	306	108	14	188	191	-	1606

1994	1508	565	156	36	349	378	-	2992

1996	1441	544	187	55	334	343	-	2904

1998	1381	565	180	9	381	316	-	2832

2000	1409	522	172	9	425	280	-	2817

2002	1358	591	133	11	338	334	-	2765

2004	1401	615	165	18	363	250	-	2812

2006	3002	608	195	12	378	315	-	4510

2008	1004	404	118	9	279	209	-	2023

2010	1014	427	147	14	252	190	-	2044

2012	998	407	156	12	206	195	-	1974

2014	1274	510	161	15	316	262	-	2538

2016	1437	535	139	6	534	216	-	2867

2018	1188	379	128	6	507	140	-	2348

2021	1987	692	16	3	930	384	20	4032

2022	1780	566	97	8	808	278	7	3544

2024	1683	578	115	8	632	286	7	3309

Total	46217	11312	2939	329	8960	5908	34	75699

See also