Highways and bridges

natroad

Source

General Social Survey https://gss.norc.org

Details

Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)? L. Highways and bridges

Values

[1] too little

[2] about right

[3] too much

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for natroad.

Counts by year:

year	iap	too little	about right	too much	don't know	no answer	skipped on web	Total

1984	484	452	419	65	37	16	-	1473

1985	-	627	705	116	76	10	-	1534

1986	-	511	760	122	66	11	-	1470

1987	-	606	917	169	103	24	-	1819

1988	-	527	758	111	63	22	-	1481

1989	-	583	739	106	96	13	-	1537

1990	-	592	586	86	87	21	-	1372

1991	-	511	749	155	82	20	-	1517

1993	-	603	764	144	80	15	-	1606

1994	-	1141	1452	222	128	49	-	2992

1996	-	1018	1398	256	169	63	-	2904

1998	-	1082	1333	272	131	14	-	2832

2000	-	954	1398	324	129	12	-	2817

2002	-	950	1377	337	98	3	-	2765

2004	-	844	1532	360	74	2	-	2812

2006	1518	1049	1507	343	90	3	-	4510

2008	-	917	847	193	66	-	-	2023

2010	-	887	888	211	56	2	-	2044

2012	-	847	825	248	54	-	-	1974

2014	-	1161	1020	301	55	1	-	2538

2016	-	1423	1112	285	44	3	-	2867

2018	-	1283	817	209	39	-	-	2348

2021	-	2281	1504	219	6	-	22	4032

2022	-	1885	1269	334	50	1	5	3544

Total	2002	22734	24676	5188	1879	305	27	56811

See also