Space exploration program

natspac

Source

General Social Survey https://gss.norc.org

Details

Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)? A. Space exploration program

Values

1 too little

2 about right

3 too much

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1973	-/-/-	full

1974	-/-/-	full

1975	-/-/-	full

1976	-/-/-	full

1977	-/-/-	full

1978	-/-/-	full

1980	-/-/-	full

1982	-/-/-	full

1983	-/-/-	full

1984	-/-/-	full

1985	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/C	full

1990	A/B/C	full

1991	A/B/C	full

1993	A/B/C	full

1994	A/B/C	full

1996	A/B/C	full

1998	A/B/C	full

2000	A/B/C	full

2002	A/B/C	full

2004	A/B/C	full

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	A/B/C	full

2018	A/B/C	full

2021	A/B/C	full

2022	A/B/C	full

2024	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	-	441	71	1	112	879	-	1504

1974	-	408	53	4	114	905	-	1484

1975	-	449	65	-	110	866	-	1490

1976	-	419	37	3	137	903	-	1499

1977	-	527	90	-	154	759	-	1530

1978	-	536	96	-	177	723	-	1532

1980	-	506	122	2	264	574	-	1468

1982	-	682	121	2	201	854	-	1860

1983	-	647	96	1	221	634	-	1599

1984	983	210	27	3	58	192	-	1473

1985	783	331	27	3	86	304	-	1534

1986	740	316	30	3	82	299	-	1470

1987	1216	204	38	1	87	273	-	1819

1988	763	300	41	4	127	246	-	1481

1989	769	338	49	2	113	266	-	1537

1990	698	294	39	3	75	263	-	1372

1991	761	321	56	6	88	285	-	1517

1993	807	295	54	3	67	380	-	1606

1994	1484	562	82	1	136	727	-	2992

1996	1463	577	103	2	158	601	-	2904

1998	1451	605	92	2	139	543	-	2832

2000	1408	529	105	2	195	578	-	2817

2002	1407	640	87	2	150	479	-	2765

2004	1411	598	75	1	186	541	-	2812

2006	3026	651	74	3	211	545	-	4510

2008	1019	459	67	3	121	354	-	2023

2010	1030	411	72	1	169	361	-	2044

2012	976	399	94	5	203	297	-	1974

2014	1264	518	121	3	298	334	-	2538

2016	1430	661	144	3	281	348	-	2867

2018	1160	528	158	1	230	271	-	2348

2021	2045	1043	11	-	488	438	7	4032

2022	1764	831	90	2	358	496	3	3544

2024	1626	803	94	1	309	468	8	3309

Total	33097	17039	2581	73	5905	16986	18	75699

See also