Space exploration:version y

natspacy

Source

General Social Survey https://gss.norc.org

Details

Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)? A. Space exploration

Values

[1] too little

[2] about right

[3] too much

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for natspacy.

Counts by year:

year	iap	too little	about right	too much	don't know	no answer	skipped on web	Total

1984	974	49	197	225	18	10	-	1473

1985	751	79	332	334	35	3	-	1534

1986	730	69	345	283	36	7	-	1470

1987	603	160	491	495	59	11	-	1819

1988	718	147	272	284	51	9	-	1481

1989	768	108	327	285	41	8	-	1537

1990	674	74	298	278	33	15	-	1372

1991	756	92	338	262	53	16	-	1517

1993	799	52	283	427	35	10	-	1606

1994	1508	125	579	680	66	34	-	2992

1996	1441	176	572	559	101	55	-	2904

1998	1381	161	609	565	109	7	-	2832

2000	1409	180	537	587	101	3	-	2817

2002	1358	163	647	521	74	2	-	2765

2004	1401	184	597	558	71	1	-	2812

2006	3002	218	645	550	94	1	-	4510

2008	1005	138	416	412	51	1	-	2023

2010	1014	155	451	348	73	3	-	2044

2012	998	222	379	288	86	1	-	1974

2014	1274	291	496	380	97	-	-	2538

2016	1437	307	628	379	115	1	-	2867

2018	1188	264	515	278	100	3	-	2348

2021	1987	461	1071	498	9	-	6	4032

2022	1780	360	836	470	96	1	1	3544

Total	28956	4235	11861	9946	1604	202	7	56811

See also