How often does r read newspaper

news

Source

General Social Survey https://gss.norc.org

Details

Question How often do you read the newspaper–every day, a few times a week, once a week, less than once a week, or never?

Values

1 every day

2 a few times a week

3 once a week

4 less than once a week

5 never

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/D	2006

Full	1972, 1975, 1977, 1978, 1982, 1983, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for news.

Counts by year:

year	every day	a few times a week	once a week	less than once a week	never	no answer	iap	skipped on web	don't know	Total

1972	1105	241	128	69	68	2	-	-	-	1613

1975	981	235	126	81	65	2	-	-	-	1490

1977	952	254	148	100	73	3	-	-	-	1530

1978	874	308	154	113	79	4	-	-	-	1532

1982	961	425	218	138	112	6	-	-	-	1860

1983	891	331	170	127	80	-	-	-	-	1599

1985	807	324	196	116	87	4	-	-	-	1534

1986	789	289	184	116	90	2	-	-	-	1470

1987	949	391	231	130	110	8	-	-	-	1819

1988	500	235	118	87	48	-	493	-	-	1481

1989	502	248	130	76	49	1	531	-	-	1537

1990	488	205	100	89	45	1	444	-	-	1372

1991	525	220	139	87	51	2	493	-	-	1517

1993	497	282	136	96	64	5	526	-	-	1606

1994	969	452	261	196	76	-	1038	-	-	2992

1996	825	476	311	225	120	3	944	-	-	2904

1998	805	420	294	202	149	1	961	-	-	2832

2000	698	459	277	277	162	-	944	-	-	2817

2002	369	219	109	121	90	-	1857	-	-	2765

2004	358	222	134	121	71	-	1906	-	-	2812

2006	945	611	418	404	350	-	1782	-	-	4510

2008	431	300	207	200	191	-	694	-	-	2023

2010	449	277	215	220	269	-	614	-	-	2044

2012	353	230	195	217	306	1	672	-	-	1974

2014	417	260	246	271	481	-	863	-	-	2538

2016	398	270	229	300	688	2	980	-	-	2867

2018	312	174	184	211	678	-	789	-	-	2348

2021	642	361	240	452	1001	-	1329	7	-	4032

2022	383	267	195	329	1191	1	1168	6	4	3544

Total	19175	8986	5693	5171	6844	48	19028	13	4	64962

See also