True-false: the sun never shines at the south pole

nosun

Source

General Social Survey https://gss.norc.org

Details

Question 1077. The next few questions are about the Arctic and the Antarctic. The Arctic is the region around the North Pole; Antarctic is the region that contains the South Pole. These questions are like ones you might see on a television game show. If you don't know or aren't sure, just tell me so, and we will skip to the next question. Remember true, false, or don't know. b. The sun never shines at the South Pole. (Is that true or false?)

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	can't choose	false	no answer	true	not available in this year	Total

1972	1613	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	2992

1996	2904	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	2812

2006	2646	493	1217	1	153	-	4510

2008	2023	-	-	-	-	-	2023

2010	1338	132	484	11	79	-	2044

2012	1974	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	4032

2022	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	3309	3309

Total	66276	625	1701	12	232	6853	75699

Values

1 true

2 false

NA(d) can't choose

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook