Number of male sex partner's since 18

nummen

Source

General Social Survey https://gss.norc.org

Details

Question Now thinking about the time since your 18th birthday (including the past 12 months) how many male partners have you had sex with?

Values

989 989 or more

990 dash or slash

991 1 or more, number unknown

992 'x'

993 garbled text

994 'several'

995 'many,' 'lots'

996 'n.a.'

997 refused

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1989	A/B/C	full

1990	A/B/C	full

1991	A/B/C	full

1993	A/B/C	full

1994	A/B/C	full

1996	A/B/C	full

1998	A/B/C	full

2000	A/B/C	full

2002	A/B/C	full

2004	A/B/C	full

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	B/C/-	partial

2018	B/C/-	partial

2021	B/C/-	partial

2022	B/C/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	'many,' 'lots'	0	1	10	11	12	14	15	16	2	20	25	3	30	300	35	4	40	5	55	6	7	8	9	dash or slash	don't know	garbled text	no answer	19	23	45	50	refused	'n.a.'	'several'	'x'	100	13	22	26	375	90	1 or more, number unknown	17	18	200	21	336	150	24	34	350	36	60	70	27	33	65	41	43	500	51	75	120	122	28	32	400	37	42	48	103	110	39	989 or more	175	52	54	600	99	62	80	700	85	84	83	125	365	38	397	88	56	57	123	140	250	29	47	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	136	2	574	324	22	1	4	2	11	1	99	9	4	75	5	1	1	50	1	34	1	27	11	10	3	6	9	1	113	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	199	2	494	254	15	-	10	1	5	-	81	7	6	67	-	-	-	43	2	32	-	21	6	9	3	1	5	-	100	1	1	1	3	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	221	-	496	280	27	2	5	-	10	-	117	11	2	84	4	-	2	40	3	54	-	20	13	9	4	2	5	-	87	-	1	-	-	5	3	1	2	2	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	114	2	615	298	33	7	9	-	10	2	123	6	4	85	2	-	-	56	-	43	-	35	20	11	5	1	9	-	85	-	-	2	3	6	1	-	-	-	-	-	-	-	-	10	2	2	3	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	201	-	1069	536	72	2	21	4	26	1	212	24	9	166	4	1	2	102	3	93	-	55	31	24	7	12	14	2	217	-	1	-	11	30	7	1	3	4	5	2	-	-	-	5	-	3	-	1	-	2	1	1	1	1	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	258	11	1033	431	85	1	16	2	23	2	187	30	7	151	14	-	1	106	2	82	-	49	25	28	14	4	11	3	215	1	1	2	3	25	2	2	-	3	4	2	1	-	-	52	1	2	1	4	-	1	-	-	-	-	1	2	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2904

1998	383	4	990	439	59	1	9	2	23	1	176	21	4	149	10	2	-	96	-	114	-	62	39	19	9	1	14	-	162	-	-	-	2	9	10	1	1	3	-	-	-	-	-	5	1	2	1	1	-	-	-	-	-	-	1	-	-	-	-	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	417	4	906	386	71	4	17	3	19	-	171	25	7	132	7	-	1	111	3	102	-	60	24	23	9	4	12	-	228	-	-	1	7	27	1	-	5	4	2	4	1	-	-	5	-	3	1	-	-	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	3	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	489	2	982	332	62	3	13	1	28	1	156	15	13	138	9	-	1	84	2	128	-	52	33	19	8	-	31	-	131	-	1	1	1	2	-	-	-	6	3	1	2	-	-	4	-	1	-	1	-	-	1	-	-	-	-	-	2	-	1	-	-	-	-	-	-	-	1	-	-	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	584	-	974	359	65	3	18	5	18	-	153	14	10	127	9	-	1	81	5	84	-	64	34	19	15	-	30	-	90	-	1	-	3	1	1	-	-	9	2	1	2	-	-	15	1	2	1	2	-	-	-	-	-	1	1	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	2096	4	1028	379	64	2	11	2	25	3	174	31	10	152	6	1	2	112	2	107	-	56	36	24	7	-	18	-	100	-	3	2	10	3	-	5	-	8	1	4	1	-	-	-	1	4	2	-	-	-	1	-	-	-	-	1	1	-	-	-	-	-	1	2	-	-	1	-	-	-	-	-	-	2	-	-	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510

2008	240	2	757	314	50	2	11	3	13	3	137	23	7	102	8	-	4	68	4	71	-	35	12	17	6	-	30	-	60	-	-	2	5	2	-	-	-	6	6	1	-	-	-	1	3	6	-	-	-	-	1	-	-	-	1	1	1	1	1	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	2	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	202	-	809	247	58	5	9	2	24	2	121	18	7	114	5	-	1	82	4	84	-	53	20	21	7	-	24	-	88	-	-	-	9	6	-	-	-	4	2	1	-	-	-	1	-	3	-	-	-	-	1	1	-	1	-	-	2	-	-	-	-	-	-	2	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	238	4	715	253	46	3	16	1	19	-	113	17	12	102	8	2	1	80	4	70	-	39	21	23	13	-	30	1	88	1	-	1	5	1	1	1	-	1	-	2	-	-	-	22	2	3	2	2	-	-	1	-	-	-	2	-	-	1	-	-	-	-	-	1	1	-	2	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1974

2014	188	1	996	337	69	8	9	2	16	5	168	23	13	158	9	1	2	89	1	126	-	50	32	36	7	-	116	-	28	-	1	1	9	1	2	2	-	5	5	2	-	-	-	5	1	2	1	2	-	1	1	-	-	-	2	-	-	3	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2016	1096	-	757	242	54	3	10	2	18	3	112	27	7	115	9	-	4	78	3	87	1	51	17	16	11	-	9	1	83	-	1	1	4	1	1	1	-	6	7	2	-	-	-	2	2	6	1	1	-	1	-	-	-	-	3	1	-	1	-	-	-	1	-	2	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	1	1	1	-	-	-	-	-	-	-	-	2867

2018	942	-	607	200	50	3	10	3	23	-	86	16	6	69	5	-	-	76	2	74	-	33	20	13	2	-	15	-	51	-	-	1	6	-	-	-	-	-	3	-	2	-	-	16	1	3	-	-	-	-	1	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	-	-	1	-	-	-	-	1	-	-	1	1	-	-	-	-	-	-	2348

2021	1705	3	965	309	98	2	30	4	30	6	137	39	12	111	23	-	-	75	1	125	-	49	36	38	14	-	24	-	113	-	1	2	13	1	1	-	-	8	7	2	-	-	-	9	3	6	4	1	-	1	-	-	-	-	1	1	2	-	1	-	-	1	-	3	1	-	2	1	-	-	1	-	-	-	-	4	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	1	1	1	1	-	4032

2022	1927	-	654	193	42	3	10	1	14	4	89	21	11	90	8	-	2	213	1	74	-	45	21	13	5	-	-	-	76	-	2	-	9	-	-	-	-	2	1	1	-	-	-	-	1	1	-	1	-	1	-	1	-	-	1	-	1	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	-	-	-	1	-	-	3544

2024	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3309	3309

Total	34992	41	15421	6113	1042	55	238	40	355	34	2612	377	151	2187	145	8	25	1642	43	1584	2	856	451	372	149	31	406	8	2115	3	14	17	103	123	30	14	11	71	49	26	10	1	1	152	19	49	17	17	1	10	8	3	1	3	15	7	12	9	5	1	2	4	2	12	3	3	8	4	1	2	3	2	1	3	1	9	1	1	1	1	2	2	3	1	2	1	1	1	1	2	1	2	1	3	1	1	1	2	1	3309	75699

See also