Number of people working in organization at all locations

numorg

Source

General Social Survey https://gss.norc.org

Details

Question About how many people work in your organization at all locations? Count part-time as well as full-time employees.

Values

[9981] 1-9 in range

[9982] 10-49 in range

[9983] 50-99 in range

[9984] 100-499 in range

[9985] 500-999 in range

[9986] 1,000-1,999 in range

[9987] 2,000-2,999 in range

[9988] 3,000-3,999 in range

[9989] 4,000-4,999 in range

[9990] 5,000-10,000 in range

[9991] 10,000-50,000 in range

[9992] 50,000-100,000 in range

[9993] above 100,000 in range

[9997] 10,000 or more

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2006, 2010, 2014, 2018

Overview

For further details see the GSS Data Explorer page for numorg.

Counts by year:

year	iap	1	10	100	1000	1001	11	110	1100	113	12	120	1200	13	130	14	140	15	150	1500	154	155	16	160	1600	163	165	17	172	175	18	180	1800	19	2	20	200	2000	207	21	22	225	23	2300	234	240	25	250	2500	258	2600	27	3	30	300	3000	309	3200	33	35	350	3500	37	370	3800	4	40	400	4000	42	425	43	4300	44	45	450	467	48	49	5	50	500	5000	532	55	550	59	6	60	600	6000	61	63	630	65	6500	66	7	70	700	7000	72	745	75	750	7500	76	8	80	800	8000	83	85	8500	9	90	900	9000	92	9800	1-9 in range	10-49 in range	50-99 in range	100-499 in range	500-999 in range	1,000-1,999 in range	2,000-2,999 in range	3,000-3,999 in range	4,000-4,999 in range	5,000-10,000 in range	10,000-50,000 in range	50,000-100,000 in range	above 100,000 in range	10,000 or more	don't know	no answer	1700	28	280	31	32	38	47	480	52	650	850	24	260	375	4500	56	73	78	99	999	105	106	116	125	1300	161	185	190	210	220	275	36	499	5100	62	670	86	95	Total

2006	2778	88	22	30	34	2	9	2	1	1	15	10	5	5	3	5	1	19	20	6	1	1	6	1	1	1	2	7	2	3	2	1	4	3	31	22	31	33	1	1	2	1	3	1	1	3	14	11	6	1	1	1	38	14	25	9	1	2	1	9	3	2	1	1	1	26	15	10	3	2	1	1	2	1	5	4	1	2	3	37	39	32	17	1	2	1	1	21	11	9	6	1	2	1	3	1	1	20	11	4	5	1	1	5	1	1	1	19	13	5	2	2	1	1	22	2	1	1	1	1	23	66	52	95	51	37	35	12	9	35	36	15	46	80	149	24	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510

2010	857	71	7	10	5	-	6	-	1	-	5	-	-	8	-	1	-	5	8	3	-	-	3	1	-	-	-	1	-	1	1	-	1	-	28	10	7	8	-	1	3	1	-	-	-	-	1	1	1	-	-	-	22	13	4	2	-	-	-	4	-	-	-	-	-	13	5	5	-	-	-	-	-	-	1	-	-	-	1	18	11	9	7	-	-	-	-	14	4	1	-	-	-	-	1	-	-	10	4	1	1	1	-	-	-	1	-	9	1	2	1	-	-	-	13	4	-	-	-	-	84	110	71	114	56	28	29	15	20	38	47	21	47	26	-	87	1	1	1	2	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2014	1765	56	5	5	4	-	7	1	-	-	3	3	2	4	2	3	-	5	5	1	-	-	-	1	-	-	-	-	-	-	2	1	-	-	29	4	6	5	-	-	-	-	1	-	-	-	2	2	-	-	-	-	14	3	5	2	-	-	-	1	-	-	-	-	1	17	3	5	-	-	-	-	-	-	-	-	-	-	2	9	11	10	3	-	-	-	-	6	1	1	1	-	1	-	3	-	-	5	1	-	-	-	-	4	-	-	-	1	-	3	-	-	-	-	10	1	1	-	-	-	30	56	28	86	29	22	15	11	11	31	29	19	46	23	-	42	-	-	-	1	-	-	-	-	-	-	-	1	1	2	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2018	900	62	25	36	33	-	4	1	1	-	14	7	3	8	2	1	2	17	16	7	-	1	5	2	-	-	1	4	-	1	4	2	1	1	26	30	24	28	-	-	-	-	-	-	-	1	12	11	4	-	-	1	26	15	20	13	-	1	-	6	9	1	1	1	-	19	11	19	7	1	-	-	-	-	4	4	-	-	-	15	50	24	15	-	-	3	-	15	6	5	5	-	-	-	3	2	-	16	10	4	3	-	-	6	1	-	-	14	4	8	3	-	1	-	13	1	2	3	-	-	9	38	31	37	30	27	16	15	9	28	29	21	67	100	124	44	2	1	-	-	2	3	-	-	-	1	-	3	1	-	1	-	-	-	1	-	2	1	1	2	1	1	1	1	1	1	1	1	1	1	1	1	1	1	2348

Total	6300	277	59	81	76	2	26	4	3	1	37	20	10	25	7	10	3	46	49	17	1	2	14	5	1	1	3	12	2	5	9	4	6	4	114	66	68	74	1	2	5	2	4	1	1	4	29	25	11	1	1	2	100	45	54	26	1	3	1	20	12	3	2	2	2	75	34	39	10	3	1	1	2	1	10	8	1	2	6	79	111	75	42	1	2	4	1	56	22	16	12	1	3	1	10	3	1	51	26	9	9	2	1	15	2	2	1	43	18	18	6	2	2	1	58	8	4	4	1	1	146	270	182	332	166	114	95	53	49	132	141	76	206	229	273	197	3	2	1	3	3	4	1	1	1	2	1	4	2	2	2	1	1	1	2	1	2	1	1	2	1	1	1	1	1	1	1	1	1	1	1	1	1	1	11440

See also