To obey

obey

Source

General Social Survey https://gss.norc.org

Details

Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO OBEY

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	fifth	first	fourth	no answer	second	third	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	1534

1986	730	96	168	235	8	116	117	-	1470

1987	-	326	433	525	19	250	266	-	1819

1988	484	152	226	286	20	122	191	-	1481

1989	502	167	199	329	35	137	168	-	1537

1990	473	147	165	285	28	108	166	-	1372

1991	500	179	195	343	35	123	142	-	1517

1993	549	159	191	361	26	146	174	-	1606

1994	981	273	357	759	33	261	328	-	2992

1996	979	279	341	689	30	250	336	-	2904

1998	921	288	355	680	31	237	320	-	2832

2000	921	283	389	662	33	230	299	-	2817

2002	1853	144	131	367	13	112	145	-	2765

2004	1936	127	107	365	8	123	146	-	2812

2006	3508	133	184	389	7	119	170	-	4510

2008	658	206	212	555	14	155	223	-	2023

2010	667	204	193	537	28	170	245	-	2044

2012	636	210	182	537	15	156	238	-	1974

2014	851	292	204	718	17	203	253	-	2538

2016	911	327	239	847	22	211	310	-	2867

2018	785	260	167	706	21	169	240	-	2348

2021	1360	674	127	1318	99	160	294	-	4032

2022	1173	594	169	1085	77	176	270	-	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	39964	5520	4934	12578	619	3734	5041	3309	75699

Values

1 first

2 second

3 third

4 fourth

5 fifth

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook