R's census occupation code (1970)

occ

Source

General Social Survey https://gss.norc.org

Details

Question A. What kind of work do you (did you normally) do? That is, what (is/was) your job called? OCCUPATION: (See REMARKS) B. What (do/did) you actually do in that job? Tell me, what (are/were) some of your main duties? C. What kind of place (do/did) you work for? INDUSTRY: (See REMARKS) D. What (do/did) they (make/do)? E. (Are/Were) you self employed or (do/did) you work for someone else? F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)? G. About how much time (does/did) it usually take you to travel to work - about how many minutes?

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990

Full	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

See also