Status of job higher than father's?

occmobil

Source

General Social Survey https://gss.norc.org

Details

Question Please think of your present job (or your last one if you don't have one now). If you compare this job with the job your father had when you were 16, would you say that the level or status of your job is (or was) . . .

Values

1 much higher than dad

2 higher

3 about equal

4 lower

5 much lower than dad

6 r never had job

7 no dad-never had job

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1987	-/-/-	full

2000	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for occmobil.

See also