Which of these procedures would be the worst?

organsw

Source

General Social Survey https://gss.norc.org

Details

Question Recently, medical science has made it possible to save lives by transplanting body parts from donors to patients who need them. But for some body organs there are not enough to go around, and some patients die before they can obtain a transplant. When only one organ is available and several patients need it for survival, the organ could be assigned to a patient by one of the following procedures: (READ PROCEDURES ON SHOWCARD) By auction: The organ is assigned to the patient who can pay the most for it. By first come, first served: the organ is assigned to the patient who has been waiting the longest. By lottery: The organ is assigned to the patient whose name is drawn at random. By merit: The organ is assigned to the patient who can make the greatest contribution to others and society. B. Which of these procedures would be the worst way to assign an organ? (CODE ONE RESPONSE ONLY)

Values

[1] auction

[2] first come-first served

[3] lottery

[4] merit

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1996

Overview

For further details see the GSS Data Explorer page for organsw.

Counts by year:

year	iap	auction	first come-first served	lottery	merit	don't know	no answer	Total

1996	1460	1102	16	110	171	37	8	2904

Total	1460	1102	16	110	171	37	8	2904

See also