R has too much work to do well

overwork

Source

General Social Survey https://gss.norc.org

Details

Question Now I”m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements. F. I have too much work to do everything well

Values

1 strongly agree

2 agree

3 disagree

4 strongly disagree

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2002	A/B/C	full

2006	A/B/C	full

2010	A/B/C	full

2014	A/B/C	full

2018	A/B/C	full

2022	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	agree	disagree	don't know	no answer	strongly agree	strongly disagree	skipped on web	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	-	-	2992

1996	2904	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	2817

2002	969	367	1083	14	13	113	206	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	2812

2006	2776	361	1063	10	13	98	189	-	-	4510

2008	2023	-	-	-	-	-	-	-	-	2023

2010	857	285	668	4	23	93	114	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	1974

2014	1288	321	732	1	7	100	89	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	2867

2018	900	369	786	5	30	127	131	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	4032

2022	1546	472	1100	14	12	185	195	20	-	3544

2024	-	-	-	-	-	-	-	-	3309	3309

Total	62977	2175	5432	48	98	716	924	20	3309	75699

See also