How much r thinks doctor makes

paydoc

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. Please write in how much they actually earn each year. B. A doctor in general practice?

Values

[1e+06] $1 million (or more, for 1987 & 2000)

[99999997] never worked

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2000

B/C/-	2021

Full	1987

Overview

For further details see the GSS Data Explorer page for paydoc.

Counts by year:

year	iap	10000	105000	109000	110000	12000	120000	12500	125000	130000	135000	140000	15000	150000	16000	165000	175000	18000	180000	1e+05	$1 million (or more, for 1987 & 2000)	2000	20000	208000	22000	225000	23040	24000	25000	250000	27000	28000	29000	2e+05	3000	30000	312000	32000	3400	34000	35000	350000	36000	38000	3e+05	40000	42000	44000	45000	46000	47000	48000	49000	4e+05	5000	50000	52000	5500	55000	56000	57600	58000	5e+05	6000	60000	6200	65000	67000	67600	68000	6e+05	70000	72800	75000	75500	7e+05	8000	80000	82000	85000	86950	87500	89000	8e+05	9000	90000	95000	98000	can't choose	no answer	1000	11000	115000	1250	13000	15	155000	160000	170000	185000	2400	240000	260000	275000	3100	360000	425000	440000	62050	63000	750000	84000	86000	96000	0	1	1.5e+07	118000	1200	123000	1250000	130	1300000	13325	1334	136000	138000	145000	146000	149000	150	1500000	1555000	156000	169000	180	1800000	189000	19000	190000	195000	1e+07	2.5e+07	200	21	210000	215000	220000	22500	230000	235000	2400000	250	2500000	270000	27500000	29	299000	2e+06	2e+07	300009	313000	320000	329000	358990	3e+06	3e+07	40	4500	450000	47274	475000	499000	4e+07	500	599000	5e+06	5e+07	6	650000	7000	72000	75	775000	7e+06	88000	9600	99000	999000	9e+05	skipped on web	Total

1987	255	10	2	1	2	2	19	1	18	3	3	4	7	94	1	1	7	1	3	280	3	1	10	4	1	1	1	2	8	15	1	1	2	52	1	30	1	1	1	1	22	1	1	3	11	44	1	1	25	1	1	3	2	5	2	134	4	1	10	1	1	2	6	1	106	1	27	1	1	1	1	58	1	87	1	1	1	92	1	14	1	1	1	3	1	32	3	2	65	182	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

2000	1545	1	1	-	10	-	38	-	32	3	3	5	4	125	-	-	11	-	7	203	12	-	4	-	-	2	-	2	7	42	-	-	1	109	-	8	-	3	-	-	6	4	-	1	26	11	-	-	11	-	-	-	-	9	-	53	1	-	4	-	-	-	13	1	46	-	13	-	-	-	2	29	-	44	-	-	1	75	-	14	-	-	1	2	-	29	1	1	68	147	1	1	1	1	1	1	1	1	1	1	1	1	1	2	1	1	1	1	1	1	2	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2021	2029	15	-	-	7	3	62	-	27	16	-	9	2	232	1	2	43	-	31	197	14	3	7	-	-	13	-	-	4	185	-	-	-	343	-	6	-	-	-	-	2	28	-	-	132	2	-	-	1	-	-	-	-	39	-	17	1	-	1	-	-	-	82	-	13	-	1	-	-	-	10	7	-	19	-	2	-	31	1	5	-	-	1	6	1	19	2	1	12	123	2	1	3	-	-	-	1	16	6	1	-	3	2	6	-	1	1	-	-	-	7	1	-	1	2	1	1	1	1	1	3	1	1	1	1	1	1	3	1	2	2	6	1	1	1	1	1	1	1	3	2	3	1	2	1	2	1	10	1	1	1	1	2	4	1	1	1	1	7	3	2	1	2	1	1	2	2	1	1	5	1	1	1	1	1	1	3	2	1	2	1	1	1	1	1	1	1	2	1	2	40	4032

Total	3829	26	3	1	19	5	119	1	77	22	6	18	13	451	2	3	61	1	41	680	29	4	21	4	1	16	1	4	19	242	1	1	3	504	1	44	1	4	1	1	30	33	1	4	169	57	1	1	37	1	1	3	2	53	2	204	6	1	15	1	1	2	101	2	165	1	41	1	1	1	13	94	1	150	1	3	2	198	2	33	1	1	3	11	2	80	6	4	145	452	3	2	4	1	1	1	2	17	7	2	1	4	3	8	1	2	2	1	1	1	9	2	1	2	2	1	1	1	1	1	3	1	1	1	1	1	1	3	1	2	2	6	1	1	1	1	1	1	1	3	2	3	1	2	1	2	1	10	1	1	1	1	2	4	1	1	1	1	7	3	2	1	2	1	1	2	2	1	1	5	1	1	1	1	1	1	3	2	1	2	1	1	1	1	1	1	1	2	1	2	40	8668

See also