How much r thinks supreme court justice makes

payjudge

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn each year. O. A Supreme Court Justice?

Values

[1e+06] $1 million (or more, for 1987 & 2000)

[99999997] never worked

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2000

Overview

For further details see the GSS Data Explorer page for payjudge.

Counts by year:

year	iap	1000	10000	110000	112000	115000	12000	120000	125000	128000	130000	135000	1400	140000	145000	15000	150000	16000	160000	165000	170000	175000	1800	180000	185000	190000	195000	1e+05	$1 million (or more, for 1987 & 2000)	20000	220000	225000	230000	24000	25000	250000	265000	29000	2e+05	30000	31000	33000	330000	35000	350000	39000	3e+05	40000	45000	450000	47000	4e+05	50000	520	55000	5e+05	6000	60000	65000	70000	72800	73000	75000	78000	7e+05	8000	80000	85000	8e+05	90000	95000	9e+05	can't choose	no answer	Total

2000	1545	1	2	9	1	1	1	18	48	1	8	4	1	5	2	4	130	2	5	2	2	24	1	7	2	1	1	190	16	5	1	2	1	2	8	34	2	2	125	8	1	1	1	1	6	1	29	11	5	1	1	6	34	1	2	13	1	30	11	26	1	1	50	1	1	1	63	6	2	36	9	1	86	153	2817

Total	1545	1	2	9	1	1	1	18	48	1	8	4	1	5	2	4	130	2	5	2	2	24	1	7	2	1	1	190	16	5	1	2	1	2	8	34	2	2	125	8	1	1	1	1	6	1	29	11	5	1	1	6	34	1	2	13	1	30	11	26	1	1	50	1	1	1	63	6	2	36	9	1	86	153	2817

See also