Birth control information

pill

Source

General Social Survey https://gss.norc.org

Details

Question A. In some places in the United States, it is not legal to supply birth control information. How do you feel about this–do you think birth control information should be available to anyone who wants it, or not?

Values

[1] available

[2] not available

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-	1974, 1975, 1977, 1982, 1983

Overview

For further details see the GSS Data Explorer page for pill.

Counts by year:

year	iap	available	not available	don't know	no answer	Total

1974	-	1354	115	15	-	1484

1975	-	1331	126	31	2	1490

1977	-	1389	112	24	5	1530

1982	-	1664	148	45	3	1860

1983	-	1441	127	29	2	1599

Total	0	7179	628	144	12	7963

See also