Citizen said vulgar or obscene things

polabuse

Source

General Social Survey https://gss.norc.org

Details

Question Are there any situations you can imagine in which you would approve of a policeman striking an adult male citizen? A. Had said vulgar and obscene things to the policeman?

Values

1 yes

2 no

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for polabuse.

Counts by year:

year	iap	yes	no	don't know	no answer	skipped on web	Total

1973	380	251	855	17	1	-	1504

1975	343	221	881	40	5	-	1490

1976	-	295	1158	42	4	-	1499

1978	-	274	1225	31	2	-	1532

1980	-	198	1232	30	8	-	1468

1983	-	235	1325	38	1	-	1599

1984	-	176	1258	35	4	-	1473

1986	-	205	1241	21	3	-	1470

1987	-	190	1576	46	7	-	1819

1988	484	118	857	18	4	-	1481

1989	502	113	899	18	5	-	1537

1990	473	108	759	31	1	-	1372

1991	500	87	902	21	7	-	1517

1993	549	73	961	19	4	-	1606

1994	981	176	1804	19	12	-	2992

1996	979	138	1756	24	7	-	2904

1998	921	125	1750	31	5	-	2832

2000	921	115	1738	32	11	-	2817

2002	1853	59	833	19	1	-	2765

2004	1936	66	783	26	1	-	2812

2006	2521	136	1787	66	-	-	4510

2008	658	126	1193	45	1	-	2023

2010	667	126	1220	29	2	-	2044

2012	636	125	1174	38	1	-	1974

2014	851	148	1508	31	-	-	2538

2016	911	198	1715	38	5	-	2867

2018	785	150	1374	37	2	-	2348

2021	2719	100	1204	1	1	7	4032

2022	2353	106	1076	4	1	4	3544

Total	22923	4438	36044	847	106	11	64369

See also