Best policy vs narrow interest: elected official

polbstgw

Source

General Social Survey https://gss.norc.org

Details

Question Please look at Card B5. When making policy recommendations about global warming, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole versus what serves their own narrow interests? INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career. B. Elected Officials

Values

[1] what is best for the country

[5] own narrow interests

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/C	2010

B/C	2006

Overview

For further details see the GSS Data Explorer page for polbstgw.

Counts by year:

year	iap	what is best for the country	2	3	4	own narrow interests	don't know	no answer	Total

2006	3582	81	88	236	222	252	47	2	4510

2010	1569	41	48	128	113	117	23	5	2044

Total	5151	122	136	364	335	369	70	7	6554

See also