Don't have any say about what the government does

poleff11

Source

General Social Survey https://gss.norc.org

Details

Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE. A. People like me don't have any say about what the government does

Values

[1] strongly agree

[2] agree

[3] neither agree nor disagree

[4] disagree

[5] strongly disagree

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	2016, 2018

A/B/C	1996, 2004, 2014

A/C/-	2012

D/-/-	2006

Overview

For further details see the GSS Data Explorer page for poleff11.

Counts by year:

year	iap	strongly agree	agree	neither agree nor disagree	disagree	strongly disagree	can't choose	no answer	Total

1996	1572	219	392	189	377	110	24	21	2904

2004	1340	200	317	263	521	163	4	4	2812

2006	2992	309	392	227	419	159	11	1	4510

2012	1312	75	193	103	213	50	26	2	1974

2014	1274	242	327	202	330	138	18	7	2538

2016	1477	250	413	265	355	86	7	14	2867

2018	1175	174	300	222	367	88	20	2	2348

Total	11142	1469	2334	1471	2582	794	110	51	19953

See also