Most people better informed about politics than i

poleff15

Source

General Social Survey https://gss.norc.org

Details

Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE. E. I think most people are better informed about politics and government than I am

Values

1 strongly agree

2 agree

3 neither agree nor disagree

4 disagree

5 strongly disagree

NA(d) can't choose / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	1996

A/C/-	2012

D/-/-	2006

Overview

For further details see the GSS Data Explorer page for poleff15.

Counts by year:

year	iap	strongly agree	agree	neither agree nor disagree	disagree	strongly disagree	can't choose	no answer	Total

1996	1572	92	323	348	423	88	34	24	2904

2006	2992	158	420	280	512	137	11	-	4510

2012	1312	37	197	143	218	53	11	3	1974

Total	5876	287	940	771	1153	278	56	27	9388

See also