Think of self as liberal or conservative

polviews

Source

General Social Survey https://gss.norc.org

Details

Question A. We hear a lot of talk these days about liberals and conservatives. I'm going to show you a seven-point scale on which the political views that people might hold are arranged from extremely liberal–point 1–to extremely conservative–point 7. Where would you place yourself on this scale?

Values

[1] extremely liberal

[2] liberal

[3] slightly liberal

[4] moderate, middle of the road

[5] slightly conservative

[6] conservative

[7] extremely conservative

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C/D	2006

Full	1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for polviews.

Counts by year:

year	iap	extremely liberal	liberal	slightly liberal	moderate, middle of the road	slightly conservative	conservative	extremely conservative	don't know	no answer	skipped on web	Total

1974	-	22	201	207	564	221	160	35	70	4	-	1484

1975	-	46	179	196	559	232	150	35	81	12	-	1490

1976	-	31	187	186	559	221	190	27	93	5	-	1499

1977	-	37	169	214	564	251	179	39	71	6	-	1530

1978	-	22	142	241	549	263	188	30	70	27	-	1532

1980	-	36	121	208	582	258	180	44	22	17	-	1468

1982	-	48	199	267	694	235	229	67	104	17	-	1860

1983	795	16	67	98	319	142	108	20	31	3	-	1599

1984	-	29	133	177	568	276	186	41	52	11	-	1473

1985	-	35	163	171	566	271	214	42	63	9	-	1534

1986	-	25	132	176	579	242	209	38	67	2	-	1470

1987	-	47	234	226	641	278	208	45	98	42	-	1819

1988	-	34	176	189	514	246	225	32	56	9	-	1481

1989	-	40	178	191	566	246	194	27	88	7	-	1537

1990	-	36	141	179	476	240	192	51	53	4	-	1372

1991	-	37	154	214	584	218	212	40	54	4	-	1517

1993	-	30	180	203	575	264	254	42	49	9	-	1606

1994	-	71	328	378	1049	472	478	103	101	12	-	2992

1996	-	59	303	334	1045	451	458	93	155	6	-	2904

1998	-	64	357	351	986	432	415	86	133	8	-	2832

2000	-	107	308	285	1054	390	411	89	149	24	-	2817

2002	1393	47	143	159	522	209	210	41	36	5	-	2765

2004	1472	46	120	153	497	214	223	56	26	5	-	2812

2006	-	139	524	517	1683	618	685	167	154	23	-	4510

2008	-	69	240	221	740	268	327	68	77	13	-	2023

2010	-	76	259	232	746	265	315	80	61	10	-	2044

2012	-	81	244	208	713	268	292	68	87	13	-	1974

2014	-	94	304	263	989	334	358	107	65	24	-	2538

2016	-	136	350	310	1032	382	426	120	81	30	-	2867

2018	-	122	278	256	855	283	354	99	79	22	-	2348

2021	15	207	623	490	1377	476	617	174	10	3	40	4032

2022	1	192	486	400	1245	430	514	159	73	27	17	3544

Total	3676	2081	7623	7900	23992	9596	9361	2165	2409	413	57	69273

See also