Materials lead to rape

pornrape

Source

General Social Survey https://gss.norc.org

Details

Question The next questions are about pornography–books, movies, magazines, and photographs that show or describe sex activities. I”m going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect. READ EACH ITEM. CODE ONE FOR EACH C. Sexual materials lead people to commit rape

Values

[1] yes

[2] no

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for pornrape.

Counts by year:

year	iap	yes	no	don't know	no answer	Total

1973	-	748	642	105	9	1504

1975	-	772	563	149	6	1490

1976	-	787	569	136	7	1499

1978	-	867	549	112	4	1532

1980	-	790	538	134	6	1468

1983	-	880	596	113	10	1599

1984	-	806	542	114	11	1473

1986	-	830	530	101	9	1470

1987	-	963	677	168	11	1819

1988	484	558	351	80	8	1481

1989	502	616	339	74	6	1537

1990	473	507	297	94	1	1372

1991	500	532	372	111	2	1517

1993	549	596	351	100	10	1606

1994	1965	491	426	104	6	2992

Total	4473	10743	7342	1695	106	24359

See also