Sex before marriage

premarsx

Source

General Social Survey https://gss.norc.org

Details

Question There's been a lot of discussion about the way morals and attitudes about sex are changing in this country. If a man and woman have sex relations before marriage, do you think it is always wrong, almost always wrong, wrong only sometimes, or not wrong at all?

Values

[1] always wrong

[2] almost always wrong

[3] wrong only sometimes

[4] not wrong at all

[5] other

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1972, 1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for premarsx.

Counts by year:

year	always wrong	almost always wrong	wrong only sometimes	not wrong at all	don't know	no answer	iap	skipped on web	Total

1972	563	182	373	419	65	11	-	-	1613

1974	472	181	337	439	48	7	-	-	1484

1975	441	176	342	468	58	5	-	-	1490

1977	459	141	340	541	39	10	-	-	1530

1978	437	175	304	578	34	4	-	-	1532

1982	520	152	359	763	55	11	-	-	1860

1983	427	155	378	601	31	7	-	-	1599

1985	422	123	294	643	48	4	-	-	1534

1986	396	127	328	574	39	6	-	-	1470

1988	251	102	214	388	27	6	493	-	1481

1989	272	84	221	394	31	4	531	-	1537

1990	229	103	205	356	32	3	444	-	1372

1991	274	98	186	435	28	3	493	-	1517

1993	283	106	218	443	27	3	526	-	1606

1994	496	192	389	830	60	10	1015	-	2992

1996	452	183	429	832	58	6	944	-	2904

1998	476	166	378	784	65	2	961	-	2832

2000	502	157	384	749	75	10	940	-	2817

2002	242	73	176	394	18	5	1857	-	2765

2004	238	79	157	409	18	5	1906	-	2812

2006	509	167	379	882	44	11	2518	-	4510

2008	305	90	206	692	30	6	694	-	2023

2010	305	116	241	742	21	5	614	-	2044

2012	273	72	206	704	33	14	672	-	1974

2014	324	111	257	956	18	9	863	-	2538

2016	389	100	263	1101	25	10	979	-	2867

2018	278	95	213	943	19	11	789	-	2348

2021	390	162	367	1761	9	2	1315	26	4032

2022	308	104	305	1607	36	8	1168	8	3544

Total	10933	3772	8449	20428	1091	198	19722	34	64627

See also