Vote for kerry, bush, nader

pres04

Source

General Social Survey https://gss.norc.org

Details

Question J. In 2004, you remember that Kerry ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?

IF VOTED: Did you vote for Kerry or Bush?

Values

[1] kerry

[2] bush

[3] nader

[4] other (specify)

[6] didnt vote

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	2008, 2010

A/B/C/D	2006

Overview

For further details see the GSS Data Explorer page for pres04.

Counts by year:

year	iap	kerry	bush	nader	didnt vote	don't know	no answer	Total

2006	1458	1434	1446	47	17	18	90	4510

2008	768	580	599	25	12	8	31	2023

2010	811	526	630	26	6	8	37	2044

Total	3037	2540	2675	98	35	34	158	8577

See also