Private enterprise will solve u.s. problems

privent

Source

General Social Survey https://gss.norc.org

Details

Question How much do you agree or disagree with each of these statements? Private enterprise is the best way to solve America”s economic problems.

Values

[1] strongly agree

[2] agree

[3] neither agree nor disagree

[4] disagree

[5] strongly disagree

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	2010, 2021

A/B/C	1993, 1994, 2000

Overview

For further details see the GSS Data Explorer page for privent.

Counts by year:

year	iap	strongly agree	agree	neither agree nor disagree	disagree	strongly disagree	don't know	no answer	skipped on web	Total

1993	49	221	541	469	189	19	87	31	-	1606

1994	1606	229	502	363	151	18	87	36	-	2992

2000	1541	175	377	409	138	31	71	75	-	2817

2010	614	185	391	352	294	71	132	5	-	2044

2021	2003	225	366	730	322	111	119	144	12	4032

Total	5813	1035	2177	2323	1094	250	496	291	12	13491

See also