Third mention of problem

probfix3

Source

General Social Survey https://gss.norc.org

Details

Question Problems often come up in life. Sometimes they're personal problems – people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage – a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it? 3. What R would do about the problem – Third mention.

Values

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

2006	A/B/C	full

Link at the GSS

For further details see the GSS Data Explorer page for probfix3.

See also