Vote on open housing law

racopen

Source

General Social Survey https://gss.norc.org

Details

Question 128. Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: a. One law says that a homeowner can decide for himself whom to sell his house to, even if he prefers not to sell to (negroes/blacks/African-Americans). b. The second law says that a homeowner cannot refuse to sell to someone because of their race or color. Which law would you vote for?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	don't know	first law	neither	no answer	second law	not available in this year	Total

1972	1613	-	-	-	-	-	-	1613

1973	183	18	826	22	7	448	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	163	21	844	10	2	450	-	1490

1976	129	17	851	24	7	471	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	-	30	863	19	3	617	-	1532

1980	-	25	806	23	2	612	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	-	35	779	30	7	748	-	1599

1984	-	22	662	26	9	754	-	1473

1985	1534	-	-	-	-	-	-	1534

1986	-	18	684	20	4	744	-	1470

1987	-	31	697	31	14	1046	-	1819

1988	484	19	388	25	3	562	-	1481

1989	502	16	380	21	7	611	-	1537

1990	-	39	549	22	5	757	-	1372

1991	500	17	352	29	14	605	-	1517

1993	549	19	314	40	4	680	-	1606

1994	981	36	652	58	14	1251	-	2992

1996	1926	15	281	20	11	651	-	2904

1998	2832	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	2765

2004	1914	13	289	-	4	592	-	2812

2006	2507	8	542	94	6	1353	-	4510

2008	671	11	342	59	3	937	-	2023

2010	763	4	307	52	2	916	-	2044

2012	666	7	334	33	4	930	-	1974

2014	824	6	411	42	3	1252	-	2538

2016	978	13	365	58	7	1446	-	2867

2018	774	14	304	27	4	1225	-	2348

2021	-	-	-	-	-	-	4032	4032

2022	2943	5	108	-	8	480	-	3544

2024	2823	17	93	-	7	369	-	3309

Total	36715	476	13023	785	161	20507	4032	75699

Values

1 first law

2 second law

3 neither

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook