Blacks shouldn't push

racpush

Source

General Social Survey https://gss.norc.org

Details

Question Here are some opinions other people have expressed in connection with (Negro/Black)-white relations. Which statement on the card comes closest to how you, yourself, feel? The first one is . . . A. (Negroes/Blacks/African-Americans) shouldn't push themselves where they're not wanted.

Values

[1] agree strongly

[2] agree slightly

[3] disagree slightly

[4] disagree strongly

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1972, 1973, 1975, 1976, 1977, 1980, 1982, 1984, 1985

A/C	1994, 1996, 1998, 2000, 2002

Overview

For further details see the GSS Data Explorer page for racpush.

Counts by year:

year	agree strongly	agree slightly	disagree slightly	disagree strongly	don't know	iap	no answer	Total

1972	561	392	170	135	88	261	6	1613

1973	575	383	194	147	14	183	8	1504

1975	597	364	193	135	32	163	6	1490

1976	571	378	224	164	26	129	7	1499

1977	584	380	234	124	27	176	5	1530

1980	492	451	280	205	36	-	4	1468

1982	461	491	375	480	47	-	6	1860

1984	373	440	311	307	33	-	9	1473

1985	193	241	167	128	17	783	5	1534

1994	215	405	376	458	55	1480	3	2992

1996	307	436	464	637	72	981	7	2904

1998	256	476	426	610	108	950	6	2832

2000	269	403	379	582	104	1067	13	2817

2002	126	178	225	317	74	1841	4	2765

Total	5580	5418	4018	4429	733	8014	89	28281

See also