Whites have right to seg. neighborhood

racseg

Source

General Social Survey https://gss.norc.org

Details

Question Here are some opinions other people have expressed in connection with (Negro/Black)-white relations. Which statement on the card comes closest to how you, yourself, feel? The first one is . . . B. White people have a right to keep (Negroes/Blacks/African-Americans) out of their neighborhoods if they want to, and (Negroes/Blacks/African-Americans) should respect that right.

Values

1 agree strongly

2 agree slightly

3 disagree slightly

4 disagree strongly

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

-/-	1972, 1976, 1977, 1980, 1982, 1984, 1985, 1987

A/C	1988, 1989, 1990, 1991, 1993, 1994, 1996

Overview

For further details see the GSS Data Explorer page for racseg.

Counts by year:

year	agree strongly	agree slightly	disagree slightly	disagree strongly	don't know	iap	no answer	Total

1972	279	224	307	443	96	261	3	1613

1976	288	237	346	466	26	129	7	1499

1977	289	266	381	373	40	176	5	1530

1980	221	220	397	576	50	-	4	1468

1982	220	222	452	923	39	-	4	1860

1984	151	211	368	705	31	-	7	1473

1985	151	224	432	680	38	-	9	1534

1987	149	217	401	1009	38	-	5	1819

1988	72	138	230	510	26	504	1	1481

1989	78	136	234	553	28	504	4	1537

1990	74	124	218	471	27	455	3	1372

1991	66	105	221	574	21	524	6	1517

1993	42	109	218	681	23	531	2	1606

1994	102	197	472	1176	44	996	5	2992

1996	50	59	177	636	27	1953	2	2904

Total	2232	2689	4854	9776	554	6033	67	26205

See also