R's self ranking of social position

rank

Source

General Social Survey https://gss.norc.org

Details

Question In our society there are groups which tend to be towards the top and those that are towards the bottom. Here we have a scale that runs from top to bottom. A. Where would you put yourself on this scale? HAND SCALE AND PENCIL TO RESPONDENT. LET RESPONDENT MARK SCALE. Be sure that the mark is within one of the boxes. Record answer below.

Values

1 1 - top

10 10 - bottom

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1983	-/-/-	full

1987	-/-/-	full

2000	A/B/C	full

2004	A/B/C	full

2006	D/-/-	partial

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	A/B/C	full

2018	A/B/C	full

2021	A/B/C	full

2022	A/B/C	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	1 - top	10 - bottom	2	3	4	5	6	7	8	9	don't know	no answer	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	804	65	15	50	112	98	241	78	60	39	9	17	11	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	255	47	48	36	170	230	460	230	149	95	32	-	67	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	2904	-	-	-	-	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	1545	33	18	33	117	174	352	188	117	64	24	7	145	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	1596	115	11	66	180	165	424	109	56	35	15	22	18	-	2812

2006	2992	101	25	95	247	204	517	125	92	55	10	43	4	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	-	114	25	72	257	290	735	213	157	70	20	65	26	-	2044

2012	-	113	31	68	236	240	714	221	151	80	16	72	32	-	1974

2014	-	156	39	85	297	367	930	253	194	105	20	71	21	-	2538

2016	-	163	45	102	366	421	1000	284	227	130	50	48	31	-	2867

2018	-	131	41	88	292	338	817	246	199	83	25	61	27	-	2348

2021	2031	76	30	89	361	365	685	142	134	64	20	3	2	30	4032

2022	1755	109	38	61	238	286	604	165	132	76	16	42	9	13	3544

2024	1656	85	37	58	223	227	584	149	131	72	17	38	9	23	3309

Total	52120	1308	403	903	3096	3405	8063	2403	1799	968	274	489	402	66	75699

See also