Relationship of 2nd person to household head

relate2

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what?

Relationship of second person to head of household

Values

[1] head of household

[2] spouse

[3] child (natural, adopted, step)

[4] son-/daughter-in-law

[5] grand/great-grandchild

[6] parent/parent-in-law

[7] other relative

[8] non-relative

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/D	2006

Full/-	1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for relate2.

Counts by year:

year	iap	spouse	child (natural, adopted, step)	grand/great-grandchild	parent/parent-in-law	other relative	non-relative	son-/daughter-in-law	no answer	head of household	Total

1975	203	1075	127	4	10	25	46	-	-	-	1490

1976	893	496	73	2	4	11	20	-	-	-	1499

1977	256	1047	151	6	11	11	42	1	5	-	1530

1978	295	1002	146	8	7	15	47	2	10	-	1532

1980	285	941	140	7	10	22	53	-	10	-	1468

1982	412	1081	223	13	8	33	76	-	14	-	1860

1983	311	1003	169	2	7	24	59	3	5	16	1599

1984	327	870	145	7	5	16	77	1	9	16	1473

1985	339	913	144	4	6	19	85	1	5	18	1534

1986	313	892	145	6	7	17	78	1	6	5	1470

1987	424	945	216	10	6	48	120	5	38	7	1819

1988	328	824	164	8	7	26	111	1	7	5	1481

1989	327	864	146	6	2	22	106	2	57	5	1537

1990	323	760	157	7	7	19	86	3	10	-	1372

1991	374	829	151	7	11	27	100	1	14	3	1517

1993	373	900	172	6	8	18	121	1	7	-	1606

1994	752	1596	339	13	17	47	206	4	18	-	2992

1996	743	1394	353	16	20	37	248	3	90	-	2904

1998	777	1389	299	8	16	38	226	4	75	-	2832

2000	741	1325	354	13	27	46	296	4	11	-	2817

2002	883	1214	328	7	11	44	259	13	6	-	2765

2004	707	1527	279	13	22	39	217	5	3	-	2812

2006	1226	2288	464	18	31	61	414	4	4	-	4510

2008	527	1002	228	15	12	35	195	5	4	-	2023

2010	595	909	225	10	19	25	243	4	14	-	2044

2012	533	937	187	15	23	32	236	1	10	-	1974

2014	678	1198	290	13	17	41	282	8	11	-	2538

2016	849	1256	315	16	22	50	351	3	5	-	2867

2018	677	1057	249	11	24	37	286	1	6	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	-	3544

Total	23047	31534	6379	271	377	885	4686	81	454	75	67789

See also