Relationship of 8th person to household head

relate8

Source

General Social Survey https://gss.norc.org

Details

Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.) Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily–on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.) Are there any people currently staying here–visitors, friends or relatives–do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.) A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD? B. Code sex. (ASK IF NOT OBVIOUS). C. How old was (HEAD/PERSON) on (his/her) last birthday? D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married? E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now? F. Where is (PERSON) staying right now? Is (PERSON) staying at another household; Is (he/she) traveling; Is (he/she) in some institution or dormitory–like at college or in a hospital or something; or what?

Relationship of eighth person to head of household

Values

1 head of household

2 spouse

3 child (natural, adopted, step)

4 son-/daughter-in-law

5 grand/great-grandchild

6 parent/parent-in-law

7 other relative

8 non-relative

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/D	2006

Full/-	1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for relate8.

Counts by year:

year	iap	child (natural, adopted, step)	grand/great-grandchild	parent/parent-in-law	non-relative	no answer	other relative	spouse	son-/daughter-in-law	Total

1975	1465	20	2	3	-	-	-	-	-	1490

1976	1490	9	-	-	-	-	-	-	-	1499

1977	1505	21	1	-	2	1	-	-	-	1530

1978	1514	17	-	-	-	-	1	-	-	1532

1980	1454	11	1	-	-	2	-	-	-	1468

1982	1835	11	10	1	1	1	1	-	-	1860

1983	1584	12	1	1	-	-	1	-	-	1599

1984	1465	5	3	-	-	-	-	-	-	1473

1985	1519	12	2	-	1	-	-	-	-	1534

1986	1456	11	2	-	1	-	-	-	-	1470

1987	1804	7	5	-	2	-	1	-	-	1819

1988	1475	5	-	-	1	-	-	-	-	1481

1989	1520	8	3	1	1	-	4	-	-	1537

1990	1368	1	2	-	-	-	1	-	-	1372

1991	1510	6	1	-	-	-	-	-	-	1517

1993	1602	2	1	-	-	-	1	-	-	1606

1994	2985	3	4	-	-	-	-	-	-	2992

1996	2891	7	3	-	-	1	2	-	-	2904

1998	2817	6	2	-	-	1	5	1	-	2832

2000	2802	8	2	-	-	-	4	-	1	2817

2002	2757	5	1	-	1	-	1	-	-	2765

2004	2803	6	1	-	1	-	1	-	-	2812

2006	4495	12	1	1	-	-	1	-	-	4510

2008	2015	4	3	-	-	-	1	-	-	2023

2010	2026	11	3	-	1	2	1	-	-	2044

2012	1967	5	1	-	-	-	1	-	-	1974

2014	2531	4	3	-	-	-	-	-	-	2538

2016	2850	13	3	-	-	-	1	-	-	2867

2018	2338	4	3	1	1	1	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	3544

Total	67419	246	64	8	13	9	28	1	1	67789

See also