Influence on funding for stem cell rsch: religious leaders

relinfsc

Source

General Social Survey https://gss.norc.org

Details

Question Please look at Card B8. How much influence should each of the following groups have in deciding about government funding for stem cell research? B. Religious leaders. (Would you say a great deal of influence, a fair amount, a little influence, or none at all?)

Values

1 a great deal of influence

2 a fair amount

3 a little influence

4 none at all

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2006	B/C	partial

2010	A/C	partial

Link at the GSS

For details see the GSS Data Explorer page for relinfsc.

See also