R accepts a person from a different religion or view from r's marrying relative

relmarry

Source

General Social Survey https://gss.norc.org

Details

Question People belong to different religions and have different religious views. A. Would you accept a person from a different religion or with a very different religious view from yours marrying a relative of yours? Would you say…

Values

[1] definitely accept

[2] probably accept

[3] probably not accept

[4] definitely not accept

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B	2018

B/C	2008

Overview

For further details see the GSS Data Explorer page for relmarry.

Counts by year:

year	iap	definitely accept	probably accept	probably not accept	definitely not accept	don't know	no answer	Total

2008	658	539	641	109	46	26	4	2023

2018	1173	517	513	72	39	23	11	2348

Total	1831	1056	1154	181	85	49	15	4371

See also