For religious reasons r has in home shrine, altar, or religious object on displa

relobjct

Source

General Social Survey https://gss.norc.org

Details

Question For religious reasons do you have in your home a shrine, altar, or a religious object on display such as an icon, menorah, or crucifix?

Values

1 yes

2 no

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

2008	B/C	partial

2018	A/B	partial

Link at the GSS

For details see the GSS Data Explorer page for relobjct.

See also