Religious tradition age 16

reltrad16

Source

General Social Survey https://gss.norc.org

Details

Question RELTRAD16 is constructed from variables RELIG16, DENOM16 and OTH16. The missing values coded as ".z: Variable-specific reserve code" represent OTH16 denominations that are not originally covered by the Steensland et al. (2000) article (i.e., OTHER values larger than 148). Please see Appendix F in the codebook for more information.

Values

[1] evangelical protestant

[2] mainline protestant

[3] black protestant

[4] catholic

[5] jewish

[6] other faith

[7] nonaffiliated

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for reltrad16.

Counts by year:

year	iap	evangelical protestant	mainline protestant	black protestant	catholic	jewish	other faith	nonaffiliated	don't know	no answer	see codebook	skipped on web	Total

1973	24	311	464	156	419	43	33	36	7	11	-	-	1504

1974	22	323	464	154	402	49	21	39	6	4	-	-	1484

1975	30	339	497	141	378	25	29	45	3	3	-	-	1490

1976	27	350	467	110	428	33	34	47	-	3	-	-	1499

1977	31	352	494	146	390	37	31	39	1	9	-	-	1530

1978	33	362	453	145	418	31	33	48	-	9	-	-	1532

1980	25	326	480	122	389	32	42	45	-	7	-	-	1468

1982	33	316	511	414	441	44	42	48	-	11	-	-	1860

1983	26	331	447	139	481	50	63	52	-	10	-	-	1599

1984	36	386	346	138	399	30	63	43	1	31	-	-	1473

1985	23	423	370	127	444	32	46	63	-	6	-	-	1534

1986	26	386	338	149	412	33	71	45	3	6	1	-	1470

1987	26	420	364	431	407	22	63	48	1	37	-	-	1819

1988	30	368	364	146	415	31	68	53	-	6	-	-	1481

1989	61	379	359	131	444	28	65	61	-	9	-	-	1537

1990	40	328	352	120	361	27	59	72	-	13	-	-	1372

1991	37	393	344	153	434	34	58	58	-	6	-	-	1517

1993	37	465	358	142	422	35	60	71	-	16	-	-	1606

1994	69	786	643	285	882	55	117	127	3	23	2	-	2992

1996	60	694	633	281	837	68	141	169	2	18	1	-	2904

1998	60	684	574	272	857	53	97	175	7	52	1	-	2832

2000	91	655	515	304	844	64	119	192	5	28	-	-	2817

2002	107	577	573	267	821	56	127	195	14	27	1	-	2765

2004	150	622	502	254	829	60	137	226	13	14	5	-	2812

2006	1618	639	510	273	960	56	168	249	27	7	3	-	4510

2008	86	429	334	196	635	39	92	168	26	14	4	-	2023

2010	143	415	352	183	647	33	89	150	18	11	3	-	2044

2012	126	395	323	183	653	28	83	162	15	4	2	-	1974

2014	156	494	375	229	872	41	135	206	15	13	2	-	2538

2016	191	566	406	287	923	55	143	248	33	12	3	-	2867

2018	151	457	339	218	732	42	124	241	29	12	3	-	2348

2021	335	581	696	232	1290	87	212	499	58	4	-	38	4032

2022	338	514	477	258	1159	55	217	404	78	18	5	21	3544

2024	218	556	468	276	1003	47	218	434	51	12	4	22	3309

Total	4466	15622	15192	7062	21428	1455	3100	4758	416	466	40	81	74086

See also