Type of place lived in when 16 yrs old

res16

Source

General Social Survey https://gss.norc.org

Details

Question Which of the categories on this card comes closest to the type of place you were living in when you were 16 years old?

Values

1 in open country but not on a farm

2 farm

3 in a small city or town (under 50,000)

4 in a medium-size city (50,000-250,000)

5 in a suburb near a large city

6 in a large city (over 250,000)

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for res16.

Counts by year:

year	in open country but not on a farm	farm	in a small city or town (under 50,000)	in a medium-size city (50,000-250,000)	in a suburb near a large city	in a large city (over 250,000)	don't know	no answer	iap	skipped on web	Total

1972	147	340	500	198	91	334	1	2	-	-	1613

1973	135	343	479	188	106	250	2	1	-	-	1504

1974	144	362	438	192	123	222	-	3	-	-	1484

1975	192	337	430	179	102	250	-	-	-	-	1490

1976	160	341	448	186	120	242	1	1	-	-	1499

1977	170	344	524	187	88	211	2	4	-	-	1530

1978	174	333	472	206	130	215	1	1	-	-	1532

1980	145	310	465	204	118	225	-	1	-	-	1468

1982	187	377	577	231	166	311	-	11	-	-	1860

1983	170	252	536	244	156	241	-	-	-	-	1599

1984	144	318	466	200	133	206	2	4	-	-	1473

1985	146	281	523	235	154	189	1	5	-	-	1534

1986	140	268	449	247	137	229	-	-	-	-	1470

1987	206	341	489	295	176	306	1	5	-	-	1819

1988	172	252	477	217	142	219	-	2	-	-	1481

1989	168	258	496	239	169	203	1	3	-	-	1537

1990	134	236	437	196	139	227	2	1	-	-	1372

1991	176	251	501	212	158	215	2	2	-	-	1517

1993	187	288	466	220	205	236	2	2	-	-	1606

1994	357	405	931	469	368	456	2	4	-	-	2992

1996	333	374	922	494	357	418	3	3	-	-	2904

1998	327	371	857	450	361	462	2	2	-	-	2832

2000	311	385	871	461	333	449	1	6	-	-	2817

2002	327	315	831	439	390	459	3	1	-	-	2765

2004	318	274	999	427	364	429	1	-	-	-	2812

2006	374	268	1005	523	385	434	2	1	1518	-	4510

2008	227	213	653	340	252	337	1	-	-	-	2023

2010	213	198	672	369	261	328	2	1	-	-	2044

2012	217	206	596	376	280	298	1	-	-	-	1974

2014	275	211	813	438	377	421	2	1	-	-	2538

2016	304	234	979	493	413	438	4	2	-	-	2867

2018	255	154	774	450	324	391	-	-	-	-	2348

2021	406	204	1208	777	742	692	1	-	-	2	4032

2022	330	189	1135	648	604	630	7	-	-	1	3544

Total	7671	9833	22419	11230	8424	11173	50	69	1518	3	72390

See also