Allow revolutionary public meetings

revspeak

Source

General Social Survey https://gss.norc.org

Details

Question There are some people whose views are considered extreme by the majority. First, consider people who want to overthrow the government by revolution. please circle a number after each question to show if you think such people should be allowed to . . . A. Hold public meetings to express their views.

Values

[1] definitely allowed

[2] probably allowed

[3] probably not allowed

[4] definitely not allowed

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	2016

A/B/C	1990, 1996

D/-/-	2006

Full	1985

Overview

For further details see the GSS Data Explorer page for revspeak.

Counts by year:

year	iap	definitely allowed	probably allowed	probably not allowed	definitely not allowed	can't choose	no answer	Total

1985	857	191	160	77	219	24	6	1534

1990	155	378	307	173	288	58	13	1372

1996	1572	358	367	188	326	61	32	2904

2006	2992	731	344	193	234	16	-	4510

2016	1477	669	375	168	151	18	9	2867

Total	7053	2327	1553	799	1218	177	60	13187

See also