If rich, continue or stop working

richwork

Source

General Social Survey https://gss.norc.org

Details

Question 182. If you were to get enough money to live as comfortably as you would like for the rest of your life, would you continue to work or would you stop working?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	continue to work	don't know	no answer	stop working	skipped on web	Total

1972	1613	-	-	-	-	-	1613

1973	652	566	12	21	253	-	1504

1974	641	532	16	6	289	-	1484

1975	1490	-	-	-	-	-	1490

1976	721	515	11	21	231	-	1499

1977	572	658	14	4	282	-	1530

1978	1532	-	-	-	-	-	1532

1980	570	674	9	12	203	-	1468

1982	703	816	18	11	312	-	1860

1983	1599	-	-	-	-	-	1599

1984	510	721	6	8	228	-	1473

1985	538	680	10	8	298	-	1534

1986	1470	-	-	-	-	-	1470

1987	622	888	8	11	290	-	1819

1988	852	440	6	3	180	-	1481

1989	883	462	9	5	178	-	1537

1990	770	426	8	8	160	-	1372

1991	891	412	8	2	204	-	1517

1993	901	474	13	5	213	-	1606

1994	1610	895	16	6	465	-	2992

1996	1513	927	22	5	437	-	2904

1998	1524	894	18	5	391	-	2832

2000	1544	833	22	6	412	-	2817

2002	2113	444	4	3	201	-	2765

2004	2187	430	2	8	185	-	2812

2006	3197	896	10	4	403	-	4510

2008	1137	629	7	5	245	-	2023

2010	1183	591	5	9	256	-	2044

2012	1129	580	4	12	249	-	1974

2014	1440	749	3	1	345	-	2538

2016	1646	860	6	11	344	-	2867

2018	1326	723	6	13	280	-	2348

2021	2367	1004	-	34	620	7	4032

2022	2082	950	4	2	502	4	3544

2024	1957	872	10	4	461	5	3309

Total	45485	20541	287	253	9117	16	75699

Values

1 continue to work

2 stop working

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook