Does gun belong to r

rowngun

Source

General Social Survey https://gss.norc.org

Details

Question Do you happen to have in your home (IF HOUSE: or garage) any guns or revolvers? B. Do any of these guns personally belong to you?

Values

1 yes

2 no

3 refused

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1980	-/-/-	full

1982	-/-/-	full

1984	-/-/-	full

1985	-/-/-	full

1987	-/-/-	full

1988	A/C/-	partial

1989	A/C/-	partial

1990	A/C/-	partial

1991	A/C/-	partial

1993	A/C/-	partial

1994	A/C/-	partial

1996	A/C/-	partial

1998	A/C/-	partial

2000	A/C/-	partial

2002	A/C/-	partial

2004	A/C/-	partial

2006	A/C/-	partial

2008	A/C/-	partial

2010	A/C/-	partial

2012	A/C/-	partial

2014	A/C/-	partial

2016	A/C/-	partial

2018	A/C/-	partial

2021	A/B/C	full

2022	A/C/-	partial

2024	A/C/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	no	no answer	refused	yes	don't know	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	761	250	40	5	412	-	-	1468

1982	1022	261	45	23	509	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	796	285	8	9	375	-	-	1473

1985	842	232	7	10	443	-	-	1534

1986	1470	-	-	-	-	-	-	1470

1987	1032	275	20	10	482	-	-	1819

1988	1075	147	9	12	238	-	-	1481

1989	1059	184	8	-	286	-	-	1537

1990	975	130	12	-	255	-	-	1372

1991	1107	134	22	-	254	-	-	1517

1993	1145	141	4	7	309	-	-	1606

1994	2156	253	20	5	555	3	-	2992

1996	2125	254	6	7	512	-	-	2904

1998	2166	222	11	6	427	-	-	2832

2000	2187	187	4	23	416	-	-	2817

2002	2446	71	-	9	239	-	-	2765

2004	2474	83	5	16	234	-	-	2812

2006	3816	214	-	30	449	1	-	4510

2008	1544	137	-	15	327	-	-	2023

2010	1608	118	-	42	275	1	-	2044

2012	1510	139	-	26	299	-	-	1974

2014	1955	147	-	55	380	1	-	2538

2016	2224	186	-	53	404	-	-	2867

2018	1772	170	-	41	365	-	-	2348

2021	2532	340	-	117	1036	-	7	4032

2022	2717	187	-	69	567	1	3	3544

2024	2454	204	-	69	578	-	4	3309

Total	59221	4951	221	659	10626	7	14	75699

See also