How often r feels rushed

rushed

Source

General Social Survey https://gss.norc.org

Details

Question In general, how do you feel about your time–would you say you always feel rushed even to do things you have to do, Only sometimes feel rushed, or almost never feel rushed.

Values

1 always

2 sometimes

3 never

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1982	-/-/-	full

1996	A/B/C	full

2004	B/C/-	partial

2016	C/-/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	always	never	no answer	sometimes	don't know	not available in this year	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	-	442	443	8	967	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	2992

1996	1444	434	261	7	757	1	-	2904

1998	2832	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	2765

2004	1828	304	160	6	513	1	-	2812

2006	4510	-	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	2538

2016	1888	215	218	4	542	-	-	2867

2018	2348	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	4032

2022	-	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	-	3309	3309

Total	63563	1395	1082	25	2779	2	6853	75699

See also