Liking for russia

russia

Source

General Social Survey https://gss.norc.org

Details

Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM. A. Russia

Values

[0] +5

[1] +4

[2] +3

[3] +2

[4] +1

[5] -1

[6] -2

[7] -3

[8] -4

[9] -5

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1974, 1975, 1977, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for russia.

Counts by year:

year	not available in this year	+5	+4	+3	+2	+1	-1	-2	-3	-4	-5	dk, na, iap	iap	Total

1974	-	42	47	180	153	254	157	75	122	54	327	73	-	1484

1975	-	39	38	182	152	240	156	78	125	68	308	104	-	1490

1977	-	31	21	123	102	204	181	81	133	81	458	115	-	1530

1982	-	46	21	89	79	182	196	103	167	110	754	113	-	1860

1983	-	11	15	56	73	188	197	118	203	117	550	71	-	1599

1985	-	24	17	52	67	175	175	116	171	101	552	84	-	1534

1986	-	19	17	102	107	229	193	103	165	83	387	65	-	1470

1988	-	27	22	87	101	199	102	52	88	39	213	551	-	1481

1989	-	40	42	111	123	209	103	64	64	31	164	586	-	1537

1990	-	57	40	136	132	212	90	35	36	20	109	505	-	1372

1991	-	57	51	162	143	225	87	48	48	36	98	562	-	1517

1993	-	74	43	159	141	299	86	49	43	25	95	592	-	1606

1994	-	38	24	64	80	121	38	28	25	12	42	2520	-	2992

Total	0	505	398	1503	1453	2737	1761	950	1390	777	4057	5941	0	21472

See also