City or place r lives in

satcity

Source

General Social Survey https://gss.norc.org

Details

Question For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. READ ITEMS A-E. CIRCLE ONE CODE FOR EACH. A. The city or place you live in.

Values

[1] very great deal

[2] great deal

[3] quite a bit

[4] a fair amount

[5] some

[6] a little

[7] none

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987

B/C	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for satcity.

Counts by year:

year	iap	very great deal	great deal	quite a bit	a fair amount	some	a little	none	don't know	no answer	Total

1973	-	346	357	254	337	91	77	40	2	-	1504

1974	-	297	394	257	343	103	59	30	1	-	1484

1975	-	311	438	222	298	93	80	41	2	5	1490

1976	-	293	446	260	307	101	57	29	4	2	1499

1977	-	295	421	262	330	95	85	37	3	2	1530

1978	-	254	460	291	305	101	71	43	1	6	1532

1980	-	347	450	224	267	82	65	27	6	-	1468

1982	-	304	442	358	446	161	96	44	5	4	1860

1983	-	253	488	280	337	119	85	28	6	3	1599

1984	-	288	478	256	262	85	72	20	6	6	1473

1986	-	232	395	274	315	117	92	28	3	14	1470

1987	-	320	557	304	383	130	81	32	1	11	1819

1988	484	176	307	187	206	58	36	23	1	3	1481

1989	502	178	283	217	196	69	63	22	2	5	1537

1990	473	136	299	161	183	69	38	12	-	1	1372

1991	500	176	302	187	207	88	44	11	-	2	1517

1993	549	182	304	195	211	84	57	21	-	3	1606

1994	2481	80	161	105	99	33	16	16	-	1	2992

Total	4989	4468	6982	4294	5032	1679	1174	504	43	68	29233

See also