Work satisfaction

satjob

Source

General Social Survey https://gss.norc.org

Details

Question On the whole, how satisfied are you with the work you do–would you say you are very satisfied, moderately satisfied, a little dissatisfied, or very dissatisfied?

Values

[1] very satisfied

[2] moderately satisfied

[3] a little dissatisfied

[4] very dissatisfied

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Full	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for satjob.

Counts by year:

year	very satisfied	moderately satisfied	a little dissatisfied	very dissatisfied	don't know	iap	no answer	skipped on web	Total

1972	458	347	106	33	3	660	6	-	1613

1973	564	436	92	49	2	231	130	-	1504

1974	586	458	122	57	5	224	32	-	1484

1975	631	381	105	48	3	295	27	-	1490

1976	617	407	108	53	3	294	17	-	1499

1977	601	496	130	35	1	213	54	-	1530

1978	651	463	108	58	1	222	29	-	1532

1980	584	444	159	59	1	214	7	-	1468

1982	686	562	160	84	3	301	64	-	1860

1983	657	499	117	60	-	221	45	-	1599

1984	555	423	146	84	-	208	57	-	1473

1985	591	467	122	55	2	263	34	-	1534

1986	570	458	104	30	1	257	50	-	1470

1987	617	559	186	77	1	324	55	-	1819

1988	537	455	117	44	2	290	36	-	1481

1989	557	472	125	52	2	287	42	-	1537

1990	481	413	109	38	-	263	68	-	1372

1991	510	482	108	49	3	296	69	-	1517

1993	532	505	133	58	1	329	48	-	1606

1994	1078	932	247	81	7	551	96	-	2992

1996	1041	918	253	101	3	488	100	-	2904

1998	1067	851	223	75	7	527	82	-	2832

2000	981	918	186	77	8	542	105	-	2817

2002	534	386	100	41	13	1689	2	-	2765

2004	702	507	125	63	16	1375	24	-	2812

2006	1073	823	192	89	36	2237	60	-	4510

2008	778	570	135	53	19	453	15	-	2023

2010	746	565	155	62	20	477	19	-	2044

2012	732	539	136	50	11	481	25	-	1974

2014	956	686	161	92	12	626	5	-	2538

2016	900	645	167	67	15	1064	9	-	2867

2018	844	656	174	65	8	574	27	-	2348

2021	1198	1119	305	112	5	1199	60	34	4032

2022	1103	1012	241	107	7	1053	8	13	3544

Total	24718	19854	5157	2158	221	18728	1507	47	72390

See also