Govt support for declining industries

savejobs

Source

General Social Survey https://gss.norc.org

Details

Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it. G. Supporting declining industries to protect jobs.

Values

1 strongly in favor of

2 in favor

3 neither in favor nor against

4 against

5 strongly against

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/-	2016

A/B/C	1990, 1996

D/-/-	2006

Full	1985

Overview

For further details see the GSS Data Explorer page for savejobs.

Counts by year:

year	iap	strongly in favor of	in favor	neither in favor nor against	against	strongly against	no answer	don't know	Total

1985	857	118	225	147	144	34	9	-	1534

1990	155	177	436	341	188	44	30	1	1372

1996	1572	178	457	326	263	66	38	4	2904

2006	2992	369	651	232	203	45	2	16	4510

2016	1477	252	595	292	189	40	9	13	2867

Total	7053	1094	2364	1338	987	229	88	34	13187

See also