Percentage of some college educ in occ10 based on acs 2010

sei10educ

Source

General Social Survey https://gss.norc.org

Details

Question Percentage of those who had some college or more education among those who had the same occupation as in OCC10 in American Community Survey between 2010 and 2012 (25-64 years old only). This percentage is not from the GSS but it is used to calculate SEI10.

Values

NA(d) don't know / NA(i) iap / NA(j) i don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1972	-/-/-/-	full

1973	-/-/-/-	full

1974	-/-/-/-	full

1975	-/-/-/-	full

1976	-/-/-/-	full

1977	-/-/-/-	full

1978	-/-/-/-	full

1980	-/-/-/-	full

1982	-/-/-/-	full

1983	-/-/-/-	full

1984	-/-/-/-	full

1985	-/-/-/-	full

1986	-/-/-/-	full

1987	-/-/-/-	full

1988	A/B/C/-	full

1989	A/B/C/-	full

1990	A/B/C/-	full

1991	A/B/C/-	full

1993	A/B/C/-	full

1994	A/B/C/-	full

1996	A/B/C/-	full

1998	A/B/C/-	full

2000	A/B/C/-	full

2002	A/B/C/-	full

2004	A/B/C/-	full

2006	A/B/C/D	full

2008	A/B/C/-	full

2010	A/B/C/-	full

2012	A/B/C/-	full

2014	A/B/C/-	full

2016	A/B/C/-	full

2018	A/B/C/-	full

2021	A/B/C/-	full

2022	A/B/C/-	full

2024	A/B/C/-	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	11.9	12.2	12.3	12.4	12.6	13.1	13.3	13.5	14.1	15.5	16	16.5	16.8	16.9	17	17.2	17.3	17.5	17.7	17.9	18	18.7	19.2	19.3	20.4	20.5	20.6	21	21.5	21.6	21.7	22	22.3	22.7	23.6	23.9	24.1	24.3	25.1	25.4	25.6	25.7	26.2	26.6	26.7	27	27.1	27.5	27.8	28	28.2	28.3	28.9	29.5	29.7	30.6	30.7	31	31.9	32.4	32.5	33.7	33.9	34.1	35	35.1	36	36.5	37	37.4	37.7	38.3	38.4	38.5	39.2	39.9	40.2	40.3	40.4	40.8	40.9	41.6	42	42.1	42.3	42.8	43.7	43.9	44	45.2	45.4	45.5	46.1	46.3	46.6	46.9	47.5	48.2	48.5	48.8	48.9	50	50.1	50.6	50.7	51	51.1	51.7	52.1	52.7	53	53.9	54.2	54.3	54.7	55.2	55.4	55.5	55.7	55.8	55.9	56.1	56.7	57.1	57.6	57.7	58.5	58.8	58.9	59.2	59.3	61.1	62.2	62.4	63.1	63.9	64.1	64.6	65.1	65.3	65.6	66.8	66.9	67.3	68.4	68.6	68.8	69.1	69.9	70.1	70.4	71.6	72.2	73	73.3	73.5	73.9	74.2	75.1	75.5	75.9	76.5	76.9	77.2	78	78.1	78.3	78.4	79.2	79.5	79.6	79.8	81.1	81.4	82.4	82.5	82.8	83	83.1	83.7	84.4	85.7	85.9	86.3	86.7	87	87.3	89.4	9.8	90.5	90.6	90.9	91.4	91.7	91.9	92	92.1	92.4	93.1	93.7	94	94.1	94.4	94.8	95.1	95.6	95.7	97.3	97.5	97.8	98.9	99	99.1	99.2	99.4	99.5	99.7	don't know	iap	19	19.1	19.6	19.7	19.8	19.9	21.1	22.9	24	24.4	26	29.1	30.4	31.1	32.1	32.8	33.5	34.2	34.8	36.2	39.6	41.8	44.4	45.1	45.7	46.4	47.3	47.4	47.9	48.3	49	49.3	49.7	52.2	54.6	54.8	56.6	62.7	63.3	63.5	64.7	66.3	67.4	67.6	69.7	71.7	72.5	74.4	78.8	79.9	80.1	80.7	81.3	84	85.4	86.8	92.9	95	95.3	99.3	no answer	18.6	21.9	29.4	35.3	37.8	42.6	47.8	58.7	60.7	66	66.7	75.6	77.5	78.2	82.3	84.5	84.6	88.2	88.3	95.5	96.3	98.3	99.6	17.8	23	23.2	25.8	31.2	35.9	38.6	44.2	55.6	65.5	68.2	79.4	83.3	92.6	13.4	14.2	16.7	22.4	34.9	58.6	69.8	81.9	96	97	20.8	35.2	35.8	46.7	60.9	61	96.5	61.3	63.7	70.3	80	85.6	93.3	14.3	48.1	55.1	58.2	62.3	62.8	70.9	96.4	96.8	24.7	28.7	33.8	34.7	35.5	83.5	84.7	90.2	98.5	22.5	69.6	89.8	90.8	91.2	81	84.8	91.3	74.6	62.9	85.3	87.5	91.1	98.7	53.3	86.2	94.3	34.3	10.9	not available in this release	Total

1972	1	4	1	1	8	3	4	2	12	3	3	32	2	5	2	5	54	1	1	1	2	3	19	1	10	1	1	4	1	6	18	3	47	5	2	3	18	6	2	45	1	13	4	3	1	1	1	1	9	7	3	7	1	4	1	1	8	4	5	3	1	4	8	1	18	6	1	1	6	3	10	4	1	2	1	2	1	2	4	14	1	4	26	2	21	2	2	10	27	38	11	5	6	1	4	2	3	3	4	2	1	1	9	1	13	1	3	1	3	1	6	1	9	1	1	2	11	3	5	1	19	53	14	8	70	1	1	6	1	6	1	1	2	4	2	8	2	6	1	11	2	1	6	3	1	3	1	10	3	16	1	2	4	12	2	2	2	1	1	2	5	2	4	6	7	1	2	1	6	4	1	1	1	3	5	3	7	7	1	2	1	1	1	2	1	2	1	4	1	2	3	1	3	1	3	9	2	1	2	8	2	2	2	1	3	3	2	1	1	23	31	19	11	3	1	1	4	1	334	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	4	3	2	-	1	-	2	3	18	2	3	29	2	2	5	5	33	1	-	1	4	-	7	6	10	-	3	2	1	3	9	8	34	3	-	2	7	15	13	27	1	20	-	2	2	1	6	1	6	4	4	5	2	5	1	1	4	1	6	4	2	13	12	1	23	7	4	5	8	5	14	-	1	1	1	-	1	1	1	23	1	2	16	1	20	1	1	7	24	20	14	2	2	2	3	3	1	2	7	7	1	-	4	1	6	-	7	1	5	-	4	2	6	1	2	1	12	4	5	2	36	55	9	1	80	2	5	4	1	3	-	1	-	3	5	7	-	-	1	11	3	-	5	-	2	5	-	3	2	7	-	1	4	18	2	1	2	-	-	2	10	4	6	4	4	-	1	2	2	3	1	-	1	1	1	4	6	3	-	2	-	6	1	6	3	-	-	1	-	4	4	1	2	-	2	7	3	1	4	6	1	-	4	1	7	2	-	-	1	13	26	23	16	-	5	1	2	-	245	2	1	1	2	1	2	2	3	1	1	2	2	2	5	1	2	1	1	1	1	4	1	1	1	2	3	2	1	1	3	3	1	1	1	1	1	2	2	1	1	2	1	1	1	2	1	1	6	1	2	2	1	3	1	1	2	1	1	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	2	1	-	-	2	2	2	6	8	1	6	28	-	1	4	5	34	-	1	1	1	1	9	6	8	-	1	5	1	2	9	2	38	8	2	2	18	15	5	20	1	20	2	1	1	2	1	2	13	4	8	-	-	9	-	1	6	2	4	2	1	13	5	-	24	5	8	4	5	5	7	-	3	3	1	-	1	2	1	25	2	1	20	2	20	3	1	5	24	22	12	2	5	6	3	2	-	1	8	2	-	1	4	2	7	1	3	-	6	-	4	3	7	1	-	1	24	4	4	7	27	40	21	2	67	1	2	6	2	2	-	1	-	5	2	2	2	3	-	12	3	-	6	1	1	7	-	8	3	10	-	3	1	11	2	2	1	2	1	2	9	3	13	6	3	-	-	-	8	4	1	-	-	4	6	4	6	1	-	3	-	3	1	2	2	1	1	5	2	4	6	1	3	-	-	12	-	-	2	6	-	2	5	1	-	1	1	1	3	17	40	30	12	-	6	-	3	2	225	-	-	1	-	-	1	4	2	2	2	1	1	1	2	1	-	-	-	2	-	5	1	-	-	-	1	1	2	-	-	2	-	1	1	2	-	1	1	-	-	2	-	-	-	4	2	-	3	1	2	-	-	-	-	-	2	1	2	-	-	-	1	2	1	1	1	1	1	1	2	1	1	1	1	2	1	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	3	4	1	2	5	1	4	5	11	1	5	49	3	8	6	1	40	1	-	2	1	-	6	5	10	-	4	3	1	6	9	5	36	6	1	2	11	6	6	28	-	21	-	1	2	-	7	2	7	1	8	3	-	13	-	1	4	3	5	2	-	15	11	-	19	11	5	2	6	4	11	3	5	2	-	-	-	1	-	23	2	2	14	-	19	5	1	10	29	18	15	2	2	3	1	-	1	1	4	2	-	1	8	1	4	-	11	1	6	-	5	6	8	-	-	-	18	2	3	5	21	47	11	4	75	3	4	7	-	3	1	1	-	2	1	2	1	2	2	11	4	1	8	1	1	4	-	13	1	6	3	1	1	11	1	-	1	-	-	-	7	2	10	8	3	1	3	-	3	1	-	-	-	3	-	6	10	3	-	6	-	2	6	3	1	1	1	3	-	2	1	-	4	-	5	7	1	-	4	9	1	1	9	-	1	2	1	1	2	19	27	16	9	-	7	2	-	-	246	-	1	1	-	-	1	1	3	2	-	-	-	-	-	-	-	-	-	1	-	-	-	1	-	2	1	3	-	1	-	1	-	-	1	1	3	-	-	-	-	1	-	1	1	2	3	-	5	1	-	-	1	-	-	-	1	-	1	1	-	-	-	-	-	-	1	-	1	-	-	-	-	3	1	-	2	-	-	-	1	1	-	-	-	1	1	1	1	1	1	1	2	2	1	1	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1	2	-	1	4	1	1	4	12	7	5	32	2	1	1	7	38	2	1	2	1	1	5	10	6	1	2	3	2	5	15	6	39	3	-	1	17	21	12	28	5	3	1	1	2	3	3	1	5	1	8	6	2	9	4	2	4	1	14	1	1	11	8	-	17	4	3	2	5	5	7	3	4	1	-	-	-	2	1	16	-	2	23	1	16	-	-	5	27	35	14	3	2	5	1	3	-	-	1	-	2	-	1	1	12	-	8	1	3	-	1	4	8	-	1	1	26	1	8	3	32	53	15	8	70	4	1	8	1	4	1	-	-	2	5	9	-	2	-	14	4	-	5	-	1	3	-	10	6	5	3	-	4	9	1	2	-	-	2	1	11	6	11	8	6	1	1	-	4	4	1	1	-	3	2	2	6	4	1	2	1	9	-	5	2	-	2	6	-	3	6	-	4	-	4	4	2	1	3	7	-	-	3	1	3	-	2	1	3	16	28	23	12	2	6	1	4	2	196	-	1	2	1	1	-	1	3	2	1	2	1	1	2	1	1	-	2	-	-	2	1	-	-	2	5	3	1	-	-	1	1	1	-	-	3	-	3	-	-	-	-	-	1	-	3	-	1	-	1	-	1	-	-	-	1	-	2	-	1	-	1	1	-	2	-	-	-	1	-	1	-	-	1	1	1	-	-	1	-	3	-	1	1	-	-	-	1	2	-	-	2	1	3	-	1	-	-	1	1	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	-	5	1	1	1	-	3	1	13	-	6	41	4	4	9	3	35	3	1	2	3	1	10	2	15	3	2	4	-	10	21	-	32	6	-	5	19	16	13	32	-	23	1	4	-	1	2	2	4	2	1	7	1	13	4	2	3	5	5	4	1	10	5	-	22	5	3	1	5	4	13	2	-	2	-	-	-	2	-	20	1	1	15	-	16	1	2	9	30	24	8	2	2	2	-	2	1	-	6	4	-	2	3	1	8	1	4	-	3	1	6	2	11	1	-	-	15	-	3	7	21	43	7	1	72	1	4	7	-	3	1	1	-	3	5	7	3	1	-	10	4	-	5	1	1	2	-	7	-	11	1	3	-	5	1	1	1	1	1	2	4	3	15	2	10	-	1	-	4	3	-	-	-	1	5	5	6	5	-	4	-	4	-	1	3	1	1	8	-	4	2	1	2	-	1	11	4	-	2	10	-	-	2	-	4	1	2	1	2	19	35	27	13	-	5	2	6	1	270	1	-	1	-	-	2	2	6	2	-	1	-	-	2	3	-	-	-	2	-	1	1	-	-	-	1	2	1	-	-	-	1	-	3	-	2	1	-	1	2	1	-	1	1	-	-	-	4	1	1	1	-	-	-	-	-	2	1	-	-	-	-	2	-	-	-	-	-	-	-	1	-	-	1	1	2	-	-	-	-	-	-	-	-	-	-	2	1	2	-	-	-	-	2	-	-	1	-	-	-	1	-	-	-	1	-	-	-	2	1	2	1	4	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1	5	2	2	5	3	3	3	9	1	1	42	1	2	11	4	31	1	-	2	1	2	7	3	5	-	2	10	1	3	14	9	33	6	2	3	16	16	11	31	1	24	-	-	6	4	8	2	2	1	9	2	1	6	3	2	6	2	8	-	2	8	12	-	15	6	5	2	8	9	14	2	1	1	-	1	2	-	1	21	-	4	20	-	28	3	1	12	33	15	13	4	2	6	-	-	-	1	5	1	1	2	2	1	5	-	6	1	1	1	7	4	13	1	1	1	25	2	5	5	34	52	13	2	87	2	3	3	4	7	-	1	-	1	2	6	-	3	3	17	11	-	4	1	1	6	1	8	3	8	3	1	4	13	1	2	2	1	-	2	10	5	14	7	8	1	3	1	2	3	2	-	1	5	3	1	2	7	-	5	-	5	1	1	1	3	2	3	-	3	4	1	3	-	3	9	-	-	4	9	-	2	-	6	4	5	4	4	1	19	32	23	11	-	6	-	8	-	131	1	-	1	1	1	1	6	4	1	-	1	-	-	1	5	-	1	1	2	1	2	2	1	-	-	3	4	1	-	1	1	-	-	-	-	1	-	1	1	-	1	-	3	-	-	-	-	3	1	1	1	1	2	-	-	3	1	1	1	-	-	-	-	2	-	3	1	-	-	-	1	-	-	4	-	1	-	-	-	1	-	-	-	-	-	1	-	-	2	-	1	2	-	2	-	1	3	-	-	-	-	-	-	-	-	1	-	-	1	-	-	-	-	1	-	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	4	-	2	2	-	1	1	3	11	1	9	29	1	8	3	1	30	-	-	-	-	-	6	10	6	1	1	4	-	7	14	3	39	5	-	1	12	13	8	27	2	16	-	-	2	2	4	1	2	2	6	2	2	9	8	1	6	1	3	2	3	7	10	-	12	8	2	3	1	6	7	2	2	1	-	1	-	3	1	24	-	4	17	-	15	-	-	5	27	25	14	-	6	2	3	1	-	-	3	3	-	6	-	1	5	2	9	3	5	1	4	9	12	-	-	1	27	4	8	4	28	42	6	2	77	-	6	4	2	5	2	2	-	3	3	8	1	5	1	19	6	-	7	-	2	5	-	6	3	5	1	1	3	7	1	-	-	3	-	2	7	3	28	5	9	-	2	-	4	5	-	2	-	3	4	5	7	-	-	3	-	6	3	3	1	4	-	-	1	4	5	-	4	-	10	10	2	-	2	10	3	-	5	-	2	2	2	1	2	19	32	18	10	3	10	-	4	-	167	1	-	1	-	-	2	4	6	2	1	-	2	-	1	-	-	-	1	3	3	4	3	2	-	1	2	4	1	1	-	1	1	-	1	2	1	1	-	2	-	-	-	-	-	-	2	1	1	1	1	-	-	1	-	-	2	1	-	1	1	-	2	-	-	2	2	1	-	-	-	-	2	1	3	1	-	1	-	-	-	-	-	-	-	1	1	1	-	1	1	1	-	-	-	1	3	1	-	-	-	2	-	-	1	1	1	-	-	1	-	1	-	2	-	1	-	1	-	-	-	2	2	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	-	3	1	4	1	-	3	8	16	4	3	33	1	10	5	5	82	-	-	3	2	2	11	9	16	2	2	5	2	5	19	9	44	6	-	2	18	22	16	31	-	16	1	2	4	3	4	1	7	4	9	1	3	16	3	2	6	2	9	2	-	16	12	1	30	14	4	15	5	8	20	2	5	2	-	-	-	3	-	42	1	3	23	3	21	4	2	11	28	22	19	2	3	5	3	3	-	1	9	8	1	7	4	7	6	1	8	-	3	-	8	6	15	2	-	-	27	4	6	4	35	51	12	6	76	1	9	7	3	7	-	1	-	7	13	4	1	2	2	15	8	-	4	-	2	5	-	10	2	7	2	1	3	8	1	4	2	2	-	3	7	5	20	6	4	-	2	-	7	3	-	1	2	4	3	7	5	7	1	5	-	8	1	9	3	1	-	5	-	7	9	-	4	1	1	7	3	1	1	8	1	-	3	6	-	1	1	3	2	27	44	19	10	-	12	1	4	-	168	1	-	3	2	1	-	4	7	2	3	2	1	1	-	2	1	1	-	2	-	3	5	-	-	1	2	1	3	1	-	-	-	-	-	4	3	-	2	-	-	1	-	2	1	2	1	1	2	-	1	1	2	-	-	-	-	1	1	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	3	1	-	-	-	-	1	-	2	-	-	-	-	-	1	3	-	-	1	-	-	1	2	4	2	-	-	-	2	-	-	1	1	2	-	2	1	-	-	3	1	-	-	-	-	-	-	1	-	1	-	-	-	2	-	-	-	1	1	1	1	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1	1	-	1	3	2	1	5	10	-	6	32	2	5	3	4	30	2	1	1	1	-	7	6	10	2	1	7	1	5	12	10	39	5	-	2	18	16	12	23	1	15	3	1	5	3	9	3	8	3	7	6	2	10	2	4	1	4	4	1	1	13	9	-	30	4	3	4	4	5	11	2	2	2	-	-	-	1	-	24	-	4	9	-	25	2	-	12	27	31	10	1	3	-	2	1	-	-	4	3	2	-	2	1	5	2	8	1	1	1	3	6	12	1	-	-	25	5	10	6	26	56	11	5	98	-	3	6	4	7	1	3	1	8	6	9	-	1	2	23	11	-	1	-	2	5	1	5	4	6	1	-	1	18	3	1	-	1	-	1	5	6	13	4	7	-	3	-	4	2	1	-	2	3	2	4	6	2	-	6	-	9	2	7	4	3	-	4	-	2	10	2	5	-	4	9	3	4	2	16	-	2	5	1	1	6	2	2	4	21	34	21	20	2	7	-	4	-	133	1	-	-	1	-	2	-	7	-	-	-	1	2	2	4	-	1	-	5	-	-	-	1	-	-	5	-	4	-	-	1	1	-	2	-	3	-	5	2	1	-	1	1	-	1	1	1	1	2	1	-	1	5	-	1	2	1	-	2	-	2	-	-	-	-	5	1	-	-	-	-	2	1	-	-	3	-	-	-	1	2	1	1	-	-	-	-	1	2	-	-	-	-	1	-	3	1	2	-	2	1	1	-	2	1	-	1	-	-	-	1	-	4	2	-	-	-	-	-	1	-	1	-	1	-	-	-	1	-	-	-	-	-	-	-	1	-	-	1	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	-	7	4	3	3	1	3	4	10	3	7	25	1	5	2	5	30	3	-	1	-	-	14	4	10	-	1	1	-	3	11	6	46	4	1	2	10	14	14	22	1	20	2	1	6	2	2	4	6	3	6	-	3	7	3	4	1	4	7	3	1	15	9	-	22	5	3	8	4	7	12	2	3	1	-	-	-	-	1	20	-	1	13	1	26	2	-	4	21	21	19	2	2	4	4	-	-	1	4	1	-	3	1	1	4	1	6	1	3	1	5	4	10	-	-	1	30	6	5	2	53	40	8	1	65	1	6	9	8	7	2	3	-	6	5	10	3	3	1	22	10	1	4	1	1	5	2	9	2	4	6	-	1	11	-	1	3	2	3	2	6	5	20	7	6	3	1	-	11	2	-	-	-	5	4	3	8	4	1	2	-	9	4	3	2	2	2	2	1	4	5	-	8	1	7	7	2	-	1	16	-	1	2	2	4	4	-	1	3	24	18	25	13	1	5	1	7	1	95	-	1	-	-	1	3	1	2	1	1	1	-	-	-	-	-	1	1	3	-	3	1	-	-	1	1	-	-	-	-	2	-	-	-	-	1	-	1	-	-	-	-	1	1	2	2	-	-	-	-	2	-	1	-	-	3	-	1	2	-	-	-	1	-	-	1	1	1	1	-	3	-	2	1	1	4	-	-	1	2	2	-	1	1	-	1	-	-	1	1	-	2	-	1	-	1	1	-	1	-	1	1	-	1	-	-	1	1	1	1	-	1	-	1	-	1	-	-	1	-	1	-	1	-	-	-	1	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	-	4	-	1	1	-	-	4	16	3	6	27	2	7	2	3	28	2	-	-	-	1	11	2	8	1	1	9	2	5	10	3	39	2	1	-	15	7	7	31	2	14	1	3	4	1	1	1	5	1	4	-	2	10	5	2	3	2	3	-	2	17	7	1	15	5	1	6	3	10	12	1	2	2	-	-	-	1	-	17	-	4	17	-	16	1	1	16	34	23	18	1	4	-	3	2	-	4	6	7	1	4	4	1	8	1	10	1	2	2	8	6	9	2	-	1	29	3	7	3	24	52	10	2	77	-	13	5	9	7	2	-	-	12	1	2	1	2	3	24	10	-	6	3	1	6	3	6	5	2	1	1	1	13	5	3	3	5	1	4	8	5	24	8	9	-	1	1	3	1	-	-	1	6	5	6	11	4	2	4	-	8	5	7	1	1	-	2	-	4	3	1	4	-	4	10	5	1	2	12	2	1	4	3	6	4	4	3	1	24	26	19	13	-	11	-	12	3	120	1	2	1	3	1	-	-	-	-	1	2	-	-	1	6	-	-	3	1	-	2	2	2	-	-	2	4	1	-	-	-	-	-	3	1	-	1	3	-	-	2	-	1	-	2	-	-	5	-	-	1	-	-	-	-	-	2	1	2	-	-	1	-	1	-	1	-	-	2	-	-	1	-	2	2	1	-	-	-	-	-	-	-	-	-	-	1	-	2	-	-	1	1	4	1	1	-	-	1	1	-	1	-	1	-	-	-	-	-	-	1	-	3	2	-	1	-	-	-	1	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	1	-	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	1534

1986	2	5	-	2	2	-	1	1	9	1	4	25	1	2	4	7	24	1	1	-	-	-	5	4	4	1	1	6	3	5	12	4	39	6	-	2	14	21	12	19	-	19	2	2	5	3	3	2	6	1	8	1	6	8	6	-	6	2	2	5	-	9	8	-	20	14	6	6	1	3	18	-	1	1	1	1	-	1	-	27	-	5	15	-	23	1	1	4	23	20	18	4	3	6	3	1	-	-	10	2	2	-	2	2	4	1	11	1	2	2	7	9	13	-	-	-	31	8	4	4	44	37	12	3	66	2	4	4	2	3	-	2	-	6	2	8	-	3	-	14	13	1	3	1	1	9	-	9	8	4	-	1	1	13	1	4	2	-	1	1	6	4	22	8	7	-	1	-	3	3	-	-	-	5	3	4	7	4	-	2	-	7	2	4	4	3	-	2	-	-	3	1	7	-	3	12	4	1	4	11	3	-	6	-	3	3	1	2	4	30	27	17	19	1	6	2	10	1	112	-	-	-	-	1	-	1	2	3	2	-	-	2	1	3	1	-	-	1	-	3	-	-	-	-	1	2	-	1	-	1	1	-	1	1	2	-	2	-	-	1	-	-	1	1	4	1	5	1	1	-	-	1	-	1	1	-	-	2	-	1	-	2	-	-	1	1	1	-	-	-	1	-	1	5	1	-	-	-	-	3	-	-	-	-	-	-	-	2	-	-	1	1	1	-	1	-	-	1	-	-	-	-	1	-	1	2	-	-	-	1	-	1	-	-	1	-	-	-	2	-	1	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	2	1	1	-	-	1	1	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	1470

1987	-	3	1	1	3	4	1	2	12	1	8	27	4	1	6	3	63	2	-	3	-	2	8	2	16	3	4	5	-	4	13	2	54	10	-	3	24	12	10	27	2	25	1	1	6	1	5	-	4	3	5	4	3	7	-	1	9	-	3	2	1	20	7	1	32	9	3	3	6	5	20	2	5	2	1	-	-	2	-	50	-	3	21	1	21	4	1	8	19	29	16	3	2	4	-	-	1	-	12	6	-	3	7	2	6	2	8	-	1	1	5	2	10	3	-	-	37	2	10	6	39	44	5	7	74	4	11	10	6	3	-	1	-	4	8	6	-	-	1	16	10	1	4	7	3	8	-	16	5	6	2	-	4	11	3	3	1	1	2	4	7	5	24	4	7	-	2	-	5	3	-	-	1	4	3	9	3	5	3	5	1	8	8	8	1	2	2	5	-	5	11	1	4	3	4	6	3	1	5	13	1	1	5	2	6	6	1	3	3	24	25	26	16	-	8	-	7	1	177	2	-	-	1	-	1	-	3	2	-	1	3	-	1	1	-	1	-	3	1	3	1	-	-	1	4	3	4	-	1	-	1	1	3	2	2	-	4	-	-	1	-	-	3	1	2	2	-	1	1	-	1	3	-	1	-	1	3	2	1	-	2	1	-	-	5	1	1	-	-	1	1	1	1	3	3	-	-	1	-	2	-	1	-	1	-	1	1	1	1	-	-	-	3	3	5	-	-	-	1	-	-	3	1	2	2	-	-	-	-	1	2	-	-	-	1	-	-	-	1	-	1	-	-	-	-	-	2	-	-	1	1	-	-	1	-	-	-	-	1	1	-	-	-	2	-	1	1	2	1	1	1	1	-	-	-	-	-	-	1819

1988	1	3	1	1	4	3	2	3	7	-	3	31	-	8	2	2	36	2	-	2	1	-	8	3	12	1	1	7	3	5	5	-	43	7	-	1	15	9	19	19	1	20	3	-	1	-	2	-	1	2	4	4	2	11	1	5	3	4	8	4	-	13	5	-	19	8	10	6	1	5	9	1	3	1	-	-	-	-	-	26	1	2	14	-	14	3	2	7	18	10	17	3	1	4	4	2	-	3	2	7	1	3	6	2	5	-	8	4	2	-	5	5	10	1	-	1	44	2	6	5	52	40	12	3	68	2	4	5	6	9	-	1	1	3	4	2	1	1	3	20	9	-	8	-	-	7	-	8	1	3	-	3	5	21	1	2	1	1	2	5	3	9	17	5	6	2	2	-	7	7	1	-	1	4	3	5	8	5	2	3	1	6	5	6	2	2	3	4	1	3	6	2	10	-	7	5	3	3	3	12	1	3	6	1	5	6	1	-	2	24	25	11	14	-	6	1	13	1	89	-	-	1	-	-	1	-	3	3	-	-	1	-	4	4	-	-	1	-	-	3	1	2	-	-	5	3	2	-	-	2	1	-	2	1	4	1	-	1	-	1	-	-	1	-	1	-	1	1	1	-	-	1	-	-	1	-	4	3	-	-	1	-	1	-	1	1	1	1	-	-	3	1	2	3	2	1	-	-	1	1	-	-	-	1	1	-	-	1	-	-	1	1	3	-	-	3	-	-	-	-	-	-	-	2	-	1	-	-	1	-	-	1	-	-	2	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	3	-	-	1	-	-	-	-	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1	2	4	2	3	1	1	2	9	3	4	19	1	5	1	7	24	3	-	1	-	2	5	2	7	1	4	6	1	5	15	7	42	5	-	2	13	18	9	29	-	18	2	2	3	1	6	1	6	3	5	1	2	9	4	3	7	1	6	-	1	14	5	1	23	9	5	6	5	4	5	2	2	2	1	-	-	3	-	29	-	3	10	1	24	1	2	10	26	15	12	1	6	4	4	-	-	-	5	5	1	-	2	1	3	1	11	1	-	-	5	1	10	1	-	-	41	2	5	4	47	42	7	2	84	1	6	4	6	7	2	1	-	3	2	5	2	1	1	24	8	-	4	2	-	8	3	10	3	7	3	2	2	15	-	-	3	2	-	4	8	12	27	4	5	1	7	-	8	1	-	1	1	4	2	11	5	5	1	4	-	9	6	8	4	4	1	2	-	4	7	2	7	-	3	11	6	5	4	9	2	2	5	6	4	5	-	5	4	24	19	23	20	-	12	-	5	-	109	1	-	-	1	-	-	-	1	1	-	-	-	-	-	2	-	1	-	2	-	-	1	-	-	1	6	2	1	-	-	2	1	-	-	-	1	-	1	-	-	-	-	-	2	-	1	-	2	1	1	-	-	3	-	-	4	1	-	1	-	1	-	-	1	-	1	2	-	1	-	-	-	-	1	2	3	1	-	1	1	-	-	-	-	-	-	1	-	2	-	-	1	-	4	1	1	1	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	2	2	1	-	3	1	-	1	1	-	2	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	1	-	-	2	-	1	-	-	-	-	-	-	-	-	-	1537

1990	-	2	-	2	3	-	3	3	11	2	8	19	-	3	4	2	22	2	1	-	-	2	8	-	13	1	3	4	-	3	5	3	31	5	-	1	11	16	8	29	1	12	1	-	5	1	4	-	10	3	5	1	2	7	4	3	2	2	9	2	-	7	10	1	16	15	6	3	5	7	10	3	2	2	-	-	-	3	-	23	-	2	8	1	14	2	3	6	19	15	13	1	4	3	-	1	1	-	2	3	3	1	3	1	2	2	10	2	5	-	3	5	16	1	1	-	31	2	5	2	28	41	3	3	76	2	6	8	3	4	-	1	-	2	4	9	1	2	3	19	11	-	5	2	-	4	1	6	3	5	1	3	2	16	3	2	3	-	2	2	6	9	17	4	13	-	1	1	2	1	-	-	1	4	4	9	5	8	2	3	-	4	1	7	2	1	1	3	-	6	7	2	10	-	2	11	5	1	2	15	2	1	10	3	2	2	3	4	3	20	23	22	13	1	6	-	5	3	73	-	1	3	1	-	1	1	-	4	1	-	-	-	-	1	-	-	1	3	1	1	2	-	-	1	5	1	1	-	-	1	-	-	2	2	3	-	1	1	1	1	1	2	-	1	-	1	-	-	-	-	2	2	-	-	1	-	1	-	2	-	-	-	-	1	-	-	3	1	-	-	3	-	2	1	5	-	-	-	2	3	1	1	-	-	-	-	-	-	-	2	2	-	1	-	2	1	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	-	3	-	-	1	1	-	-	-	-	-	-	-	2	-	3	2	-	1	2	1	-	-	-	1	-	-	1	-	1	-	-	-	2	1	1	-	-	-	-	-	-	-	-	-	1372

1991	-	5	1	1	-	1	-	2	12	-	4	29	-	3	6	3	15	2	2	1	2	2	6	3	8	1	2	7	-	1	13	8	42	5	1	3	15	25	8	22	2	16	1	-	3	2	4	1	6	1	4	-	3	8	2	-	7	6	3	2	-	14	8	1	24	7	4	3	2	2	11	1	1	1	-	-	-	3	-	38	-	1	13	-	21	2	2	7	25	12	12	-	5	-	2	1	-	-	7	6	4	5	4	1	5	-	10	2	2	1	5	3	10	1	-	-	37	2	15	4	36	45	10	2	66	5	7	7	8	2	2	3	-	5	2	3	1	5	2	16	8	1	4	2	1	6	2	12	3	1	4	-	2	12	1	1	-	3	1	3	5	2	32	5	8	-	5	1	13	2	-	-	2	5	3	10	8	4	2	4	-	7	2	8	4	2	2	-	-	2	7	3	10	-	3	7	2	4	1	9	7	-	2	5	2	4	-	4	-	18	30	20	15	1	8	4	4	1	100	1	-	2	1	-	-	2	3	-	-	3	1	-	2	2	1	-	1	2	-	2	1	-	-	1	5	2	3	-	-	-	2	-	2	1	3	1	3	-	1	-	1	-	2	1	2	2	3	2	2	1	-	3	-	-	-	-	3	1	3	2	-	2	2	-	-	-	1	1	-	1	-	4	-	2	4	-	1	1	-	1	-	1	-	-	2	1	-	3	-	1	3	-	4	-	3	3	-	2	-	-	1	-	-	-	1	-	-	1	-	-	-	1	1	-	2	1	-	-	2	-	1	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	2	-	2	1	4	3	-	2	7	3	1	27	-	3	3	3	25	1	-	2	-	-	12	5	10	-	1	6	2	3	12	4	40	3	-	3	18	18	12	30	-	14	-	2	3	1	3	2	8	3	4	1	1	10	4	1	6	1	9	1	1	10	7	-	24	8	5	3	4	5	16	3	-	4	-	1	-	1	1	25	1	2	12	-	22	1	1	12	24	20	16	2	4	5	2	1	-	-	4	6	1	1	2	-	4	1	7	1	3	-	6	5	10	2	-	-	38	6	7	4	31	39	9	1	74	3	13	8	8	5	-	5	-	3	1	10	4	4	2	25	9	-	6	4	2	14	3	5	2	2	3	2	3	13	4	3	-	2	5	1	8	7	24	6	12	1	3	-	5	3	2	1	1	9	5	7	5	6	-	8	-	10	7	13	1	5	7	4	-	3	9	2	9	2	2	13	7	1	2	20	1	4	2	5	1	6	4	2	1	23	25	15	10	1	16	-	8	3	85	1	-	1	-	-	2	1	2	1	-	2	1	-	2	3	-	-	-	1	-	1	4	1	-	1	2	4	1	1	1	-	2	-	2	2	1	-	5	3	-	-	1	1	-	1	4	1	1	1	1	-	-	3	-	1	1	-	3	3	2	-	2	1	-	-	-	-	-	1	-	2	3	2	3	5	4	-	1	1	1	-	-	3	-	-	-	-	-	1	-	-	-	1	2	-	4	2	1	-	-	-	-	-	1	1	-	-	-	-	1	-	-	3	-	-	1	-	-	1	3	-	1	-	-	-	1	-	1	1	-	2	-	1	3	-	-	-	-	1	-	1	1	1	-	-	-	-	2	-	-	-	-	-	2	1	1	-	-	-	1606

1994	7	5	2	1	4	4	2	8	13	2	8	26	2	6	6	4	50	6	1	3	1	3	19	7	18	2	8	11	2	7	17	7	57	6	1	1	31	39	18	58	-	23	2	1	10	3	7	2	17	1	14	3	12	21	9	2	11	3	16	3	-	24	7	1	44	16	13	10	4	10	23	3	10	4	-	2	1	4	-	59	-	3	24	2	32	5	1	23	52	24	38	4	4	7	4	3	2	2	8	9	2	2	9	6	7	1	17	2	7	1	6	10	22	-	1	-	61	2	12	8	74	64	11	4	143	3	20	21	25	11	4	7	-	8	3	14	2	5	4	31	12	-	18	1	2	13	3	17	9	6	3	2	7	27	10	6	7	5	1	3	8	11	58	17	15	-	5	1	11	5	-	-	1	14	6	13	15	11	5	8	3	26	12	15	3	3	3	8	-	14	20	4	20	3	10	16	9	2	9	35	4	6	8	3	8	17	6	6	3	40	61	34	27	2	22	3	11	5	147	-	-	1	2	1	2	2	-	3	-	1	2	2	1	4	-	-	2	-	1	5	6	1	-	3	6	3	4	-	3	4	-	1	2	2	10	-	11	2	3	-	-	1	4	5	6	1	7	1	2	-	-	8	-	2	4	-	4	2	2	-	-	-	-	1	4	5	3	1	-	4	3	1	-	7	9	1	3	2	-	2	2	2	1	-	-	-	1	3	-	3	2	1	8	1	4	-	1	3	1	1	-	-	-	1	3	3	1	-	1	-	-	2	1	1	5	-	-	3	2	1	1	-	3	-	-	1	-	1	1	-	-	-	-	-	1	-	3	-	-	-	1	-	-	3	3	-	1	-	-	-	1	-	-	-	-	-	-	-	2992

1996	3	5	2	2	7	1	2	7	11	4	2	33	1	5	6	8	46	8	-	-	2	11	15	8	13	6	5	17	-	8	20	11	87	5	1	3	27	35	28	55	-	29	3	4	9	1	9	3	21	4	14	5	3	23	5	1	7	3	12	1	1	35	16	-	36	18	14	12	2	6	25	2	9	5	-	-	-	3	2	58	-	3	37	1	34	3	3	9	40	16	34	4	2	2	3	-	3	1	13	7	7	3	9	3	6	-	34	3	7	1	16	7	23	5	1	1	63	1	6	10	63	77	8	4	102	5	15	17	23	13	4	5	-	4	1	18	4	3	3	35	15	-	9	3	-	11	4	16	10	-	1	2	5	19	4	1	6	2	5	4	7	21	42	10	11	1	6	-	18	4	-	2	3	13	12	18	9	12	4	5	1	24	13	17	6	4	3	12	-	8	21	2	21	1	9	14	13	3	8	29	2	2	13	9	9	18	1	4	6	40	50	41	20	3	15	2	25	3	123	1	-	2	2	1	1	1	-	5	-	1	2	2	5	5	2	-	1	3	1	3	3	2	-	2	5	4	3	2	-	1	3	-	3	2	6	1	6	2	-	3	4	-	1	1	3	-	10	-	1	3	1	10	-	1	2	1	2	3	1	-	-	3	4	-	3	3	1	1	-	3	1	4	3	5	12	1	1	1	-	1	-	4	-	-	-	-	-	1	-	2	-	1	5	-	3	2	-	-	-	-	1	-	-	-	4	3	-	2	2	-	1	3	3	1	5	-	3	2	3	1	1	2	2	1	-	-	4	2	-	1	1	-	2	-	-	-	-	-	-	2	2	-	1	1	3	-	2	1	-	-	1	-	-	-	-	-	-	-	2904

1998	1	3	2	1	1	2	2	3	12	3	5	17	3	3	9	10	39	4	1	3	-	3	19	3	17	2	4	17	1	8	17	7	69	6	1	1	17	25	32	54	1	33	-	2	6	-	11	3	10	2	12	1	4	18	8	3	8	7	10	2	1	27	4	1	53	19	9	14	4	2	19	1	12	3	1	3	-	3	-	56	-	4	28	1	35	5	-	22	42	15	39	2	8	8	8	5	2	1	11	10	7	5	5	6	5	3	20	2	2	2	10	5	22	2	-	-	56	6	10	6	62	59	7	2	119	11	16	17	23	10	1	7	-	2	6	11	3	5	3	36	15	-	9	3	3	17	7	14	5	3	3	4	3	29	8	1	8	6	7	7	9	11	41	15	23	2	9	-	9	2	-	1	2	7	4	13	19	14	3	2	-	16	5	22	6	7	6	11	-	4	11	1	23	1	24	29	6	6	5	22	3	3	17	10	3	20	6	5	6	57	38	28	33	5	22	1	11	5	166	2	-	-	3	-	-	2	-	5	1	1	2	-	3	6	-	-	-	-	-	3	4	1	-	2	5	8	4	-	1	1	1	-	4	2	2	1	4	2	1	1	3	1	3	1	2	1	7	-	1	2	-	11	-	-	-	-	2	3	1	2	2	-	-	1	3	5	-	-	-	2	-	1	2	4	6	3	1	1	2	6	-	3	1	1	-	1	-	2	-	-	-	-	7	1	3	4	-	-	1	1	-	1	1	-	1	1	-	1	2	-	2	-	-	2	4	-	2	1	2	1	1	1	-	1	1	2	1	-	2	2	1	1	-	1	1	-	5	-	-	-	-	3	2	2	2	-	1	2	1	-	-	-	-	-	-	-	-	-	2832

2000	-	2	2	1	1	4	1	4	14	2	9	29	1	4	6	9	52	9	-	2	-	3	17	5	21	2	3	11	2	7	20	9	73	5	-	2	36	28	22	45	1	22	2	1	7	4	10	-	17	1	9	3	2	17	4	5	10	4	12	5	-	17	16	2	55	9	9	10	6	8	18	7	14	-	-	2	-	5	-	60	1	-	19	4	45	3	1	20	45	17	35	4	3	5	9	4	2	4	11	7	5	5	7	1	2	2	17	1	9	1	5	12	17	3	1	2	62	8	10	10	63	46	8	1	100	6	14	14	27	6	-	9	-	4	4	11	7	7	1	41	15	-	13	6	3	18	3	18	8	4	1	1	2	13	8	2	5	2	3	5	6	15	42	12	19	2	7	1	13	2	2	1	4	9	5	14	21	10	-	6	1	18	7	22	4	7	8	7	-	6	13	3	27	1	10	20	21	7	3	43	3	1	8	10	5	14	2	5	4	49	56	33	31	2	13	1	10	2	143	-	-	1	1	1	-	1	1	3	-	1	-	2	3	3	-	1	-	2	-	4	4	1	1	2	7	5	4	-	1	1	1	-	4	4	4	-	5	2	1	1	3	-	4	8	3	2	6	-	-	2	-	11	-	3	1	-	1	1	1	1	3	1	-	-	2	1	3	1	-	2	2	-	2	8	9	1	-	1	1	2	1	4	-	1	-	-	1	2	-	-	-	3	9	1	5	3	1	1	1	1	-	-	-	3	-	-	-	-	-	-	-	4	2	1	5	3	1	-	2	-	-	-	3	-	-	1	2	3	-	-	-	1	1	3	1	1	1	-	-	-	-	1	6	1	1	1	3	-	-	2	2	-	2	-	1	1	-	-	2817

2002	4	4	2	1	2	3	7	2	16	2	3	22	-	2	5	4	35	4	2	4	1	-	15	10	20	-	1	12	-	10	11	5	81	7	1	-	27	22	23	48	1	25	3	1	12	1	9	2	11	5	7	5	1	16	6	4	14	1	11	1	1	21	10	-	48	18	12	13	3	8	13	2	14	3	1	1	-	5	2	60	-	5	24	3	30	8	1	14	43	9	45	2	6	4	1	2	2	-	14	16	6	6	4	4	4	4	21	2	9	3	6	6	21	3	1	-	49	4	12	7	72	61	9	1	101	6	17	16	36	6	-	1	1	5	2	8	4	4	3	35	21	1	8	3	5	14	3	7	5	4	4	4	9	33	3	8	6	1	-	5	8	10	52	11	19	5	7	5	15	5	1	1	1	11	7	13	17	7	1	5	-	22	7	18	7	5	3	8	1	6	19	7	31	2	8	27	10	3	9	29	6	4	11	4	6	15	2	3	4	35	50	36	30	3	13	1	12	2	112	-	2	2	-	-	1	2	-	4	-	3	1	-	-	1	-	-	1	1	1	7	1	-	-	3	6	3	10	-	-	1	1	2	1	1	3	-	5	-	2	-	3	-	3	6	5	3	4	-	4	1	1	19	-	3	-	1	2	1	1	-	-	-	2	1	5	2	4	-	-	1	-	1	8	4	14	1	-	1	1	1	2	2	1	1	-	-	-	-	-	1	1	-	10	3	4	3	1	-	-	-	1	-	1	-	1	1	-	1	1	1	-	2	-	1	2	-	5	2	5	-	-	3	3	3	1	1	3	1	-	-	2	-	3	-	1	-	3	2	2	-	4	3	1	1	3	-	2	-	1	1	3	-	2	1	-	-	-	-	2765

2004	1	1	4	1	2	1	5	7	15	2	4	28	-	5	4	3	32	1	2	2	2	3	19	8	23	2	7	10	2	6	17	9	63	6	-	2	25	30	13	53	2	34	4	-	7	1	11	2	11	3	2	4	5	16	8	4	6	4	5	2	2	12	12	1	38	18	13	11	5	4	14	6	7	-	-	-	-	3	-	73	-	1	14	-	30	7	2	10	48	19	21	2	10	1	5	2	-	-	9	9	5	8	3	4	2	2	11	1	2	4	5	9	19	2	1	1	59	2	9	16	60	55	2	2	102	4	25	24	27	7	1	4	-	5	9	11	3	5	2	30	13	-	12	3	6	21	2	22	12	3	3	2	8	30	6	7	4	5	2	3	13	13	60	9	13	2	6	2	18	4	1	2	3	6	9	13	15	14	2	9	1	22	6	13	4	6	8	12	-	7	19	4	32	2	7	24	14	5	8	42	3	3	9	8	6	30	1	3	5	30	59	42	31	5	19	3	19	3	154	-	-	3	4	1	2	-	1	-	-	1	3	1	3	5	-	-	-	1	1	5	4	-	-	3	1	4	4	1	-	1	-	-	2	2	4	2	3	1	2	-	2	-	1	3	4	2	6	-	1	3	1	19	1	1	3	3	1	5	3	-	-	-	2	-	2	3	1	2	-	4	2	1	8	6	11	4	2	1	1	-	1	3	-	-	1	1	-	1	-	-	-	1	11	1	-	3	2	-	1	-	1	-	3	1	2	2	2	-	-	1	-	2	1	-	4	1	2	1	4	2	-	-	-	-	2	1	1	1	2	-	2	1	1	-	-	1	3	1	1	-	1	-	1	1	2	1	5	-	-	-	2	-	2	-	-	-	1	-	2812

2006	4	3	6	3	5	9	4	15	25	4	4	38	2	8	6	4	74	9	3	4	3	6	29	8	56	8	4	25	-	12	22	7	137	13	-	2	41	40	33	96	1	42	7	1	10	4	10	1	20	5	10	4	5	32	13	10	9	7	13	3	-	38	19	2	71	27	11	20	6	12	32	6	19	3	1	-	-	4	-	93	-	3	33	1	54	5	1	22	63	28	54	4	9	5	5	3	5	2	16	11	5	11	8	5	3	1	21	4	7	6	26	13	22	6	1	4	83	11	16	10	85	84	14	5	145	3	24	30	53	16	3	6	-	5	13	13	4	8	1	52	15	-	16	10	1	29	2	11	9	12	5	4	9	43	5	13	7	3	4	8	21	33	84	25	34	4	12	4	21	5	3	1	1	17	14	34	26	16	6	10	1	25	16	25	7	10	9	15	-	10	31	7	34	1	23	32	14	4	13	47	5	8	17	13	10	44	6	7	7	51	114	48	52	4	36	3	29	4	223	-	1	3	3	-	-	1	3	6	1	3	2	-	2	5	1	1	3	4	1	3	4	1	1	5	10	14	8	-	1	2	1	-	4	2	9	1	8	7	3	2	4	-	7	12	8	1	1	1	2	-	3	27	-	3	5	4	9	5	5	1	1	-	-	5	10	5	2	1	2	2	-	1	1	6	10	2	2	3	5	5	1	6	-	-	-	-	1	6	-	1	1	1	14	1	13	1	4	-	2	-	1	1	3	1	4	-	-	1	2	1	1	5	4	1	6	-	6	1	-	-	-	-	3	-	1	1	5	1	2	-	1	2	2	1	2	-	6	4	-	1	-	2	1	1	1	-	5	-	4	2	5	1	-	1	2	-	-	-	4510

2008	3	1	-	1	3	4	4	7	17	-	2	16	3	8	2	4	23	4	-	4	4	3	14	3	18	1	3	9	4	4	15	6	54	9	-	1	21	19	8	46	-	27	2	3	11	4	7	1	4	1	3	4	3	14	3	2	4	2	10	1	1	13	6	2	30	15	6	8	1	5	12	5	8	-	-	-	-	1	-	48	-	1	16	-	26	4	-	8	33	13	22	-	3	2	1	1	2	-	11	8	3	6	6	-	3	1	16	-	3	-	3	8	12	2	-	1	30	4	8	8	32	22	5	4	61	3	19	6	25	5	2	4	-	1	3	6	3	1	-	22	13	-	10	2	3	13	1	7	3	4	3	1	3	23	5	2	2	6	6	2	7	5	37	8	10	3	3	1	13	6	1	-	1	4	4	13	18	8	2	3	1	21	1	22	3	3	2	15	-	5	17	3	25	1	7	16	9	8	4	26	2	3	11	4	7	22	2	3	3	32	31	26	23	4	10	1	6	4	72	1	2	3	-	-	-	-	-	1	2	2	-	-	2	1	-	1	2	3	-	6	-	1	-	-	6	4	3	-	-	-	1	-	3	4	8	-	10	2	1	-	2	-	3	2	8	1	5	-	-	1	1	11	-	-	2	-	2	-	2	-	1	1	-	1	2	4	-	3	-	3	1	-	1	7	2	1	1	-	1	1	-	-	-	-	-	3	-	1	-	1	1	1	10	-	1	4	1	-	-	-	3	-	2	-	2	2	1	-	-	-	-	-	-	-	1	-	-	3	3	-	-	1	-	-	-	1	-	-	2	-	-	1	4	-	-	-	1	2	2	-	2	1	-	-	1	1	2	-	1	2	4	-	-	-	1	-	-	-	2023

2010	-	1	1	-	2	4	4	1	12	2	4	20	1	-	4	3	26	1	-	1	1	2	14	3	16	-	4	15	1	8	21	5	64	10	-	-	27	11	11	39	-	23	-	2	13	-	5	-	9	3	1	3	2	7	4	1	3	2	5	1	2	12	5	-	37	9	6	11	4	3	13	1	16	2	-	1	-	-	2	43	-	3	12	2	12	2	-	9	35	10	28	2	4	5	4	3	3	-	9	2	4	1	6	-	-	2	8	-	2	1	4	6	13	3	-	1	32	8	3	4	44	33	4	2	54	-	14	9	32	4	1	4	1	3	5	5	1	-	2	28	12	-	15	5	1	15	3	4	3	9	1	2	2	17	1	7	5	3	2	5	8	24	42	13	14	1	1	3	13	2	2	1	1	7	5	13	11	12	3	5	-	15	14	14	5	1	5	10	-	5	8	3	19	4	3	10	4	8	8	22	3	-	7	9	1	13	1	2	8	30	47	17	27	5	13	3	11	1	117	1	-	4	-	-	-	-	1	3	-	2	-	-	1	4	-	1	-	4	-	2	4	-	-	2	4	4	9	-	-	-	1	-	5	1	2	1	4	-	-	1	2	3	1	1	4	-	5	-	3	-	2	12	-	2	3	2	2	-	3	-	-	2	2	-	1	-	-	3	-	1	-	4	3	4	2	1	-	1	-	3	2	-	-	-	-	-	-	2	-	1	-	1	11	1	4	5	1	-	-	-	1	-	1	-	1	-	1	-	-	-	-	3	-	2	2	-	2	1	3	-	-	-	-	-	-	-	-	-	-	1	1	1	2	-	2	-	-	-	-	-	-	2	1	3	-	1	2	-	1	1	2	-	-	-	1	-	-	-	2044

2012	-	1	3	1	-	1	2	5	7	1	3	13	-	3	4	5	31	4	2	-	-	5	8	4	26	-	2	23	1	4	5	5	69	5	-	3	21	26	9	39	-	9	-	3	9	1	15	-	9	-	4	2	3	11	3	2	2	2	8	-	-	12	9	-	32	5	6	8	2	2	16	2	15	2	-	-	-	-	-	43	-	2	10	-	18	5	-	8	25	5	31	3	1	-	3	1	4	2	9	1	1	1	4	-	-	-	14	1	2	1	6	6	11	2	-	1	32	4	2	7	35	29	1	-	52	5	9	19	34	6	1	2	1	1	5	2	1	2	3	23	7	-	10	1	3	14	1	5	4	-	-	1	6	11	3	5	3	6	3	7	10	17	49	7	10	1	1	2	18	1	-	-	2	4	9	10	12	7	1	6	-	11	9	8	7	2	3	5	1	5	11	2	19	-	4	18	12	5	6	27	5	2	7	9	3	19	1	3	2	46	41	29	27	-	10	6	10	1	103	-	-	2	1	-	-	-	2	1	-	1	-	1	3	5	-	2	-	1	1	5	2	-	-	1	3	2	3	-	-	-	1	-	3	1	3	-	7	-	1	-	-	1	2	5	2	3	2	-	2	1	-	7	-	1	20	3	2	3	1	1	2	1	1	-	4	6	-	5	-	-	1	1	3	2	6	1	-	-	1	1	1	2	-	-	-	2	1	-	-	-	2	-	7	-	3	2	2	-	-	1	1	-	-	1	2	2	-	-	-	-	-	1	1	-	2	-	1	2	2	1	-	1	-	1	-	-	-	-	1	-	1	-	2	-	1	-	2	-	2	-	1	-	1	1	1	1	4	-	-	-	3	-	-	1	-	-	-	-	1974

2014	3	1	1	1	-	5	1	3	12	2	8	15	3	3	4	3	42	-	2	-	6	2	13	8	28	-	5	13	1	9	8	6	69	10	-	1	27	30	17	50	-	16	3	2	22	1	13	-	10	2	6	2	1	17	6	2	1	4	9	-	-	26	11	1	40	6	11	12	1	2	18	1	8	2	-	3	-	2	-	72	-	3	7	-	26	3	1	10	35	11	20	4	8	2	3	1	4	1	4	8	7	4	4	-	4	2	14	1	3	2	5	14	16	3	-	1	50	5	3	8	58	53	2	1	60	7	16	10	29	5	1	8	1	1	4	12	4	3	1	40	10	1	9	3	1	7	-	11	7	22	5	3	4	15	1	6	8	4	9	3	10	18	33	17	22	2	4	2	19	2	-	3	3	10	12	8	27	15	3	5	-	25	6	19	7	3	9	7	-	6	16	3	24	2	4	10	12	7	8	34	4	6	11	7	7	20	2	1	5	31	51	32	27	3	20	3	21	7	111	1	-	1	1	1	1	-	2	1	-	1	-	2	2	7	-	-	1	2	1	7	4	1	-	2	5	3	2	-	-	2	1	-	3	-	10	1	5	1	2	3	2	1	-	3	4	2	4	-	1	2	1	15	-	-	1	1	5	-	3	-	-	3	-	-	3	7	2	2	-	6	-	2	4	4	7	4	-	2	1	-	1	3	-	-	-	1	-	3	-	2	-	-	15	-	3	2	1	1	-	-	3	-	1	1	2	-	2	1	-	-	-	7	-	1	3	1	10	3	1	-	-	1	3	1	-	-	1	1	1	-	1	-	1	-	-	-	1	2	-	2	2	-	1	2	1	-	1	1	-	1	4	-	-	1	-	-	-	-	2538

2016	-	5	1	2	-	4	2	8	17	1	3	11	-	6	7	3	34	4	-	-	4	4	20	2	34	-	2	21	3	10	17	6	85	6	-	2	26	30	17	52	-	21	2	2	11	5	13	-	6	2	4	1	7	19	2	1	5	5	3	1	-	13	16	-	44	11	14	6	5	2	21	2	34	4	-	1	-	-	-	72	-	1	14	3	30	5	-	18	37	4	38	3	7	3	3	1	2	2	9	2	5	5	9	3	3	-	18	3	2	5	8	7	11	3	-	3	54	8	6	8	54	57	2	3	69	4	19	23	56	8	3	4	-	4	5	6	-	3	1	31	10	-	7	3	5	26	3	13	7	10	1	5	6	17	3	6	4	4	5	8	13	18	56	15	5	4	4	2	21	1	2	3	1	8	16	18	16	13	13	10	2	31	16	10	8	3	9	6	-	4	23	7	21	2	6	29	18	5	6	29	4	5	9	6	8	29	6	8	10	51	54	40	40	1	18	3	14	6	119	-	-	-	1	4	-	1	-	1	-	2	3	3	4	5	-	1	3	3	2	6	3	1	-	3	2	6	14	1	1	2	-	-	5	1	13	1	7	-	2	-	2	-	1	2	4	4	2	-	2	1	3	24	-	1	3	-	5	1	2	-	-	-	-	-	3	2	-	2	-	4	2	1	2	5	17	7	1	-	5	1	-	2	-	-	-	1	1	5	-	1	2	4	11	1	-	7	2	-	-	-	2	2	-	3	1	2	3	-	-	-	2	6	2	1	2	-	4	2	4	-	-	-	1	1	-	-	1	1	1	-	3	1	6	-	1	-	4	1	3	-	1	-	-	2	-	-	3	1	-	1	2	-	-	1	1	-	1	-	2867

2018	3	-	3	1	-	1	2	3	16	-	1	22	2	2	11	2	33	7	3	1	1	4	11	4	23	-	5	20	4	2	10	5	73	2	-	3	34	14	12	38	1	14	-	2	14	3	8	-	7	5	3	1	4	8	6	6	1	6	12	2	-	25	8	-	24	18	13	16	1	2	12	1	14	3	1	1	1	1	1	60	-	1	5	-	30	1	-	13	28	12	22	4	4	2	4	-	3	3	7	6	3	6	6	5	-	2	19	2	5	5	10	2	11	1	2	-	30	7	5	7	44	26	2	-	52	5	18	12	15	12	1	8	-	2	4	12	1	8	2	26	10	-	6	3	1	9	1	12	5	9	2	1	10	14	5	5	10	6	3	8	12	17	28	14	17	2	3	-	15	1	3	-	-	3	12	10	23	17	6	9	-	26	6	24	4	4	3	7	-	2	15	3	21	1	4	12	13	4	5	32	1	6	7	7	5	30	2	8	4	53	40	29	33	3	22	1	13	3	100	-	4	-	2	3	-	1	2	1	-	-	-	-	3	3	-	1	2	1	1	12	2	1	1	1	6	1	11	-	1	-	-	-	-	-	3	1	8	-	2	-	-	-	2	5	5	6	1	-	1	2	2	11	-	2	2	3	2	4	3	-	-	3	1	2	1	3	1	2	-	5	-	3	2	8	13	3	1	1	4	2	2	3	-	-	-	-	-	1	-	1	1	-	15	1	1	4	3	-	-	1	1	-	1	-	5	2	1	-	-	2	1	5	1	4	2	1	2	1	2	-	-	-	-	1	-	-	2	1	-	-	1	-	1	-	1	1	1	1	3	2	1	-	2	1	5	-	2	1	2	-	3	-	-	-	-	-	1	-	2348

2021	1	1	2	1	-	6	3	8	13	1	-	9	1	6	4	2	36	7	-	-	2	1	11	6	30	2	1	15	2	7	15	3	104	3	1	2	39	43	15	63	-	14	2	1	7	1	16	-	8	1	4	4	4	18	5	2	4	3	5	-	1	17	22	-	53	10	10	8	2	12	20	2	46	4	-	-	-	4	1	40	-	3	6	-	27	4	1	19	32	10	27	1	11	-	4	1	2	-	19	4	4	11	12	2	4	2	31	4	4	3	10	22	28	4	-	2	61	13	7	7	73	92	4	6	98	4	20	18	71	9	6	14	-	7	6	16	4	7	3	39	30	1	21	2	1	28	3	26	9	42	-	5	7	36	6	10	11	7	7	6	20	26	109	29	33	3	2	1	27	3	1	2	3	19	17	26	41	16	9	8	1	25	18	27	11	10	7	25	-	14	35	5	62	4	16	58	27	24	13	65	7	6	17	16	16	63	5	10	11	102	132	61	76	10	38	9	45	8	160	-	1	-	1	2	-	1	-	1	1	3	2	1	4	3	2	1	-	-	-	8	-	-	1	6	6	3	8	-	-	2	3	-	2	2	7	3	10	-	1	1	-	1	5	7	14	3	2	1	4	-	3	28	-	2	-	8	18	4	6	4	-	1	-	1	2	2	3	4	2	3	3	3	4	7	10	21	2	3	2	6	-	2	2	-	-	1	-	1	-	4	-	-	12	3	3	3	2	-	1	-	2	-	-	2	3	2	-	1	1	-	-	5	6	3	5	-	3	6	5	2	-	1	1	1	-	1	4	1	-	-	-	-	11	-	1	1	3	-	-	4	3	6	3	3	5	1	2	-	2	4	9	-	-	1	1	-	-	-	4032

2022	-	-	2	4	-	4	1	5	11	1	5	18	-	6	9	4	35	12	-	1	2	4	23	7	53	5	3	23	4	10	14	7	80	9	-	2	42	37	27	68	-	33	-	1	9	-	25	-	14	3	4	3	3	27	6	3	3	3	9	1	-	21	17	-	52	19	15	10	2	2	27	1	45	3	3	-	-	3	-	61	1	-	14	-	24	2	-	22	34	8	28	2	23	5	5	2	-	1	18	7	2	6	12	2	5	-	23	1	4	2	3	10	19	4	-	3	65	7	4	6	64	62	1	5	71	8	21	9	69	6	3	6	-	-	4	13	3	11	1	19	23	-	9	5	8	17	3	24	7	21	-	10	7	24	4	9	16	5	2	5	18	29	73	30	18	1	9	2	21	1	-	4	3	11	16	26	39	21	7	10	2	27	12	20	6	5	11	7	-	9	24	3	37	4	10	34	19	14	8	39	2	8	8	10	14	54	4	7	3	91	104	38	34	4	26	6	36	4	218	-	1	1	-	1	-	2	2	-	1	2	-	-	7	-	1	1	1	2	-	8	1	-	1	5	1	1	11	-	-	1	-	-	3	-	4	2	14	-	-	-	-	5	7	3	5	3	3	-	3	1	1	14	-	5	-	2	11	4	4	-	1	1	3	-	4	2	-	6	-	4	1	3	2	4	9	13	-	-	5	3	-	8	-	1	-	-	-	2	1	1	1	1	11	2	4	8	1	1	-	-	-	-	1	1	2	3	2	2	-	-	-	4	2	3	4	1	4	3	3	-	-	-	4	1	-	-	-	1	1	-	4	-	4	-	2	-	1	2	-	2	2	1	3	3	3	-	4	-	-	1	-	1	3	1	-	1	-	-	3544

2024	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3309	3309

Total	53	92	56	49	80	79	77	149	424	63	153	893	46	150	166	143	1262	109	25	50	48	73	411	169	580	49	94	340	48	198	465	190	1922	202	15	65	730	715	482	1324	28	694	55	52	218	61	244	41	289	87	205	96	97	435	143	84	180	103	253	63	27	542	336	18	1062	368	243	252	132	182	510	77	314	71	13	20	7	69	19	1385	12	83	569	30	835	100	34	387	1077	616	753	81	167	113	104	54	44	36	271	182	87	120	172	69	160	39	432	49	124	49	220	220	466	62	14	30	1315	152	234	200	1516	1692	275	103	2751	109	375	363	619	218	46	118	8	134	147	275	67	118	58	811	365	9	267	82	66	364	56	368	162	264	67	72	134	588	105	122	129	92	81	122	302	365	1147	339	397	44	120	33	353	99	26	28	44	219	211	347	430	277	81	174	17	455	204	379	130	111	115	218	8	171	391	78	519	37	213	516	258	130	163	727	81	85	240	169	166	471	78	114	123	1142	1475	931	768	69	440	62	383	78	5143	19	18	42	35	22	26	44	71	64	19	42	31	23	69	96	12	17	28	61	17	123	69	21	6	54	127	106	125	10	14	36	27	7	68	44	122	22	141	31	27	26	32	27	59	84	106	43	108	18	44	28	29	255	2	31	70	40	93	59	49	16	21	29	23	18	71	60	30	43	6	56	33	42	72	111	174	68	17	24	41	53	18	54	8	8	8	18	11	56	5	24	29	21	188	24	81	73	28	12	12	11	24	8	23	24	42	32	15	17	14	12	11	72	33	23	67	12	47	35	56	14	11	14	26	12	9	14	33	16	18	10	23	13	49	9	16	5	37	22	16	16	26	22	29	31	33	9	47	10	16	16	42	3	11	8	8	2	3	3309	75699

See also