Person #2 is r's sibling

sibling2

Source

General Social Survey https://gss.norc.org

Details

Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer. When I read you a name, please tell me all the ways that person is connected to you. How is (NAME) connected to you? PROBE: What other ways? ASK FOR EACH NAME. C2. Second Person is Sibling

Values

1 mentioned

2 not mentioned

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	2004

Full	1985, 1987

Overview

For further details see the GSS Data Explorer page for sibling2.

Counts by year:

year	iap	mentioned	not mentioned	no answer	Total

1985	364	113	1054	3	1534

1987	385	121	1288	25	1819

2004	1977	76	712	47	2812

Total	2726	310	3054	75	6165

See also