Number of brothers and sisters

sibs

Source

General Social Survey https://gss.norc.org

Details

Question How many brothers and sisters did you have? Please count those born alive, but no longer living, as well as those alive now. Also include stepbrothers and stepsisters, and children adopted by your parents.

Values

[6] 6 or more

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

Full	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for sibs.

Counts by year:

year	0	1	2	3	4	5	6	7	don't know	no answer	10	11	12	13	14	15	16	18	20	27	8	9	17	19	21	31	22	26	23	24	68	25	35	30	32	33	29	34	iap	37	55	28	43	skipped on web	36	Total

1972	103	249	249	217	169	130	107	382	1	6	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	86	221	231	197	159	129	112	81	3	-	36	44	25	14	10	7	2	2	1	1	65	78	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	76	218	214	224	181	131	94	92	1	-	39	21	27	14	5	6	4	1	1	1	76	55	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	95	206	243	212	145	125	110	93	2	3	40	30	18	15	7	4	1	2	-	-	75	60	2	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	89	194	222	222	175	153	98	98	-	1	47	30	22	10	5	8	3	-	-	-	64	54	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	80	194	249	188	183	138	124	105	1	2	40	29	21	15	9	2	2	2	1	-	75	67	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	98	216	268	210	168	133	131	91	-	-	37	22	23	14	8	4	6	-	1	-	58	40	1	1	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	73	212	230	195	189	147	107	84	1	-	32	27	18	16	7	5	3	1	3	1	70	45	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	103	240	269	273	211	179	140	104	2	3	48	47	29	18	13	7	-	6	2	-	92	69	1	1	1	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	98	237	274	253	185	154	112	80	1	-	22	27	18	9	4	3	4	-	3	-	68	43	2	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	64	222	266	234	158	122	90	92	-	8	36	28	14	17	5	2	1	2	1	-	63	46	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	84	233	240	215	177	149	120	92	4	4	21	20	13	7	6	4	9	1	1	-	73	58	1	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	60	236	268	228	166	146	98	64	-	3	29	23	15	15	3	3	3	-	-	-	70	38	-	1	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	89	261	273	292	217	157	130	80	3	6	65	32	29	17	9	2	4	-	2	1	76	69	2	1	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	63	226	292	216	170	135	85	86	5	4	34	23	10	9	6	4	4	2	-	-	61	43	-	2	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	77	237	311	264	167	130	102	67	-	4	30	16	16	12	4	3	2	-	-	-	53	40	1	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	64	224	261	243	164	119	88	60	-	1	23	22	10	4	5	1	-	-	-	-	49	31	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	74	236	276	236	209	118	80	81	4	8	34	22	11	9	5	3	1	1	-	-	58	47	2	-	1	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	89	260	341	282	170	117	93	67	3	3	23	15	11	11	3	4	2	-	-	-	60	47	1	2	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	162	537	581	483	354	232	185	131	5	6	30	33	28	20	16	8	5	1	2	2	89	76	1	1	-	-	-	-	2	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	140	501	570	497	350	228	154	136	2	5	47	50	20	17	5	8	7	1	-	-	84	53	1	2	13	5	1	-	2	-	-	2	-	1	1	1	-	-	-	-	-	-	-	-	-	2904

1998	105	493	558	483	350	207	157	118	5	12	43	47	26	10	9	5	6	4	2	-	84	68	2	1	25	-	4	-	5	1	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	2832

2000	142	507	569	475	301	237	162	123	5	6	40	40	19	20	10	7	3	-	2	-	90	51	3	1	-	-	3	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	140	505	583	457	314	224	157	115	2	7	40	28	22	7	8	11	6	2	2	-	77	49	2	3	1	-	-	1	1	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	144	553	556	472	344	198	157	108	2	2	47	22	18	15	11	7	6	3	1	-	80	63	1	1	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	2812

2006	145	560	605	439	345	234	189	141	1	3	47	39	30	19	10	4	5	4	1	-	81	81	3	1	-	-	-	-	1	-	-	-	-	-	1	1	1	1	1518	-	-	-	-	-	-	4510

2008	89	394	400	315	250	160	117	100	-	2	30	19	14	8	6	3	3	4	1	-	75	27	3	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	2023

2010	94	401	412	320	239	167	122	92	1	5	28	26	17	7	8	5	4	-	-	-	62	32	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	96	368	410	331	208	147	113	91	2	1	35	18	12	7	5	10	3	1	1	-	64	45	1	-	1	-	2	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	1974

2014	95	461	496	436	287	214	154	114	1	6	43	26	14	14	11	4	6	1	1	-	83	61	6	1	1	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2016	130	550	596	452	319	206	161	128	2	3	39	39	27	23	7	2	3	3	2	1	90	72	4	-	3	-	2	-	1	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	2867

2018	95	472	483	384	244	176	147	125	4	1	34	19	13	14	5	2	4	-	-	-	76	44	1	1	1	-	-	-	1	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	2348

2021	216	886	935	679	406	287	186	116	49	-	29	33	25	12	9	6	1	1	-	-	81	55	-	1	-	-	-	-	-	-	-	1	1	-	1	-	-	-	4	-	-	1	-	11	-	4032

2022	174	743	790	565	400	243	175	110	-	12	49	33	27	15	8	7	6	2	1	-	90	65	5	1	3	-	2	1	-	-	-	1	-	-	1	1	-	-	-	-	-	-	-	13	1	3544

Total	3532	12253	13521	11189	8074	5772	4357	3647	112	127	1217	950	642	434	242	161	119	47	32	7	2412	1772	58	28	53	6	19	7	16	6	1	9	2	4	4	3	2	1	1522	1	1	2	1	24	1	72390

See also