Help if sick #1

sick1

Source

General Social Survey https://gss.norc.org

Details

Question Suppose you had the flu and you had to stay in bed for a few days and needed help around the home, with shopping and such. A. Who would you turn to first for help? B. Who would you turn to second? Please CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER. A. First Choice

Values

[1] spouse

[2] mother

[3] father

[4] daughter

[5] son

[6] sister

[7] brother

[8] other relative

[9] closest friend

[10] other friend

[11] neighbor

[12] coworker

[13] nurse

[14] clergy

[15] paid help

[16] other

[17] no one

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-	1986

Overview

For further details see the GSS Data Explorer page for sick1.

Counts by year:

year	iap	spouse	mother	father	daughter	son	sister	brother	other relative	closest friend	other friend	neighbor	coworker	nurse	clergy	paid help	other	no one	no answer	Total

1986	42	737	142	10	128	54	62	23	37	122	22	34	2	3	2	11	7	12	20	1470

Total	42	737	142	10	128	54	62	23	37	122	22	34	2	3	2	11	7	12	20	1470

See also