Spend evening at bar

socbar

Source

General Social Survey https://gss.norc.org

Details

Question Would you use this card and tell me which answer comes closest to how often you do the following things… D. Go to a bar or tavern?

Values

[1] almost daily

[2] once or twice a week

[3] several times a month

[4] about once a month

[5] several times a year

[6] about once a year

[7] never

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

Overview

For further details see the GSS Data Explorer page for socbar.

Counts by year:

year	iap	almost daily	once or twice a week	several times a month	about once a month	several times a year	about once a year	never	don't know	no answer	skipped on web	Total

1974	-	33	139	92	131	178	112	777	18	4	-	1484

1975	-	22	114	98	136	134	140	832	10	4	-	1490

1977	-	39	139	124	152	164	144	763	1	4	-	1530

1978	-	29	133	140	112	186	153	775	1	3	-	1532

1982	-	30	178	122	162	224	193	935	9	7	-	1860

1983	-	26	155	139	130	182	161	800	2	4	-	1599

1985	-	28	119	127	169	175	164	747	1	4	-	1534

1986	-	16	118	88	142	172	145	782	2	5	-	1470

1988	493	10	82	65	97	105	123	501	1	4	-	1481

1989	531	10	82	56	85	103	109	559	-	2	-	1537

1990	444	6	74	59	94	106	106	479	1	3	-	1372

1991	493	18	70	62	109	101	119	539	2	4	-	1517

1993	526	7	85	78	110	120	120	556	1	3	-	1606

1994	1015	20	142	145	180	261	243	971	5	10	-	2992

1996	944	25	144	136	198	250	250	948	8	1	-	2904

1998	961	26	125	124	199	232	239	916	3	7	-	2832

2000	940	20	142	140	180	223	233	922	11	6	-	2817

2002	1857	11	78	74	114	108	119	401	2	1	-	2765

2004	1906	6	65	61	91	108	142	432	1	-	-	2812

2006	2518	14	113	122	200	254	259	1027	1	2	-	4510

2008	694	12	95	86	151	177	158	649	1	-	-	2023

2010	614	14	86	135	159	192	172	668	3	1	-	2044

2012	672	11	75	107	145	146	167	649	1	1	-	1974

2014	863	11	94	128	209	243	233	757	-	-	-	2538

2016	979	16	118	154	221	254	271	852	1	1	-	2867

2018	789	10	93	131	195	223	216	689	1	1	-	2348

2021	1320	5	91	190	271	513	461	1171	-	1	9	4032

2022	1168	11	105	199	272	406	361	1014	3	1	4	3544

Total	19727	486	3054	3182	4414	5540	5313	21111	90	84	13	63014

See also