Spend evening with friends

socfrend

Source

General Social Survey https://gss.norc.org

Details

Question Would you use this card and tell me which answer comes closest to how often you do the following things… C. Spend a social evening with friends who live outside the neighborhood?

Values

[1] almost daily

[2] once or twice a week

[3] several times a month

[4] about once a month

[5] several times a year

[6] about once a year

[7] never

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

Overview

For further details see the GSS Data Explorer page for socfrend.

Counts by year:

year	iap	almost daily	once or twice a week	several times a month	about once a month	several times a year	about once a year	never	don't know	no answer	skipped on web	Total

1974	-	35	290	271	327	270	121	164	5	1	-	1484

1975	-	51	266	259	337	272	114	186	3	2	-	1490

1977	-	43	287	301	338	291	110	153	3	4	-	1530

1978	-	33	289	319	252	324	130	179	3	3	-	1532

1982	-	63	341	378	389	361	143	173	4	8	-	1860

1983	-	41	301	339	354	292	98	169	2	3	-	1599

1985	-	44	291	304	320	281	139	150	1	4	-	1534

1986	-	24	299	269	331	292	79	171	1	4	-	1470

1988	493	29	174	205	232	184	62	100	-	2	-	1481

1989	531	40	184	180	229	181	73	115	2	2	-	1537

1990	444	23	160	193	219	175	58	97	1	2	-	1372

1991	493	38	214	187	242	154	74	110	1	4	-	1517

1993	526	35	227	220	214	218	65	97	1	3	-	1606

1994	1015	48	391	408	445	354	139	181	1	10	-	2992

1996	944	71	410	400	427	328	142	177	4	1	-	2904

1998	961	64	347	399	383	379	126	158	10	5	-	2832

2000	940	69	357	379	401	338	137	181	7	8	-	2817

2002	1857	45	178	221	196	139	60	67	2	-	-	2765

2004	1906	31	162	179	216	177	54	85	2	-	-	2812

2006	2518	64	390	370	462	343	155	205	1	2	-	4510

2008	694	59	245	263	277	235	101	145	4	-	-	2023

2010	614	41	289	311	308	251	90	135	3	2	-	2044

2012	672	48	246	255	280	219	110	141	2	1	-	1974

2014	863	52	278	367	385	300	140	153	-	-	-	2538

2016	979	70	276	407	448	350	168	167	1	1	-	2867

2018	789	32	244	338	378	277	131	157	1	1	-	2348

2021	1318	34	253	467	554	777	273	343	-	1	12	4032

2022	1168	68	343	447	494	488	221	305	2	-	8	3544

Total	19725	1295	7732	8636	9438	8250	3313	4464	67	74	20	63014

See also