Spend evening with neighbor

socommun

Source

General Social Survey https://gss.norc.org

Details

Question Would you use this card and tell me which answer comes closest to how often you do the following things… B. Spend a social evening with someone who lives in your neighborhood?

Values

[1] almost daily

[2] once or twice a week

[3] several times a month

[4] about once a month

[5] several times a year

[6] about once a year

[7] never

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

Overview

For further details see the GSS Data Explorer page for socommun.

Counts by year:

year	iap	almost daily	once or twice a week	several times a month	about once a month	several times a year	about once a year	never	don't know	no answer	skipped on web	Total

1974	-	110	347	189	249	168	91	322	6	2	-	1484

1975	-	89	334	196	218	181	117	350	3	2	-	1490

1977	-	86	331	184	239	191	123	370	2	4	-	1530

1978	-	92	359	167	189	206	130	379	6	4	-	1532

1982	-	122	347	234	305	254	182	407	1	8	-	1860

1983	-	91	316	194	230	198	148	415	3	4	-	1599

1985	-	85	276	200	211	199	161	395	3	4	-	1534

1986	-	104	308	165	215	161	136	378	-	3	-	1470

1988	493	46	203	110	149	128	69	279	2	2	-	1481

1989	531	51	170	96	149	122	106	305	5	2	-	1537

1990	444	52	165	105	134	103	88	274	5	2	-	1372

1991	493	55	191	119	140	125	78	311	2	3	-	1517

1993	526	49	181	130	156	138	102	320	2	2	-	1606

1994	1015	103	315	227	278	248	188	606	3	9	-	2992

1996	944	101	330	223	307	248	182	561	6	2	-	2904

1998	961	98	299	196	292	261	193	523	6	3	-	2832

2000	940	97	313	206	255	220	197	573	9	7	-	2817

2002	1857	82	133	113	108	103	94	272	3	-	-	2765

2004	1906	59	138	109	140	125	81	252	2	-	-	2812

2006	2518	95	326	208	288	275	210	588	-	2	-	4510

2008	694	83	209	140	183	171	143	398	2	-	-	2023

2010	614	97	254	151	198	188	138	400	3	1	-	2044

2012	672	67	191	148	202	135	127	428	3	1	-	1974

2014	863	82	272	189	243	204	163	522	-	-	-	2538

2016	979	102	281	216	295	217	164	611	1	1	-	2867

2018	789	70	219	174	243	196	145	510	-	2	-	2348

2021	1318	39	235	272	322	438	323	1076	-	-	9	4032

2022	1168	83	320	264	303	312	226	864	2	1	1	3544

Total	19725	2290	7363	4925	6241	5515	4105	12689	80	71	10	63014

See also