Spend evening with parents

socpars

Source

General Social Survey https://gss.norc.org

Details

Question Would you use this card and tell me which answer comes closest to how often you do the following things… E. Spend a social evening with your parents?

Values

[0] no such people

[1] almost daily

[2] sev times a week

[3] sev times a mnth

[4] once a month

[5] sev times a year

[6] once a year

[7] never

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1978, 1982, 1983, 1985, 1986

A/B	1988, 1989, 1990, 1991, 1993, 1994

Overview

For further details see the GSS Data Explorer page for socpars.

Counts by year:

year	iap	no such people	almost daily	sev times a week	sev times a mnth	once a month	sev times a year	once a year	never	don't know	no answer	Total

1978	-	551	49	216	166	113	204	118	109	2	4	1532

1982	-	545	155	267	161	189	203	134	181	11	14	1860

1983	-	494	89	248	157	161	203	129	113	1	4	1599

1985	-	470	92	173	199	160	190	132	111	2	5	1534

1986	-	443	100	220	113	150	201	128	106	2	7	1470

1988	493	244	69	133	92	100	134	78	124	7	7	1481

1989	531	236	66	160	92	99	145	80	124	2	2	1537

1990	444	262	53	134	103	88	135	78	68	2	5	1372

1991	493	259	66	154	113	102	133	68	122	4	3	1517

1993	526	283	73	154	107	121	148	92	95	2	5	1606

1994	2485	127	27	66	46	47	65	51	75	2	1	2992

Total	4972	3914	839	1925	1349	1330	1761	1088	1228	37	57	18500

See also