Favor spanking to discipline child

spanking

Source

General Social Survey https://gss.norc.org

Details

Question Do you strongly agree, agree, disagree, or strongly disagree that it is sometimes necessary to discipline a child with a good, hard spanking?

Values

1 strongly agree

2 agree

3 disagree

4 strongly disagree

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1986	-/-	full

1988	A/B	partial

1989	A/B	partial

1990	A/B	partial

1991	A/B	partial

1993	A/B	partial

1994	A/B	partial

1996	A/B	partial

1998	A/B	partial

2000	A/B	partial

2002	A/B	partial

2004	A/B	partial

2006	A/B	partial

2008	A/B	partial

2010	A/B	partial

2012	A/B	partial

2014	A/B	partial

2016	A/B	partial

2018	A/B	partial

2021	A/B	partial

2022	A/B	partial

2024	A/B	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	agree	disagree	don't know	no answer	strongly agree	strongly disagree	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	1534

1986	-	816	195	9	1	404	45	-	1470

1987	1819	-	-	-	-	-	-	-	1819

1988	493	476	145	8	2	302	55	-	1481

1989	531	455	169	9	3	315	55	-	1537

1990	444	466	154	8	3	261	36	-	1372

1991	493	485	194	15	2	263	65	-	1517

1993	526	538	210	15	4	243	70	-	1606

1994	1015	909	357	20	7	522	162	-	2992

1996	944	890	357	33	4	512	164	-	2904

1998	961	880	336	34	2	486	133	-	2832

2000	940	780	328	33	4	588	144	-	2817

2002	1857	387	176	6	3	272	64	-	2765

2004	1906	433	182	8	1	222	60	-	2812

2006	2518	893	402	20	2	532	143	-	4510

2008	694	608	280	18	3	342	78	-	2023

2010	614	647	324	10	3	336	110	-	2044

2012	672	598	277	21	4	302	100	-	1974

2014	863	777	365	12	4	378	139	-	2538

2016	979	817	440	15	7	457	152	-	2867

2018	789	700	368	23	7	340	121	-	2348

2021	1317	977	827	4	1	424	456	26	4032

2022	1168	887	634	23	2	434	389	7	3544

2024	1126	804	587	29	9	381	349	24	3309

Total	41255	15223	7307	373	78	8316	3090	57	75699

See also