Allow anti-religionist to speak

spkath

Source

General Social Survey https://gss.norc.org

Details

Question There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against all churches and religion . . . A. If such a person wanted to make a speech in your (city/town/community) against churches and religion, should he be allowed to speak, or not?

Values

1 yes, allowed to speak

2 not allowed

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1972	-/-	full

1973	-/-	full

1974	-/-	full

1976	-/-	full

1977	-/-	full

1980	-/-	full

1982	-/-	full

1984	-/-	full

1985	-/-	full

1987	-/-	full

1988	A/C	partial

1989	A/C	partial

1990	A/C	partial

1991	A/C	partial

1993	A/C	partial

1994	A/C	partial

1996	A/C	partial

1998	A/C	partial

2000	A/C	partial

2002	A/C	partial

2004	A/C	partial

2006	A/C	partial

2008	A/C	partial

2010	A/C	partial

2012	A/C	partial

2014	A/C	partial

2016	A/C	partial

2018	A/C	partial

2021	A/C	partial

2022	A/C	partial

2024	A/C	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	don't know	not allowed	yes, allowed to speak	no answer	iap	skipped on web	Total

1972	37	527	1049	-	-	-	1613

1973	10	511	982	1	-	-	1504

1974	20	547	915	2	-	-	1484

1975	-	-	-	-	1490	-	1490

1976	16	524	959	-	-	-	1499

1977	9	566	952	3	-	-	1530

1978	-	-	-	-	1532	-	1532

1980	12	485	971	-	-	-	1468

1982	35	684	1140	1	-	-	1860

1983	-	-	-	-	1599	-	1599

1984	8	459	1002	4	-	-	1473

1985	22	519	991	2	-	-	1534

1986	-	-	-	-	1470	-	1470

1987	13	564	1236	6	-	-	1819

1988	15	280	681	1	504	-	1481

1989	11	280	740	2	504	-	1537

1990	15	236	663	3	455	-	1372

1991	11	266	713	3	524	-	1517

1993	19	290	764	2	531	-	1606

1994	21	520	1451	4	996	-	2992

1996	27	483	1409	4	981	-	2904

1998	29	457	1393	3	950	-	2832

2000	25	450	1381	5	956	-	2817

2002	8	206	710	-	1841	-	2765

2004	7	200	689	2	1914	-	2812

2006	12	438	1553	-	2507	-	4510

2008	11	315	1021	5	671	-	2023

2010	11	312	956	2	763	-	2044

2012	15	281	1011	1	666	-	1974

2014	12	351	1345	6	824	-	2538

2016	11	378	1493	8	977	-	2867

2018	8	309	1253	4	774	-	2348

2021	1	244	1068	-	2714	5	4032

2022	11	238	909	1	2380	5	3544

2024	9	234	899	-	2163	4	3309

Total	471	12154	32299	75	30686	14	75699

See also