Percentage of some college educ in spocc10 based on acs 2010

spsei10educ

Source

General Social Survey https://gss.norc.org

Details

Question Percentage of those who had some college or more education among those who had the same occupation as in SPOCC10 in American Community Survey between 2010 and 2012 (25-64 years old only). This percentage is not from the GSS but it is used to calculate SPSEI10.

Values

NA(d) don't know / NA(i) iap / NA(j) i don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1972	-/-/-/-	full

1973	-/-/-/-	full

1974	-/-/-/-	full

1975	-/-/-/-	full

1976	-/-/-/-	full

1977	-/-/-/-	full

1978	-/-/-/-	full

1980	-/-/-/-	full

1982	-/-/-/-	full

1983	-/-/-/-	full

1984	-/-/-/-	full

1985	-/-/-/-	full

1986	-/-/-/-	full

1987	-/-/-/-	full

1988	A/B/C/-	full

1989	A/B/C/-	full

1990	A/B/C/-	full

1991	A/B/C/-	full

1993	A/B/C/-	full

1994	A/B/C/-	full

1996	A/B/C/-	full

1998	A/B/C/-	full

2000	A/B/C/-	full

2002	A/B/C/-	full

2004	A/B/C/-	full

2006	A/B/C/D	full

2008	A/B/C/-	full

2010	A/B/C/-	full

2012	A/B/C/-	full

2014	A/B/C/-	full

2016	A/B/C/-	full

2018	A/B/C/-	full

2021	A/B/C/-	full

2022	A/B/C/-	full

2024	A/B/C/-	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	11.9	12.2	12.6	13.1	13.3	13.5	14.1	14.2	15.5	16	16.5	16.7	16.8	16.9	17.2	17.3	17.5	17.7	17.9	18	18.6	18.7	19.2	19.3	19.8	20.4	20.6	21	21.6	21.7	22	22.3	22.5	22.7	22.9	23.6	23.9	24.1	24.3	25.1	25.4	25.6	25.7	26.2	26.6	27.1	27.5	27.8	28	28.2	28.3	29.1	29.5	29.7	30.7	31	31.2	31.9	32.1	32.4	32.5	33.7	33.9	34.2	34.3	34.8	35	35.1	36	36.2	36.5	37	37.4	37.7	37.8	38.5	39.2	39.6	40.2	40.4	40.8	40.9	41.6	41.8	42	42.3	42.8	43.7	43.9	44	44.2	45.2	45.4	45.5	46.1	46.3	46.4	47.9	48.5	48.8	49	49.3	50	50.1	50.6	50.7	51.1	51.7	52.1	53	53.9	54.2	54.6	54.8	55.2	55.4	55.5	55.6	55.7	55.8	55.9	56.1	56.7	57.1	57.6	57.7	58.5	58.7	58.8	58.9	59.2	62.4	62.7	63.1	63.9	64.1	64.6	65.1	65.3	65.5	65.6	66.8	66.9	67.3	68.6	69.1	69.8	69.9	70.1	71.6	71.7	72.2	73	73.3	73.9	75.5	75.6	75.9	76.9	77.2	78	79.2	79.5	79.9	80.1	81.4	82.3	82.4	82.5	82.8	83	85.7	86.3	86.7	87	87.3	90.2	90.5	90.6	91.9	92	92.1	93.1	93.3	93.7	94	94.1	94.4	94.8	95.1	95.5	97.3	97.5	97.8	98.9	99	99.1	99.2	99.4	99.5	99.6	99.7	iap	no answer	12.3	17	19.7	20.5	20.8	21.1	21.5	21.9	23.2	24	26	26.7	29.4	30.4	30.6	32.8	33.8	35.5	35.9	38.3	38.4	40.3	42.6	44.4	46.6	47.3	48.9	54.7	61	63.3	66	68.4	69.7	70.4	74.2	75.1	76.5	78.2	78.3	78.4	81	81.1	83.3	83.7	85.9	86.8	88.3	89.4	89.8	91.4	91.7	95	95.3	95.6	96.8	don't know	12.4	14.3	19	19.6	24.4	28.7	35.8	46.9	47.4	48.2	61.1	67.4	72.5	74.4	77.5	79.6	81.9	83.1	83.5	90.9	91.2	91.3	92.4	95.7	98.3	28.9	31.1	38.6	47.5	48.1	51	58.6	59.3	64.7	67.6	70.3	79.4	81.3	85.6	88.2	17.8	27	34.1	35.3	42.1	45.1	45.7	53.3	54.3	60.9	62.2	66.3	85.4	96.4	98.5	22.4	39.9	52.2	61.3	68.8	78.1	78.8	92.9	99.3	19.1	24.7	33.5	35.2	69.6	73.5	85.3	90.8	96	25.8	70.9	84.5	91.1	92.6	96.5	98.7	13.4	23	47.8	48.3	60.7	62.8	68.2	79.8	96.3	19.9	49.7	56.6	80.7	84.4	74.6	84	84.6	97	52.7	66.7	63.7	80	62.3	63.5	55.1	84.8	34.9	34.7	9.8	58.2	86.2	10.9	87.5	62.9	84.7	not available in this release	Total

1972	1	2	2	2	1	3	11	1	1	6	19	2	1	3	3	25	1	1	1	2	1	1	7	2	1	7	1	4	2	7	3	33	1	4	1	1	2	14	10	6	25	2	12	1	2	3	2	5	2	1	2	2	7	1	4	3	1	1	1	2	1	5	7	1	1	1	7	2	3	1	3	1	2	17	2	4	1	4	1	1	9	2	1	1	14	12	5	1	5	16	1	32	8	2	2	2	3	2	3	2	1	1	2	4	1	8	5	3	7	5	1	10	1	1	1	12	1	1	3	3	19	40	8	3	70	2	1	1	4	2	3	4	1	1	4	3	4	1	5	1	2	1	1	1	6	7	1	3	7	1	2	2	6	5	1	1	1	6	8	6	5	1	2	1	1	2	2	2	2	4	3	4	5	1	1	1	1	2	4	1	1	1	1	2	9	1	1	4	2	2	1	1	3	11	28	9	8	1	1	2	1	3	695	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	3	2	-	1	1	1	7	-	1	4	29	1	1	4	1	13	1	2	-	1	-	-	6	2	-	4	-	1	2	13	3	21	1	3	2	-	1	10	11	9	25	2	12	-	-	2	-	5	4	1	-	-	9	2	2	4	3	4	-	-	-	3	7	-	1	2	12	10	2	3	2	8	5	7	-	1	-	1	-	1	12	-	4	2	12	12	1	-	5	13	-	13	1	-	5	1	3	-	3	3	2	1	1	1	1	6	2	2	2	6	2	9	1	-	-	15	2	-	3	1	23	23	6	1	45	2	4	-	5	-	4	3	1	1	4	-	3	3	10	-	5	1	5	1	2	6	-	4	4	-	2	2	9	-	1	2	-	9	9	3	2	4	4	2	-	1	-	3	3	1	4	9	2	-	-	1	-	1	2	1	5	1	2	-	6	-	-	4	-	4	-	2	2	13	25	14	8	-	7	-	-	7	619	-	1	3	1	1	2	1	2	3	1	2	1	1	1	1	3	1	1	1	1	1	1	1	1	1	1	2	3	1	2	2	1	2	4	2	1	1	5	1	2	2	1	1	1	2	2	2	2	2	1	5	1	1	2	3	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1	1	3	1	1	-	6	-	-	2	26	-	-	3	4	12	1	-	2	-	-	1	10	6	-	8	1	3	2	14	6	19	-	4	1	-	-	10	7	6	17	1	16	-	1	-	1	6	2	4	-	-	6	2	2	3	2	2	2	1	-	10	12	-	-	-	14	10	3	-	3	1	1	5	3	1	-	1	1	-	7	-	-	-	21	15	-	1	7	14	-	12	6	1	4	-	5	-	2	3	-	1	1	5	-	3	7	-	4	3	1	5	-	-	1	17	-	-	6	4	29	30	6	1	51	1	1	-	2	-	2	4	-	2	2	-	2	4	10	-	6	1	-	-	3	8	-	4	4	-	-	2	13	2	1	2	1	4	11	9	3	7	3	-	-	2	-	1	1	3	3	2	3	1	-	-	-	2	-	4	3	1	5	-	6	-	-	4	1	3	-	-	3	19	25	16	3	1	4	-	-	5	607	-	-	6	-	-	-	1	-	-	-	1	-	1	-	-	3	-	-	-	-	-	-	1	-	-	2	3	-	1	1	-	-	-	1	2	1	1	1	-	1	-	-	-	1	1	1	-	2	3	-	3	-	-	1	-	-	-	2	1	1	3	1	1	1	1	1	2	1	5	1	3	1	1	1	1	1	2	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1	1	1	-	3	-	9	1	1	4	15	-	3	8	1	11	1	-	2	-	1	-	9	2	-	8	-	2	4	9	-	19	-	3	8	-	-	10	7	8	23	1	23	4	3	3	1	6	1	4	1	1	9	1	3	4	1	1	2	1	-	8	6	-	-	1	10	3	-	-	7	-	4	5	3	2	1	1	-	2	15	-	4	3	11	7	3	1	5	10	-	20	4	1	3	1	4	-	3	4	-	-	2	2	-	7	5	2	2	1	-	6	2	1	-	18	-	-	4	5	15	33	6	3	41	4	2	-	2	-	3	3	1	2	6	2	-	-	7	-	2	1	-	1	2	5	-	3	7	1	-	-	5	1	1	-	1	9	5	3	7	1	1	-	-	2	-	3	2	5	3	4	1	-	2	2	1	1	1	2	2	-	1	-	8	3	1	2	1	3	-	-	-	12	24	15	6	-	2	2	-	8	653	-	1	4	-	1	-	2	1	-	1	-	-	-	-	-	3	-	-	-	-	1	1	1	-	-	3	2	1	-	-	-	1	2	2	-	-	-	2	-	-	-	1	-	2	3	2	2	1	1	-	2	1	1	-	4	-	1	2	-	-	1	-	-	-	1	2	-	1	1	-	-	-	-	1	1	-	-	-	-	-	1	-	2	1	1	1	1	1	1	1	2	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1	4	-	1	3	5	10	-	-	2	19	-	1	6	3	9	2	-	3	2	-	3	5	1	-	3	6	5	3	8	3	31	-	2	4	-	-	4	10	10	20	-	18	1	1	-	2	3	2	1	1	-	9	3	3	2	-	3	-	2	-	10	12	-	-	-	9	4	3	-	2	3	3	6	2	3	-	-	-	-	11	-	1	-	18	18	-	1	6	12	1	27	6	2	4	2	4	-	7	1	1	2	3	1	1	3	4	-	1	3	4	6	-	1	-	24	2	-	2	1	18	30	6	2	45	2	1	1	2	-	2	-	-	2	1	-	1	-	8	1	4	-	-	1	3	2	-	-	9	1	1	4	3	-	-	2	-	5	12	1	6	1	1	-	-	1	1	2	1	4	2	3	3	-	-	1	1	2	2	1	7	2	4	-	10	1	1	5	2	2	-	1	4	10	17	13	8	-	5	-	-	2	646	-	-	5	-	1	1	-	1	-	-	2	1	-	-	-	4	-	-	1	-	1	1	2	1	-	-	2	-	-	-	-	-	1	1	2	1	1	1	1	-	-	1	-	-	-	2	1	2	4	-	-	1	1	-	-	-	-	-	-	-	1	1	-	1	-	1	-	1	2	-	2	-	-	-	-	1	1	-	-	-	1	1	-	1	-	-	-	1	-	-	-	1	-	1	-	-	-	2	4	1	1	1	1	1	1	2	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1	4	3	1	-	5	5	-	2	2	18	-	-	2	1	17	1	-	-	-	1	3	12	3	-	5	1	3	1	14	3	19	-	1	3	-	-	7	7	4	37	-	11	1	-	5	2	4	1	-	1	-	12	1	3	-	-	4	1	-	-	5	4	-	-	3	6	7	3	-	3	4	4	3	-	3	-	1	-	-	10	-	2	-	13	16	3	-	5	14	1	18	6	1	-	1	2	-	6	-	1	1	2	2	-	3	5	-	4	4	1	4	-	-	-	11	5	-	2	-	13	36	6	3	46	1	2	-	2	1	4	3	1	2	1	2	-	2	8	1	7	-	1	-	1	3	1	4	7	-	2	2	2	2	2	2	-	5	9	3	3	4	1	-	-	-	-	1	2	5	1	5	5	-	2	-	-	2	4	-	4	4	1	1	8	1	-	6	1	5	-	2	-	12	17	14	5	-	2	-	-	3	733	1	-	3	-	4	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	1	2	1	-	1	1	-	-	-	1	2	-	-	4	2	1	1	3	-	-	1	-	-	2	-	1	-	1	2	-	1	-	-	-	1	-	1	2	-	-	-	-	-	-	2	-	-	1	-	-	1	-	1	-	-	-	1	-	-	-	2	-	-	1	-	-	-	-	-	1	-	-	-	1	-	-	-	-	1	1	-	-	-	-	-	-	1	-	-	-	-	1	1	1	1	1	1	1	1	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1	-	-	2	2	1	5	-	1	3	20	-	-	3	-	11	-	-	3	1	1	1	9	-	-	5	-	3	5	13	8	25	-	4	5	-	-	7	6	11	18	-	15	2	1	1	-	8	2	4	3	-	11	6	4	1	1	2	4	-	1	4	10	2	-	1	7	12	5	1	3	1	4	6	-	-	-	-	-	1	11	-	1	1	16	10	-	1	5	10	-	16	8	-	1	1	3	1	-	2	1	2	1	3	-	-	2	1	-	2	2	6	1	-	1	26	1	-	3	3	25	20	3	-	45	-	4	-	6	-	4	1	-	1	3	-	1	1	7	1	8	-	1	1	2	4	-	3	6	1	3	-	10	-	1	1	-	6	22	2	4	-	3	-	-	3	1	1	1	4	4	4	4	3	1	-	-	3	1	1	5	3	2	-	6	2	1	2	-	2	2	-	1	14	19	5	9	-	5	1	1	2	660	-	-	7	1	2	1	3	1	-	-	2	-	1	-	1	3	-	-	-	-	-	-	1	2	-	2	3	1	-	-	-	-	2	1	5	-	2	4	-	1	-	1	-	2	-	1	1	1	1	1	9	1	2	1	1	-	1	1	1	1	2	-	2	1	1	1	1	2	-	1	2	-	2	-	1	-	-	-	-	-	1	-	1	-	-	-	-	-	1	1	-	2	-	2	1	2	-	1	-	-	1	-	-	-	1	-	-	-	1	-	-	-	-	-	1	-	2	1	1	-	1	2	1	1	1	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	-	3	2	-	3	1	8	-	2	-	12	-	-	2	-	10	1	-	3	1	-	1	7	5	1	3	2	2	3	7	1	23	-	5	4	-	2	8	16	6	22	-	6	-	-	1	-	4	4	1	-	-	11	2	2	-	-	3	-	-	1	5	10	-	1	2	1	9	3	1	-	3	-	6	-	4	-	1	-	2	14	-	-	1	19	13	1	-	3	5	1	14	6	-	1	2	3	-	1	6	1	-	-	1	-	-	8	2	-	4	3	13	2	-	-	21	1	-	5	4	17	22	3	1	38	-	4	-	-	1	2	1	1	2	4	2	1	2	12	-	6	-	3	1	3	5	1	4	6	1	4	-	7	1	-	2	1	5	16	7	3	1	5	1	-	1	1	3	-	2	-	6	3	1	2	3	-	2	2	-	6	1	2	-	4	-	1	3	-	2	1	1	2	8	19	12	6	-	5	-	-	1	686	-	-	3	-	-	1	1	1	1	1	2	-	1	-	1	3	1	-	-	-	-	2	-	-	1	1	4	1	-	-	-	-	-	-	1	1	1	7	1	-	-	-	-	-	1	3	2	-	5	-	1	1	1	1	1	-	2	1	-	-	2	-	-	-	1	-	-	-	1	-	-	1	1	-	1	-	-	-	1	-	1	-	1	-	-	-	-	-	-	-	-	-	-	1	-	1	-	-	1	-	1	-	-	-	-	-	1	-	-	-	-	-	-	1	1	-	-	-	-	-	-	2	-	-	-	-	-	-	-	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	-	2	1	3	1	1	7	-	1	4	17	-	3	4	3	12	1	-	-	1	-	2	16	3	-	7	-	3	2	11	4	23	-	3	2	-	1	7	15	6	22	1	10	4	1	1	2	9	1	3	3	-	9	7	1	5	1	1	1	-	1	12	8	-	-	1	13	4	7	-	3	2	7	10	2	3	-	3	-	-	16	-	1	3	8	11	2	-	8	9	1	14	8	2	2	-	5	-	6	3	3	-	3	7	4	2	4	-	3	5	3	4	1	-	-	22	2	-	4	5	18	20	3	2	37	3	7	-	5	1	2	3	1	4	3	-	2	-	14	-	4	1	2	-	5	2	1	2	4	1	-	-	9	1	2	1	1	8	12	4	4	6	2	1	-	2	-	1	3	6	2	-	-	1	-	3	1	2	2	2	9	4	1	1	6	-	-	1	1	2	-	1	1	10	25	9	8	-	4	-	-	8	962	1	1	2	1	4	-	1	-	1	-	1	-	1	-	1	4	-	-	-	-	2	2	-	-	1	1	1	-	-	-	-	-	-	1	1	-	-	3	-	1	-	-	-	-	1	-	-	-	2	-	1	-	1	-	1	-	-	-	-	-	-	1	-	1	1	-	-	4	1	1	3	1	-	-	-	-	1	-	-	1	2	-	2	-	-	2	1	-	1	2	1	-	-	-	-	-	-	-	2	-	1	-	-	-	-	1	-	-	3	-	-	1	-	-	-	-	-	-	1	1	1	2	-	-	2	-	1	-	-	-	-	-	-	-	-	-	-	1	1	1	1	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1	1	2	-	1	2	8	-	1	1	17	1	1	4	-	8	2	-	-	1	-	1	4	3	-	7	1	2	1	6	2	18	-	-	2	-	1	9	5	9	25	1	11	-	1	2	-	4	2	2	1	1	8	6	2	1	2	4	2	1	-	4	4	-	-	1	10	7	1	-	4	1	4	7	2	3	-	2	1	-	16	1	1	4	9	16	-	-	5	8	-	18	6	-	1	2	4	-	4	1	-	2	2	4	-	6	6	-	2	5	3	2	-	2	1	22	2	-	4	3	18	29	5	3	51	1	4	-	3	-	5	6	2	2	5	-	-	-	8	1	1	-	1	-	3	7	-	2	2	-	5	1	14	1	1	1	-	8	15	7	9	4	5	1	-	-	1	2	5	7	2	6	7	1	2	2	-	3	3	2	5	3	1	1	10	-	1	4	1	2	-	1	1	17	9	12	9	-	6	1	-	7	722	1	-	2	-	2	1	1	1	-	-	3	1	3	-	-	2	-	1	-	-	3	3	2	-	1	1	2	1	-	1	-	-	1	2	-	-	-	2	-	-	-	1	-	-	1	3	1	1	3	-	5	-	3	-	1	-	1	1	-	-	1	-	-	-	1	-	-	1	-	-	4	1	1	-	1	-	-	-	1	-	1	3	1	-	-	1	-	-	1	1	1	1	-	1	-	-	1	-	1	-	1	1	-	2	-	-	1	-	1	-	-	1	1	1	-	1	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	1	1	1	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	-	2	3	1	1	2	8	-	-	2	21	-	1	1	1	7	2	-	2	-	1	-	6	1	-	3	2	-	4	8	3	20	1	4	1	-	1	9	8	8	16	1	7	-	1	1	-	1	-	2	1	-	7	3	2	-	1	2	2	-	-	6	8	-	-	1	11	9	3	-	4	2	3	5	1	2	1	3	-	-	14	-	-	1	12	11	-	-	2	12	-	14	3	3	2	2	2	-	6	1	1	-	2	1	-	2	4	1	2	5	-	3	1	-	-	28	3	-	6	1	18	18	5	1	32	-	6	-	1	1	2	1	3	2	5	1	1	1	13	-	5	2	3	1	4	5	-	5	11	-	2	3	7	1	-	6	-	1	14	3	4	3	4	1	-	3	2	1	-	7	5	5	4	1	1	2	1	2	1	2	4	2	2	1	5	-	1	8	2	1	-	1	1	11	9	10	4	-	6	-	-	-	695	-	1	-	1	-	1	-	-	2	1	1	1	1	-	1	4	-	-	-	-	2	1	1	-	-	2	1	-	1	-	2	-	-	1	1	-	-	4	1	-	2	-	-	1	2	1	2	-	2	-	3	1	1	2	3	-	-	-	-	1	2	4	-	-	1	1	2	1	-	-	-	-	-	-	1	-	-	-	1	1	2	-	1	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	2	-	-	-	-	-	1	-	1	-	-	1	1	-	-	1	1	-	-	-	2	-	-	-	-	-	-	-	-	3	1	-	1	-	-	-	-	-	1	-	-	-	-	1	-	-	-	-	-	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	-	4	1	-	-	2	7	1	-	-	18	-	1	2	2	8	-	-	-	1	-	2	3	3	-	2	-	1	5	7	3	16	2	4	-	-	1	9	7	6	20	-	9	3	-	2	1	4	1	4	-	-	5	2	2	3	2	4	4	-	1	9	5	1	1	2	7	9	1	-	3	1	3	7	1	1	-	1	-	-	12	-	1	-	15	12	-	-	6	13	-	14	10	2	1	2	1	-	1	4	-	1	1	4	-	1	1	1	3	9	-	5	1	2	1	21	3	-	4	-	32	34	4	2	34	-	4	-	4	-	2	5	2	2	6	-	1	1	13	-	8	-	-	-	2	4	-	5	1	-	-	2	9	3	-	-	1	5	10	2	1	1	5	-	-	1	1	2	2	9	3	6	2	2	2	-	-	3	2	4	9	1	2	1	11	-	-	3	2	2	-	1	2	11	15	9	5	1	2	-	-	4	739	-	1	5	-	3	-	-	-	-	-	-	1	1	-	-	2	-	-	1	-	-	3	-	2	-	1	2	1	-	-	-	1	-	1	2	-	-	1	-	-	-	1	-	1	2	1	2	-	3	-	-	-	-	1	1	-	1	3	-	-	1	-	-	-	-	1	-	2	2	-	4	1	1	-	-	-	-	-	1	-	3	1	2	-	-	-	-	-	1	-	-	-	-	-	2	-	-	-	-	-	-	2	-	1	-	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	1	-	-	-	2	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	1	-	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	-	4	1	2	2	-	5	-	1	2	11	-	1	4	2	9	1	-	-	1	1	1	10	4	-	4	1	3	1	8	4	12	-	5	1	-	1	11	6	5	24	-	14	1	-	1	-	2	1	3	-	-	4	1	2	1	1	2	2	1	-	6	3	2	-	3	8	6	5	-	2	-	5	5	1	-	-	1	1	-	9	1	4	-	13	10	1	1	2	13	-	13	6	3	2	1	1	-	1	1	1	-	-	4	-	2	5	2	1	4	2	4	-	1	-	17	2	-	3	1	20	17	2	2	45	2	2	-	4	3	-	1	1	1	2	1	1	-	10	-	10	-	1	3	3	3	-	4	6	-	1	4	7	-	2	2	-	2	8	3	2	3	3	1	1	4	2	3	6	1	1	3	5	1	-	-	1	1	-	1	9	5	1	-	11	-	-	5	1	3	1	3	1	18	22	9	3	1	7	-	-	5	712	-	2	2	1	-	-	-	1	1	-	-	-	1	-	-	1	1	-	1	-	-	1	-	-	1	1	2	1	1	-	-	-	-	1	2	-	1	1	2	-	-	-	1	-	4	-	1	-	5	-	4	-	1	-	-	-	2	-	-	-	1	-	1	-	-	-	1	1	-	1	1	1	1	-	2	-	1	-	-	1	2	-	2	-	-	-	-	-	-	-	-	1	-	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3	-	1	-	-	-	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	1	-	-	-	1	1	-	1	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	-	-	1	1	-	2	8	-	2	2	13	-	1	3	1	11	1	-	1	-	1	3	11	-	-	7	1	3	3	6	2	20	-	5	3	-	1	10	14	4	21	1	11	-	-	-	2	1	4	3	1	-	9	3	1	1	1	3	1	1	2	5	10	-	-	3	12	6	2	-	4	1	4	6	3	2	-	-	-	-	13	-	1	-	11	9	-	-	2	6	-	20	10	4	2	-	2	2	3	3	-	1	2	4	-	1	6	-	2	3	-	6	-	1	2	31	2	-	5	3	26	15	4	2	25	1	6	-	4	4	3	-	-	2	3	-	1	-	6	-	10	-	1	-	2	3	-	3	7	-	1	1	6	1	-	1	1	5	15	1	3	3	-	1	-	3	2	1	2	7	1	11	2	1	1	2	-	2	3	2	1	3	2	1	11	-	-	-	2	1	-	-	4	9	20	13	7	1	3	-	-	6	1005	-	2	6	-	1	1	-	-	2	-	-	1	1	-	-	2	-	-	-	-	1	1	-	2	1	2	-	-	-	1	2	-	1	1	3	-	-	3	1	-	-	1	-	1	1	1	-	1	2	-	5	1	-	-	4	1	-	1	-	-	1	-	-	1	2	1	3	1	-	-	1	2	1	-	2	-	-	1	-	-	-	1	1	1	-	-	-	1	2	2	-	-	1	1	3	-	-	1	-	-	1	1	-	1	-	2	-	-	-	1	-	-	-	1	-	-	1	-	-	-	-	-	-	-	-	-	3	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	1	2	2	1	1	2	2	-	1	2	8	-	1	-	-	9	1	1	-	-	-	-	8	5	-	2	1	5	2	5	2	13	-	2	2	1	-	8	11	8	19	2	12	1	2	2	1	6	-	1	-	-	4	5	1	-	2	1	1	-	-	6	4	-	-	1	7	9	4	-	6	1	4	9	1	2	-	-	-	1	9	-	1	3	6	10	1	-	7	10	-	8	8	-	4	1	1	-	2	1	-	-	-	4	1	3	5	2	-	3	1	5	-	1	-	19	2	-	1	5	20	17	3	1	35	-	4	-	3	5	4	3	-	1	2	2	2	2	9	-	10	-	2	-	4	1	1	3	2	2	1	1	7	2	-	1	1	5	18	5	9	3	2	2	-	2	4	3	5	3	1	5	5	-	1	2	-	2	1	3	6	1	1	1	7	1	1	1	2	1	-	1	2	9	12	15	4	-	8	-	-	6	751	1	-	1	-	-	-	1	-	-	-	1	2	2	-	-	2	-	-	-	-	-	1	-	-	-	-	2	2	-	-	-	-	-	2	-	-	-	2	1	1	-	1	-	-	3	4	2	-	1	-	3	-	3	1	3	1	-	-	-	1	-	-	2	-	-	1	-	1	1	-	-	1	-	-	-	1	1	-	-	-	1	-	1	-	-	-	-	1	-	1	1	-	-	-	-	-	-	1	2	1	-	-	-	-	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	1	-	-	-	-	1	-	-	-	-	-	1	-	-	-	-	1	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1	-	-	-	1	2	5	1	-	4	16	-	2	5	1	3	1	-	-	-	1	1	3	2	-	9	3	4	4	8	5	22	2	2	1	1	-	7	15	8	18	1	11	1	-	2	-	3	1	2	1	1	6	-	4	-	2	7	1	-	-	3	6	1	-	2	11	5	5	1	3	1	3	3	1	2	-	1	-	-	11	-	-	-	14	8	1	-	1	10	-	13	7	1	-	1	2	-	3	2	2	-	2	5	-	1	6	-	-	3	2	5	3	-	-	20	2	-	4	6	15	17	2	3	43	1	5	-	3	4	3	3	2	-	2	2	-	-	13	-	9	-	1	1	6	4	1	-	10	-	1	3	12	-	2	2	-	5	21	5	2	3	1	-	2	3	1	1	2	7	4	3	9	-	2	-	1	5	4	5	2	2	3	-	2	1	-	1	-	2	1	2	3	12	13	6	9	-	8	1	-	6	751	-	1	2	1	-	-	-	-	-	-	-	-	1	-	-	-	-	1	-	-	1	1	2	1	-	-	1	1	-	-	-	-	-	-	2	-	2	1	1	2	1	-	-	1	2	7	1	1	2	-	3	-	1	-	-	-	-	-	-	-	-	1	1	-	1	1	1	1	2	-	-	-	-	-	-	-	-	-	1	1	1	1	1	-	-	1	-	1	-	-	-	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	1	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	2	-	-	-	1	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	1	-	-	1	1	-	-	-	-	-	-	-	-	-	-	1537

1990	1	2	1	-	1	1	2	-	-	-	12	1	-	1	-	7	1	1	-	-	-	1	4	-	-	8	-	1	1	5	5	11	-	-	-	-	1	4	7	4	11	1	21	-	1	2	-	2	1	1	-	1	2	3	1	1	-	1	2	-	-	4	2	1	-	-	6	10	4	-	-	4	3	8	3	4	1	-	-	2	8	-	1	1	8	10	1	-	5	8	1	11	10	-	-	1	1	-	3	2	1	-	1	1	-	1	6	-	-	2	-	5	-	2	-	14	2	-	1	1	15	21	3	1	31	1	5	1	3	3	3	4	-	1	3	2	1	-	8	-	8	-	2	-	5	1	-	2	4	2	3	-	11	1	-	3	2	3	14	3	10	7	1	2	-	1	1	1	2	4	1	11	2	1	-	-	1	2	1	2	6	1	3	-	8	1	2	2	1	3	1	2	3	7	13	4	11	1	11	2	-	3	695	-	-	3	-	-	-	-	-	-	-	-	-	1	2	-	2	-	2	-	-	-	1	-	-	-	-	2	-	1	-	-	-	3	2	1	1	-	3	-	1	-	-	1	1	4	2	-	2	5	1	1	-	-	1	2	-	-	-	-	1	-	-	-	-	-	1	1	2	1	-	1	-	-	1	1	-	1	-	-	-	-	1	2	-	-	-	-	-	-	1	-	2	-	-	-	1	1	-	1	1	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	1	-	-	-	1	-	-	-	-	-	2	-	-	-	1	-	-	-	-	-	-	-	-	1	-	1	-	-	1	-	-	-	-	-	-	-	-	-	1372

1991	1	1	-	2	2	4	-	-	-	3	8	1	-	3	2	12	-	-	-	1	-	3	7	3	-	5	1	5	3	9	4	19	-	5	1	3	3	6	8	3	20	1	7	-	-	3	2	7	1	-	-	-	12	3	4	-	-	-	-	1	-	9	8	-	-	1	12	4	4	-	1	2	2	5	2	3	-	1	-	-	13	1	1	1	13	10	-	-	3	8	-	5	2	2	2	-	1	-	2	3	-	-	-	3	1	1	3	-	3	4	2	5	-	1	-	20	1	-	2	-	21	22	-	2	36	3	2	-	2	2	4	2	-	1	3	2	2	1	11	-	2	-	2	1	6	3	-	3	3	-	2	3	6	1	1	1	-	1	23	3	2	8	-	3	-	-	-	3	4	4	3	8	6	-	1	-	1	7	4	-	4	3	2	-	10	-	2	1	3	-	-	-	2	17	17	13	7	-	5	-	-	5	757	-	2	1	1	1	-	-	-	-	-	1	-	1	1	-	2	-	-	-	-	-	1	1	1	1	-	1	-	-	1	1	-	2	1	1	3	-	4	-	-	-	-	-	1	-	4	1	1	2	1	2	2	1	1	3	-	-	-	-	-	2	-	-	-	-	1	-	6	-	-	3	1	-	1	-	1	1	-	-	1	1	1	1	-	-	-	-	-	1	1	-	1	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	1	1	1	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	1517

1993	-	-	1	2	1	2	3	-	1	4	19	-	1	3	-	5	1	-	1	1	1	2	7	-	-	6	-	4	6	3	-	13	-	3	-	-	-	7	10	9	17	1	11	1	1	-	-	10	2	2	1	1	3	2	1	1	2	5	-	2	-	4	8	-	-	1	8	5	3	-	3	1	3	4	-	1	-	-	-	1	9	-	1	1	7	6	-	2	8	8	-	12	6	2	2	-	3	-	-	5	-	1	3	3	1	2	7	-	3	2	1	9	-	2	-	22	-	-	2	1	20	19	5	1	40	3	6	1	4	2	2	3	-	1	2	1	1	-	9	2	13	-	4	1	6	3	1	3	3	-	3	2	9	-	1	1	2	4	19	4	2	2	-	1	-	5	6	5	3	4	3	16	8	2	1	3	1	3	1	4	9	4	2	1	9	1	-	2	-	3	-	-	4	19	13	6	5	-	11	1	-	4	792	-	2	-	-	-	-	-	-	-	1	1	-	-	-	-	1	-	-	-	-	1	4	1	-	-	1	4	-	-	2	-	-	2	-	-	-	1	4	3	1	-	-	1	1	2	1	2	-	6	1	3	-	1	1	3	1	1	-	-	-	1	-	-	-	-	1	1	1	1	-	-	-	1	-	3	-	1	1	-	-	1	2	1	2	-	-	-	-	-	-	-	-	-	-	1	1	-	-	2	1	-	-	-	-	-	2	1	-	2	2	-	-	-	1	-	-	-	2	-	-	1	-	-	-	-	1	-	-	-	1	-	-	-	-	-	2	-	-	-	2	-	-	-	-	1	-	1	-	-	-	-	-	-	1	-	-	1	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	1606

1994	-	3	2	2	3	2	6	2	2	2	15	1	1	3	2	14	3	-	-	-	-	1	11	2	-	9	3	8	2	7	2	31	-	2	-	-	3	12	16	10	34	1	15	1	-	3	2	8	4	4	1	2	11	4	2	1	1	5	-	3	-	10	9	1	-	1	13	17	6	-	8	3	3	5	2	3	2	1	-	-	22	-	2	1	17	18	1	-	8	15	-	14	15	2	2	2	3	2	4	7	-	1	4	3	2	-	7	-	3	8	1	13	1	4	2	41	4	-	5	5	34	43	8	1	68	1	9	-	9	11	6	4	5	5	3	2	2	2	21	2	9	-	1	-	12	7	-	7	6	2	5	2	34	3	2	1	4	10	27	4	12	5	2	1	1	3	1	6	6	4	8	23	16	1	4	1	-	4	9	8	17	8	3	-	12	-	-	12	1	4	2	6	4	31	27	19	17	3	10	3	-	6	1533	-	-	3	-	1	-	-	-	2	1	3	1	4	-	1	1	1	-	-	-	3	2	2	3	-	2	3	3	1	1	1	1	-	4	2	-	1	13	6	2	-	-	-	1	2	5	3	1	5	-	3	-	4	2	11	-	-	1	-	-	1	-	2	-	2	3	1	1	1	1	3	-	1	1	-	-	3	-	2	2	2	-	3	1	1	-	-	1	-	3	-	2	-	4	2	-	-	-	2	-	-	-	-	-	-	1	1	-	2	-	-	-	1	-	-	3	1	1	-	-	1	-	-	1	-	1	1	-	-	1	1	-	-	-	-	1	-	-	1	1	-	-	-	-	3	-	-	-	-	-	2	-	-	-	1	1	-	1	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	2992

1996	3	1	1	-	1	-	13	-	-	5	21	-	-	1	-	11	2	-	1	1	-	1	11	4	-	8	-	3	3	9	1	32	1	2	1	3	3	9	21	7	36	-	8	1	1	4	-	9	1	4	-	-	10	4	3	-	2	8	1	2	-	11	6	-	-	1	11	11	6	-	8	3	6	8	1	1	-	1	-	1	13	-	1	2	16	24	1	-	9	12	-	4	13	1	4	1	4	1	4	6	1	1	2	3	-	-	8	1	2	1	4	13	1	4	1	36	4	-	3	6	35	28	7	-	51	4	12	2	5	8	10	1	4	5	4	2	9	2	14	2	7	-	4	2	7	8	1	6	3	2	2	1	13	3	2	4	-	9	24	5	9	10	-	-	-	6	-	3	4	5	5	18	3	3	1	4	-	6	3	8	17	7	3	-	24	1	1	16	5	4	-	3	2	23	29	16	8	2	14	-	-	16	1580	-	1	3	1	3	1	-	1	2	-	2	-	3	2	-	5	1	-	-	-	2	4	2	-	-	2	3	3	1	-	2	2	-	2	1	1	-	9	4	4	-	1	-	2	4	2	1	1	7	-	7	1	6	1	6	1	1	1	-	-	1	1	-	-	-	2	1	-	1	1	1	2	-	-	1	-	1	-	1	1	3	3	3	-	1	2	-	1	2	1	-	2	-	3	4	1	-	-	1	-	-	-	-	1	1	1	4	-	1	-	-	2	1	-	-	-	3	-	-	-	-	-	-	-	-	-	-	1	-	2	-	1	1	1	-	-	2	-	-	2	-	-	-	-	1	1	-	-	-	-	-	1	-	-	2	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	2904

1998	2	2	1	2	-	-	3	1	2	4	4	-	1	1	1	9	4	1	1	3	1	1	8	2	-	8	5	5	4	9	4	24	-	3	-	1	2	9	10	8	36	1	10	2	1	2	-	6	3	8	1	-	9	5	4	-	-	4	2	1	1	8	13	-	1	2	12	14	4	-	6	1	5	12	4	-	-	3	-	-	19	-	3	1	15	17	1	1	9	12	1	12	22	4	4	-	1	1	5	5	1	1	1	4	2	2	6	-	1	4	2	11	2	4	2	33	4	-	3	2	33	25	4	-	53	4	14	1	8	6	3	3	-	1	8	-	-	-	14	4	8	-	3	2	10	8	2	8	-	1	3	4	18	1	2	2	1	6	37	16	9	10	1	1	1	2	1	6	5	10	2	14	8	4	4	3	-	1	8	6	9	10	4	-	11	-	-	12	1	5	1	4	5	16	26	15	16	1	9	1	-	8	1549	1	-	1	-	1	2	-	-	-	1	4	-	1	1	-	5	-	1	-	1	-	-	1	-	-	-	3	-	-	1	1	1	2	3	4	2	-	11	3	4	2	1	2	1	-	6	3	1	4	1	13	-	-	1	9	1	2	-	-	-	-	-	-	1	1	2	-	2	1	1	3	3	-	1	2	1	-	-	-	2	4	3	-	2	-	-	-	1	-	-	-	-	-	5	3	1	1	-	-	-	1	1	-	1	-	1	2	-	1	1	-	-	-	-	-	2	1	1	-	2	1	-	-	1	-	-	-	1	-	1	-	3	1	-	-	-	-	-	-	1	1	-	-	-	-	1	1	-	-	-	-	1	-	-	-	-	-	2	1	-	-	-	1	-	-	1	1	-	-	-	-	-	-	2832

2000	-	2	-	1	2	1	7	-	3	1	5	-	1	3	-	9	3	1	1	-	-	3	4	2	-	9	-	3	3	9	5	22	-	5	1	1	3	16	12	15	37	1	12	2	-	-	1	8	1	4	4	2	6	2	4	-	1	5	2	2	1	8	12	1	-	-	13	11	5	1	5	2	1	5	2	3	-	1	-	-	10	-	2	1	19	20	-	-	10	8	-	10	7	-	2	2	-	-	4	3	1	2	3	1	3	-	12	-	-	2	3	5	-	5	-	33	-	1	4	2	22	28	4	3	36	1	9	1	7	16	3	6	2	2	6	4	2	1	17	2	5	1	4	1	6	6	-	9	4	-	2	-	7	7	1	1	1	7	28	3	5	9	4	-	-	6	3	6	4	9	7	14	8	2	4	2	2	1	9	3	6	11	4	1	21	2	2	2	4	1	1	2	-	12	29	13	11	3	12	1	-	10	1648	-	1	2	1	2	-	3	1	-	-	3	1	-	1	2	3	-	-	-	-	2	4	2	-	1	1	2	3	-	1	-	1	2	1	2	1	5	9	3	2	-	-	2	2	2	6	-	-	1	1	12	-	3	-	7	2	1	1	-	1	-	-	-	-	-	1	-	2	-	1	3	4	-	-	-	-	-	-	-	3	3	-	3	-	-	2	-	2	-	1	-	-	-	2	1	1	1	-	-	1	-	-	-	1	-	3	-	-	3	-	1	1	-	-	2	2	-	-	-	-	1	1	-	-	-	-	2	1	-	-	-	-	1	-	-	1	-	-	-	-	-	-	-	-	2	-	1	-	-	-	-	-	-	-	1	-	1	1	-	-	-	1	1	-	2	-	-	1	-	-	-	-	-	2817

2002	-	1	-	2	2	1	2	-	-	2	6	-	1	1	1	15	3	-	-	3	-	1	6	3	1	12	-	3	2	3	3	17	1	6	-	2	1	13	18	11	29	1	10	-	-	3	-	3	2	5	1	-	12	4	2	-	5	1	2	-	1	4	7	-	-	2	14	14	2	1	4	1	3	8	1	2	1	4	1	-	21	1	5	1	13	15	3	-	10	12	1	10	18	-	4	4	3	1	7	4	-	1	3	5	-	-	5	-	2	3	3	8	-	5	1	25	1	-	4	5	28	28	-	1	50	3	6	-	9	13	2	2	1	2	4	2	3	2	21	2	9	-	4	-	9	8	-	7	4	-	3	5	11	-	4	4	2	2	29	6	6	11	4	1	1	3	5	6	9	7	7	11	5	3	2	4	-	1	5	3	14	3	5	-	16	1	1	3	1	6	3	2	4	12	27	21	11	1	12	-	1	13	1565	1	1	2	-	-	2	2	-	-	1	-	1	1	-	1	-	-	-	1	-	1	2	-	4	1	1	5	-	1	-	-	-	2	5	1	2	3	4	-	3	-	-	2	-	4	5	2	1	5	1	11	-	2	2	9	-	3	-	-	1	1	-	-	-	1	-	1	1	-	1	-	-	-	-	3	-	1	-	-	3	1	2	-	3	1	1	-	-	-	1	-	1	-	3	7	2	1	-	-	1	-	1	-	1	-	1	2	-	-	1	-	1	-	-	2	1	2	3	-	1	-	-	-	1	-	-	1	-	-	-	1	2	1	-	-	-	-	-	-	-	1	-	-	1	1	1	-	-	-	1	1	-	-	-	-	2	3	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	2765

2004	1	1	-	3	3	2	1	-	1	3	8	-	-	4	1	16	1	-	-	-	2	3	18	1	-	10	-	6	4	5	3	25	1	2	-	-	-	8	15	6	38	-	16	1	1	2	1	5	-	5	1	-	11	6	2	1	1	5	4	1	-	7	3	-	1	1	12	8	4	1	8	-	2	13	-	1	-	2	1	-	19	-	-	1	17	20	1	1	12	5	-	9	15	1	1	-	2	-	7	6	-	-	4	5	-	-	13	1	2	5	2	8	-	5	1	41	4	-	7	6	33	24	2	2	46	3	17	-	2	12	5	1	3	2	10	1	3	-	21	1	11	-	4	-	5	6	-	3	3	2	4	3	14	4	4	1	3	5	35	7	13	12	1	-	1	3	4	5	8	13	6	9	11	2	2	1	2	4	9	3	16	5	4	1	27	1	-	11	4	4	1	2	1	22	41	19	18	4	15	1	-	15	1448	1	-	4	1	-	-	-	3	-	-	-	1	2	1	-	5	-	-	-	-	1	5	1	3	-	1	-	2	-	4	2	1	2	3	2	4	2	12	7	6	-	1	2	-	2	4	1	-	6	3	5	1	5	3	5	1	2	-	-	1	2	-	2	-	1	3	2	2	1	1	3	2	-	3	-	-	-	-	-	2	2	3	1	-	1	1	-	2	-	-	1	1	1	2	8	2	-	-	2	-	1	-	2	-	-	-	3	-	1	-	-	1	-	-	1	3	2	1	-	1	-	-	1	-	-	-	1	-	-	2	-	2	-	-	-	-	-	-	-	1	1	-	-	-	3	1	1	-	-	-	1	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	2812

2006	3	1	-	2	-	4	13	-	2	2	19	-	-	1	2	21	2	1	-	1	-	3	19	3	1	25	4	10	10	15	3	53	1	9	-	-	-	12	30	13	64	2	26	4	-	4	-	15	3	3	2	-	17	9	11	1	-	8	3	-	-	12	13	-	-	2	24	15	7	-	8	-	7	12	2	-	-	2	-	1	34	-	4	2	23	26	-	1	13	9	-	10	18	5	6	6	6	-	8	6	1	1	3	4	1	1	14	-	2	9	7	4	1	10	3	41	7	-	5	4	27	37	2	4	63	8	18	1	12	20	5	2	5	3	5	3	5	1	14	4	14	-	4	-	15	5	1	5	12	3	6	3	21	3	3	5	-	9	42	14	12	14	-	1	1	3	6	4	10	16	4	20	7	5	3	4	1	5	11	11	20	20	7	2	30	-	1	9	7	7	1	3	5	36	63	23	24	5	13	1	1	22	2508	-	1	4	-	-	1	1	4	1	-	-	-	2	1	-	3	-	1	-	-	4	9	-	4	-	3	5	3	-	-	1	1	2	3	2	-	5	20	6	2	1	-	-	1	4	6	4	-	11	2	10	-	6	-	22	2	6	2	-	-	1	-	-	-	3	3	1	6	1	1	4	1	-	1	3	-	4	3	2	7	4	3	3	1	1	1	-	-	2	1	1	1	1	2	9	2	1	-	4	1	-	1	-	2	-	1	2	-	1	3	-	3	1	-	-	2	2	4	-	1	3	-	-	1	1	1	2	1	-	1	1	-	4	1	1	1	1	2	-	2	-	-	-	-	2	-	-	1	-	1	-	-	-	-	-	2	3	-	-	-	-	1	-	1	-	1	-	-	-	-	-	-	-	4510

2008	1	-	-	-	1	-	5	1	1	1	4	-	-	3	-	11	1	-	-	3	-	-	8	-	-	8	2	8	2	9	3	20	-	3	1	-	3	10	11	7	31	-	7	1	1	3	-	6	1	2	1	2	4	3	-	-	1	-	-	1	-	5	1	-	-	-	7	8	3	-	2	-	2	2	1	1	-	2	-	-	16	1	1	-	7	15	1	-	8	7	-	8	9	1	2	1	2	1	2	4	-	-	1	3	-	-	8	1	2	5	-	7	-	6	-	14	7	-	1	-	21	18	1	-	41	3	10	1	11	11	5	-	2	3	4	2	-	-	7	-	5	-	5	2	4	3	-	3	5	2	2	3	8	1	3	1	2	1	21	8	3	5	1	-	-	7	2	4	6	12	5	13	5	2	1	3	-	3	7	3	9	2	-	-	11	2	2	5	3	4	-	2	2	16	14	15	14	1	9	1	-	3	1095	-	-	2	-	2	-	-	-	-	-	1	1	1	1	-	1	-	-	-	-	2	5	1	1	-	1	-	1	2	-	1	-	1	-	-	3	2	5	-	-	2	1	-	1	3	3	1	1	6	1	12	-	3	2	10	1	3	2	-	1	-	-	-	-	1	5	-	1	1	-	2	1	-	-	2	-	-	-	-	7	1	-	1	-	1	-	-	1	-	1	-	-	2	2	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	1	1	1	-	1	1	-	-	-	-	-	-	-	1	1	1	-	4	3	-	-	1	-	-	-	-	-	-	-	1	-	-	-	1	-	-	2	-	-	2	1	1	-	1	-	1	1	-	-	-	-	-	-	-	-	-	-	-	2023

2010	-	2	-	1	1	3	4	-	-	2	8	-	1	3	-	7	2	1	-	2	-	2	10	3	-	8	3	9	2	5	1	24	2	2	-	-	1	5	4	3	26	-	6	1	2	1	-	3	1	3	-	1	5	1	2	3	1	2	-	-	-	6	6	-	-	2	12	4	1	-	3	1	4	2	1	1	1	2	-	-	15	-	-	-	7	10	1	-	6	3	-	3	8	2	2	-	-	-	1	5	-	-	2	1	1	1	7	1	2	4	3	5	1	2	1	11	2	-	1	4	16	18	-	3	25	3	9	1	3	7	1	1	5	2	3	1	1	-	15	3	7	-	3	1	4	2	-	3	2	-	2	2	9	1	2	1	1	4	19	4	8	4	-	-	-	4	1	7	3	12	4	7	3	-	1	3	-	1	6	3	14	11	-	1	9	-	-	5	-	1	-	2	4	12	15	13	7	2	8	1	-	11	1205	-	1	3	-	-	-	-	2	-	-	-	1	2	-	-	-	-	-	-	-	1	6	-	3	-	1	-	1	-	1	-	2	-	-	3	2	2	8	1	2	1	-	2	-	2	4	3	2	5	2	5	1	1	2	6	-	-	-	1	1	-	-	-	-	-	1	1	-	-	1	1	-	-	1	2	-	3	-	-	2	1	1	3	-	1	-	-	1	-	-	-	1	-	4	7	1	-	-	1	-	1	-	-	1	-	1	-	-	-	1	-	-	-	1	-	1	-	1	-	-	-	-	-	-	-	-	1	-	-	1	-	-	1	2	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	-	-	-	1	1	1	-	-	2	-	-	-	2	-	-	-	-	-	-	1	-	-	-	2044

2012	-	-	-	1	-	1	1	-	-	2	4	1	-	4	1	11	1	-	-	-	-	-	4	-	-	12	-	6	3	6	-	23	-	-	1	1	-	6	10	6	27	-	5	-	1	-	-	3	-	-	2	-	7	1	3	-	1	3	-	-	-	3	3	-	-	-	6	5	4	-	7	1	3	10	3	1	1	-	-	-	6	-	-	-	4	8	1	-	6	2	-	3	6	1	1	5	1	-	7	2	-	-	1	2	-	-	4	-	1	2	1	9	2	3	-	11	-	1	1	1	17	16	-	-	26	2	15	1	6	8	1	1	2	1	4	-	-	1	5	3	4	-	1	1	4	1	-	-	-	-	-	2	9	1	1	1	1	5	41	3	8	5	1	-	-	1	5	3	4	10	1	12	9	2	-	4	2	3	3	6	8	6	10	-	12	1	2	1	2	5	-	1	-	8	16	8	15	3	11	-	-	14	1149	1	-	1	1	-	-	-	1	-	-	1	-	-	1	-	1	1	-	-	-	2	5	-	4	-	2	1	1	-	1	1	-	-	1	-	-	1	3	1	-	1	-	-	-	2	4	18	1	6	-	10	4	1	-	7	1	1	1	-	-	1	1	-	-	1	1	-	2	-	2	1	-	-	-	2	-	-	1	-	5	-	3	-	1	1	-	-	-	-	-	-	2	1	-	4	1	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	1	-	2	1	3	1	-	1	-	-	1	-	-	-	-	-	1	-	-	1	2	2	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	-	-	1	-	2	-	-	-	-	-	-	1974

2014	-	1	-	1	1	2	4	-	1	1	10	-	-	2	-	11	-	1	2	-	-	1	7	-	-	16	-	9	2	6	-	28	1	1	3	1	-	14	11	8	30	-	9	-	1	5	-	5	3	1	-	2	7	9	4	1	2	3	1	-	-	5	4	2	-	-	4	3	7	-	5	-	5	5	2	2	-	2	-	-	17	-	-	-	8	5	-	-	10	8	-	1	6	1	4	-	2	-	4	3	-	-	2	5	-	-	5	-	1	3	-	8	-	3	-	21	4	-	2	2	20	21	-	-	31	4	14	1	6	10	2	-	1	2	4	-	1	4	12	5	7	-	1	1	10	5	-	4	13	2	7	2	15	3	2	3	1	6	20	6	9	9	2	1	-	4	3	6	5	14	8	10	5	-	1	1	-	2	13	2	12	8	3	-	17	-	1	6	4	3	-	2	4	16	17	20	14	3	12	3	-	15	1455	-	1	2	-	-	-	-	-	1	-	1	1	3	-	1	1	1	-	-	-	1	6	2	1	1	1	2	1	-	2	-	4	2	1	2	3	1	6	3	2	2	2	-	-	3	8	2	-	3	-	10	-	3	1	15	1	5	-	-	-	1	-	-	1	1	4	-	4	2	1	2	2	1	-	-	-	-	-	-	3	1	4	1	-	-	-	-	-	1	-	-	3	2	1	10	1	-	-	1	-	-	-	-	-	-	-	3	-	1	1	1	1	1	-	-	-	2	2	-	-	2	-	-	-	-	-	2	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	1	-	1	-	-	-	-	1	-	-	-	-	1	1	-	2	1	-	-	2538

2016	-	-	-	1	-	2	6	-	1	1	-	-	-	3	2	9	2	1	1	1	1	3	13	2	-	13	5	5	4	12	3	34	1	1	-	-	-	17	8	3	20	-	13	1	1	3	-	6	1	1	1	-	11	5	3	1	3	5	-	-	1	5	8	-	-	2	16	5	4	1	7	1	1	8	-	2	-	3	-	-	11	-	-	2	5	11	-	-	9	6	1	5	5	3	3	-	3	1	1	5	1	1	2	3	-	-	8	-	3	5	3	12	-	5	1	22	4	-	-	3	19	16	-	4	23	5	13	1	11	5	3	4	1	1	4	-	-	-	13	2	3	-	4	3	11	7	-	3	5	-	-	3	16	1	2	-	1	5	34	3	11	11	-	-	-	4	2	8	10	5	9	12	5	6	1	3	-	1	6	2	13	5	3	-	19	1	1	12	9	4	2	1	1	27	29	23	19	1	22	1	-	7	1723	-	1	2	-	-	-	2	-	-	1	-	-	2	1	-	2	-	-	-	-	-	8	2	3	-	1	4	-	-	3	1	1	2	-	3	1	1	12	3	-	1	1	1	1	1	10	1	-	4	-	15	-	2	1	10	-	3	1	-	-	-	-	1	-	1	1	-	-	1	1	1	-	-	-	-	-	1	-	1	2	-	2	2	-	-	-	-	1	-	1	-	2	1	3	4	2	-	-	1	-	2	1	-	2	-	2	2	-	-	1	1	1	1	1	-	-	1	1	-	1	1	1	-	-	-	-	5	2	-	1	1	-	6	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	1	3	-	-	-	-	1	-	-	1	1	-	-	-	2	1	-	-	-	-	-	1	-	2867

2018	-	-	-	-	-	1	5	-	-	2	4	-	1	2	1	5	1	1	-	2	2	-	5	4	1	11	-	6	2	6	3	27	-	2	-	-	1	5	7	5	32	1	9	1	1	1	-	2	1	-	2	-	2	2	-	1	2	2	-	-	-	2	6	2	-	-	9	7	2	-	6	1	4	8	1	2	1	4	-	-	13	1	-	-	4	10	2	-	9	1	1	6	11	2	2	1	-	1	4	4	-	-	4	3	1	-	3	-	4	1	-	3	1	4	-	13	2	2	-	2	13	11	1	3	22	6	11	-	4	1	6	4	4	4	9	-	-	-	10	1	5	-	2	-	4	5	-	2	2	2	-	1	7	4	5	3	1	7	21	5	10	7	-	-	-	3	5	5	12	9	10	10	9	4	1	2	-	1	6	5	8	7	5	-	12	-	1	5	4	3	2	3	-	18	21	22	14	4	12	-	1	5	1396	2	-	-	1	2	-	2	1	1	-	-	-	5	1	1	2	-	-	-	-	-	4	-	2	1	-	1	1	-	1	1	-	2	2	2	2	1	4	2	-	1	1	1	-	2	4	1	1	6	-	7	1	4	1	14	-	-	-	-	1	2	-	1	-	1	4	1	1	1	3	1	-	-	-	3	-	-	1	-	3	2	2	1	-	1	-	-	1	1	-	-	2	-	3	8	1	-	-	1	-	1	-	-	2	-	-	1	-	-	1	-	1	-	-	-	-	-	2	-	3	2	-	-	1	1	-	2	1	-	1	-	-	2	3	1	-	-	1	-	-	1	-	-	-	1	-	-	-	2	3	-	-	-	-	1	-	2	1	1	-	3	-	1	-	1	-	-	-	-	1	-	1	-	2348

2021	-	-	-	2	1	1	11	-	-	2	5	-	-	5	-	6	2	-	-	2	-	-	8	-	-	22	2	8	6	11	6	29	-	4	-	-	4	20	21	11	40	-	21	-	1	7	-	9	-	3	1	-	10	8	4	1	2	4	3	-	-	12	12	1	-	1	15	11	5	-	11	-	6	10	-	3	-	4	-	-	9	-	3	2	5	8	2	-	11	6	2	6	16	2	18	2	3	-	5	6	-	-	1	3	1	-	11	1	1	4	3	8	1	9	1	28	2	-	2	4	44	41	-	1	48	3	15	2	9	18	5	2	8	3	4	1	3	2	9	2	5	-	7	2	12	6	1	1	22	2	5	5	20	3	11	3	2	3	56	15	10	13	4	2	-	7	7	12	22	21	14	25	15	4	3	4	1	4	14	7	28	11	7	1	39	1	1	6	7	8	1	7	4	52	75	17	31	7	22	3	-	25	2173	-	2	1	-	3	1	-	-	1	1	-	4	4	1	-	1	-	-	-	-	2	14	-	4	-	1	1	-	-	3	1	2	1	5	-	2	2	9	3	2	3	2	3	1	5	7	-	1	4	1	24	1	9	2	29	-	5	-	-	1	-	1	-	1	1	6	-	4	2	-	5	2	1	3	2	1	-	-	-	15	3	10	2	-	1	1	1	1	-	3	-	4	6	2	7	-	-	-	-	-	-	1	-	2	2	2	4	-	1	1	2	1	-	-	-	2	2	2	-	3	3	-	1	-	1	-	6	-	1	3	-	-	14	2	1	1	-	-	-	1	-	-	1	-	-	1	-	-	-	-	-	4	-	-	-	1	1	-	5	3	-	-	3	-	2	-	-	-	-	1	-	1	-	4032

2022	-	-	-	4	-	1	6	-	-	-	7	-	1	1	1	9	5	-	-	-	-	1	3	3	-	15	-	8	7	5	3	30	-	2	-	-	2	11	9	8	28	-	11	1	1	1	-	4	1	1	-	-	9	4	2	1	3	4	3	1	-	4	8	-	1	2	11	7	8	1	8	4	2	4	3	2	1	3	-	-	18	-	-	1	11	13	-	-	10	12	1	7	9	1	-	-	3	-	8	1	-	-	3	9	-	-	12	-	2	4	1	4	-	3	1	15	4	-	1	3	23	18	-	1	31	4	8	-	11	12	-	2	2	1	3	-	3	1	9	3	15	-	4	1	12	5	-	7	19	2	2	7	10	4	6	4	1	9	47	14	8	9	-	3	-	6	4	6	15	11	7	10	12	1	1	3	2	6	9	6	18	4	5	1	26	-	4	7	3	6	1	5	1	36	51	20	17	2	12	1	-	23	2182	1	1	2	-	2	-	-	-	-	-	-	1	-	2	-	5	-	2	-	-	1	8	-	-	-	3	2	5	-	1	1	3	5	2	-	3	1	12	2	5	1	1	-	2	2	9	-	2	10	1	13	-	6	1	18	-	4	-	-	-	-	-	1	-	-	3	1	2	2	-	1	2	-	-	3	3	3	3	-	4	-	1	3	1	3	2	1	-	-	-	-	2	-	2	11	2	1	-	1	-	-	-	1	-	-	1	2	-	-	2	1	-	-	-	1	3	1	5	-	4	4	2	-	-	-	1	1	1	2	2	1	1	4	1	-	2	-	-	-	1	-	-	1	-	2	-	-	1	1	1	-	-	-	-	-	-	2	-	-	-	-	1	1	-	2	-	1	-	-	1	-	-	-	3544

2024	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3309	3309

Total	24	49	28	42	40	57	203	8	28	77	438	8	25	98	37	363	51	12	24	31	15	46	279	74	5	289	45	151	110	278	101	796	15	103	48	15	38	324	383	251	908	23	415	36	27	70	20	182	54	83	33	16	274	120	90	41	47	109	47	23	11	220	245	15	7	42	350	271	129	12	152	55	118	236	50	65	11	55	6	13	462	8	46	36	411	436	33	11	230	317	13	402	299	52	93	44	83	13	127	114	20	21	64	113	21	56	214	21	67	133	61	230	23	87	21	765	82	5	103	96	767	835	109	57	1404	81	250	16	172	187	111	83	61	67	136	38	56	34	383	43	234	8	81	29	193	158	12	125	203	30	76	75	364	61	66	65	33	184	742	187	214	196	63	27	9	98	74	126	169	245	143	319	197	55	48	64	20	90	156	113	306	160	101	17	433	22	29	170	77	108	22	64	76	576	822	468	361	48	295	27	5	278	38089	12	24	87	12	36	15	21	21	18	11	33	21	47	17	11	76	7	9	5	2	36	108	27	42	12	39	66	36	10	28	22	22	39	57	51	35	38	188	56	44	21	19	19	27	67	119	60	27	134	18	208	18	73	31	209	14	47	22	3	12	28	11	14	8	26	52	21	55	31	19	56	29	13	14	37	9	26	11	12	67	48	49	45	15	14	15	4	17	14	23	7	33	16	50	97	26	9	5	29	8	14	10	4	21	5	23	32	1	21	17	7	19	11	7	14	26	29	31	3	22	26	12	5	6	6	6	36	11	6	25	9	13	43	21	7	10	5	4	4	15	7	2	4	5	21	10	7	3	6	13	8	14	1	4	10	10	19	6	14	5	7	7	10	2	13	5	6	2	1	6	1	3	3309	75699

See also