Spouse labor force status

spwrksta

Source

General Social Survey https://gss.norc.org

Details

Question Last week was your (wife/husband) working full time, part time, going to school, keeping house, or what?

Values

[1] working full time

[2] working part time

[3] with a job, but not at work because of temporary illness, vacation, strike

[4] unemployed, laid off, looking for work

[5] retired

[6] in school

[7] keeping house

[8] other

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C/D	2006

Full/-	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for spwrksta.

Counts by year:

year	working full time	working part time	with a job, but not at work because of temporary illness, vacation, strike	unemployed, laid off, looking for work	retired	in school	keeping house	other	iap	no answer	skipped on web	don't know	Total

1972	569	91	15	17	68	9	368	21	453	2	-	-	1613

1973	580	71	21	8	61	12	303	20	428	-	-	-	1504

1974	565	70	18	19	71	6	304	12	419	-	-	-	1484

1975	513	81	22	27	74	7	263	15	488	-	-	-	1490

1976	482	75	19	21	91	11	259	16	525	-	-	-	1499

1977	503	73	22	17	89	10	250	10	555	1	-	-	1530

1978	505	83	26	17	78	8	224	19	571	1	-	-	1532

1980	468	75	26	18	92	9	188	11	579	2	-	-	1468

1982	532	81	16	23	93	7	217	23	867	1	-	-	1860

1983	505	86	20	39	91	9	199	16	634	-	-	-	1599

1984	465	61	22	18	84	7	164	7	644	1	-	-	1473

1985	454	81	18	17	99	14	175	13	663	-	-	-	1534

1986	429	86	19	11	121	5	142	12	644	1	-	-	1470

1987	527	86	16	16	101	10	142	14	903	4	-	-	1819

1988	409	70	14	15	128	7	136	7	693	2	-	-	1481

1989	474	70	20	16	117	7	126	14	690	3	-	-	1537

1990	391	65	16	13	93	12	126	8	645	3	-	-	1372

1991	436	68	15	15	126	8	120	16	713	-	-	-	1517

1993	484	83	18	21	112	6	120	12	747	3	-	-	1606

1994	867	155	26	33	208	20	211	21	1450	1	-	-	2992

1996	841	145	22	13	146	14	182	24	1513	4	-	-	2904

1998	813	129	23	15	176	18	138	32	1486	2	-	-	2832

2000	764	122	15	13	149	16	164	28	1538	8	-	-	2817

2002	704	142	30	33	164	13	159	22	1496	2	-	-	2765

2004	840	133	35	34	194	19	178	35	1333	11	-	-	2812

2006	1237	187	25	28	311	25	293	60	2334	10	-	-	4510

2008	533	100	24	14	150	8	112	27	1046	9	-	-	2023

2010	438	91	17	31	145	21	111	30	1152	8	-	-	2044

2012	487	83	18	21	129	17	127	16	1074	2	-	-	1974

2014	610	118	24	27	197	12	147	22	1376	5	-	-	2538

2016	658	96	23	23	222	11	146	24	1654	10	-	-	2867

2018	552	90	23	19	161	10	122	16	1348	7	-	-	2348

2021	953	152	38	80	511	14	184	63	2024	2	11	-	4032

2022	736	132	28	26	329	16	154	31	2080	5	4	3	3544

Total	20324	3331	734	758	4981	398	6254	717	34765	110	15	3	72390

See also