How often does r find work stressful

stress

Source

General Social Survey https://gss.norc.org

Details

Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often... C. Do you find your work stressful?

Values

1 always

2 often

3 sometimes

4 hardly ever

5 never

NA(d) can't choose / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1989	A/B/C/-	full

1998	A/B/C/-	full

2002	A/B/C/-	full

2004	A/B/C/-	full

2006	A/B/C/D	full

2010	A/B/C/-	full

2012	A/B/C/-	full

2014	A/B/C/-	full

2016	B/C/-/-	partial

2018	A/B/C/-	full

2022	A/B/C/-	full

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	not available in this year	always	can't choose	hardly ever	never	no answer	often	sometimes	skipped on web	Total

1972	3	1610	-	-	-	-	-	-	-	-	1613

1973	-	1504	-	-	-	-	-	-	-	-	1504

1974	-	1484	-	-	-	-	-	-	-	-	1484

1975	-	1490	-	-	-	-	-	-	-	-	1490

1976	-	1499	-	-	-	-	-	-	-	-	1499

1977	-	1530	-	-	-	-	-	-	-	-	1530

1978	-	1532	-	-	-	-	-	-	-	-	1532

1980	-	1468	-	-	-	-	-	-	-	-	1468

1982	-	1860	-	-	-	-	-	-	-	-	1860

1983	-	1599	-	-	-	-	-	-	-	-	1599

1984	-	1473	-	-	-	-	-	-	-	-	1473

1985	-	1534	-	-	-	-	-	-	-	-	1534

1986	-	1470	-	-	-	-	-	-	-	-	1470

1987	-	1819	-	-	-	-	-	-	-	-	1819

1988	-	1481	-	-	-	-	-	-	-	-	1481

1989	-	-	82	3	109	37	663	261	382	-	1537

1990	-	1372	-	-	-	-	-	-	-	-	1372

1991	-	1517	-	-	-	-	-	-	-	-	1517

1993	-	1606	-	-	-	-	-	-	-	-	1606

1994	-	2992	-	-	-	-	-	-	-	-	2992

1996	-	2904	-	-	-	-	-	-	-	-	2904

1998	-	-	89	1	91	42	2002	229	378	-	2832

2000	-	2817	-	-	-	-	-	-	-	-	2817

2002	-	-	161	6	329	137	985	391	756	-	2765

2004	-	-	213	5	242	95	985	475	797	-	2812

2006	-	-	353	3	449	185	1771	595	1154	-	4510

2008	-	2023	-	-	-	-	-	-	-	-	2023

2010	-	-	110	1	215	78	881	284	475	-	2044

2012	-	-	118	5	162	66	825	320	478	-	1974

2014	-	-	113	1	225	99	1294	247	559	-	2538

2016	911	-	96	2	130	32	1022	207	467	-	2867

2018	-	-	107	3	264	110	932	304	628	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	4032

2022	1557	-	171	5	339	128	14	409	906	15	3544

2024	-	3309	-	-	-	-	-	-	-	-	3309

Total	6503	41893	1613	35	2555	1009	11374	3722	6980	15	75699

See also