What was r's first religious preference

switch1

Source

General Social Survey https://gss.norc.org

Details

Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/ DENOM16). Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION menTIONED IN RELIG16/DENOM16). Yes................(ASK A) No..............(GO TO MARELIG) A. Starting when you were 16 years old, what other religious preferences have you had? Please mention your earliest preference first. Any others?

First Mention

Values

110000 amer baptist assoc

111000 amer bapt ch in us

112000 natl bapt con of am

113000 natl bapt con, usa

114000 southern bapt con

115000 other baptist church

118000 baptist, dk which

120000 afrcn meth epis ch

121000 afrcn meth epis zion

122000 united methodist ch

123000 other methodist ch

128000 methodist, dk which

130000 amer lutheran church

131000 lutheran ch in amer

132000 luth ch-missouri syn

133000 wisconsin ev luth sn

134000 other lutheran churc

135000 evangelical lutheran

138000 lutheran, dk which

140000 presbyterian ch n us

141000 united pres ch in us

142000 other presbyterian

148000 presbyterian, dk whi

150000 episcopal church

160000 other protestant

170000 no denomination

2e+05 catholic

3e+05 jewish

4e+05 none

5e+05 other religion

999999 missing

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1988	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for switch1.

See also