Why did r make first switch

switwhy1

Source

General Social Survey https://gss.norc.org

Details

Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/DENOM16). Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION MENTIONED IN RELIG16/DENOM16). Yes................(ASK A) No..............(GO TO MARELIG) C2. Why did you become a (FIRST/SECOND MENTION)? (1) Why first switch made

Values

[0] na

[10] marriage

[14] marr & location

[15] marr & theology

[16] marr. & clergy

[17] marr & positive blf

[20] friends

[23] frnds & family

[24] frnds & location

[25] frnds & theo dis

[26] frnds & clergy

[28] frnds & other pos

[30] family

[34] fam & location

[35] fam & theo diss

[37] fam & positive blf

[39] fam & misc

[40] location

[41] loc & to change

[45] loc & theo dissat

[46] loc and clergy

[47] loc & positive blf

[50] theological diss

[56] theo diss & clergy

[60] clergy

[68] clergy & oth pos

[69] clergy & missing

[70] to change

[71] other, to change

[76] other, misc

[77] other, pos blfs

[78] other, pos other

[79] other, missing

[85] loc, theo & clrgy

[86] frnds,fam,location

[87] marr,loc,theo dis

[99] missing values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1988

Overview

For further details see the GSS Data Explorer page for switwhy1.

Counts by year:

year	iap	marriage	marr & location	marr & theology	marr. & clergy	marr & positive blf	friends	frnds & family	frnds & location	frnds & clergy	frnds & other pos	family	fam & location	fam & theo diss	fam & positive blf	fam & misc	location	loc & theo dissat	loc and clergy	loc & positive blf	theological diss	theo diss & clergy	clergy	clergy & missing	other, to change	other, misc	other, pos blfs	other, pos other	other, missing	no answer	Total

1988	945	71	1	2	1	3	47	2	4	2	1	46	1	2	1	1	30	1	2	3	63	1	8	1	3	31	14	8	1	185	1481

Total	945	71	1	2	1	3	47	2	4	2	1	46	1	2	1	1	30	1	2	3	63	1	8	1	3	31	14	8	1	185	1481

See also