People with secret should be subject to drug tests

testdrug

Source

General Social Survey https://gss.norc.org

Details

Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET of TOP SECRET clearance. Do you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree that people with SECRET or TOP SECRET clearance should be subject to... B. Random drug tests

Values

[1] strongly agree

[2] agree

[3] neither agree nor disagree

[4] disagree

[5] strongly disagree

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1998, 2000

Overview

For further details see the GSS Data Explorer page for testdrug.

Counts by year:

year	iap	strongly agree	agree	neither agree nor disagree	disagree	strongly disagree	don't know	no answer	Total

1998	1445	618	624	49	46	10	25	15	2832

2000	1398	687	544	92	45	12	22	17	2817

Total	2843	1305	1168	141	91	22	47	32	5649

See also