Time r would spend on paid work

timepdwk

Source

General Social Survey https://gss.norc.org

Details

Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now? A. Time in a paid job?

Values

[1] spend much more

[2] spend a bit more

[3] spend same

[4] spend a bit less

[5] spend much less

[NA(d)] can't choose [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1989, 1998

D/-/-	2006

Overview

For further details see the GSS Data Explorer page for timepdwk.

Counts by year:

year	iap	spend much more	spend a bit more	spend same	spend a bit less	spend much less	can't choose	no answer	Total

1989	84	114	186	570	263	122	146	52	1537

1998	1609	110	151	350	204	105	227	76	2832

2006	2992	140	162	578	283	115	237	3	4510

Total	4685	364	499	1498	750	342	610	131	8879

See also