How much money in the largest of these decisions

totmoney

Source

General Social Survey https://gss.norc.org

Details

Question About how much money was involved in the largest of these decisions?

Values

[9999996] $9999996 or more

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1991

Overview

For further details see the GSS Data Explorer page for totmoney.

Counts by year:

year	iap	100	1000	10000	1200	12000	120000	13000	14000	1500	15000	150000	1500000	1600000	17000	1800	18000	180000	1e+05	1e+06	20	200	2000	20000	23000	2500	25000	250000	2e+05	2e+06	30	300	3000	30000	31000	35000	350000	3e+05	3e+06	4000	40000	4e+05	50	500	5000	50000	55000	5e+05	5e+06	600	6000	60000	6e+05	6e+06	7000	70000	750	7500	750000	7e+05	7e+06	800	8000	80000	8e+06	900	99	$9999996 or more	9e+06	don't know	no answer	Total

1991	1172	3	16	21	4	3	2	1	1	5	7	2	2	1	1	1	1	2	12	16	1	8	13	9	1	2	10	8	8	4	1	5	13	7	1	2	1	5	1	5	1	2	1	6	17	12	2	8	2	3	1	5	1	2	4	1	3	1	2	1	1	2	1	1	2	1	1	2	1	26	26	1517

Total	1172	3	16	21	4	3	2	1	1	5	7	2	2	1	1	1	1	2	12	16	1	8	13	9	1	2	10	8	8	4	1	5	13	7	1	2	1	5	1	5	1	2	1	6	17	12	2	8	2	3	1	5	1	2	4	1	3	1	2	1	1	2	1	1	2	1	1	2	1	26	26	1517

See also