How likely to mention problems to relatives

transrel

Source

General Social Survey https://gss.norc.org

Details

Question Now imagine that one of your relatives sees the car and wants to buy it. How likely would you be to mention the car's history of transmission problems (if this person did not ask)?

Values

1 definitely would not mention it

2 probably would not

3 probably would

4 definitely would mention it

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

Link at the GSS

For details see the GSS Data Explorer page for transrel.

See also