How well business leader knows effect of reducing fed inc tax

txbiz

Source

General Social Survey https://gss.norc.org

Details

Question The first issue is federal income taxes. Some people say that reducing income taxes almost always helps the economy grow. Others say that reducing income taxes has very little long-term impact on economic growth but could cause a big increase in the deficit. Please look at Card B10. (On a scale of 1 to 5, where 1 means "Very Well" and 5 means "Not at All"), how well do the following groups understand the likely effects of reducing federal income taxes? B. Business leaders

Values

[1] very well

[5] not at all

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/C	2010

B/C	2006

Overview

For further details see the GSS Data Explorer page for txbiz.

Counts by year:

year	iap	very well	2	3	4	not at all	don't know	no answer	Total

2006	3574	157	262	295	111	65	46	-	4510

2010	1578	66	112	152	63	43	27	3	2044

Total	5152	223	374	447	174	108	73	3	6554

See also