Don't know response to unemployment items

undk

Source

General Social Survey https://gss.norc.org

Details

Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst? Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank. H. Respondent can't choose

Values

1 dk chosen

2 not chosen

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1989	A/B/C	full

Link at the GSS

For further details see the GSS Data Explorer page for undk.

See also