Is r us citizen

uscitzn

Source

General Social Survey https://gss.norc.org

Details

Question If not born in US, is R a US citizen As of 2024, USCITZN has been recoded to simplify responses. It is now a synthetic variable which represents the original USCITZN, as well as the experimental USCITZNV and USCITZNNV. This recode is retroactive, covering all previous years of USCITZN. The original data have been preserved in USCITZN_0822. USCITZNV and USCITZNNV will not be made available in 2024.

Values

1 a u.s. citizen

2 not a u.s. citizen

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

ballots	years

A/B/C	2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for uscitzn.

Counts by year:

year	iap	a u.s. citizen	not a u.s. citizen	a u.s. citizen born in puerto rico, the u.s. virgin islands, or the northern marianas islands (if volunteered)	born outside of the u.s. to parents who were u.s. citizens at that time (if volunteered)	don't know	no answer	not available in this year	Total

2008	1758	118	133	2	6	1	5	-	2023

2010	1795	116	122	3	3	3	2	-	2044

2012	1700	141	123	1	2	2	5	-	1974

2014	2184	187	157	3	4	-	3	-	2538

2016	2507	180	148	8	9	5	10	-	2867

2018	2045	163	131	2	2	1	4	-	2348

2021	-	-	-	-	-	-	-	4032	4032

2022	3277	151	92	4	6	4	10	-	3544

Total	15266	1056	906	23	32	16	39	4032	21370

See also