Usual reason r work at home

whywkhme

Source

General Social Survey https://gss.norc.org

Details

Question Is it usually because you want to, you have to in order to keep up with your job, or for some other reason?

Values

[1] worker wants to work at home

[2] worker has to work at home to keep up with job

[3] worker is operating home-based business

[4] other combinations and other reasons

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	2002, 2006, 2010, 2014, 2018, 2022

Overview

For further details see the GSS Data Explorer page for whywkhme.

Counts by year:

year	iap	worker wants to work at home	worker has to work at home to keep up with job	worker is operating home-based business	no answer	don't know	other combinations and other reasons	skipped on web	Total

2002	2077	254	224	197	13	-	-	-	2765

2006	3803	257	225	213	11	1	-	-	4510

2010	1531	160	189	60	27	3	74	-	2044

2014	2033	142	203	84	7	-	69	-	2538

2018	900	171	200	76	896	3	102	-	2348

2022	2641	476	160	106	3	8	130	20	3544

Total	12985	1460	1201	736	957	15	375	20	17749

See also