Rich - poor

wlthasns

Source

General Social Survey https://gss.norc.org

Details

Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich." A score of 7 means that you think almost everyone in the group are "poor." A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand. 4. Asian Americans?

Values

[1] 1 - rich

[7] 7 - poor

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C	1990, 1994, 2000

Overview

For further details see the GSS Data Explorer page for wlthasns.

Counts by year:

year	iap	1 - rich	2	3	4	5	6	7 - poor	don't know	no answer	Total

1990	-	17	73	214	437	286	143	57	128	17	1372

1994	1518	40	95	244	484	277	147	42	132	13	2992

2000	1419	41	143	270	476	198	86	24	146	14	2817

Total	2937	98	311	728	1397	761	376	123	406	44	7181

See also