Word g

wordg

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words–you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters – like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone – just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW. G. Word G

Values

[0] incorrect answer

[1] correct answer

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024

A/B/C	1998, 2000, 2004, 2006

Full	1974, 1976, 1978, 1982, 1984, 1987

Overview

For further details see the GSS Data Explorer page for wordg.

Counts by year:

year	iap	incorrect answer	correct answer	no answer	not available in this year	Total

1974	-	808	521	155	-	1484

1976	-	789	513	197	-	1499

1978	-	827	504	201	-	1532

1982	-	1001	528	331	-	1860

1984	-	787	483	203	-	1473

1987	-	951	521	347	-	1819

1988	493	539	281	168	-	1481

1989	531	514	340	152	-	1537

1990	444	411	327	190	-	1372

1991	493	535	324	165	-	1517

1993	526	583	341	156	-	1606

1994	1015	1032	618	327	-	2992

1996	944	1045	609	306	-	2904

1998	1445	746	396	245	-	2832

2000	1398	676	437	306	-	2817

2004	1340	840	497	135	-	2812

2006	3041	774	482	213	-	4510

2008	845	696	356	126	-	2023

2010	614	837	449	144	-	2044

2012	672	784	378	140	-	1974

2014	863	1027	519	129	-	2538

2016	979	1199	572	117	-	2867

2018	789	1000	467	92	-	2348

2022	2377	664	400	103	-	3544

Total	18809	19065	10863	4648	0	53385

See also