Word i

wordi

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words–you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters – like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone – just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW. I. Word I

Values

[0] incorrect answer

[1] correct answer

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024

A/B/C	1998, 2000, 2004, 2006

Full	1974, 1976, 1978, 1982, 1984, 1987

Overview

For further details see the GSS Data Explorer page for wordi.

Counts by year:

year	iap	incorrect answer	correct answer	no answer	not available in this year	Total

1974	-	296	1118	70	-	1484

1976	-	248	1158	93	-	1499

1978	-	252	1201	79	-	1532

1982	-	433	1251	176	-	1860

1984	-	287	1087	99	-	1473

1987	-	391	1228	200	-	1819

1988	493	224	672	92	-	1481

1989	531	206	730	70	-	1537

1990	444	185	642	101	-	1372

1991	493	187	747	90	-	1517

1993	526	211	786	83	-	1606

1994	1015	390	1414	173	-	2992

1996	944	401	1425	134	-	2904

1998	1445	244	1018	125	-	2832

2000	1398	289	969	161	-	2817

2004	1340	327	1094	51	-	2812

2006	3041	337	1018	114	-	4510

2008	845	289	848	41	-	2023

2010	614	334	1033	63	-	2044

2012	672	301	953	48	-	1974

2014	863	404	1230	41	-	2538

2016	979	464	1381	43	-	2867

2018	789	417	1115	27	-	2348

2022	2378	261	861	44	-	3544

Total	18810	7378	24979	2218	0	53385

See also