Word j

wordj

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words–you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters – like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone – just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW. J. Word J

Values

[0] incorrect answer

[1] correct answer

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024

A/B/C	1998, 2000, 2004, 2006

Full	1974, 1976, 1978, 1982, 1984, 1987

Overview

For further details see the GSS Data Explorer page for wordj.

Counts by year:

year	iap	incorrect answer	correct answer	no answer	not available in this year	Total

1974	-	1039	340	105	-	1484

1976	-	1022	329	148	-	1499

1978	-	1073	336	123	-	1532

1982	-	1257	343	260	-	1860

1984	-	1018	297	158	-	1473

1987	-	1188	343	288	-	1819

1988	493	669	180	139	-	1481

1989	531	659	225	122	-	1537

1990	444	534	226	168	-	1372

1991	493	667	214	143	-	1517

1993	526	692	250	138	-	1606

1994	1015	1276	431	270	-	2992

1996	944	1225	456	279	-	2904

1998	1445	793	355	239	-	2832

2000	1398	858	297	264	-	2817

2004	1340	993	397	82	-	2812

2006	3041	937	357	175	-	4510

2008	845	812	274	92	-	2023

2010	614	979	345	106	-	2044

2012	672	895	311	96	-	1974

2014	863	1183	414	78	-	2538

2016	979	1320	496	72	-	2867

2018	789	1099	409	51	-	2348

2022	1179	1384	838	143	-	3544

Total	17611	23572	8463	3739	0	53385

See also