Number words correct in vocabulary test (7 common items)

wordsum7

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words–you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters – like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone – just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.

K. Total number of correct words. NOTE: This variable only covers the 7 common vocabulary items from the original and experimental versions of WORDSUM. For more details, see the Codebook.

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-/-	2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2024

A/B/C/D	2006

Full	1974, 1976, 1978, 1982, 1984, 1987

Overview

For further details see the GSS Data Explorer page for wordsum7.

Counts by year:

year	no answer	0	1	2	3	4	5	6	7	iap	Total

1974	35	20	54	159	277	386	249	152	152	-	1484

1976	64	23	77	165	241	388	250	153	138	-	1499

1978	46	29	75	180	246	409	256	155	136	-	1532

1982	132	39	93	214	298	504	273	188	119	-	1860

1984	77	23	63	149	243	375	261	157	125	-	1473

1987	150	44	100	185	300	483	263	176	118	-	1819

1988	569	19	42	95	157	284	153	97	65	-	1481

1989	553	35	51	87	149	298	182	122	60	-	1537

1990	519	12	39	70	137	234	177	102	82	-	1372

1991	556	14	38	95	153	302	171	111	77	-	1517

1993	576	20	56	77	178	316	178	137	68	-	1606

1994	888	29	68	145	295	614	337	221	151	244	2992

1996	564	26	69	159	330	571	386	201	129	469	2904

1998	1028	15	55	102	199	423	246	162	104	498	2832

2000	1049	22	67	112	209	396	273	139	96	454	2817

2002	2301	-	-	-	-	-	-	-	-	464	2765

2004	939	33	37	91	190	456	335	189	108	434	2812

2006	2608	24	38	82	218	448	312	195	74	511	4510

2008	169	25	40	92	196	381	232	128	66	694	2023

2010	43	16	57	107	228	447	247	197	88	614	2044

2012	22	30	42	116	203	409	253	149	78	672	1974

2014	18	24	64	123	271	508	390	196	81	863	2538

2016	23	25	59	125	313	600	425	227	91	979	2867

2018	12	32	57	121	261	465	327	203	81	789	2348

2021	2717	-	-	-	-	-	-	-	-	1315	4032

2022	75	27	72	142	292	628	526	419	195	1168	3544

2024	1151	23	73	143	267	568	494	433	157	-	3309

Total	16884	629	1486	3136	5851	10893	7196	4609	2639	10168	63491

See also