Number words correct in vocabulary test (original items)

wordsumold

Source

General Social Survey https://gss.norc.org

Details

Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words–you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters – like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone – just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.

K. Total number of correct words. NOTE: This variable contains the 10 original vocabulary items from WORDSUM. For more details, see the Codebook. For more details, see the Codebook.

Values

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/-/-	2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2024

A/B/C/D	2006

Full	1974, 1976, 1978, 1982, 1984, 1987

Overview

For further details see the GSS Data Explorer page for wordsumold.

Counts by year:

year	no answer	0	1	10	2	3	4	5	6	7	8	9	iap	Total

1974	35	10	21	112	53	96	160	260	282	196	147	112	-	1484

1976	64	6	28	109	38	114	159	229	283	209	133	127	-	1499

1978	46	10	37	107	48	109	164	226	307	229	141	108	-	1532

1982	132	17	37	86	87	135	192	301	342	260	140	131	-	1860

1984	77	11	25	90	53	86	158	227	259	217	153	117	-	1473

1987	150	24	38	85	54	144	209	257	370	219	151	118	-	1819

1988	569	6	21	46	39	59	90	166	211	137	79	58	-	1481

1989	553	27	21	51	43	52	95	145	219	147	113	71	-	1537

1990	519	5	16	63	27	55	86	124	168	143	96	70	-	1372

1991	556	1	18	55	38	44	98	158	206	170	88	85	-	1517

1993	576	6	33	50	38	59	114	143	227	157	113	90	-	1606

1994	888	16	35	113	43	96	181	309	415	304	196	152	244	2992

1996	564	10	29	95	68	103	186	310	416	296	212	146	469	2904

1998	1028	8	18	78	32	81	113	212	316	201	136	111	498	2832

2000	1049	9	27	72	41	74	132	220	289	206	147	97	454	2817

2002	2301	-	-	-	-	-	-	-	-	-	-	-	464	2765

2004	939	20	19	74	27	54	137	219	324	270	170	125	434	2812

2006	2608	10	22	48	38	59	121	224	315	241	195	118	511	4510

2008	169	11	20	44	36	56	115	222	266	196	113	81	694	2023

2010	43	4	22	67	53	79	138	213	337	211	151	112	614	2044

2012	22	12	20	55	45	69	142	215	306	193	137	86	672	1974

2014	18	7	34	53	46	90	153	277	400	284	195	118	863	2538

2016	23	14	20	63	44	100	174	356	414	339	226	115	979	2867

2018	12	14	26	52	45	100	151	272	338	266	191	92	789	2348

2021	2717	-	-	-	-	-	-	-	-	-	-	-	1315	4032

2022	35	7	21	52	40	46	91	153	257	202	156	111	2373	3544

2024	564	9	19	47	25	59	96	142	218	172	157	118	1683	3309

Total	16257	274	627	1767	1101	2019	3455	5580	7485	5465	3736	2669	13056	63491

See also