To work hard

workhard

Source

General Social Survey https://gss.norc.org

Details

Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO WORK HARD

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	fifth	first	fourth	no answer	second	third	not available in this release	Total

1972	1613	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	1534

1986	730	40	81	129	8	244	238	-	1470

1987	-	65	214	304	19	621	596	-	1819

1988	484	36	140	174	20	303	324	-	1481

1989	502	22	138	158	35	335	347	-	1537

1990	473	21	139	116	28	301	294	-	1372

1991	500	15	147	144	35	355	321	-	1517

1993	549	22	143	129	26	368	369	-	1606

1994	981	30	302	255	33	768	623	-	2992

1996	979	31	327	229	30	697	611	-	2904

1998	921	28	325	238	31	691	598	-	2832

2000	921	34	315	256	33	662	596	-	2817

2002	1853	17	190	90	13	314	288	-	2765

2004	1936	10	180	95	8	287	296	-	2812

2006	3508	15	186	124	7	366	304	-	4510

2008	658	25	295	157	14	454	420	-	2023

2010	667	23	338	121	28	467	400	-	2044

2012	636	20	326	127	15	451	399	-	1974

2014	851	14	425	137	17	609	485	-	2538

2016	911	22	510	146	22	696	560	-	2867

2018	785	15	415	114	21	515	483	-	2348

2021	1360	5	624	116	99	934	894	-	4032

2022	1173	23	554	166	77	827	724	-	3544

2024	-	-	-	-	-	-	-	3309	3309

Total	39964	533	6314	3525	619	11265	10170	3309	75699

Values

1 first

2 second

3 third

4 fourth

5 fifth

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook