World image:much evil vs. much good

world1

Source

General Social Survey https://gss.norc.org

Details

Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world and human nature between the two contrasting images? (HAND CARD AH) Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .? (READ EACH SET AND CIRCLE ONE CODE FOR EACH)

Values

1 world is evil

7 world is good

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1985	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/-	partial

1990	A/B/-	partial

1991	A/B/C	full

1994	A/B/-	partial

1996	A/B/-	partial

1998	A/B/C	full

2000	A/B/-	partial

2002	A/B/-	partial

Overview

For further details see the GSS Data Explorer page for this variable.

Counts by year:

year	iap	2	3	4	5	6	don't know	no answer	world is evil	world is good	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	1473

1985	-	41	105	447	377	194	7	21	100	242	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	1470

1987	-	68	154	461	372	225	-	31	143	365	-	1819

1988	-	54	112	428	340	180	7	20	103	237	-	1481

1989	531	38	77	290	229	134	-	21	66	151	-	1537

1990	444	28	76	279	243	126	2	11	47	116	-	1372

1991	158	52	73	339	326	237	3	86	74	169	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	-	1606

1994	1522	61	155	418	345	196	1	20	78	196	-	2992

1996	944	65	179	553	467	298	5	18	107	268	-	2904

1998	463	92	196	687	542	384	7	29	125	307	-	2832

2000	940	74	135	503	465	327	4	20	71	278	-	2817

2002	1857	26	75	253	212	139	14	3	38	148	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	-	4032

2022	-	-	-	-	-	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	-	-	-	-	-	3309	3309

Total	52135	599	1337	4658	3918	2440	50	280	952	2477	6853	75699

See also