Labor force status

wrkstat

Source

General Social Survey https://gss.norc.org

Details

Question Last week were you working full time, part time, going to school, keeping house, or what?

Values

[1] working full time

[2] working part time

[3] with a job, but not at work because of temporary illness, vacation, strike

[4] unemployed, laid off, looking for work

[5] retired

[6] in school

[7] keeping house

[8] other

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

A/B/C/-	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

A/B/C/D	2006

Full	1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987

Overview

For further details see the GSS Data Explorer page for wrkstat.

Counts by year:

year	working full time	working part time	with a job, but not at work because of temporary illness, vacation, strike	unemployed, laid off, looking for work	retired	in school	keeping house	other	no answer	skipped on web	don't know	Total

1972	750	121	38	46	144	58	432	24	-	-	-	1613

1973	651	134	31	35	154	59	421	19	-	-	-	1504

1974	635	108	46	54	167	47	417	10	-	-	-	1484

1975	622	145	26	61	165	49	402	20	-	-	-	1490

1976	619	131	28	63	176	46	427	9	-	-	-	1499

1977	775	107	39	37	152	38	359	23	-	-	-	1530

1978	716	139	39	39	157	35	377	30	-	-	-	1532

1980	687	137	35	39	149	40	356	25	-	-	-	1468

1982	839	185	41	92	211	44	402	46	-	-	-	1860

1983	741	170	34	84	163	31	349	27	-	-	-	1599

1984	717	162	34	49	154	41	302	14	-	-	-	1473

1985	741	171	39	45	204	44	275	15	-	-	-	1534

1986	698	146	31	33	196	38	305	23	-	-	-	1470

1987	934	201	19	42	248	47	299	29	-	-	-	1819

1988	728	166	31	33	213	62	233	15	-	-	-	1481

1989	760	153	41	25	217	48	271	22	-	-	-	1537

1990	704	147	26	33	194	48	199	21	-	-	-	1372

1991	702	182	28	39	218	57	270	21	-	-	-	1517

1993	809	167	33	52	243	47	216	39	-	-	-	1606

1994	1594	315	60	88	425	79	382	47	2	-	-	2992

1996	1641	302	60	75	350	84	338	54	-	-	-	2904

1998	1568	304	46	58	394	79	328	54	1	-	-	2832

2000	1524	306	60	60	404	77	325	61	-	-	-	2817

2002	1432	312	52	121	414	78	268	88	-	-	-	2765

2004	1466	320	80	99	403	115	266	62	1	-	-	2812

2006	2322	440	90	148	715	140	496	155	4	-	-	4510

2008	1003	211	53	74	336	57	227	60	2	-	-	2023

2010	917	234	33	145	319	93	235	65	3	-	-	2044

2012	912	226	40	104	357	70	210	54	1	-	-	1974

2014	1230	273	40	104	460	90	263	76	2	-	-	2538

2016	1321	345	57	118	574	76	284	89	3	-	-	2867

2018	1134	259	53	84	445	81	242	48	2	-	-	2348

2021	1776	365	104	265	993	88	315	118	3	5	-	4032

2022	1599	346	89	177	772	101	273	180	2	4	1	3544

Total	35267	7430	1556	2621	10886	2187	10764	1643	26	9	1	72390

See also