Denomination in which r was raised

denom16

Source

General Social Survey https://gss.norc.org

Details

Question In what religion were you raised? A. What specific denomination is that, if any? As of 2024, DENOM16 has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of DENOM16. The original DENOM16 containing data 1973 to 2022 has been renamed DENOM16_7322.

Values

10 american baptist association

11 american baptist churches in the u.s.a.

12 national baptist convention of america

13 national baptist convention u.s.a., inc.

14 southern baptist convention

15 other baptist churches

18 baptist, don't know which

20 african methodist episcopal church

21 african methodist episcopal zion church

22 united methodist church

23 other methodist churches

28 methodist, don't know which

30 american lutheran church

31 lutheran church in america

32 lutheran church - missouri synod

33 wisconsin evangelical lutheran synod

34 other lutheran churches

35 evangelical lutheran

38 lutheran, don't know which

40 presbyterian church in the u.s.

41 united presbyterian church in the u.s.a.

42 other presbyterian churches

43 presbyterian, merged

48 presbyterian, don't know which

50 episcopal church

60 other, specify

70 no denomination or non-denominational church

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1973	-/-/-	full

1974	-/-/-	full

1975	-/-/-	full

1976	-/-/-	full

1977	-/-/-	full

1978	-/-/-	full

1980	-/-/-	full

1982	-/-/-	full

1983	-/-/-	full

1984	-/-/-	full

1985	-/-/-	full

1986	-/-/-	full

1987	-/-/-	full

1988	A/B/C	full

1989	A/B/C	full

1990	A/B/C	full

1991	A/B/C	full

1993	A/B/C	full

1994	A/B/C	full

1996	A/B/C	full

1998	A/B/C	full

2000	A/B/C	full

2002	A/B/C	full

2004	A/B/C	full

2006	A/B/C	full

2008	A/B/C	full

2010	A/B/C	full

2012	A/B/C	full

2014	A/B/C	full

2016	A/B/C	full

2018	A/B/C	full

2021	A/B/C	full

2022	A/B/C	full

2024	A/B/C	full

Link at the GSS

For further details see the GSS Data Explorer page for denom16.

See also