Materials lead to rape

pornrape

Source

General Social Survey https://gss.norc.org

Details

Question 221. The next questions are about pornography - books, movies, magazines, and photographs that show or describe sex activities. I'm going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect: c. Sexual materials lead people to commit rape.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	don't know	no	no answer	yes	not available in this year	Total

1972	1613	-	-	-	-	-	1613

1973	-	105	642	9	748	-	1504

1974	1484	-	-	-	-	-	1484

1975	-	149	563	6	772	-	1490

1976	-	136	569	7	787	-	1499

1977	1530	-	-	-	-	-	1530

1978	-	112	549	4	867	-	1532

1980	-	134	538	6	790	-	1468

1982	1860	-	-	-	-	-	1860

1983	-	113	596	10	880	-	1599

1984	-	114	542	11	806	-	1473

1985	1534	-	-	-	-	-	1534

1986	-	101	530	9	830	-	1470

1987	-	168	677	11	963	-	1819

1988	484	80	351	8	558	-	1481

1989	502	74	339	6	616	-	1537

1990	473	94	297	1	507	-	1372

1991	500	111	372	2	532	-	1517

1993	549	100	351	10	596	-	1606

1994	1965	104	426	6	491	-	2992

1996	2904	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	4032

2022	-	-	-	-	-	3544	3544

2024	-	-	-	-	-	3309	3309

Total	48960	1695	7342	106	10743	6853	75699

Values

1 yes

2 no

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook