If rich, continue or stop working

richwork

Source

General Social Survey https://gss.norc.org

Details

Question If you were to get enough money to live as comfortably as you would like for the rest of your life, would you continue to work or would you stop working?

Values

[1] continue to work

[2] stop working

[NA(d)] don't know [NA(i)] iap [NA(j)] I don't have a job [NA(m)] dk, na, iap [NA(n)] no answer [NA(p)] not imputable [NA(r)] refused [NA(s)] skipped on web [NA(u)] uncodeable [NA(x)] not available in this release [NA(y)] not available in this year [NA(z)] see codebook

Question Years and Ballots

ballots	years

-/-	1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987

A/C	1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024

Overview

For further details see the GSS Data Explorer page for richwork.

Counts by year:

year	iap	continue to work	stop working	don't know	no answer	skipped on web	Total

1973	652	566	253	12	21	-	1504

1974	641	532	289	16	6	-	1484

1976	721	515	231	11	21	-	1499

1977	572	658	282	14	4	-	1530

1980	570	674	203	9	12	-	1468

1982	703	816	312	18	11	-	1860

1984	510	721	228	6	8	-	1473

1985	538	680	298	10	8	-	1534

1987	622	888	290	8	11	-	1819

1988	852	440	180	6	3	-	1481

1989	883	462	178	9	5	-	1537

1990	770	426	160	8	8	-	1372

1991	891	412	204	8	2	-	1517

1993	901	474	213	13	5	-	1606

1994	1610	895	465	16	6	-	2992

1996	1513	927	437	22	5	-	2904

1998	1524	894	391	18	5	-	2832

2000	1544	833	412	22	6	-	2817

2002	2113	444	201	4	3	-	2765

2004	2187	430	185	2	8	-	2812

2006	3197	896	403	10	4	-	4510

2008	1137	629	245	7	5	-	2023

2010	1183	591	256	5	9	-	2044

2012	1129	580	249	4	12	-	1974

2014	1440	749	345	3	1	-	2538

2016	1646	860	344	6	11	-	2867

2018	1326	723	280	6	13	-	2348

2021	2367	1004	620	-	34	7	4032

2022	2082	950	502	4	2	4	3544

Total	35824	19669	8656	277	249	11	64686

See also