Vignette version number

vigversn

Source

General Social Survey https://gss.norc.org

Details

Question TURN TO FACESHEET INSIDE FRONT COVER AND TEAR OFF VIGNETTE SECTION. WRITE VIGNETTE ID HERE. 648. Next I'm going to describe a person – lets call him/her (REFER TO NAME IN VIGNETTE). After I read a description of him/her I'll ask you some questions about how you think and feel about him/her. There are no right or wrong answers. I'm only interested in what you think of NAME.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	10: alcohol dependence, mary, white, eighth grade	11: alcohol dependence, mary, african american, eighth grade	12: alcohol dependence, mary, hispanic, eighth grade	13: alcohol dependence, mary, white, high school	14: alcohol dependence, mary, african american, high school	15: alcohol dependence, mary, hispanic, high school	16: alcohol dependence, mary, white, college	17: alcohol dependence, mary, african american, college	18: alcohol dependence, mary, hispanic, college	19: major depression, john, white, eighth grade	1: alcohol dependence, john, white, eighth grade	20: major depression, john, african american, eighth grade	21: major depression, john, hispanic, eighth grade	22: major depression, john, white, high school	23: major depression, john, african american, high school	24: major depression, john, hispanic, high school	25: major depression, john, white, college	26: major depression, john, african american, college	27: major depression, john, hispanic, college	28: major depression, mary, white, eighth grade	29: major depression, mary, african american, eighth grade	2: alcohol dependence, john, african american, eighth grade	30: major depression, mary, hispanic, eighth grade	31: major depression, mary, white, high school	32: major depression, mary, african american, high school	33: major depression, mary, hispanic, high school	34: major depression, mary, white, college	35: major depression, mary, african american, college	36: major depression, mary, hispanic, college	37: schizophrenia, john, white, eighth grade	38: schizophrenia, john, african american, eighth grade	39: schizophrenia, john, hispanic, eighth grade	3: alcohol dependence, john, hispanic, eighth grade	40: schizophrenia, john, white, high school	41: schizophrenia, john, african american, high school	42: schizophrenia, john, hispanic, high school	43: schizophrenia, john, white, college	44: schizophrenia, john, african american, college	45: schizophrenia, john, hispanic, college	46: schizophrenia, mary, white, eighth grade	47: schizophrenia, mary, african american, eighth grade	48: schizophrenia, mary, hispanic, eighth grade	49: schizophrenia, mary, white, high school	4: alcohol dependence, john, white, high school	50: schizophrenia, mary, african american, high school	51: schizophrenia, mary, hispanic, high school	52: schizophrenia, mary, white, college	53: schizophrenia, mary, african american, college	54: schizophrenia, mary, hispanic, college	55: drug problem, john, white, eighth grade	56: drug problem, john, african american, eighth grade	57: drug problem, john, hispanic, eighth grade	58: drug problem, john, white, high school	59: drug problem, john, african american, high school	5: alcohol dependence, john, african american, high school	60: drug problem, john, hispanic, high school	61: drug problem, john, white, college	62: drug problem, john, african american, college	63: drug problem, john, hispanic, college	64: drug problem, mary, white, eighth grade	65: drug problem, mary, african american, eighth grade	66: drug problem, mary, hispanic, eighth grade	67: drug problem, mary, white, high school	68: drug problem, mary, african american, high school	69: drug problem, mary, hispanic, high school	6: alcohol dependence, john, hispanic, high school	70: drug problem, mary, white, college	71: drug problem, mary, african american, college	72: drug problem, mary, hispanic, college	73: no problem, john, white, eighth grade	74: no problem, john, african american, eighth grade	75: no problem, john, hispanic, eighth grade	76: no problem, john, white, high school	77: no problem, john, african american, high school	78: no problem, john, hispanic, high school	79: no problem, john, white, college	7: alcohol dependence, john, white, college	80: no problem, john, african american, college	81: no problem, john, hispanic, college	82: no problem, mary, white, eighth grade	83: no problem, mary, african american, eighth grade	84: no problem, mary, hispanic, eighth grade	85: no problem, mary, white, high school	86: no problem, mary, african american, high school	87: no problem, mary, hispanic, high school	88: no problem, mary, white, college	89: no problem, mary, african american, college	8: alcohol dependence, john, african american, college	90: no problem, mary, hispanic, college	9: alcohol dependence, john, hispanic, college	not available in this year	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	2992	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	1460	22	18	18	14	12	11	18	20	22	14	15	10	16	21	17	19	15	15	14	23	21	11	12	19	17	8	24	16	19	19	19	16	13	17	20	20	16	12	12	19	19	21	10	17	11	18	18	17	17	15	14	8	17	18	16	18	15	15	21	17	16	21	13	17	14	17	16	19	18	9	13	10	20	25	18	10	13	12	14	16	19	21	13	11	19	12	15	15	14	8	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	2987	20	22	26	18	15	19	22	15	20	22	22	22	21	19	22	20	25	26	27	17	23	23	22	19	25	21	23	21	22	20	19	26	21	19	21	18	19	19	21	22	22	16	17	24	21	19	24	24	26	-	-	-	-	-	17	-	-	-	-	-	-	-	-	-	-	25	-	-	-	25	17	18	24	15	15	27	26	22	20	25	20	23	20	22	21	23	18	20	21	22	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2867

2018	1175	12	14	12	14	15	12	17	14	12	11	12	18	12	12	17	10	16	17	13	11	13	16	12	12	14	18	19	12	14	14	14	11	13	12	9	14	12	14	8	8	11	15	14	19	12	17	12	14	18	14	10	12	9	11	15	16	16	12	12	9	15	11	13	15	15	16	12	15	10	18	12	13	13	16	11	12	14	9	10	9	12	12	9	9	9	12	12	9	19	13	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4032

2022	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3544	3544

2024	1670	18	14	15	15	16	16	17	12	20	23	19	17	14	21	18	15	21	15	13	26	10	28	19	15	28	21	10	17	19	22	17	19	16	16	19	17	13	13	15	23	21	19	18	24	26	21	28	17	16	20	20	21	25	19	14	21	15	21	18	10	21	21	20	15	14	23	14	16	20	14	22	13	17	18	18	27	17	15	21	21	15	18	15	15	19	24	16	14	15	25	-	3309

Total	66376	72	68	71	61	58	58	74	61	74	70	68	67	63	73	74	64	77	73	67	77	67	78	65	65	84	68	76	66	74	75	69	72	63	64	69	69	60	58	56	72	73	71	59	84	70	75	82	72	77	49	44	41	51	48	62	55	46	48	51	36	52	53	46	47	43	81	42	50	48	66	64	54	74	74	62	76	70	58	65	71	66	74	57	57	68	71	61	58	69	68	3544	75699

Values

1 1: alcohol dependence, john, white, eighth grade

2 2: alcohol dependence, john, african american, eighth grade

3 3: alcohol dependence, john, hispanic, eighth grade

4 4: alcohol dependence, john, white, high school

5 5: alcohol dependence, john, african american, high school

6 6: alcohol dependence, john, hispanic, high school

7 7: alcohol dependence, john, white, college

8 8: alcohol dependence, john, african american, college

9 9: alcohol dependence, john, hispanic, college

10 10: alcohol dependence, mary, white, eighth grade

11 11: alcohol dependence, mary, african american, eighth grade

12 12: alcohol dependence, mary, hispanic, eighth grade

13 13: alcohol dependence, mary, white, high school

14 14: alcohol dependence, mary, african american, high school

15 15: alcohol dependence, mary, hispanic, high school

16 16: alcohol dependence, mary, white, college

17 17: alcohol dependence, mary, african american, college

18 18: alcohol dependence, mary, hispanic, college

19 19: major depression, john, white, eighth grade

20 20: major depression, john, african american, eighth grade

21 21: major depression, john, hispanic, eighth grade

22 22: major depression, john, white, high school

23 23: major depression, john, african american, high school

24 24: major depression, john, hispanic, high school

25 25: major depression, john, white, college

26 26: major depression, john, african american, college

27 27: major depression, john, hispanic, college

28 28: major depression, mary, white, eighth grade

29 29: major depression, mary, african american, eighth grade

30 30: major depression, mary, hispanic, eighth grade

31 31: major depression, mary, white, high school

32 32: major depression, mary, african american, high school

33 33: major depression, mary, hispanic, high school

34 34: major depression, mary, white, college

35 35: major depression, mary, african american, college

36 36: major depression, mary, hispanic, college

37 37: schizophrenia, john, white, eighth grade

38 38: schizophrenia, john, african american, eighth grade

39 39: schizophrenia, john, hispanic, eighth grade

40 40: schizophrenia, john, white, high school

41 41: schizophrenia, john, african american, high school

42 42: schizophrenia, john, hispanic, high school

43 43: schizophrenia, john, white, college

44 44: schizophrenia, john, african american, college

45 45: schizophrenia, john, hispanic, college

46 46: schizophrenia, mary, white, eighth grade

47 47: schizophrenia, mary, african american, eighth grade

48 48: schizophrenia, mary, hispanic, eighth grade

49 49: schizophrenia, mary, white, high school

50 50: schizophrenia, mary, african american, high school

51 51: schizophrenia, mary, hispanic, high school

52 52: schizophrenia, mary, white, college

53 53: schizophrenia, mary, african american, college

54 54: schizophrenia, mary, hispanic, college

55 55: drug problem, john, white, eighth grade

56 56: drug problem, john, african american, eighth grade

57 57: drug problem, john, hispanic, eighth grade

58 58: drug problem, john, white, high school

59 59: drug problem, john, african american, high school

60 60: drug problem, john, hispanic, high school

61 61: drug problem, john, white, college

62 62: drug problem, john, african american, college

63 63: drug problem, john, hispanic, college

64 64: drug problem, mary, white, eighth grade

65 65: drug problem, mary, african american, eighth grade

66 66: drug problem, mary, hispanic, eighth grade

67 67: drug problem, mary, white, high school

68 68: drug problem, mary, african american, high school

69 69: drug problem, mary, hispanic, high school

70 70: drug problem, mary, white, college

71 71: drug problem, mary, african american, college

72 72: drug problem, mary, hispanic, college

73 73: no problem, john, white, eighth grade

74 74: no problem, john, african american, eighth grade

75 75: no problem, john, hispanic, eighth grade

76 76: no problem, john, white, high school

77 77: no problem, john, african american, high school

78 78: no problem, john, hispanic, high school

79 79: no problem, john, white, college

80 80: no problem, john, african american, college

81 81: no problem, john, hispanic, college

82 82: no problem, mary, white, eighth grade

83 83: no problem, mary, african american, eighth grade

84 84: no problem, mary, hispanic, eighth grade

85 85: no problem, mary, white, high school

86 86: no problem, mary, african american, high school

87 87: no problem, mary, hispanic, high school

88 88: no problem, mary, white, college

89 89: no problem, mary, african american, college

90 90: no problem, mary, hispanic, college

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook