Vignette version number

vigversn

Source

General Social Survey https://gss.norc.org

Details

Question Next I'm going to describe a person – lets call him/her (REFER TO NAME IN VIGNETTE). After I read a description of him/her I'll ask you some questions about how you think and feel about him/her. There are no right or wrong answers. I'm only interested in what you think of NAME.

Values

1 1: alcohol dependence, john, white, eighth grade

2 2: alcohol dependence, john, african american, eighth grade

3 3: alcohol dependence, john, hispanic, eighth grade

4 4: alcohol dependence, john, white, high school

5 5: alcohol dependence, john, african american, high school

6 6: alcohol dependence, john, hispanic, high school

7 7: alcohol dependence, john, white, college

8 8: alcohol dependence, john, african american, college

9 9: alcohol dependence, john, hispanic, college

10 10: alcohol dependence, mary, white, eighth grade

11 11: alcohol dependence, mary, african american, eighth grade

12 12: alcohol dependence, mary, hispanic, eighth grade

13 13: alcohol dependence, mary, white, high school

14 14: alcohol dependence, mary, african american, high school

15 15: alcohol dependence, mary, hispanic, high school

16 16: alcohol dependence, mary, white, college

17 17: alcohol dependence, mary, african american, college

18 18: alcohol dependence, mary, hispanic, college

19 19: major depression, john, white, eighth grade

20 20: major depression, john, african american, eighth grade

21 21: major depression, john, hispanic, eighth grade

22 22: major depression, john, white, high school

23 23: major depression, john, african american, high school

24 24: major depression, john, hispanic, high school

25 25: major depression, john, white, college

26 26: major depression, john, african american, college

27 27: major depression, john, hispanic, college

28 28: major depression, mary, white, eighth grade

29 29: major depression, mary, african american, eighth grade

30 30: major depression, mary, hispanic, eighth grade

31 31: major depression, mary, white, high school

32 32: major depression, mary, african american, high school

33 33: major depression, mary, hispanic, high school

34 34: major depression, mary, white, college

35 35: major depression, mary, african american, college

36 36: major depression, mary, hispanic, college

37 37: schizophrenia, john, white, eighth grade

38 38: schizophrenia, john, african american, eighth grade

39 39: schizophrenia, john, hispanic, eighth grade

40 40: schizophrenia, john, white, high school

41 41: schizophrenia, john, african american, high school

42 42: schizophrenia, john, hispanic, high school

43 43: schizophrenia, john, white, college

44 44: schizophrenia, john, african american, college

45 45: schizophrenia, john, hispanic, college

46 46: schizophrenia, mary, white, eighth grade

47 47: schizophrenia, mary, african american, eighth grade

48 48: schizophrenia, mary, hispanic, eighth grade

49 49: schizophrenia, mary, white, high school

50 50: schizophrenia, mary, african american, high school

51 51: schizophrenia, mary, hispanic, high school

52 52: schizophrenia, mary, white, college

53 53: schizophrenia, mary, african american, college

54 54: schizophrenia, mary, hispanic, college

55 55: drug problem, john, white, eighth grade

56 56: drug problem, john, african american, eighth grade

57 57: drug problem, john, hispanic, eighth grade

58 58: drug problem, john, white, high school

59 59: drug problem, john, african american, high school

60 60: drug problem, john, hispanic, high school

61 61: drug problem, john, white, college

62 62: drug problem, john, african american, college

63 63: drug problem, john, hispanic, college

64 64: drug problem, mary, white, eighth grade

65 65: drug problem, mary, african american, eighth grade

66 66: drug problem, mary, hispanic, eighth grade

67 67: drug problem, mary, white, high school

68 68: drug problem, mary, african american, high school

69 69: drug problem, mary, hispanic, high school

70 70: drug problem, mary, white, college

71 71: drug problem, mary, african american, college

72 72: drug problem, mary, hispanic, college

73 73: no problem, john, white, eighth grade

74 74: no problem, john, african american, eighth grade

75 75: no problem, john, hispanic, eighth grade

76 76: no problem, john, white, high school

77 77: no problem, john, african american, high school

78 78: no problem, john, hispanic, high school

79 79: no problem, john, white, college

80 80: no problem, john, african american, college

81 81: no problem, john, hispanic, college

82 82: no problem, mary, white, eighth grade

83 83: no problem, mary, african american, eighth grade

84 84: no problem, mary, hispanic, eighth grade

85 85: no problem, mary, white, high school

86 86: no problem, mary, african american, high school

87 87: no problem, mary, hispanic, high school

88 88: no problem, mary, white, college

89 89: no problem, mary, african american, college

90 90: no problem, mary, hispanic, college

NA(d) don't know / NA(i) iap / NA(j) I don't have a job / NA(m) dk, na, iap / NA(n) no answer / NA(p) not imputable / NA(r) refused / NA(s) skipped on web / NA(u) uncodeable / NA(x) not available in this release / NA(y) not available in this year / NA(z) see codebook

Question Years and Ballots

year	ballots	availability

1996	A/B/C	full

2006	A/B/C	full

2018	B/C/-	partial

2024	A/B/-	partial

Link at the GSS

For details see the GSS Data Explorer page for vigversn.

See also